as_id {googledrive}	R Documentation

Extract and/or mark as file id

Description

Gets file ids from various inputs and marks them as such, to distinguish them from file names or paths.

This is a generic function.

Usage

as_id(x, ...)

Arguments

x
A character vector of file or shared drive ids or URLs, a dribble, or a suitable data frame.

...
Other arguments passed down to methods. (Not used.)

Value

A character vector bearing the S3 class drive_id.

Examples

as_id("123abc")
as_id("https://docs.google.com/spreadsheets/d/qawsedrf16273849/edit#gid=12345")

x <- drive_find(n_max = 3)
as_id(x)

[Package googledrive version 2.0.0 Index]