Identify Team Drives

Converts various representations of Team Drive into a dribble, the object used by googledrive to hold Drive file metadata. Team Drives can be specified via

Name.

Team Drive id. Mark with as_id() to distinguish from name.

Data frame or dribble consisting solely of Team Drives.

List representing Team Drive resource objects. Mostly for internal use.

Note: Team Drives are only available to users of certain enhanced Google services, such as G Suite Enterprise, G Suite Business, or G Suite for Education.

This is a generic function.

as_team_drive(x, ...)

Arguments

x
A vector of Team Drive names, a vector of Team Drive ids marked with as_id(), a list of Team Drive Resource objects, or a suitable data frame.

...
Other arguments passed down to methods. (Not used.)

Examples

# NOT RUN {
## specify the name
as_team_drive("abc")

## specify the id (substitute one of your own!)
as_team_drive(as_id("0AOPK1X2jaNckUk9PVA"))
# }