dribble-checks {googledrive}	R Documentation

Check facts about a dribble

Description

Sometimes you need to check things about a dribble' or about the files it represents, such as:

Is it even a dribble?

Size: Does the dribble hold exactly one file? At least one file? No file?

File type: Is this file a folder?

File ownership and access: Is it mine? Published? Shared?

Usage

is_dribble(d)

no_file(d)

single_file(d)

some_files(d)

confirm_dribble(d)

confirm_single_file(d)

confirm_some_files(d)

is_folder(d)

is_shortcut(d)

is_folder_shortcut(d)

is_native(d)

is_parental(d)

is_mine(d)

is_shared_drive(d)

Arguments

d
A dribble.

Examples

## most of us have multiple files or folders on Google Drive
d <- drive_find()
is_dribble(d)
no_file(d)
single_file(d)
some_files(d)

# this will error
# confirm_single_file(d)

confirm_some_files(d)
is_folder(d)
is_mine(d)

[Package googledrive version 2.1.1 Index]