drive_browse {googledrive}	R Documentation

Visit Drive file in browser

Description

Visits a file on Google Drive in your default browser.

Usage

drive_browse(file = .Last.value)

Arguments

file
Something that identifies the file of interest on your Google Drive. Can be a name or path, a file id or URL marked with as_id(), or a dribble.

Value

Character vector of file hyperlinks, from drive_link(), invisibly.

Examples

drive_find(n_max = 1) %>% drive_browse()

[Package googledrive version 2.1.1 Index]