drive_empty_trash {googledrive}	R Documentation

Empty Drive Trash

Description

Caution, this will permanently delete files in your Drive trash.

Usage

drive_empty_trash(verbose = deprecated())

Arguments

verbose
This logical argument to individual googledrive functions is deprecated. To globally suppress googledrive messaging, use options(googledrive_quiet = TRUE) (the default behaviour is to emit informational messages). To suppress messaging in a more limited way, use the helpers local_drive_quiet() or with_drive_quiet().

[Package googledrive version 2.1.1 Index]