List Drive endpoints

Returns a list of selected Drive API v3 endpoints, as stored inside the googledrive package. The names of this list (or the id sub-elements) are the nicknames that can be used to specify an endpoint in generate_request(). For each endpoint, we store its nickname or id, the associated HTTP verb, the path, and details about the parameters. This list is derived from the Drive API v3 Discovery Document.

drive_endpoints()

Value

A list containing the subset of the Drive API v3 endpoints that are used internally by googledrive.

Examples

str(drive_endpoints(), max.level = 2)

#> List of 22
#>  $ drive.about.get                   :List of 4
#>   ..$ id        : chr "drive.about.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/about"
#>   ..$ parameters:List of 1
#>  $ drive.files.copy                  :List of 4
#>   ..$ id        : chr "drive.files.copy"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "drive/v3/files/{fileId}/copy"
#>   ..$ parameters:List of 57
#>  $ drive.files.create                :List of 4
#>   ..$ id        : chr "drive.files.create"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "drive/v3/files"
#>   ..$ parameters:List of 57
#>  $ drive.files.delete                :List of 4
#>   ..$ id        : chr "drive.files.delete"
#>   ..$ method    : chr "DELETE"
#>   ..$ path      : chr "drive/v3/files/{fileId}"
#>   ..$ parameters:List of 3
#>  $ drive.files.emptyTrash            :List of 4
#>   ..$ id        : chr "drive.files.emptyTrash"
#>   ..$ method    : chr "DELETE"
#>   ..$ path      : chr "drive/v3/files/trash"
#>   ..$ parameters:List of 1
#>  $ drive.files.export                :List of 4
#>   ..$ id        : chr "drive.files.export"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/export"
#>   ..$ parameters:List of 3
#>  $ drive.files.generateIds           :List of 4
#>   ..$ id        : chr "drive.files.generateIds"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/generateIds"
#>   ..$ parameters:List of 3
#>  $ drive.files.get                   :List of 4
#>   ..$ id        : chr "drive.files.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}"
#>   ..$ parameters:List of 4
#>  $ drive.files.list                  :List of 4
#>   ..$ id        : chr "drive.files.list"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files"
#>   ..$ parameters:List of 11
#>  $ drive.files.update                :List of 4
#>   ..$ id        : chr "drive.files.update"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "drive/v3/files/{fileId}"
#>   ..$ parameters:List of 59
#>  $ drive.files.watch                 :List of 4
#>   ..$ id        : chr "drive.files.watch"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "drive/v3/files/{fileId}/watch"
#>   ..$ parameters:List of 4
#>  $ drive.permissions.create          :List of 4
#>   ..$ id        : chr "drive.permissions.create"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions"
#>   ..$ parameters:List of 18
#>  $ drive.permissions.delete          :List of 4
#>   ..$ id        : chr "drive.permissions.delete"
#>   ..$ method    : chr "DELETE"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions/{permissionId}"
#>   ..$ parameters:List of 4
#>  $ drive.permissions.get             :List of 4
#>   ..$ id        : chr "drive.permissions.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions/{permissionId}"
#>   ..$ parameters:List of 4
#>  $ drive.permissions.list            :List of 4
#>   ..$ id        : chr "drive.permissions.list"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions"
#>   ..$ parameters:List of 5
#>  $ drive.permissions.update          :List of 4
#>   ..$ id        : chr "drive.permissions.update"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions/{permissionId}"
#>   ..$ parameters:List of 18
#>  $ drive.revisions.delete            :List of 4
#>   ..$ id        : chr "drive.revisions.delete"
#>   ..$ method    : chr "DELETE"
#>   ..$ path      : chr "drive/v3/files/{fileId}/revisions/{revisionId}"
#>   ..$ parameters:List of 3
#>  $ drive.revisions.get               :List of 4
#>   ..$ id        : chr "drive.revisions.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/revisions/{revisionId}"
#>   ..$ parameters:List of 4
#>  $ drive.revisions.list              :List of 4
#>   ..$ id        : chr "drive.revisions.list"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/revisions"
#>   ..$ parameters:List of 4
#>  $ drive.revisions.update            :List of 4
#>   ..$ id        : chr "drive.revisions.update"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "drive/v3/files/{fileId}/revisions/{revisionId}"
#>   ..$ parameters:List of 15
#>  $ drive.files.update.media          :List of 4
#>   ..$ id        : chr "drive.files.update.media"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "/upload/drive/v3/files/{fileId}"
#>   ..$ parameters:List of 9
#>  $ drive.files.update.media.resumable:List of 4
#>   ..$ id        : chr "drive.files.update.media.resumable"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "/resumable/upload/drive/v3/files/{fileId}"
#>   ..$ parameters:List of 9

drive_endpoints()$`drive.files.delete`

#> $id
#> [1] "drive.files.delete"
#> 
#> $method
#> [1] "DELETE"
#> 
#> $path
#> [1] "drive/v3/files/{fileId}"
#> 
#> $parameters
#> $parameters$fileId
#> $parameters$fileId$location
#> [1] "path"
#> 
#> $parameters$fileId$required
#> [1] TRUE
#> 
#> $parameters$fileId$type
#> [1] "string"
#> 
#> $parameters$fileId$repeated
#> [1] NA
#> 
#> $parameters$fileId$format
#> [1] NA
#> 
#> $parameters$fileId$enum
#> [1] NA
#> 
#> $parameters$fileId$description
#> [1] "The ID of the file."
#> 
#> 
#> $parameters$supportsTeamDrives
#> $parameters$supportsTeamDrives$location
#> [1] "query"
#> 
#> $parameters$supportsTeamDrives$required
#> [1] NA
#> 
#> $parameters$supportsTeamDrives$type
#> [1] "boolean"
#> 
#> $parameters$supportsTeamDrives$repeated
#> [1] NA
#> 
#> $parameters$supportsTeamDrives$format
#> [1] NA
#> 
#> $parameters$supportsTeamDrives$enum
#> [1] NA
#> 
#> $parameters$supportsTeamDrives$description
#> [1] "Whether the requesting application supports Team Drives."
#> 
#> 
#> $parameters$fields
#> $parameters$fields$location
#> [1] "query"
#> 
#> $parameters$fields$required
#> [1] NA
#> 
#> $parameters$fields$type
#> [1] "string"
#> 
#> $parameters$fields$repeated
#> [1] NA
#> 
#> $parameters$fields$format
#> [1] NA
#> 
#> $parameters$fields$enum
#> [1] NA
#> 
#> $parameters$fields$description
#> [1] NA
#> 
#> 
#>