Request partial resources

You may be able to improve the performance of your API calls by requesting only the metadata that you actually need. This function is primarily for internal use and is currently focused on the Files resource. Assuming that resource = "files" (the default), drive_fields() returns a character vector of googledrive's default fields. If input is provided via fields, it's checked for validity against the known fields. To see a tibble containing all possible fields and a variable indicating which are googledrive defaults, call drive_fields(expose()). prep_fields() prepares fields for inclusion as query parameters.

drive_fields(fields = NULL, resource = "files")

prep_fields(fields, resource = "files")

Arguments

fields
Character vector of field names. Optional. If provided, and resource = "files", they are checked for validity.

resource
Character, naming the API resource of interest. Currently, only the Files resource is anticipated.

Value

drive_fields(): Character vector of field names. prep_fields(): a string.

See also

Working with partialresources, in the Drive API documentation

Examples

## to get default fields for the Files resource
drive_fields()

#>  [1] "appProperties"         "capabilities"          "contentHints"         
#>  [4] "createdTime"           "description"           "explicitlyTrashed"    
#>  [7] "fileExtension"         "folderColorRgb"        "fullFileExtension"    
#> [10] "headRevisionId"        "iconLink"              "id"                   
#> [13] "imageMediaMetadata"    "kind"                  "lastModifyingUser"    
#> [16] "md5Checksum"           "mimeType"              "modifiedByMeTime"     
#> [19] "modifiedTime"          "name"                  "originalFilename"     
#> [22] "ownedByMe"             "owners"                "parents"              
#> [25] "permissions"           "properties"            "quotaBytesUsed"       
#> [28] "shared"                "sharedWithMeTime"      "sharingUser"          
#> [31] "size"                  "spaces"                "starred"              
#> [34] "thumbnailLink"         "trashed"               "version"              
#> [37] "videoMediaMetadata"    "viewedByMe"            "viewedByMeTime"       
#> [40] "viewersCanCopyContent" "webContentLink"        "webViewLink"          
#> [43] "writersCanShare"      

## get a tibble of all fields for the Files resource + indicator of defaults
drive_fields(expose())

#> # A tibble: 51 x 3
#>                       name default
#>                      <chr>   <lgl>
#>  1           appProperties    TRUE
#>  2            capabilities    TRUE
#>  3            contentHints    TRUE
#>  4             createdTime    TRUE
#>  5             description    TRUE
#>  6       explicitlyTrashed    TRUE
#>  7           fileExtension    TRUE
#>  8          folderColorRgb    TRUE
#>  9       fullFileExtension    TRUE
#> 10 hasAugmentedPermissions   FALSE
#> # ... with 41 more rows, and 1 more variables: desc <chr>

## invalid fields are removed and throw warning
drive_fields(c("name", "parents", "ownedByMe", "pancakes!"))

#> Warning: Ignoring fields that are non-standard for the Files resource:
#> pancakes!

#> [1] "name"      "parents"   "ownedByMe"

## prepare fields for query
prep_fields(c("name", "parents", "ownedByMe"))

#> [1] "files/name,files/parents,files/ownedByMe"