Find files on Google Drive.

This is the closest googledrive function to what you get from https://drive.google.com: by default, you just get a listing of your files. You can also narrow the search in various ways, such as by file type, whether it's yours or shared with you, starred status, etc.

drive_find(pattern = NULL, type = NULL, n_max = Inf, ...,
  verbose = TRUE)

Arguments

pattern
Character. If provided, only the files whose names match this regular expression are returned. This is implemented locally on the results returned by the API.

type
Character. If provided, only files of this type will be returned. Can be anything that drive_mime_type() knows how to handle. This is processed by googledrive and sent as a query parameter.

n_max
Integer. An upper bound on the number of files to return. This applies to the results requested from the API, which may be further filtered locally, via the pattern argument.

...
Query parameters to pass along to the API query.

verbose
Logical, indicating whether to print informative messages (default TRUE).

Value

An object of class dribble, a tibble with one row per file.

See also

Helpful links for forming queries:

https://developers.google.com/drive/v3/web/search-parameters

https://developers.google.com/drive/v3/reference/files/list

Examples

# NOT RUN {
## list "My Drive" w/o regard for folder hierarchy
drive_find()

## search for files located directly in your root folder
drive_find(q = "'root' in parents")

## filter for folders, the easy way and the hard way
drive_find(type = "folder")
drive_find(q = "mimeType = 'application/vnd.google-apps.folder'")

## filter for Google Sheets, the easy way and the hard way
drive_find(type = "spreadsheet")
drive_find(q = "mimeType='application/vnd.google-apps.spreadsheet'")

## files whose names match a regex
drive_find(pattern = "jt")

## control page size or cap the number of files returned
drive_find(pageSize = 50)
drive_find(n_max = 58)
drive_find(pageSize = 5, n_max = 15)
# }