drive_link {googledrive}	R Documentation

Retrieve Drive file links

Description

Returns the "webViewLink" for one or more files, which is the "link for opening the file in a relevant Google editor or viewer in a browser".

Usage

drive_link(file)

Arguments

file
Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

Value

Character vector of file hyperlinks.

Examples

# get a few files into a dribble
three_files <- drive_find(n_max = 3)

# get their browser links
drive_link(three_files)

[Package googledrive version 2.1.1 Index]