Rename a Drive file

This is a wrapper for drive_mv() that only renames a file. If you would like to rename AND move the file, see drive_mv().

drive_rename(file, name = NULL, verbose = TRUE)

Arguments

file
Something that identifies the file of interest on your Google Drive. Can be a name or path, a file id or URL marked with as_id(), or a dribble.

name
Character. Name you would like the file to have.

verbose
Logical, indicating whether to print informative messages (default TRUE).

Value

An object of class dribble, a tibble with one row per item.

Examples

# NOT RUN {
## Create a folder to rename
folder <- drive_mkdir("folder-to-rename")

## Rename folder
folder <- folder %>%
  drive_rename(name = "renamed-folder")

## Clean up
drive_rm(folder)
# }