Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file
Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what
Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via what appears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 x 4
#>    name        path          id                                 drive_resource  
#>    <chr>       <chr>         <drv_id>                           <list>          
#>  1 name-squat… ~/name-squat… 1bBSvAy-2I8Enpe_dCHPmALvKkm2rwX20  <named list [39…
#>  2 name-squat… ~/name-squat… 1nfNqerD-OouvZN_oqjSps5GuQiGjaU5f  <named list [39…
#>  3 VisiCalc    VisiCalc      1vPkfgYOHi6fKYDDhj5kvsshW_y9Krdxt  <named list [39…
#>  4 Lotus 1-2-3 Lotus 1-2-3   1Wr2sQdKbPBSlebu6SC7E5cNmvEujFIy3… <named list [36…
#>  5 SuperCalc   SuperCalc     1keOgBzrqffJka8Z4kkXfiKAvBd0-vo7F… <named list [36…
#>  6 VisiCalc    VisiCalc      1w1PowbeGHweXqezKW3ujlB5ICRt0C23I… <named list [36…
#>  7 name-squat… ~/name-squat… 1xOSEWBVWQZnOMWKp-BYyyQVhoECOkxaJ  <named list [39…
#>  8 name-squat… ~/name-squat… 1rxkJntT8AQjuRhSkoPG1MCLZArXjDf2S  <named list [39…
#>  9 name-squat… ~/name-squat… 1IaGjF5NT7kjgfNBAh6_R7L-2plgJjoaD  <named list [39…
#> 10 name-squat… ~/name-squat… 11XNwHIINBH9U5gIzkdWEA0_En0Tq4ndZ  <named list [39…
drive_reveal(files, "permissions")
#> Error: Client error: (404) Not Found
#> File not found: 1vPkfgYOHi6fKYDDhj5kvsshW_y9Krdxt.
#> • domain: global
#> • reason: notFound
#> • message: File not found: 1vPkfgYOHi6fKYDDhj5kvsshW_y9Krdxt.
#> • locationType: parameter
#> • location: fileId
drive_reveal(files, "published")
#> Error: Client error: (404) Not Found
#> File not found: 1vPkfgYOHi6fKYDDhj5kvsshW_y9Krdxt.
#> • domain: global
#> • reason: notFound
#> • message: File not found: 1vPkfgYOHi6fKYDDhj5kvsshW_y9Krdxt.
#> • locationType: parameter
#> • location: fileId

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 x 4
#>    name       id_parent               id                         drive_resource 
#>    <chr>      <drv_id>                <drv_id>                   <list>         
#>  1 name-squa… 0AO_RMaBzcP63Uk9PVA     1bBSvAy-2I8Enpe_dCHPmALvK… <named list [3…
#>  2 name-squa… 0AO_RMaBzcP63Uk9PVA     1nfNqerD-OouvZN_oqjSps5Gu… <named list [3…
#>  3 VisiCalc   17-uSrDfcTW3b1mkVHNOVT… 1vPkfgYOHi6fKYDDhj5kvsshW… <named list [3…
#>  4 Lotus 1-2… 17-uSrDfcTW3b1mkVHNOVT… 1Wr2sQdKbPBSlebu6SC7E5cNm… <named list [3…
#>  5 SuperCalc  17-uSrDfcTW3b1mkVHNOVT… 1keOgBzrqffJka8Z4kkXfiKAv… <named list [3…
#>  6 VisiCalc   17-uSrDfcTW3b1mkVHNOVT… 1w1PowbeGHweXqezKW3ujlB5I… <named list [3…
#>  7 name-squa… 0AO_RMaBzcP63Uk9PVA     1xOSEWBVWQZnOMWKp-BYyyQVh… <named list [3…
#>  8 name-squa… 0AO_RMaBzcP63Uk9PVA     1rxkJntT8AQjuRhSkoPG1MCLZ… <named list [3…
#>  9 name-squa… 0AO_RMaBzcP63Uk9PVA     1IaGjF5NT7kjgfNBAh6_R7L-2… <named list [3…
#> 10 name-squa… 0AO_RMaBzcP63Uk9PVA     11XNwHIINBH9U5gIzkdWEA0_E… <named list [3…

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 x 4
#>    name         trashed id                                      drive_resource  
#>    <chr>        <lgl>   <drv_id>                                <list>          
#>  1 name-squatt… TRUE    1bBSvAy-2I8Enpe_dCHPmALvKkm2rwX20       <named list [39…
#>  2 name-squatt… TRUE    1nfNqerD-OouvZN_oqjSps5GuQiGjaU5f       <named list [39…
#>  3 VisiCalc     FALSE   1vPkfgYOHi6fKYDDhj5kvsshW_y9Krdxt       <named list [39…
#>  4 Lotus 1-2-3  FALSE   1Wr2sQdKbPBSlebu6SC7E5cNmvEujFIy32a5SJ… <named list [36…
#>  5 SuperCalc    FALSE   1keOgBzrqffJka8Z4kkXfiKAvBd0-vo7FJCieZ… <named list [36…
#>  6 VisiCalc     TRUE    1w1PowbeGHweXqezKW3ujlB5ICRt0C23IP564w… <named list [36…
#>  7 name-squatt… TRUE    1xOSEWBVWQZnOMWKp-BYyyQVhoECOkxaJ       <named list [39…
#>  8 name-squatt… TRUE    1rxkJntT8AQjuRhSkoPG1MCLZArXjDf2S       <named list [39…
#>  9 name-squatt… TRUE    1IaGjF5NT7kjgfNBAh6_R7L-2plgJjoaD       <named list [39…
#> 10 name-squatt… TRUE    11XNwHIINBH9U5gIzkdWEA0_En0Tq4ndZ       <named list [39…
drive_reveal(files, "mime_type")
#> # A dribble: 10 x 4
#>    name       mime_type                 id                       drive_resource 
#>    <chr>      <chr>                     <drv_id>                 <list>         
#>  1 name-squa… application/octet-stream  1bBSvAy-2I8Enpe_dCHPmAL… <named list [3…
#>  2 name-squa… application/octet-stream  1nfNqerD-OouvZN_oqjSps5… <named list [3…
#>  3 VisiCalc   application/octet-stream  1vPkfgYOHi6fKYDDhj5kvss… <named list [3…
#>  4 Lotus 1-2… application/vnd.google-a… 1Wr2sQdKbPBSlebu6SC7E5c… <named list [3…
#>  5 SuperCalc  application/vnd.google-a… 1keOgBzrqffJka8Z4kkXfiK… <named list [3…
#>  6 VisiCalc   application/vnd.google-a… 1w1PowbeGHweXqezKW3ujlB… <named list [3…
#>  7 name-squa… application/octet-stream  1xOSEWBVWQZnOMWKp-BYyyQ… <named list [3…
#>  8 name-squa… application/octet-stream  1rxkJntT8AQjuRhSkoPG1MC… <named list [3…
#>  9 name-squa… application/octet-stream  1IaGjF5NT7kjgfNBAh6_R7L… <named list [3…
#> 10 name-squa… application/octet-stream  11XNwHIINBH9U5gIzkdWEA0… <named list [3…
drive_reveal(files, "starred")
#> # A dribble: 10 x 4
#>    name         starred id                                      drive_resource  
#>    <chr>        <lgl>   <drv_id>                                <list>          
#>  1 name-squatt… FALSE   1bBSvAy-2I8Enpe_dCHPmALvKkm2rwX20       <named list [39…
#>  2 name-squatt… FALSE   1nfNqerD-OouvZN_oqjSps5GuQiGjaU5f       <named list [39…
#>  3 VisiCalc     FALSE   1vPkfgYOHi6fKYDDhj5kvsshW_y9Krdxt       <named list [39…
#>  4 Lotus 1-2-3  FALSE   1Wr2sQdKbPBSlebu6SC7E5cNmvEujFIy32a5SJ… <named list [36…
#>  5 SuperCalc    FALSE   1keOgBzrqffJka8Z4kkXfiKAvBd0-vo7FJCieZ… <named list [36…
#>  6 VisiCalc     FALSE   1w1PowbeGHweXqezKW3ujlB5ICRt0C23IP564w… <named list [36…
#>  7 name-squatt… FALSE   1xOSEWBVWQZnOMWKp-BYyyQVhoECOkxaJ       <named list [39…
#>  8 name-squatt… FALSE   1rxkJntT8AQjuRhSkoPG1MCLZArXjDf2S       <named list [39…
#>  9 name-squatt… FALSE   1IaGjF5NT7kjgfNBAh6_R7L-2plgJjoaD       <named list [39…
#> 10 name-squatt… FALSE   11XNwHIINBH9U5gIzkdWEA0_En0Tq4ndZ       <named list [39…
drive_reveal(files, "description")
#> # A dribble: 10 x 4
#>    name         description id                                  drive_resource  
#>    <chr>        <list>      <drv_id>                            <list>          
#>  1 name-squatt… <NULL>      1bBSvAy-2I8Enpe_dCHPmALvKkm2rwX20   <named list [39…
#>  2 name-squatt… <NULL>      1nfNqerD-OouvZN_oqjSps5GuQiGjaU5f   <named list [39…
#>  3 VisiCalc     <NULL>      1vPkfgYOHi6fKYDDhj5kvsshW_y9Krdxt   <named list [39…
#>  4 Lotus 1-2-3  <NULL>      1Wr2sQdKbPBSlebu6SC7E5cNmvEujFIy32… <named list [36…
#>  5 SuperCalc    <NULL>      1keOgBzrqffJka8Z4kkXfiKAvBd0-vo7FJ… <named list [36…
#>  6 VisiCalc     <NULL>      1w1PowbeGHweXqezKW3ujlB5ICRt0C23IP… <named list [36…
#>  7 name-squatt… <NULL>      1xOSEWBVWQZnOMWKp-BYyyQVhoECOkxaJ   <named list [39…
#>  8 name-squatt… <NULL>      1rxkJntT8AQjuRhSkoPG1MCLZArXjDf2S   <named list [39…
#>  9 name-squatt… <NULL>      1IaGjF5NT7kjgfNBAh6_R7L-2plgJjoaD   <named list [39…
#> 10 name-squatt… <NULL>      11XNwHIINBH9U5gIzkdWEA0_En0Tq4ndZ   <named list [39…
drive_reveal(files, "version")
#> # A dribble: 10 x 4
#>    name         version id                                      drive_resource  
#>    <chr>        <chr>   <drv_id>                                <list>          
#>  1 name-squatt… 2       1bBSvAy-2I8Enpe_dCHPmALvKkm2rwX20       <named list [39…
#>  2 name-squatt… 2       1nfNqerD-OouvZN_oqjSps5GuQiGjaU5f       <named list [39…
#>  3 VisiCalc     1       1vPkfgYOHi6fKYDDhj5kvsshW_y9Krdxt       <named list [39…
#>  4 Lotus 1-2-3  4       1Wr2sQdKbPBSlebu6SC7E5cNmvEujFIy32a5SJ… <named list [36…
#>  5 SuperCalc    4       1keOgBzrqffJka8Z4kkXfiKAvBd0-vo7FJCieZ… <named list [36…
#>  6 VisiCalc     5       1w1PowbeGHweXqezKW3ujlB5ICRt0C23IP564w… <named list [36…
#>  7 name-squatt… 2       1xOSEWBVWQZnOMWKp-BYyyQVhoECOkxaJ       <named list [39…
#>  8 name-squatt… 2       1rxkJntT8AQjuRhSkoPG1MCLZArXjDf2S       <named list [39…
#>  9 name-squatt… 2       1IaGjF5NT7kjgfNBAh6_R7L-2plgJjoaD       <named list [39…
#> 10 name-squatt… 2       11XNwHIINBH9U5gIzkdWEA0_En0Tq4ndZ       <named list [39…
drive_reveal(files, "web_view_link")
#> # A dribble: 10 x 4
#>    name     web_view_link                      id                 drive_resource
#>    <chr>    <chr>                              <drv_id>           <list>        
#>  1 name-sq… https://drive.google.com/file/d/1… 1bBSvAy-2I8Enpe_d… <named list […
#>  2 name-sq… https://drive.google.com/file/d/1… 1nfNqerD-OouvZN_o… <named list […
#>  3 VisiCalc https://drive.google.com/file/d/1… 1vPkfgYOHi6fKYDDh… <named list […
#>  4 Lotus 1… https://docs.google.com/spreadshe… 1Wr2sQdKbPBSlebu6… <named list […
#>  5 SuperCa… https://docs.google.com/spreadshe… 1keOgBzrqffJka8Z4… <named list […
#>  6 VisiCalc https://docs.google.com/spreadshe… 1w1PowbeGHweXqezK… <named list […
#>  7 name-sq… https://drive.google.com/file/d/1… 1xOSEWBVWQZnOMWKp… <named list […
#>  8 name-sq… https://drive.google.com/file/d/1… 1rxkJntT8AQjuRhSk… <named list […
#>  9 name-sq… https://drive.google.com/file/d/1… 1IaGjF5NT7kjgfNBA… <named list […
#> 10 name-sq… https://drive.google.com/file/d/1… 11XNwHIINBH9U5gIz… <named list […
drive_reveal(files, "modified_time")
#> # A dribble: 10 x 4
#>    name       modified_time       id                             drive_resource 
#>    <chr>      <dttm>              <drv_id>                       <list>         
#>  1 name-squa… 2021-07-21 13:44:14 1bBSvAy-2I8Enpe_dCHPmALvKkm2r… <named list [3…
#>  2 name-squa… 2021-07-21 13:43:58 1nfNqerD-OouvZN_oqjSps5GuQiGj… <named list [3…
#>  3 VisiCalc   2021-07-21 13:43:39 1vPkfgYOHi6fKYDDhj5kvsshW_y9K… <named list [3…
#>  4 Lotus 1-2… 2021-07-21 13:43:37 1Wr2sQdKbPBSlebu6SC7E5cNmvEuj… <named list [3…
#>  5 SuperCalc  2021-07-21 13:43:35 1keOgBzrqffJka8Z4kkXfiKAvBd0-… <named list [3…
#>  6 VisiCalc   2021-07-21 13:43:33 1w1PowbeGHweXqezKW3ujlB5ICRt0… <named list [3…
#>  7 name-squa… 2021-07-20 13:44:56 1xOSEWBVWQZnOMWKp-BYyyQVhoECO… <named list [3…
#>  8 name-squa… 2021-07-20 13:44:21 1rxkJntT8AQjuRhSkoPG1MCLZArXj… <named list [3…
#>  9 name-squa… 2021-07-20 13:44:06 1IaGjF5NT7kjgfNBAh6_R7L-2plgJ… <named list [3…
#> 10 name-squa… 2021-07-19 13:45:16 11XNwHIINBH9U5gIzkdWEA0_En0Tq… <named list [3…
drive_reveal(files, "created_time")
#> # A dribble: 10 x 4
#>    name       created_time        id                             drive_resource 
#>    <chr>      <dttm>              <drv_id>                       <list>         
#>  1 name-squa… 2021-07-21 13:44:14 1bBSvAy-2I8Enpe_dCHPmALvKkm2r… <named list [3…
#>  2 name-squa… 2021-07-21 13:43:58 1nfNqerD-OouvZN_oqjSps5GuQiGj… <named list [3…
#>  3 VisiCalc   2021-07-21 13:43:39 1vPkfgYOHi6fKYDDhj5kvsshW_y9K… <named list [3…
#>  4 Lotus 1-2… 2021-07-21 13:43:36 1Wr2sQdKbPBSlebu6SC7E5cNmvEuj… <named list [3…
#>  5 SuperCalc  2021-07-21 13:43:34 1keOgBzrqffJka8Z4kkXfiKAvBd0-… <named list [3…
#>  6 VisiCalc   2021-07-21 13:43:32 1w1PowbeGHweXqezKW3ujlB5ICRt0… <named list [3…
#>  7 name-squa… 2021-07-20 13:44:56 1xOSEWBVWQZnOMWKp-BYyyQVhoECO… <named list [3…
#>  8 name-squa… 2021-07-20 13:44:21 1rxkJntT8AQjuRhSkoPG1MCLZArXj… <named list [3…
#>  9 name-squa… 2021-07-20 13:44:06 1IaGjF5NT7kjgfNBAh6_R7L-2plgJ… <named list [3…
#> 10 name-squa… 2021-07-19 13:45:16 11XNwHIINBH9U5gIzkdWEA0_En0Tq… <named list [3…
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 x 4
#>    name         owned_by_me id                                  drive_resource  
#>    <chr>        <lgl>       <drv_id>                            <list>          
#>  1 name-squatt… TRUE        1bBSvAy-2I8Enpe_dCHPmALvKkm2rwX20   <named list [39…
#>  2 name-squatt… TRUE        1nfNqerD-OouvZN_oqjSps5GuQiGjaU5f   <named list [39…
#>  3 VisiCalc     TRUE        1vPkfgYOHi6fKYDDhj5kvsshW_y9Krdxt   <named list [39…
#>  4 Lotus 1-2-3  TRUE        1Wr2sQdKbPBSlebu6SC7E5cNmvEujFIy32… <named list [36…
#>  5 SuperCalc    TRUE        1keOgBzrqffJka8Z4kkXfiKAvBd0-vo7FJ… <named list [36…
#>  6 VisiCalc     TRUE        1w1PowbeGHweXqezKW3ujlB5ICRt0C23IP… <named list [36…
#>  7 name-squatt… TRUE        1xOSEWBVWQZnOMWKp-BYyyQVhoECOkxaJ   <named list [39…
#>  8 name-squatt… TRUE        1rxkJntT8AQjuRhSkoPG1MCLZArXjDf2S   <named list [39…
#>  9 name-squatt… TRUE        1IaGjF5NT7kjgfNBAh6_R7L-2plgJjoaD   <named list [39…
#> 10 name-squatt… TRUE        11XNwHIINBH9U5gIzkdWEA0_En0Tq4ndZ   <named list [39…
drive_reveal(files, "size")
#> # A dribble: 10 x 4
#>    name          size  id                                       drive_resource  
#>    <chr>         <chr> <drv_id>                                 <list>          
#>  1 name-squatter 0     1bBSvAy-2I8Enpe_dCHPmALvKkm2rwX20        <named list [39…
#>  2 name-squatter 0     1nfNqerD-OouvZN_oqjSps5GuQiGjaU5f        <named list [39…
#>  3 VisiCalc      0     1vPkfgYOHi6fKYDDhj5kvsshW_y9Krdxt        <named list [39…
#>  4 Lotus 1-2-3   1024  1Wr2sQdKbPBSlebu6SC7E5cNmvEujFIy32a5SJn… <named list [36…
#>  5 SuperCalc     1024  1keOgBzrqffJka8Z4kkXfiKAvBd0-vo7FJCieZK… <named list [36…
#>  6 VisiCalc      1024  1w1PowbeGHweXqezKW3ujlB5ICRt0C23IP564ww… <named list [36…
#>  7 name-squatter 0     1xOSEWBVWQZnOMWKp-BYyyQVhoECOkxaJ        <named list [39…
#>  8 name-squatter 0     1rxkJntT8AQjuRhSkoPG1MCLZArXjDf2S        <named list [39…
#>  9 name-squatter 0     1IaGjF5NT7kjgfNBAh6_R7L-2plgJjoaD        <named list [39…
#> 10 name-squatter 0     11XNwHIINBH9U5gIzkdWEA0_En0Tq4ndZ        <named list [39…
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 x 4
#>    name        quota_bytes_used id                              drive_resource  
#>    <chr>       <chr>            <drv_id>                        <list>          
#>  1 name-squat… 0                1bBSvAy-2I8Enpe_dCHPmALvKkm2rw… <named list [39…
#>  2 name-squat… 0                1nfNqerD-OouvZN_oqjSps5GuQiGja… <named list [39…
#>  3 VisiCalc    0                1vPkfgYOHi6fKYDDhj5kvsshW_y9Kr… <named list [39…
#>  4 Lotus 1-2-3 1024             1Wr2sQdKbPBSlebu6SC7E5cNmvEujF… <named list [36…
#>  5 SuperCalc   1024             1keOgBzrqffJka8Z4kkXfiKAvBd0-v… <named list [36…
#>  6 VisiCalc    1024             1w1PowbeGHweXqezKW3ujlB5ICRt0C… <named list [36…
#>  7 name-squat… 0                1xOSEWBVWQZnOMWKp-BYyyQVhoECOk… <named list [39…
#>  8 name-squat… 0                1rxkJntT8AQjuRhSkoPG1MCLZArXjD… <named list [39…
#>  9 name-squat… 0                1IaGjF5NT7kjgfNBAh6_R7L-2plgJj… <named list [39…
#> 10 name-squat… 0                11XNwHIINBH9U5gIzkdWEA0_En0Tq4… <named list [39…

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 x 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>