Move Drive files to or from trash

Move Drive files to or from trash

drive_trash(file, verbose = deprecated())

drive_untrash(file, verbose = deprecated())

Arguments

file
Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

verbose
This logical argument to individual googledrive functions is deprecated. To globally suppress googledrive messaging, use options(googledrive_quiet = TRUE) (the default behaviour is to emit informational messages). To suppress messaging in a more limited way, use the helpers local_drive_quiet() or with_drive_quiet().

Value

An object of class dribble, a tibble with one row per file.

Examples

# Create a file and put it in the trash.
file <- drive_example_remote("chicken.txt") %>%
  drive_cp("chicken-trash.txt")
#> Original file:
#> • chicken.txt <id: 1wOLeWVRkTb6lDmLRiOhg9iKM7DlN762Y>
#> Copied to file:
#> • chicken-trash.txt <id: 1Jdu1ZQrvqfePoIhbwKeOj8k8SQ0pENY4>
drive_trash("chicken-trash.txt")
#> File trashed:
#> • chicken-trash.txt <id: 1Jdu1ZQrvqfePoIhbwKeOj8k8SQ0pENY4>

# Confirm it's in the trash
drive_find(trashed = TRUE)
#> # A dribble: 28 x 3
#>    name              id                                drive_resource   
#>  * <chr>             <chr>                             <list>           
#>  1 chicken-trash.txt 1Jdu1ZQrvqfePoIhbwKeOj8k8SQ0pENY4 <named list [39]>
#>  2 name-squatter     1W8MJks58Nxiwuyc9oso4zerZ0yNw8UKQ <named list [39]>
#>  3 name-squatter     1pQwT1fWEE063bhXtOxQWbL9Rqn37OHXb <named list [39]>
#>  4 name-squatter     1cYp9k8dvvmV_MURbAkSJcPYEAUXSEZTb <named list [39]>
#>  5 name-squatter     1mI7T4eyhUBtF8_4cpXSXcRbJKmlR9GpW <named list [39]>
#>  6 name-squatter     1QuF0uUCfG5alGGPFYIBDNmdS5jKSsRUK <named list [39]>
#>  7 name-squatter     1R6rA96E9tWqzCSO5nw24fvnJAqXaivbU <named list [39]>
#>  8 name-squatter     1QOXS8DT0AR0JyEs2qa8m3nTQjxrm-dDi <named list [39]>
#>  9 name-squatter     1zJxEe_Q5HQk5wtZ6-GgtborglLyXTZYw <named list [39]>
#> 10 name-squatter     1ezJcWFbJP-t1XSfyXtDH7OiyoJKKaQ4l <named list [39]>
#> # … with 18 more rows

# Remove it from the trash and confirm
drive_untrash("chicken-trash.txt")
#> File untrashed:
#> • chicken-trash.txt <id: 1Jdu1ZQrvqfePoIhbwKeOj8k8SQ0pENY4>
drive_find(trashed = TRUE)
#> # A dribble: 27 x 3
#>    name          id                                drive_resource   
#>  * <chr>         <chr>                             <list>           
#>  1 name-squatter 1W8MJks58Nxiwuyc9oso4zerZ0yNw8UKQ <named list [39]>
#>  2 name-squatter 1pQwT1fWEE063bhXtOxQWbL9Rqn37OHXb <named list [39]>
#>  3 name-squatter 1cYp9k8dvvmV_MURbAkSJcPYEAUXSEZTb <named list [39]>
#>  4 name-squatter 1mI7T4eyhUBtF8_4cpXSXcRbJKmlR9GpW <named list [39]>
#>  5 name-squatter 1QuF0uUCfG5alGGPFYIBDNmdS5jKSsRUK <named list [39]>
#>  6 name-squatter 1R6rA96E9tWqzCSO5nw24fvnJAqXaivbU <named list [39]>
#>  7 name-squatter 1QOXS8DT0AR0JyEs2qa8m3nTQjxrm-dDi <named list [39]>
#>  8 name-squatter 1zJxEe_Q5HQk5wtZ6-GgtborglLyXTZYw <named list [39]>
#>  9 name-squatter 1ezJcWFbJP-t1XSfyXtDH7OiyoJKKaQ4l <named list [39]>
#> 10 name-squatter 16MWbiyz7iD3wqzS8O1G-24-980Y3LpxX <named list [39]>
#> # … with 17 more rows

# Clean up
drive_rm("chicken-trash.txt")
#> File deleted:
#> • chicken-trash.txt <id: 1Jdu1ZQrvqfePoIhbwKeOj8k8SQ0pENY4>