Upload into a new Drive file

Uploads a local file into a new Drive file. To update the content or metadata of an existing Drive file, use drive_update().

drive_upload(media, path = NULL, name = NULL, type = NULL, ...,
  verbose = TRUE)

Arguments

media
Character, path to the local file to upload.

path
Specifies target destination for the new file on Google Drive. Can be an actual path (character), a file id marked with as_id(), or a dribble. If specified as an actual path, it is best to explicitly indicate if it's a folder by including a trailing slash, since it cannot always be worked out from the context of the call. Will default to its local name.

name
Character, new file name if not specified as part of path. This will force path to be treated as a folder, even if it is character and lacks a trailing slash. Will default to its local name.

type
Character. If type = NULL, a MIME type is automatically determined from the file extension, if possible. If the source file is of a suitable type, you can request conversion to Google Doc, Sheet or Slides by setting type to document, spreadsheet, or presentation, respectively. All non-NULL values for type are pre-processed with drive_mime_type().

...
Named parameters to pass along to the Drive API. You can affect the metadata of the target file by specifying properties of the Files resource via .... See the "Request body" section of the Drive API docs for the associated endpoint.

verbose
Logical, indicating whether to print informative messages (default TRUE).

Value

An object of class dribble, a tibble with one row per item.

See also

Wraps the files.create endpoint:

https://developers.google.com/drive/v3/reference/files/create

MIME types that can be converted to native Google formats:

https://developers.google.com/drive/v3/web/manage-uploads#importing_to_google_docs_types_wzxhzdk18wzxhzdk19

Examples

# NOT RUN {
## upload a csv file
mirrors_csv <- drive_upload(R.home('doc/BioC_mirrors.csv'))

## or convert it to a Google Sheet
mirrors_sheet <- drive_upload(
  R.home('doc/BioC_mirrors.csv'),
  name = "BioC_mirrors",
  type = "spreadsheet"
)

## check out the new Sheet!
drive_browse(mirrors_sheet)

## clean-up
drive_find("BioC_mirrors") %>% drive_rm()

## Upload a file and, at the same time, star it
logo <- drive_upload(
  R.home('doc/html/logo.jpg'),
  starred = "true"
)

## Clean up
drive_rm(logo)
# }