drive_user {googledrive}	R Documentation

Get info on current user

Description

Reveals information about the user associated with the current token. This is a thin wrapper around drive_about() that just extracts the most useful information (the information on current user) and prints it nicely.

Usage

drive_user(verbose = deprecated())

Arguments

verbose
This logical argument to individual googledrive functions is deprecated. To globally suppress googledrive messaging, use options(googledrive_quiet = TRUE) (the default behaviour is to emit informational messages). To suppress messaging in a more limited way, use the helpers local_drive_quiet() or with_drive_quiet().

Value

A list of class drive_user.

See Also

Wraps the about.get endpoint:

https://developers.google.com/drive/api/v3/reference/about/get

Examples

drive_user()

# more info is returned than is printed
user <- drive_user()
str(user)

[Package googledrive version 2.1.1 Index]