Build a request for the Google Drive v3 API

Build a request, using some knowledge of the Drive v3 API. Most users should, instead, use higher-level wrappers that facilitate common tasks, such as uploading or downloading Drive files. The functions here are intended for internal use and for programming around the Drive API.

generate_request(endpoint = character(), params = list(),
  token = drive_token(), .api_key = NULL)

build_request(path, method, params = list(), body = NULL, token = NULL,
  .api_key = NULL)

Arguments

endpoint
Character. Nickname for one of the selected Drive v3 API endpoints built into googledrive. Inspect via drive_endpoints().

params
Named list. Parameters destined for endpoint URL substitution, the query, or body.

token
Drive token, obtained from drive_auth().

.api_key
not in use yet

path
Character, e.g., "drive/v3/files/{fileId}". It can include variables inside curly brackets, as the example does, which are substituted using named parameters found in the params argument.

method
Character, should match an HTTP verb, e.g., GET, POST, PATCH or PUT

body
List, values to pass to the API request body.

Value

list() Components are method, path, query, body, and url, suitable as input for make_request(). The path is post-substitution and the query is a named list of all the non-body params that were not used during this substitution. url is the full URL after prepending the base URL for the Drive v3 API and appending an API key to the query.

Details

There are two functions:

generate_request() takes a nickname for an endpoint and uses the API spec to look up the path and method. The params are checked for validity and completeness with respect to the endpoint. It then passes things along to gs_build_request(). Use drive_endpoints() to see which endpoints can be accessed this way.

build_request() builds a request from explicit parts. It is quite dumb, only doing URL endpoint substitution and URL formation. It's up to the caller to make sure the path, method, params, and body are valid. Use this to call a Drive API endpoint that doesn't appear in the list returned by drive_endpoints().

Examples

req <- generate_request(
  "drive.files.get",
  list(fileId = "abc"),
  token = NULL
)
req

#> $method
#> [1] "GET"
#> 
#> $path
#> drive/v3/files/abc
#> 
#> $query
#> NULL
#> 
#> $body
#> NULL
#> 
#> $token
#> NULL
#> 
#> $.api_key
#> NULL
#> 
#> $url
#> [1] "https://www.googleapis.com/drive/v3/files/abc"
#> 

## re-create the previous request, but the hard way, i.e. "by hand"
req <- build_request(
  path = "drive/v3/files/{fileId}",
  method = "GET",
  list(fileId = "abc"),
  token = NULL
)
req

#> $method
#> [1] "GET"
#> 
#> $path
#> drive/v3/files/abc
#> 
#> $query
#> NULL
#> 
#> $body
#> NULL
#> 
#> $token
#> NULL
#> 
#> $.api_key
#> NULL
#> 
#> $url
#> [1] "https://www.googleapis.com/drive/v3/files/abc"
#> 

## call an endpoint not used by googledrive
## List a file's comments
## https://developers.google.com/drive/v3/reference/comments/list
not_run({
  req <- build_request(
    path = "drive/v3/files/{fileId}/comments",
    method = "GET",
    params = list(
      fileId = "file-id-goes-here",
      fields = "*"
    ),
    token = googledrive:::drive_token()
  )
  make_request(x)
})