shared_drive_create {googledrive}	R Documentation

Create a new shared drive

Description

A shared drive supports files owned by an organization rather than an individual user. Shared drives follow different sharing and ownership models from a specific user's "My Drive". Shared drives are the successors to the earlier concept of Team Drives. Learn more about shared drives.

Usage

shared_drive_create(name)

Arguments

name
Character. Name of the new shared drive. Must be non-empty and not entirely whitespace.

Value

An object of class dribble, a tibble with one row per shared drive.

See Also

Wraps the drives.create endpoint:

https://developers.google.com/drive/api/v3/reference/drives/create

Examples

## Not run: 
shared_drive_create("my-awesome-shared-drive")

# Clean up
shared_drive_rm("my-awesome-shared-drive")

## End(Not run)

[Package googledrive version 2.1.1 Index]