Resolve shortcuts to their targets

Retrieves the metadata for the Drive file that a shortcut refers to, i.e. the shortcut's target. The returned dribble has the usual columns (name, id, drive_resource), which refer to the target. It will also include the columns name_shortcut and id_shortcut, which refer to the original shortcut. There are 3 possible scenarios:

file is a shortcut and user can drive_get() the target. All is simple and well.

file is a shortcut, but drive_get() fails for the target. This can happen if the user can see the shortcut, but does not have read access to the target. It can also happen if the target has been trashed or deleted. In such cases, all of the target's metadata, except for id, will be missing. Call drive_get() on a problematic id to see the specific error.

file is not a shortcut. name_shortcut and id_shortcut will both be NA.

shortcut_resolve(file)

Arguments

file
Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

Value

An object of class dribble, a tibble with one row per item.

Examples

# Create a file to make a shortcut to
file <- drive_upload(
  drive_example("chicken.csv"),
  name = "chicken-sheet-for-shortcut",
  type = "spreadsheet"
)
#> Local file:
#> • /home/runner/work/_temp/Library/googledrive/extdata/chicken.csv
#> Uploaded into Drive file:
#> • chicken-sheet-for-shortcut <id: 1ElQ1dw9p_C-J4zFs07RYGKo-AjdqtNRYEeRZaY6Zwdo>
#> With MIME type:
#> • application/vnd.google-apps.spreadsheet

# Create a shortcut
sc1 <- file %>%
  shortcut_create(name = "shortcut-1")
#> Created Drive file:
#> • shortcut-1 <id: 1bbYdOdchY2V92inZYZkVGoHp5nrg3V5F>
#> With MIME type:
#> • application/vnd.google-apps.shortcut

# Create a second shortcut by copying the first
sc1 <- sc1 %>%
  drive_cp(name = "shortcut-2")
#> Original file:
#> • shortcut-1 <id: 1bbYdOdchY2V92inZYZkVGoHp5nrg3V5F>
#> Copied to file:
#> • shortcut-2 <id: 1upwYfbJxQxicz-DEC4HQL9ZTfPfR3oi9>

# Get the shortcuts
(sc_dat <- drive_find("-[12]$", type = "shortcut"))
#> # A dribble: 2 x 3
#>   name       id                                drive_resource   
#> * <chr>      <chr>                             <list>           
#> 1 shortcut-2 1upwYfbJxQxicz-DEC4HQL9ZTfPfR3oi9 <named list [30]>
#> 2 shortcut-1 1bbYdOdchY2V92inZYZkVGoHp5nrg3V5F <named list [30]>

# Resolve them
(resolved <- shortcut_resolve(sc_dat))
#> Resolved 2 shortcuts found in 2 files:
#> • chicken-sheet-for-shortcut <id: 1ElQ1dw9p_C-J4zFs07RYGKo-AjdqtNRYEeRZaY6Zwdo>
#> • chicken-sheet-for-shortcut <id: 1ElQ1dw9p_C-J4zFs07RYGKo-AjdqtNRYEeRZaY6Zwdo>
#> # A dribble: 2 x 5
#>   name         id                name_shortcut id_shortcut        drive_resource
#> * <chr>        <chr>             <chr>         <chr>              <list>        
#> 1 chicken-she… 1ElQ1dw9p_C-J4zF… shortcut-2    1upwYfbJxQxicz-DE… <named list […
#> 2 chicken-she… 1ElQ1dw9p_C-J4zF… shortcut-1    1bbYdOdchY2V92inZ… <named list […

resolved$id
#> [1] "1ElQ1dw9p_C-J4zFs07RYGKo-AjdqtNRYEeRZaY6Zwdo"
#> [2] "1ElQ1dw9p_C-J4zFs07RYGKo-AjdqtNRYEeRZaY6Zwdo"
file$id
#> [1] "1ElQ1dw9p_C-J4zFs07RYGKo-AjdqtNRYEeRZaY6Zwdo"

# Delete the target file
drive_rm(file)
#> File deleted:
#> • chicken-sheet-for-shortcut <id: 1ElQ1dw9p_C-J4zFs07RYGKo-AjdqtNRYEeRZaY6Zwdo>

# (Try to) resolve the shortcuts again
shortcut_resolve(sc_dat)
#> Resolved 0 of 2 shortcuts found in 2 files:
#> • NA <id: 1ElQ1dw9p_C-J4zFs07RYGKo-AjdqtNRYEeRZaY6Zwdo>
#> • NA <id: 1ElQ1dw9p_C-J4zFs07RYGKo-AjdqtNRYEeRZaY6Zwdo>
# No error, but resolution is unsuccessful due to non-existent target

# Clean-up
drive_rm(sc_dat)
#> Files deleted:
#> • shortcut-2 <id: 1upwYfbJxQxicz-DEC4HQL9ZTfPfR3oi9>
#> • shortcut-1 <id: 1bbYdOdchY2V92inZYZkVGoHp5nrg3V5F>