team_drive_create {googledrive}	R Documentation

Create a new Team Drive

Description

Note: Team Drives are only available to users of certain enhanced Google services, such as G Suite Enterprise, G Suite Business, or G Suite for Education.

Usage

team_drive_create(name, verbose = TRUE)

Arguments

name
Character. Name of the new Team Drive. Must be non-empty and not entirely whitespace.

verbose
Logical, indicating whether to print informative messages (default TRUE).

Value

An object of class dribble, a tibble with one row per item.

See Also

Wraps the teamdrives.create endpoint:

https://developers.google.com/drive/v3/reference/teamdrives/create

Examples

## Not run: 
team_drive_create("my-awesome-team-drive")

## clean up
team_drive_rm("my-awesome-team-drive")

## End(Not run)

[Package googledrive version 1.0.0 Index]