Get Team Drives by name or id

Retrieve metadata for Team Drives specified via name or id. Note that Google Drive does NOT behave like your local file system:

You can get zero, one, or more Team Drives back for each name! Team Drive names need not be unique.

Note: Team Drives are only available to users of certain enhanced Google services, such as G Suite Enterprise, G Suite Business, or G Suite for Education.

team_drive_get(name = NULL, id = NULL, verbose = TRUE)

Arguments

name
Character vector of names. A character vector marked with as_id() is treated as if it was provided via the id argument.

id
Character vector of Team Drive ids or URLs (it is first processed with as_id()). If both name and id are non-NULL, id is silently ignored.

verbose
Logical, indicating whether to print informative messages (default TRUE).

Value

An object of class dribble, a tibble with one row per item.

Examples

# NOT RUN {
team_drive_get("my-awesome-team-drive")
team_drive_get(c("apple", "orange", "banana"))
team_drive_get(as_id("KCmiHLXUk9PVA-0AJNG"))
team_drive_get(as_id("https://drive.google.com/drive/u/0/folders/KCmiHLXUk9PVA-0AJNG"))
team_drive_get(id = "KCmiHLXUk9PVA-0AJNG")
team_drive_get(id = "https://drive.google.com/drive/u/0/folders/KCmiHLXUk9PVA-0AJNG")
# }