as_shared_drive {googledrive}	R Documentation

Coerce to shared drive

Description

Converts various representations of a shared drive into a dribble, the object used by googledrive to hold Drive file metadata. Shared drives can be specified via

Name

Shared drive id, marked with as_id() to distinguish from name

Data frame or dribble consisting solely of shared drives

List representing Drives resource objects (mostly for internal use)

A shared drive supports files owned by an organization rather than an individual user. Shared drives follow different sharing and ownership models from a specific user's "My Drive". Shared drives are the successors to the earlier concept of Team Drives. Learn more about shared drives.

This is a generic function.

Usage

as_shared_drive(x, ...)

Arguments

x
A vector of shared drive names, a vector of shared drive ids marked with as_id(), a list of Drives resource objects, or a suitable data frame.

...
Other arguments passed down to methods. (Not used.)

Examples

## Not run: 
# specify the name
as_shared_drive("abc")

# specify the id (substitute one of your own!)
as_shared_drive(as_id("0AOPK1X2jaNckUk9PVA"))

## End(Not run)

[Package googledrive version 2.0.0 Index]