dribble object

googledrive stores the metadata for one or more Drive files or Team Drives as a dribble. It is a "Drive tibble" with one row per file or Team Drive and, at a minimum, these variables:

name: a character variable containing file or Team Drive names

id: a character variable of file or Team Drive ids

drive_resource: a list-column, each element of which is either a Files resource or Team Drive resource object. Note there is no guarantee that all documented fields are always present. We do check if the kind field is present and equal to one of drive#file or drive#teamDrive.

The dribble format is handy because it exposes the file name, which is good for humans, but keeps it bundled with the file's unique id and other metadata, which are needed for API calls.

In general, the dribble class will be retained even after subsetting, as long as the required variables are present and of the correct type.

See also

as_dribble()