Check facts about a dribble

Sometimes you need to check things about a dribble` or about the files it represents, such as:

Is it even a dribble?

Size: Does the dribble hold exactly one file? At least one file? No file?

File type: Is this file a folder?

File ownership and access: Is it mine? Published? Shared?

is_dribble(d)

no_file(d)

single_file(d)

some_files(d)

confirm_dribble(d)

confirm_single_file(d)

confirm_some_files(d)

is_folder(d)

is_native(d)

is_parental(d)

is_mine(d)

is_team_drive(d)

is_team_drivy(d)

Arguments

d
A dribble.

Examples

# NOT RUN {
## most of us have multiple files or folders on Google Drive
d <- drive_find()
is_dribble(d)
no_file(d)
single_file(d)
some_files(d)
confirm_single_file(d)
confirm_some_files(d)
is_folder(d)
is_mine(d)
# }