Check facts about a dribble

Sometimes you need to check things about a dribble` or about the files it represents, such as:

Is it even a dribble?

Size: Does the dribble hold exactly one file? At least one file? No file?

File type: Is this file a folder?

File ownership and access: Is it mine? Published? Shared?

is_dribble(d)

no_file(d)

single_file(d)

some_files(d)

confirm_dribble(d)

confirm_single_file(d)

confirm_some_files(d)

is_folder(d)

is_shortcut(d)

is_folder_shortcut(d)

is_native(d)

is_parental(d)

is_mine(d)

is_shared_drive(d)

Arguments

d
A dribble.

Examples

## most of us have multiple files or folders on Google Drive
d <- drive_find()
is_dribble(d)
#> [1] TRUE
no_file(d)
#> [1] FALSE
single_file(d)
#> [1] FALSE
some_files(d)
#> [1] TRUE

# this will error
# confirm_single_file(d)

confirm_some_files(d)
#> # A dribble: 17 × 3
#>    name                id                                           drive_resource
#>    <chr>               <drv_id>                                     <list>        
#>  1 bravo               1Lbnr5CXFtJocrr-u3MIRBO8zbkvkUfXJ            <named list […
#>  2 chicken_sheet       1s0kEHcqG2PyciERoGq52L_Qwzp4y3__rBVKSx7E_j88 <named list […
#>  3 r_logo.jpg          1wFAZdmBiSRu4GShsqurxD7wIDSCZvPud            <named list […
#>  4 THANKS              19URV7BT0_E1KhYdfDODszK5aiELOwTSz            <named list […
#>  5 googledrive-NEWS.md 1h1lhFfQrDZevE2OEX10-rbi2BfvGogFm            <named list […
#>  6 def                 1ALSW_Nqs7FsPOcrJ6MqyBoRm03gansmn            <named list […
#>  7 abc                 1o89YN5n4325GbUA86Wp6pRH3dsTsE5iC            <named list […
#>  8 BioC_mirrors.csv    13tMFbhAHoeHLFS5xu19GbDjf6GWJSxyN            <named list […
#>  9 Rlogo.svg           1lCQGxjyoc9mQz719I8sKil_m2Nuhw0Fq            <named list […
#> 10 DESCRIPTION         1KKYhtcdJMKh4WYeri5TOPEeAtzdN_cqV            <named list […
#> 11 r_about.html        1mHtQhvJyDk5dX9ktKbeIoVW-wwWK0__N            <named list […
#> 12 imdb_latin1.csv     1S5HxY7a-Jb_fV4C3T6fkGyPpXfI_yb4w            <named list […
#> 13 chicken.txt         1xMvlJHia_qYNZmucaStDcOF9A9PD4BOT            <named list […
#> 14 chicken.pdf         1au0aK6YCTra2sucTRus8ZaUhbaLpinTn            <named list […
#> 15 chicken.jpg         1-BF1c4kWCkkByQbcLT-b2Hv6vnVsbqa_            <named list […
#> 16 chicken.csv         12212CXY_TopUMIKYu_l8hU5UXI8lrzQF            <named list […
#> 17 chicken_doc         11GY4Q4BUG3m5U4CnZP564lYvGydvZe2XZOkwCfxnTag <named list […
is_folder(d)
#>  [1] FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [13] FALSE FALSE FALSE FALSE FALSE
is_mine(d)
#>  [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> [16] TRUE TRUE