Check facts about a dribble

Sometimes you need to check things about a dribble` or about the files it represents, such as:

Is it even a dribble?

Size: Does the dribble hold exactly one file? At least one file? No file?

File type: Is this file a folder?

File ownership and access: Is it mine? Published? Shared?

is_dribble(d)

no_file(d)

single_file(d)

some_files(d)

confirm_dribble(d)

confirm_single_file(d)

confirm_some_files(d)

is_folder(d)

is_native(d)

is_parental(d)

is_mine(d)

is_shared_drive(d)

Arguments

d
A dribble.

Examples

## most of us have multiple files or folders on Google Drive
d <- drive_find()
is_dribble(d)

#> [1] TRUE

no_file(d)

#> [1] FALSE

single_file(d)

#> [1] FALSE

some_files(d)

#> [1] TRUE

# this will error
# confirm_single_file(d)

confirm_some_files(d)

#> # A tibble: 14 x 3
#>    name                id                                drive_resource   
#>  * <chr>               <chr>                             <list>           
#>  1 bravo               1HEIvO1zWgQvAbIJ6lRT2McAsA_0sp6bM <named list [38]>
#>  2 Rlogo.pdf           1opPZyo4WUiue56qlF8SO5ipaTJwPLzZg <named list [39]>
#>  3 THANKS              1JCQ06Wj6AjntiyjnHsXmmcHZWCkBNgBJ <named list [39]>
#>  4 googledrive-NEWS.md 1wMnFwFSIG0eQ4UUUEbn6-udF3GoKY3xN <named list [38]>
#>  5 def                 12xjsE3eIf84VSFF5hNhSEOnQF8DDSHkc <named list [32]>
#>  6 abc                 1qkS667DEnvDzTuVBSyj5NeGzIGbrAy9Q <named list [32]>
#>  7 BioC_mirrors.csv    1f-Ua7sUZq1g5YycA1mXwlVRrYn1lN15o <named list [38]>
#>  8 logo.jpg            1yhwnYA0TfXZ5lykclxIPyjvSBLL3h_Y0 <named list [40]>
#>  9 Rlogo.svg           1vhSTljYBXwOqILwgZ8hC3XKzVC98I5x3 <named list [40]>
#> 10 DESCRIPTION         1U4iwGpIa1SVPuiki3bs2fS-EltIvL3JD <named list [39]>
#> 11 chicken.txt         1af6w2ZU-JjwI7jrKW1OGiiRbM2GMTbil <named list [39]>
#> 12 chicken.pdf         1nL7iro5YQLt1d77acGiHoawmPTWonh6p <named list [39]>
#> 13 chicken.jpg         1QQhNDl_f2W-UxacIfCBaftUopfqi1Tw1 <named list [40]>
#> 14 chicken.csv         1OaBk2TpNJXVafrO5IoRlJj7ICFj3irxG <named list [38]>

is_folder(d)

#>  [1] FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#> [13] FALSE FALSE

is_mine(d)

#>  [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE