Copy a Drive file

Copies an existing Drive file into a new file id.

drive_cp(
  file,
  path = NULL,
  name = NULL,
  ...,
  overwrite = NA,
  verbose = deprecated()
)

Arguments

file
Something that identifies the file of interest on your Google Drive. Can be a name or path, a file id or URL marked with as_id(), or a dribble.

path
Specifies target destination for the new file on Google Drive. Can be an actual path (character), a file id marked with as_id(), or a dribble.

If path is a shortcut to a folder, it is automatically resolved to its target folder.

If path is given as a path (as opposed to a dribble or an id), it is best to explicitly indicate if it's a folder by including a trailing slash, since it cannot always be worked out from the context of the call. By default, the new file has the same parent folder as the source file.

name
Character, new file name if not specified as part of path. This will force path to be interpreted as a folder, even if it is character and lacks a trailing slash. Defaults to "Copy of FILE-NAME".

...
Named parameters to pass along to the Drive API. Has dynamic dots semantics. You can affect the metadata of the target file by specifying properties of the Files resource via .... Read the "Request body" section of the Drive API docs for the associated endpoint to learn about relevant parameters.

overwrite
Logical, indicating whether to check for a pre-existing file at the targetted "filepath". The quotes around "filepath" refer to the fact that Drive does not impose a 1-to-1 relationship between filepaths and files, like a typical file system; read more about that in drive_get().

NA (default): Just do the operation, even if it results in multiple files with the same filepath.

TRUE: Check for a pre-existing file at the filepath. If there is zero or one, move a pre-existing file to the trash, then carry on. Note that the new file does not inherit any properties from the old one, such as sharing or publishing settings. It will have a new file ID. An error is thrown if two or more pre-existing files are found.

FALSE: Error if there is any pre-existing file at the filepath.

Note that existence checks, based on filepath, are expensive operations, i.e. they require additional API calls.

verbose
This logical argument to individual googledrive functions is deprecated. To globally suppress googledrive messaging, use options(googledrive_quiet = TRUE) (the default behaviour is to emit informational messages). To suppress messaging in a more limited way, use the helpers local_drive_quiet() or with_drive_quiet().

Value

An object of class dribble, a tibble with one row per file.

See also

Examples

# Create a file to copy
file <- drive_upload(drive_example("chicken.txt"), "chicken-cp.txt")
#> Local file:
#> • /home/runner/work/_temp/Library/googledrive/extdata/chicken.txt
#> Uploaded into Drive file:
#> • chicken-cp.txt <id: 1NbuayGmpYXrD3w9CAeprWTZw-3AH7pFS>
#> With MIME type:
#> • text/plain

# Make a "Copy of" copy in same folder as the original
drive_cp("chicken-cp.txt")
#> Original file:
#> • chicken-cp.txt <id: 1NbuayGmpYXrD3w9CAeprWTZw-3AH7pFS>
#> Copied to file:
#> • Copy of chicken-cp.txt <id: 1UfXPiT-nfh9ouEHIocA6qyr4nv4ymqW5>

# Make an explicitly named copy in same folder as the original
drive_cp("chicken-cp.txt", "chicken-cp-two.txt")
#> Original file:
#> • chicken-cp.txt <id: 1NbuayGmpYXrD3w9CAeprWTZw-3AH7pFS>
#> Copied to file:
#> • chicken-cp-two.txt <id: 1eXdRhVe4trrePLi7FzEx4mR0-a6eCezd>

# Make an explicitly named copy in a different folder
folder <- drive_mkdir("new-folder")
#> Created Drive file:
#> • new-folder <id: 1QAtAkphdYDaVwb8kPKVgPpudXdLZ3u9f>
#> With MIME type:
#> • application/vnd.google-apps.folder
drive_cp("chicken-cp.txt", path = folder, name = "chicken-cp-three.txt")
#> Original file:
#> • chicken-cp.txt <id: 1NbuayGmpYXrD3w9CAeprWTZw-3AH7pFS>
#> Copied to file:
#> • new-folder/chicken-cp-three.txt <id: 1CLYD1IfZw6mQlOCrfiS7VWU5W-aY3dp3>

# Make an explicitly named copy and star it.
# The starring is an example of providing metadata via `...`.
# `starred` is not an actual argument to `drive_cp()`,
# it just gets passed through to the API.
x <- drive_cp("chicken-cp.txt", name = "chicken-cp-starred.txt", starred = TRUE)
#> Original file:
#> • chicken-cp.txt <id: 1NbuayGmpYXrD3w9CAeprWTZw-3AH7pFS>
#> Copied to file:
#> • chicken-cp-starred.txt <id: 10F8IglqUpkwR_wu2bb5HIN0QAabwijzd>
drive_reveal(x, "starred")
#> # A dribble: 1 x 4
#>   name                  starred id                              drive_resource  
#> * <chr>                 <lgl>   <chr>                           <list>          
#> 1 chicken-cp-starred.t… TRUE    10F8IglqUpkwR_wu2bb5HIN0QAabwi… <named list [39…

# `overwrite = FALSE` errors if file already exists at target filepath
# THIS WILL ERROR!
# drive_cp("chicken-cp.txt", name = "chicken-cp.txt", overwrite = FALSE)

# `overwrite = TRUE` moves an existing file to trash, then proceeds
drive_cp("chicken-cp.txt", name = "chicken-cp.txt", overwrite = TRUE)
#> File trashed:
#> • chicken-cp.txt <id: 1NbuayGmpYXrD3w9CAeprWTZw-3AH7pFS>
#> Original file:
#> • chicken-cp.txt <id: 1NbuayGmpYXrD3w9CAeprWTZw-3AH7pFS>
#> Copied to file:
#> • chicken-cp.txt <id: 1zi-8_Me4it9-mAeBqVTf8lW7RE7GgwE8>

# Behold all of our copies!
drive_find("chicken-cp")
#> # A dribble: 5 x 3
#>   name                   id                                drive_resource   
#> * <chr>                  <chr>                             <list>           
#> 1 chicken-cp.txt         1zi-8_Me4it9-mAeBqVTf8lW7RE7GgwE8 <named list [39]>
#> 2 chicken-cp-starred.txt 10F8IglqUpkwR_wu2bb5HIN0QAabwijzd <named list [40]>
#> 3 chicken-cp-three.txt   1CLYD1IfZw6mQlOCrfiS7VWU5W-aY3dp3 <named list [40]>
#> 4 chicken-cp-two.txt     1eXdRhVe4trrePLi7FzEx4mR0-a6eCezd <named list [40]>
#> 5 Copy of chicken-cp.txt 1UfXPiT-nfh9ouEHIocA6qyr4nv4ymqW5 <named list [40]>

# Delete all of our copies and the new folder!
drive_find("chicken-cp") %>% drive_rm()
#> Files deleted:
#> • chicken-cp.txt <id: 1zi-8_Me4it9-mAeBqVTf8lW7RE7GgwE8>
#> • chicken-cp-starred.txt <id: 10F8IglqUpkwR_wu2bb5HIN0QAabwijzd>
#> • chicken-cp-three.txt <id: 1CLYD1IfZw6mQlOCrfiS7VWU5W-aY3dp3>
#> • chicken-cp-two.txt <id: 1eXdRhVe4trrePLi7FzEx4mR0-a6eCezd>
#> • Copy of chicken-cp.txt <id: 1UfXPiT-nfh9ouEHIocA6qyr4nv4ymqW5>
drive_find("new-folder") %>% drive_rm(folder)
#> File deleted:
#> • new-folder <id: 1QAtAkphdYDaVwb8kPKVgPpudXdLZ3u9f>

# upload a csv file to copy
csv_file <- drive_upload(drive_example("chicken.csv"))
#> Local file:
#> • /home/runner/work/_temp/Library/googledrive/extdata/chicken.csv
#> Uploaded into Drive file:
#> • chicken.csv <id: 1zShIlUrabaOw-AoVdNpEqAbUDRQ7kn_5>
#> With MIME type:
#> • text/csv

# copy AND AT THE SAME TIME convert it to a Google Sheet
chicken_sheet <- drive_cp(
  csv_file,
  name = "chicken-cp",
  mime_type = drive_mime_type("spreadsheet")
)
#> Original file:
#> • chicken.csv <id: 1zShIlUrabaOw-AoVdNpEqAbUDRQ7kn_5>
#> Copied to file:
#> • chicken-cp <id: 13m4vSA5wJE0y2hGutMn4WrPb2FPriin395l_PMrQ3AI>
# is it really a Google Sheet?
drive_reveal(chicken_sheet, "mime_type")$mime_type
#> [1] "application/vnd.google-apps.spreadsheet"

# go see the new Sheet in the browser
# drive_browse(chicken_sheet)

# clean up
drive_rm(csv_file, chicken_sheet)
#> Files deleted:
#> • chicken.csv <id: 1zShIlUrabaOw-AoVdNpEqAbUDRQ7kn_5>
#> • chicken-cp <id: 13m4vSA5wJE0y2hGutMn4WrPb2FPriin395l_PMrQ3AI>