Create a new blank Drive file

Creates a new blank Drive file. Note there are better options for these special cases:

Creating a folder? Use drive_mkdir().

Want to upload existing local content into a new Drive file? Use drive_upload().

Usage

drive_create(
  name,
  path = NULL,
  type = NULL,
  ...,
  overwrite = NA,
  verbose = deprecated()
)

Arguments

name

Name for the new file or, optionally, a path that specifies an existing parent folder, as well as the new file name.

path

Target destination for the new item, i.e. a folder or a shared drive. Can be given as an actual path (character), a file id or URL marked with as_id(), or a dribble. Defaults to your "My Drive" root folder. If path is a shortcut to a folder, it is automatically resolved to its target folder.

type

Character. Create a blank Google Doc, Sheet or Slides by setting type to document, spreadsheet, or presentation, respectively. All non-NULL values for type are pre-processed with drive_mime_type().

...

Named parameters to pass along to the Drive API. Has dynamic dots semantics. You can affect the metadata of the target file by specifying properties of the Files resource via .... Read the "Request body" section of the Drive API docs for the associated endpoint to learn about relevant parameters.

overwrite

Logical, indicating whether to check for a pre-existing file at the targetted "filepath". The quotes around "filepath" refer to the fact that Drive does not impose a 1-to-1 relationship between filepaths and files, like a typical file system; read more about that in drive_get().

NA (default): Just do the operation, even if it results in multiple files with the same filepath.

TRUE: Check for a pre-existing file at the filepath. If there is zero or one, move a pre-existing file to the trash, then carry on. Note that the new file does not inherit any properties from the old one, such as sharing or publishing settings. It will have a new file ID. An error is thrown if two or more pre-existing files are found.

FALSE: Error if there is any pre-existing file at the filepath.

Note that existence checks, based on filepath, are expensive operations, i.e. they require additional API calls.

verbose

This logical argument to individual googledrive functions is deprecated. To globally suppress googledrive messaging, use options(googledrive_quiet = TRUE) (the default behaviour is to emit informational messages). To suppress messaging in a more limited way, use the helpers local_drive_quiet() or with_drive_quiet().

Value

An object of class dribble, a tibble with one row per file.

See also

Examples

# Create a blank Google Doc named 'WordStar' in
# your 'My Drive' root folder and star it
wordstar <- drive_create("WordStar", type = "document", starred = TRUE)
#> Created Drive file:
#> • WordStar <id: 1PEu0soXmU34yEjl7AUPk-ndCBlgS7tahvmPjK0FO8ws>
#> With MIME type:
#> • application/vnd.google-apps.document

# is 'WordStar' really starred? YES
purrr::pluck(wordstar, "drive_resource", 1, "starred")
#> [1] TRUE

# Create a blank Google Slides presentation in
# the root folder, and set its description
execuvision <- drive_create(
  "ExecuVision",
  type = "presentation",
  description = "deeply nested bullet lists FTW"
)
#> Created Drive file:
#> • ExecuVision <id: 1uGDTgSNmzVGZBtXli88RVd3ceGsCYIidCw2q0jyHkn0>
#> With MIME type:
#> • application/vnd.google-apps.presentation

# Did we really set the description? YES
purrr::pluck(execuvision, "drive_resource", 1, "description")
#> [1] "deeply nested bullet lists FTW"

# check out the new presentation
drive_browse(execuvision)

# Create folder 'b4xl' in the root folder,
# then create an empty new Google Sheet in it
b4xl <- drive_mkdir("b4xl")
#> Error in confirm_clear_path(path, name): Unclear if `path` specifies parent folder or full path to the new
#> file, including its name.
#> See `?as_dribble()` for advice on how to make this clear.
drive_create("VisiCalc", path = b4xl, type = "spreadsheet")
#> Error in drive_create("VisiCalc", path = b4xl, type = "spreadsheet"): object 'b4xl' not found

# Another way to create a Google Sheet in the folder 'b4xl'
drive_create("b4xl/SuperCalc", type = "spreadsheet")
#> Created Drive file:
#> • SuperCalc <id: 1JwUckd6JV_Xmt3hij5uA-iHfhPy-tyi41AgEoJDPyNo>
#> With MIME type:
#> • application/vnd.google-apps.spreadsheet

# Yet another way to create a new file in a folder,
# this time specifying parent `path` as a character
drive_create("Lotus 1-2-3", path = "b4xl", type = "spreadsheet")
#> Created Drive file:
#> • Lotus 1-2-3 <id: 1_4FhKeNbjUNTHakqjRKLPLwho0UYbPBNebulgr1y1cE>
#> With MIME type:
#> • application/vnd.google-apps.spreadsheet

# Did we really create those Sheets in the intended folder? YES
drive_ls("b4xl")
#> # A dribble: 341 × 3
#>    name        id                                           drive_resource
#>    <chr>       <drv_id>                                     <list>        
#>  1 Lotus 1-2-3 1_4FhKeNbjUNTHakqjRKLPLwho0UYbPBNebulgr1y1cE <named list>  
#>  2 SuperCalc   1JwUckd6JV_Xmt3hij5uA-iHfhPy-tyi41AgEoJDPyNo <named list>  
#>  3 Lotus 1-2-3 18lorjZOu3pH7fgjHb2eBJqOpwXgOj-UYeRLHZAFJep0 <named list>  
#>  4 SuperCalc   1uRVveS6BgHpVS2We-K003ROzzag_bsc14OhrHn0MFwM <named list>  
#>  5 Lotus 1-2-3 1B-kfjyLNPjFAuudFQeRf_Ihu6VSCHeacp6anrbRwXCs <named list>  
#>  6 SuperCalc   1qinh1_Fqdq3Gi2qyjo_QcJED_2KIErkDWbvj9FqAOs0 <named list>  
#>  7 Lotus 1-2-3 18dabsDHcqyu1QwXTJveHjQgkuzzFZZEKyOKD-9TTCAo <named list>  
#>  8 SuperCalc   1LySsaxTXaDw1Soe_1q9SLADG0tqL1vs-LJYAh84IMiE <named list>  
#>  9 Lotus 1-2-3 1LYNOBotPEYbeZQ2in_RBxkWVcazQU-mUI77cOUMsFVo <named list>  
#> 10 SuperCalc   1AaCt508FT5sfqxybwdt5Dv9EI4dcNHESnV7m53876HY <named list>  
#> # … with 331 more rows

# `overwrite = FALSE` errors if file already exists at target filepath
# THIS WILL ERROR!
drive_create("VisiCalc", path = b4xl, overwrite = FALSE)
#> Error in drive_create("VisiCalc", path = b4xl, overwrite = FALSE): object 'b4xl' not found

# `overwrite = TRUE` moves an existing file to trash, then proceeds
drive_create("VisiCalc", path = b4xl, overwrite = TRUE)
#> Error in drive_create("VisiCalc", path = b4xl, overwrite = TRUE): object 'b4xl' not found

# Clean up
drive_rm(wordstar, b4xl, execuvision)
#> Error in drive_rm(wordstar, b4xl, execuvision): object 'b4xl' not found