Download a file from Drive.

This function downloads files from Google Drive. Native Google files, such as Google Docs, Google Sheets, Google Slides, must be exported to a conventional local file type. This can be specified explicitly via type or, otherwise, implicitly via the file extension of path, if provided. If all else fails, a valid default is used. Native Google files can be downloaded to types specified in the Drive API documentation.

drive_download(file, path = NULL, type = NULL, overwrite = FALSE,
  verbose = TRUE)

Arguments

file
Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

path
Character. Path for output file. If absent, the default file name is the file's name on Google Drive and the default location is working directory, possibly with an added file extension.

type
Character. Only consulted if file is a native Google file. Specifies the desired type of the downloaded file. Will be processed via drive_mime_type(), so either a file extension like "pdf" or a full MIME type like "application/pdf" is acceptable.

overwrite
A logical scalar. If path already exists, do you want to overwrite it?

verbose
Logical, indicating whether to print informative messages (default TRUE).

Examples

# NOT RUN {
## Save "chickwts.csv" to the working directory as "chickwts.csv".
drive_download(file = "chickwts.csv")

## Export a Google Sheet named "chickwts" to the working directory as
## "chickwts.csv".
drive_download(file = "chickwts", path = "chickwts.csv")

## This will also export a Google Sheet named "chickwts" to the working
## directory as "chickwts.csv".
drive_download(file = "chickwts", type = "csv")

## Export a Google Document named "foobar" to the working directory as
## "foobar.docx".
drive_download(file = "foobar", path = "foobar.docx")

## This will also export a Google Document named "foobar" to the working
## directory as "foobar.docx".
drive_download(file = "foobar", type = "docx")
# }