List Drive endpoints

Returns a list of selected Drive API v3 endpoints, as stored inside the googledrive package. The names of this list (or the id sub-elements) are the nicknames that can be used to specify an endpoint in generate_request(). For each endpoint, we store its nickname or id, the associated HTTP verb, the path, and details about the parameters. This list is derived programmatically from the Drive API v3 Discovery Document.

drive_endpoints(i = NULL)

Arguments

i
The name(s) or integer index(ices) of the endpoints to return. Optional. By default, the entire list is returned.

Value

A list containing some or all of the subset of the Drive API v3 endpoints that are used internally by googledrive.

Examples

str(drive_endpoints(), max.level = 2)

#> List of 22
#>  $ drive.about.get                   :List of 4
#>   ..$ id        : chr "drive.about.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/about"
#>   ..$ parameters:List of 7
#>  $ drive.files.copy                  :List of 4
#>   ..$ id        : chr "drive.files.copy"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "drive/v3/files/{fileId}/copy"
#>   ..$ parameters:List of 63
#>  $ drive.files.create                :List of 4
#>   ..$ id        : chr "drive.files.create"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "drive/v3/files"
#>   ..$ parameters:List of 63
#>  $ drive.files.delete                :List of 4
#>   ..$ id        : chr "drive.files.delete"
#>   ..$ method    : chr "DELETE"
#>   ..$ path      : chr "drive/v3/files/{fileId}"
#>   ..$ parameters:List of 9
#>  $ drive.files.emptyTrash            :List of 4
#>   ..$ id        : chr "drive.files.emptyTrash"
#>   ..$ method    : chr "DELETE"
#>   ..$ path      : chr "drive/v3/files/trash"
#>   ..$ parameters:List of 7
#>  $ drive.files.export                :List of 4
#>   ..$ id        : chr "drive.files.export"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/export"
#>   ..$ parameters:List of 9
#>  $ drive.files.generateIds           :List of 4
#>   ..$ id        : chr "drive.files.generateIds"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/generateIds"
#>   ..$ parameters:List of 9
#>  $ drive.files.get                   :List of 4
#>   ..$ id        : chr "drive.files.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}"
#>   ..$ parameters:List of 10
#>  $ drive.files.list                  :List of 4
#>   ..$ id        : chr "drive.files.list"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files"
#>   ..$ parameters:List of 17
#>  $ drive.files.update                :List of 4
#>   ..$ id        : chr "drive.files.update"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "drive/v3/files/{fileId}"
#>   ..$ parameters:List of 65
#>  $ drive.files.watch                 :List of 4
#>   ..$ id        : chr "drive.files.watch"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "drive/v3/files/{fileId}/watch"
#>   ..$ parameters:List of 10
#>  $ drive.permissions.create          :List of 4
#>   ..$ id        : chr "drive.permissions.create"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions"
#>   ..$ parameters:List of 24
#>  $ drive.permissions.delete          :List of 4
#>   ..$ id        : chr "drive.permissions.delete"
#>   ..$ method    : chr "DELETE"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions/{permissionId}"
#>   ..$ parameters:List of 10
#>  $ drive.permissions.get             :List of 4
#>   ..$ id        : chr "drive.permissions.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions/{permissionId}"
#>   ..$ parameters:List of 10
#>  $ drive.permissions.list            :List of 4
#>   ..$ id        : chr "drive.permissions.list"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions"
#>   ..$ parameters:List of 11
#>  $ drive.permissions.update          :List of 4
#>   ..$ id        : chr "drive.permissions.update"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions/{permissionId}"
#>   ..$ parameters:List of 24
#>  $ drive.revisions.delete            :List of 4
#>   ..$ id        : chr "drive.revisions.delete"
#>   ..$ method    : chr "DELETE"
#>   ..$ path      : chr "drive/v3/files/{fileId}/revisions/{revisionId}"
#>   ..$ parameters:List of 9
#>  $ drive.revisions.get               :List of 4
#>   ..$ id        : chr "drive.revisions.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/revisions/{revisionId}"
#>   ..$ parameters:List of 10
#>  $ drive.revisions.list              :List of 4
#>   ..$ id        : chr "drive.revisions.list"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/revisions"
#>   ..$ parameters:List of 10
#>  $ drive.revisions.update            :List of 4
#>   ..$ id        : chr "drive.revisions.update"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "drive/v3/files/{fileId}/revisions/{revisionId}"
#>   ..$ parameters:List of 21
#>  $ drive.files.update.media          :List of 4
#>   ..$ id        : chr "drive.files.update.media"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "/upload/drive/v3/files/{fileId}"
#>   ..$ parameters:List of 15
#>  $ drive.files.update.media.resumable:List of 4
#>   ..$ id        : chr "drive.files.update.media.resumable"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "/resumable/upload/drive/v3/files/{fileId}"
#>   ..$ parameters:List of 15

drive_endpoints("drive.files.delete")

#> $drive.files.delete
#> $drive.files.delete$id
#> [1] "drive.files.delete"
#> 
#> $drive.files.delete$method
#> [1] "DELETE"
#> 
#> $drive.files.delete$path
#> [1] "drive/v3/files/{fileId}"
#> 
#> $drive.files.delete$parameters
#> $drive.files.delete$parameters$fileId
#> $drive.files.delete$parameters$fileId$location
#> [1] "path"
#> 
#> $drive.files.delete$parameters$fileId$required
#> [1] TRUE
#> 
#> $drive.files.delete$parameters$fileId$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$fileId$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$fileId$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$fileId$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$fileId$description
#> [1] "The ID of the file."
#> 
#> 
#> $drive.files.delete$parameters$supportsTeamDrives
#> $drive.files.delete$parameters$supportsTeamDrives$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$type
#> [1] "boolean"
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$description
#> [1] "Whether the requesting application supports Team Drives."
#> 
#> 
#> $drive.files.delete$parameters$alt
#> $drive.files.delete$parameters$alt$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$alt$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$alt$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$alt$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$alt$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$alt$enum
#> [1] "json"
#> 
#> $drive.files.delete$parameters$alt$description
#> [1] "Data format for the response."
#> 
#> 
#> $drive.files.delete$parameters$fields
#> $drive.files.delete$parameters$fields$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$fields$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$fields$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$fields$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$fields$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$fields$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$fields$description
#> [1] "Selector specifying which fields to include in a partial response."
#> 
#> 
#> $drive.files.delete$parameters$key
#> $drive.files.delete$parameters$key$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$key$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$key$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$key$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$key$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$key$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$key$description
#> [1] "API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token."
#> 
#> 
#> $drive.files.delete$parameters$oauth_token
#> $drive.files.delete$parameters$oauth_token$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$oauth_token$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$oauth_token$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$oauth_token$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$oauth_token$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$oauth_token$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$oauth_token$description
#> [1] "OAuth 2.0 token for the current user."
#> 
#> 
#> $drive.files.delete$parameters$prettyPrint
#> $drive.files.delete$parameters$prettyPrint$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$prettyPrint$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$prettyPrint$type
#> [1] "boolean"
#> 
#> $drive.files.delete$parameters$prettyPrint$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$prettyPrint$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$prettyPrint$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$prettyPrint$description
#> [1] "Returns response with indentations and line breaks."
#> 
#> 
#> $drive.files.delete$parameters$quotaUser
#> $drive.files.delete$parameters$quotaUser$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$quotaUser$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$quotaUser$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$quotaUser$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$quotaUser$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$quotaUser$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$quotaUser$description
#> [1] "Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided."
#> 
#> 
#> $drive.files.delete$parameters$userIp
#> $drive.files.delete$parameters$userIp$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$userIp$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$userIp$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$userIp$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$userIp$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$userIp$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$userIp$description
#> [1] "IP address of the site where the request originates. Use this if you want to enforce per-user limits."
#> 
#> 
#> 
#> 

drive_endpoints(4)

#> $drive.files.delete
#> $drive.files.delete$id
#> [1] "drive.files.delete"
#> 
#> $drive.files.delete$method
#> [1] "DELETE"
#> 
#> $drive.files.delete$path
#> [1] "drive/v3/files/{fileId}"
#> 
#> $drive.files.delete$parameters
#> $drive.files.delete$parameters$fileId
#> $drive.files.delete$parameters$fileId$location
#> [1] "path"
#> 
#> $drive.files.delete$parameters$fileId$required
#> [1] TRUE
#> 
#> $drive.files.delete$parameters$fileId$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$fileId$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$fileId$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$fileId$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$fileId$description
#> [1] "The ID of the file."
#> 
#> 
#> $drive.files.delete$parameters$supportsTeamDrives
#> $drive.files.delete$parameters$supportsTeamDrives$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$type
#> [1] "boolean"
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$description
#> [1] "Whether the requesting application supports Team Drives."
#> 
#> 
#> $drive.files.delete$parameters$alt
#> $drive.files.delete$parameters$alt$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$alt$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$alt$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$alt$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$alt$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$alt$enum
#> [1] "json"
#> 
#> $drive.files.delete$parameters$alt$description
#> [1] "Data format for the response."
#> 
#> 
#> $drive.files.delete$parameters$fields
#> $drive.files.delete$parameters$fields$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$fields$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$fields$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$fields$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$fields$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$fields$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$fields$description
#> [1] "Selector specifying which fields to include in a partial response."
#> 
#> 
#> $drive.files.delete$parameters$key
#> $drive.files.delete$parameters$key$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$key$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$key$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$key$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$key$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$key$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$key$description
#> [1] "API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token."
#> 
#> 
#> $drive.files.delete$parameters$oauth_token
#> $drive.files.delete$parameters$oauth_token$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$oauth_token$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$oauth_token$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$oauth_token$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$oauth_token$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$oauth_token$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$oauth_token$description
#> [1] "OAuth 2.0 token for the current user."
#> 
#> 
#> $drive.files.delete$parameters$prettyPrint
#> $drive.files.delete$parameters$prettyPrint$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$prettyPrint$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$prettyPrint$type
#> [1] "boolean"
#> 
#> $drive.files.delete$parameters$prettyPrint$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$prettyPrint$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$prettyPrint$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$prettyPrint$description
#> [1] "Returns response with indentations and line breaks."
#> 
#> 
#> $drive.files.delete$parameters$quotaUser
#> $drive.files.delete$parameters$quotaUser$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$quotaUser$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$quotaUser$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$quotaUser$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$quotaUser$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$quotaUser$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$quotaUser$description
#> [1] "Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided."
#> 
#> 
#> $drive.files.delete$parameters$userIp
#> $drive.files.delete$parameters$userIp$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$userIp$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$userIp$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$userIp$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$userIp$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$userIp$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$userIp$description
#> [1] "IP address of the site where the request originates. Use this if you want to enforce per-user limits."
#> 
#> 
#> 
#>