List Drive endpoints

Returns a list of selected Drive API v3 endpoints, as stored inside the googledrive package. The names of this list (or the id sub-elements) are the nicknames that can be used to specify an endpoint in generate_request(). For each endpoint, we store its nickname or id, the associated HTTP verb, the path, and details about the parameters. This list is derived programmatically from the Drive API v3 Discovery Document.

drive_endpoints(i = NULL)

Arguments

i
The name(s) or integer index(ices) of the endpoints to return. Optional. By default, the entire list is returned.

Value

A list containing some or all of the subset of the Drive API v3 endpoints that are used internally by googledrive.

Examples

str(drive_endpoints(), max.level = 2)

#> List of 27
#>  $ drive.about.get         :List of 4
#>   ..$ id        : chr "drive.about.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/about"
#>   ..$ parameters:List of 7
#>  $ drive.files.copy        :List of 4
#>   ..$ id        : chr "drive.files.copy"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "drive/v3/files/{fileId}/copy"
#>   ..$ parameters:List of 63
#>  $ drive.files.create      :List of 4
#>   ..$ id        : chr "drive.files.create"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "drive/v3/files"
#>   ..$ parameters:List of 63
#>  $ drive.files.create.media:List of 4
#>   ..$ id        : chr "drive.files.create.media"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "/upload/drive/v3/files"
#>   ..$ parameters:List of 64
#>  $ drive.files.delete      :List of 4
#>   ..$ id        : chr "drive.files.delete"
#>   ..$ method    : chr "DELETE"
#>   ..$ path      : chr "drive/v3/files/{fileId}"
#>   ..$ parameters:List of 9
#>  $ drive.files.emptyTrash  :List of 4
#>   ..$ id        : chr "drive.files.emptyTrash"
#>   ..$ method    : chr "DELETE"
#>   ..$ path      : chr "drive/v3/files/trash"
#>   ..$ parameters:List of 7
#>  $ drive.files.export      :List of 4
#>   ..$ id        : chr "drive.files.export"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/export"
#>   ..$ parameters:List of 9
#>  $ drive.files.generateIds :List of 4
#>   ..$ id        : chr "drive.files.generateIds"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/generateIds"
#>   ..$ parameters:List of 9
#>  $ drive.files.get         :List of 4
#>   ..$ id        : chr "drive.files.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}"
#>   ..$ parameters:List of 10
#>  $ drive.files.list        :List of 4
#>   ..$ id        : chr "drive.files.list"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files"
#>   ..$ parameters:List of 17
#>  $ drive.files.update      :List of 4
#>   ..$ id        : chr "drive.files.update"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "drive/v3/files/{fileId}"
#>   ..$ parameters:List of 65
#>  $ drive.files.update.media:List of 4
#>   ..$ id        : chr "drive.files.update.media"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "/upload/drive/v3/files/{fileId}"
#>   ..$ parameters:List of 66
#>  $ drive.files.watch       :List of 4
#>   ..$ id        : chr "drive.files.watch"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "drive/v3/files/{fileId}/watch"
#>   ..$ parameters:List of 10
#>  $ drive.permissions.create:List of 4
#>   ..$ id        : chr "drive.permissions.create"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions"
#>   ..$ parameters:List of 24
#>  $ drive.permissions.delete:List of 4
#>   ..$ id        : chr "drive.permissions.delete"
#>   ..$ method    : chr "DELETE"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions/{permissionId}"
#>   ..$ parameters:List of 10
#>  $ drive.permissions.get   :List of 4
#>   ..$ id        : chr "drive.permissions.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions/{permissionId}"
#>   ..$ parameters:List of 10
#>  $ drive.permissions.list  :List of 4
#>   ..$ id        : chr "drive.permissions.list"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions"
#>   ..$ parameters:List of 11
#>  $ drive.permissions.update:List of 4
#>   ..$ id        : chr "drive.permissions.update"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "drive/v3/files/{fileId}/permissions/{permissionId}"
#>   ..$ parameters:List of 24
#>  $ drive.revisions.delete  :List of 4
#>   ..$ id        : chr "drive.revisions.delete"
#>   ..$ method    : chr "DELETE"
#>   ..$ path      : chr "drive/v3/files/{fileId}/revisions/{revisionId}"
#>   ..$ parameters:List of 9
#>  $ drive.revisions.get     :List of 4
#>   ..$ id        : chr "drive.revisions.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/revisions/{revisionId}"
#>   ..$ parameters:List of 10
#>  $ drive.revisions.list    :List of 4
#>   ..$ id        : chr "drive.revisions.list"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/files/{fileId}/revisions"
#>   ..$ parameters:List of 10
#>  $ drive.revisions.update  :List of 4
#>   ..$ id        : chr "drive.revisions.update"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "drive/v3/files/{fileId}/revisions/{revisionId}"
#>   ..$ parameters:List of 21
#>  $ drive.teamdrives.create :List of 4
#>   ..$ id        : chr "drive.teamdrives.create"
#>   ..$ method    : chr "POST"
#>   ..$ path      : chr "drive/v3/teamdrives"
#>   ..$ parameters:List of 16
#>  $ drive.teamdrives.delete :List of 4
#>   ..$ id        : chr "drive.teamdrives.delete"
#>   ..$ method    : chr "DELETE"
#>   ..$ path      : chr "drive/v3/teamdrives/{teamDriveId}"
#>   ..$ parameters:List of 8
#>  $ drive.teamdrives.get    :List of 4
#>   ..$ id        : chr "drive.teamdrives.get"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/teamdrives/{teamDriveId}"
#>   ..$ parameters:List of 8
#>  $ drive.teamdrives.list   :List of 4
#>   ..$ id        : chr "drive.teamdrives.list"
#>   ..$ method    : chr "GET"
#>   ..$ path      : chr "drive/v3/teamdrives"
#>   ..$ parameters:List of 9
#>  $ drive.teamdrives.update :List of 4
#>   ..$ id        : chr "drive.teamdrives.update"
#>   ..$ method    : chr "PATCH"
#>   ..$ path      : chr "drive/v3/teamdrives/{teamDriveId}"
#>   ..$ parameters:List of 16

drive_endpoints("drive.files.delete")

#> $drive.files.delete
#> $drive.files.delete$id
#> [1] "drive.files.delete"
#> 
#> $drive.files.delete$method
#> [1] "DELETE"
#> 
#> $drive.files.delete$path
#> [1] "drive/v3/files/{fileId}"
#> 
#> $drive.files.delete$parameters
#> $drive.files.delete$parameters$fileId
#> $drive.files.delete$parameters$fileId$location
#> [1] "path"
#> 
#> $drive.files.delete$parameters$fileId$required
#> [1] TRUE
#> 
#> $drive.files.delete$parameters$fileId$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$fileId$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$fileId$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$fileId$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$fileId$description
#> [1] "The ID of the file."
#> 
#> 
#> $drive.files.delete$parameters$supportsTeamDrives
#> $drive.files.delete$parameters$supportsTeamDrives$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$type
#> [1] "boolean"
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$description
#> [1] "Whether the requesting application supports Team Drives."
#> 
#> 
#> $drive.files.delete$parameters$alt
#> $drive.files.delete$parameters$alt$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$alt$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$alt$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$alt$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$alt$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$alt$enum
#> [1] "json"
#> 
#> $drive.files.delete$parameters$alt$description
#> [1] "Data format for the response."
#> 
#> 
#> $drive.files.delete$parameters$fields
#> $drive.files.delete$parameters$fields$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$fields$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$fields$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$fields$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$fields$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$fields$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$fields$description
#> [1] "Selector specifying which fields to include in a partial response."
#> 
#> 
#> $drive.files.delete$parameters$key
#> $drive.files.delete$parameters$key$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$key$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$key$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$key$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$key$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$key$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$key$description
#> [1] "API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token."
#> 
#> 
#> $drive.files.delete$parameters$oauth_token
#> $drive.files.delete$parameters$oauth_token$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$oauth_token$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$oauth_token$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$oauth_token$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$oauth_token$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$oauth_token$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$oauth_token$description
#> [1] "OAuth 2.0 token for the current user."
#> 
#> 
#> $drive.files.delete$parameters$prettyPrint
#> $drive.files.delete$parameters$prettyPrint$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$prettyPrint$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$prettyPrint$type
#> [1] "boolean"
#> 
#> $drive.files.delete$parameters$prettyPrint$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$prettyPrint$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$prettyPrint$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$prettyPrint$description
#> [1] "Returns response with indentations and line breaks."
#> 
#> 
#> $drive.files.delete$parameters$quotaUser
#> $drive.files.delete$parameters$quotaUser$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$quotaUser$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$quotaUser$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$quotaUser$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$quotaUser$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$quotaUser$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$quotaUser$description
#> [1] "Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided."
#> 
#> 
#> $drive.files.delete$parameters$userIp
#> $drive.files.delete$parameters$userIp$location
#> [1] "query"
#> 
#> $drive.files.delete$parameters$userIp$required
#> [1] NA
#> 
#> $drive.files.delete$parameters$userIp$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$userIp$repeated
#> [1] NA
#> 
#> $drive.files.delete$parameters$userIp$format
#> [1] NA
#> 
#> $drive.files.delete$parameters$userIp$enum
#> [1] NA
#> 
#> $drive.files.delete$parameters$userIp$description
#> [1] "IP address of the site where the request originates. Use this if you want to enforce per-user limits."
#> 
#> 
#> 
#> 

drive_endpoints(4)

#> $drive.files.create.media
#> $drive.files.create.media$id
#> [1] "drive.files.create.media"
#> 
#> $drive.files.create.media$method
#> [1] "POST"
#> 
#> $drive.files.create.media$path
#> [1] "/upload/drive/v3/files"
#> 
#> $drive.files.create.media$parameters
#> $drive.files.create.media$parameters$ignoreDefaultVisibility
#> $drive.files.create.media$parameters$ignoreDefaultVisibility$location
#> [1] "query"
#> 
#> $drive.files.create.media$parameters$ignoreDefaultVisibility$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$ignoreDefaultVisibility$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$ignoreDefaultVisibility$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$ignoreDefaultVisibility$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$ignoreDefaultVisibility$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$ignoreDefaultVisibility$description
#> [1] "Whether to ignore the domain's default visibility settings for the created file. Domain administrators can choose to make all uploaded files visible to the domain by default; this parameter bypasses that behavior for the request. Permissions are still inherited from parent folders."
#> 
#> 
#> $drive.files.create.media$parameters$keepRevisionForever
#> $drive.files.create.media$parameters$keepRevisionForever$location
#> [1] "query"
#> 
#> $drive.files.create.media$parameters$keepRevisionForever$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$keepRevisionForever$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$keepRevisionForever$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$keepRevisionForever$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$keepRevisionForever$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$keepRevisionForever$description
#> [1] "Whether to set the 'keepForever' field in the new head revision. This is only applicable to files with binary content in Drive."
#> 
#> 
#> $drive.files.create.media$parameters$ocrLanguage
#> $drive.files.create.media$parameters$ocrLanguage$location
#> [1] "query"
#> 
#> $drive.files.create.media$parameters$ocrLanguage$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$ocrLanguage$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$ocrLanguage$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$ocrLanguage$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$ocrLanguage$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$ocrLanguage$description
#> [1] "A language hint for OCR processing during image import (ISO 639-1 code)."
#> 
#> 
#> $drive.files.create.media$parameters$supportsTeamDrives
#> $drive.files.create.media$parameters$supportsTeamDrives$location
#> [1] "query"
#> 
#> $drive.files.create.media$parameters$supportsTeamDrives$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$supportsTeamDrives$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$supportsTeamDrives$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$supportsTeamDrives$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$supportsTeamDrives$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$supportsTeamDrives$description
#> [1] "Whether the requesting application supports Team Drives."
#> 
#> 
#> $drive.files.create.media$parameters$useContentAsIndexableText
#> $drive.files.create.media$parameters$useContentAsIndexableText$location
#> [1] "query"
#> 
#> $drive.files.create.media$parameters$useContentAsIndexableText$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$useContentAsIndexableText$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$useContentAsIndexableText$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$useContentAsIndexableText$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$useContentAsIndexableText$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$useContentAsIndexableText$description
#> [1] "Whether to use the uploaded content as indexable text."
#> 
#> 
#> $drive.files.create.media$parameters$appProperties
#> $drive.files.create.media$parameters$appProperties$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$appProperties$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$appProperties$type
#> [1] "object"
#> 
#> $drive.files.create.media$parameters$appProperties$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$appProperties$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$appProperties$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$appProperties$description
#> [1] "A collection of arbitrary key-value pairs which are private to the requesting app.\nEntries with null values are cleared in update and copy requests."
#> 
#> 
#> $drive.files.create.media$parameters$capabilities
#> $drive.files.create.media$parameters$capabilities$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$capabilities$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$capabilities$type
#> [1] "object"
#> 
#> $drive.files.create.media$parameters$capabilities$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$capabilities$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$capabilities$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$capabilities$description
#> [1] "Capabilities the current user has on this file. Each capability corresponds to a fine-grained action that a user may take."
#> 
#> 
#> $drive.files.create.media$parameters$contentHints
#> $drive.files.create.media$parameters$contentHints$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$contentHints$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$contentHints$type
#> [1] "object"
#> 
#> $drive.files.create.media$parameters$contentHints$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$contentHints$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$contentHints$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$contentHints$description
#> [1] "Additional information about the content of the file. These fields are never populated in responses."
#> 
#> 
#> $drive.files.create.media$parameters$createdTime
#> $drive.files.create.media$parameters$createdTime$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$createdTime$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$createdTime$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$createdTime$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$createdTime$format
#> [1] "date-time"
#> 
#> $drive.files.create.media$parameters$createdTime$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$createdTime$description
#> [1] "The time at which the file was created (RFC 3339 date-time)."
#> 
#> 
#> $drive.files.create.media$parameters$description
#> $drive.files.create.media$parameters$description$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$description$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$description$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$description$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$description$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$description$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$description$description
#> [1] "A short description of the file."
#> 
#> 
#> $drive.files.create.media$parameters$explicitlyTrashed
#> $drive.files.create.media$parameters$explicitlyTrashed$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$explicitlyTrashed$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$explicitlyTrashed$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$explicitlyTrashed$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$explicitlyTrashed$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$explicitlyTrashed$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$explicitlyTrashed$description
#> [1] "Whether the file has been explicitly trashed, as opposed to recursively trashed from a parent folder."
#> 
#> 
#> $drive.files.create.media$parameters$fileExtension
#> $drive.files.create.media$parameters$fileExtension$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$fileExtension$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$fileExtension$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$fileExtension$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$fileExtension$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$fileExtension$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$fileExtension$description
#> [1] "The final component of fullFileExtension. This is only available for files with binary content in Drive."
#> 
#> 
#> $drive.files.create.media$parameters$folderColorRgb
#> $drive.files.create.media$parameters$folderColorRgb$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$folderColorRgb$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$folderColorRgb$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$folderColorRgb$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$folderColorRgb$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$folderColorRgb$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$folderColorRgb$description
#> [1] "The color for a folder as an RGB hex string. The supported colors are published in the folderColorPalette field of the About resource.\nIf an unsupported color is specified, the closest color in the palette will be used instead."
#> 
#> 
#> $drive.files.create.media$parameters$fullFileExtension
#> $drive.files.create.media$parameters$fullFileExtension$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$fullFileExtension$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$fullFileExtension$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$fullFileExtension$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$fullFileExtension$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$fullFileExtension$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$fullFileExtension$description
#> [1] "The full file extension extracted from the name field. May contain multiple concatenated extensions, such as \"tar.gz\". This is only available for files with binary content in Drive.\nThis is automatically updated when the name field changes, however it is not cleared if the new name does not contain a valid extension."
#> 
#> 
#> $drive.files.create.media$parameters$hasAugmentedPermissions
#> $drive.files.create.media$parameters$hasAugmentedPermissions$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$hasAugmentedPermissions$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$hasAugmentedPermissions$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$hasAugmentedPermissions$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$hasAugmentedPermissions$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$hasAugmentedPermissions$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$hasAugmentedPermissions$description
#> [1] "Whether any users are granted file access directly on this file. This field is only populated for Team Drive files."
#> 
#> 
#> $drive.files.create.media$parameters$hasThumbnail
#> $drive.files.create.media$parameters$hasThumbnail$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$hasThumbnail$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$hasThumbnail$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$hasThumbnail$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$hasThumbnail$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$hasThumbnail$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$hasThumbnail$description
#> [1] "Whether this file has a thumbnail. This does not indicate whether the requesting app has access to the thumbnail. To check access, look for the presence of the thumbnailLink field."
#> 
#> 
#> $drive.files.create.media$parameters$headRevisionId
#> $drive.files.create.media$parameters$headRevisionId$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$headRevisionId$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$headRevisionId$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$headRevisionId$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$headRevisionId$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$headRevisionId$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$headRevisionId$description
#> [1] "The ID of the file's head revision. This is currently only available for files with binary content in Drive."
#> 
#> 
#> $drive.files.create.media$parameters$iconLink
#> $drive.files.create.media$parameters$iconLink$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$iconLink$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$iconLink$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$iconLink$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$iconLink$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$iconLink$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$iconLink$description
#> [1] "A static, unauthenticated link to the file's icon."
#> 
#> 
#> $drive.files.create.media$parameters$id
#> $drive.files.create.media$parameters$id$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$id$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$id$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$id$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$id$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$id$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$id$description
#> [1] "The ID of the file."
#> 
#> 
#> $drive.files.create.media$parameters$imageMediaMetadata
#> $drive.files.create.media$parameters$imageMediaMetadata$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$imageMediaMetadata$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$imageMediaMetadata$type
#> [1] "object"
#> 
#> $drive.files.create.media$parameters$imageMediaMetadata$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$imageMediaMetadata$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$imageMediaMetadata$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$imageMediaMetadata$description
#> [1] "Additional metadata about image media, if available."
#> 
#> 
#> $drive.files.create.media$parameters$isAppAuthorized
#> $drive.files.create.media$parameters$isAppAuthorized$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$isAppAuthorized$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$isAppAuthorized$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$isAppAuthorized$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$isAppAuthorized$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$isAppAuthorized$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$isAppAuthorized$description
#> [1] "Whether the file was created or opened by the requesting app."
#> 
#> 
#> $drive.files.create.media$parameters$kind
#> $drive.files.create.media$parameters$kind$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$kind$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$kind$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$kind$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$kind$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$kind$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$kind$description
#> [1] "Identifies what kind of resource this is. Value: the fixed string \"drive#file\"."
#> 
#> 
#> $drive.files.create.media$parameters$lastModifyingUser
#> $drive.files.create.media$parameters$lastModifyingUser$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$lastModifyingUser$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$lastModifyingUser$type
#> [1] NA
#> 
#> $drive.files.create.media$parameters$lastModifyingUser$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$lastModifyingUser$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$lastModifyingUser$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$lastModifyingUser$description
#> [1] "The last user to modify the file."
#> 
#> 
#> $drive.files.create.media$parameters$md5Checksum
#> $drive.files.create.media$parameters$md5Checksum$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$md5Checksum$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$md5Checksum$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$md5Checksum$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$md5Checksum$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$md5Checksum$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$md5Checksum$description
#> [1] "The MD5 checksum for the content of the file. This is only applicable to files with binary content in Drive."
#> 
#> 
#> $drive.files.create.media$parameters$mimeType
#> $drive.files.create.media$parameters$mimeType$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$mimeType$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$mimeType$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$mimeType$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$mimeType$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$mimeType$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$mimeType$description
#> [1] "The MIME type of the file.\nDrive will attempt to automatically detect an appropriate value from uploaded content if no value is provided. The value cannot be changed unless a new revision is uploaded.\nIf a file is created with a Google Doc MIME type, the uploaded content will be imported if possible. The supported import formats are published in the About resource."
#> 
#> 
#> $drive.files.create.media$parameters$modifiedByMe
#> $drive.files.create.media$parameters$modifiedByMe$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$modifiedByMe$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$modifiedByMe$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$modifiedByMe$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$modifiedByMe$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$modifiedByMe$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$modifiedByMe$description
#> [1] "Whether the file has been modified by this user."
#> 
#> 
#> $drive.files.create.media$parameters$modifiedByMeTime
#> $drive.files.create.media$parameters$modifiedByMeTime$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$modifiedByMeTime$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$modifiedByMeTime$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$modifiedByMeTime$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$modifiedByMeTime$format
#> [1] "date-time"
#> 
#> $drive.files.create.media$parameters$modifiedByMeTime$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$modifiedByMeTime$description
#> [1] "The last time the file was modified by the user (RFC 3339 date-time)."
#> 
#> 
#> $drive.files.create.media$parameters$modifiedTime
#> $drive.files.create.media$parameters$modifiedTime$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$modifiedTime$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$modifiedTime$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$modifiedTime$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$modifiedTime$format
#> [1] "date-time"
#> 
#> $drive.files.create.media$parameters$modifiedTime$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$modifiedTime$description
#> [1] "The last time the file was modified by anyone (RFC 3339 date-time).\nNote that setting modifiedTime will also update modifiedByMeTime for the user."
#> 
#> 
#> $drive.files.create.media$parameters$name
#> $drive.files.create.media$parameters$name$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$name$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$name$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$name$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$name$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$name$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$name$description
#> [1] "The name of the file. This is not necessarily unique within a folder. Note that for immutable items such as the top level folders of Team Drives, My Drive root folder, and Application Data folder the name is constant."
#> 
#> 
#> $drive.files.create.media$parameters$originalFilename
#> $drive.files.create.media$parameters$originalFilename$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$originalFilename$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$originalFilename$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$originalFilename$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$originalFilename$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$originalFilename$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$originalFilename$description
#> [1] "The original filename of the uploaded content if available, or else the original value of the name field. This is only available for files with binary content in Drive."
#> 
#> 
#> $drive.files.create.media$parameters$ownedByMe
#> $drive.files.create.media$parameters$ownedByMe$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$ownedByMe$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$ownedByMe$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$ownedByMe$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$ownedByMe$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$ownedByMe$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$ownedByMe$description
#> [1] "Whether the user owns the file. Not populated for Team Drive files."
#> 
#> 
#> $drive.files.create.media$parameters$owners
#> $drive.files.create.media$parameters$owners$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$owners$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$owners$type
#> [1] "array"
#> 
#> $drive.files.create.media$parameters$owners$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$owners$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$owners$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$owners$description
#> [1] "The owners of the file. Currently, only certain legacy files may have more than one owner. Not populated for Team Drive files."
#> 
#> 
#> $drive.files.create.media$parameters$parents
#> $drive.files.create.media$parameters$parents$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$parents$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$parents$type
#> [1] "array"
#> 
#> $drive.files.create.media$parameters$parents$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$parents$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$parents$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$parents$description
#> [1] "The IDs of the parent folders which contain the file.\nIf not specified as part of a create request, the file will be placed directly in the My Drive folder. Update requests must use the addParents and removeParents parameters to modify the values."
#> 
#> 
#> $drive.files.create.media$parameters$permissions
#> $drive.files.create.media$parameters$permissions$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$permissions$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$permissions$type
#> [1] "array"
#> 
#> $drive.files.create.media$parameters$permissions$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$permissions$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$permissions$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$permissions$description
#> [1] "The full list of permissions for the file. This is only available if the requesting user can share the file. Not populated for Team Drive files."
#> 
#> 
#> $drive.files.create.media$parameters$properties
#> $drive.files.create.media$parameters$properties$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$properties$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$properties$type
#> [1] "object"
#> 
#> $drive.files.create.media$parameters$properties$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$properties$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$properties$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$properties$description
#> [1] "A collection of arbitrary key-value pairs which are visible to all apps.\nEntries with null values are cleared in update and copy requests."
#> 
#> 
#> $drive.files.create.media$parameters$quotaBytesUsed
#> $drive.files.create.media$parameters$quotaBytesUsed$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$quotaBytesUsed$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$quotaBytesUsed$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$quotaBytesUsed$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$quotaBytesUsed$format
#> [1] "int64"
#> 
#> $drive.files.create.media$parameters$quotaBytesUsed$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$quotaBytesUsed$description
#> [1] "The number of storage quota bytes used by the file. This includes the head revision as well as previous revisions with keepForever enabled."
#> 
#> 
#> $drive.files.create.media$parameters$shared
#> $drive.files.create.media$parameters$shared$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$shared$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$shared$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$shared$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$shared$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$shared$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$shared$description
#> [1] "Whether the file has been shared. Not populated for Team Drive files."
#> 
#> 
#> $drive.files.create.media$parameters$sharedWithMeTime
#> $drive.files.create.media$parameters$sharedWithMeTime$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$sharedWithMeTime$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$sharedWithMeTime$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$sharedWithMeTime$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$sharedWithMeTime$format
#> [1] "date-time"
#> 
#> $drive.files.create.media$parameters$sharedWithMeTime$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$sharedWithMeTime$description
#> [1] "The time at which the file was shared with the user, if applicable (RFC 3339 date-time)."
#> 
#> 
#> $drive.files.create.media$parameters$sharingUser
#> $drive.files.create.media$parameters$sharingUser$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$sharingUser$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$sharingUser$type
#> [1] NA
#> 
#> $drive.files.create.media$parameters$sharingUser$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$sharingUser$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$sharingUser$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$sharingUser$description
#> [1] "The user who shared the file with the requesting user, if applicable."
#> 
#> 
#> $drive.files.create.media$parameters$size
#> $drive.files.create.media$parameters$size$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$size$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$size$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$size$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$size$format
#> [1] "int64"
#> 
#> $drive.files.create.media$parameters$size$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$size$description
#> [1] "The size of the file's content in bytes. This is only applicable to files with binary content in Drive."
#> 
#> 
#> $drive.files.create.media$parameters$spaces
#> $drive.files.create.media$parameters$spaces$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$spaces$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$spaces$type
#> [1] "array"
#> 
#> $drive.files.create.media$parameters$spaces$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$spaces$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$spaces$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$spaces$description
#> [1] "The list of spaces which contain the file. The currently supported values are 'drive', 'appDataFolder' and 'photos'."
#> 
#> 
#> $drive.files.create.media$parameters$starred
#> $drive.files.create.media$parameters$starred$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$starred$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$starred$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$starred$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$starred$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$starred$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$starred$description
#> [1] "Whether the user has starred the file."
#> 
#> 
#> $drive.files.create.media$parameters$teamDriveId
#> $drive.files.create.media$parameters$teamDriveId$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$teamDriveId$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$teamDriveId$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$teamDriveId$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$teamDriveId$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$teamDriveId$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$teamDriveId$description
#> [1] "ID of the Team Drive the file resides in."
#> 
#> 
#> $drive.files.create.media$parameters$thumbnailLink
#> $drive.files.create.media$parameters$thumbnailLink$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$thumbnailLink$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$thumbnailLink$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$thumbnailLink$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$thumbnailLink$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$thumbnailLink$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$thumbnailLink$description
#> [1] "A short-lived link to the file's thumbnail, if available. Typically lasts on the order of hours. Only populated when the requesting app can access the file's content."
#> 
#> 
#> $drive.files.create.media$parameters$thumbnailVersion
#> $drive.files.create.media$parameters$thumbnailVersion$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$thumbnailVersion$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$thumbnailVersion$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$thumbnailVersion$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$thumbnailVersion$format
#> [1] "int64"
#> 
#> $drive.files.create.media$parameters$thumbnailVersion$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$thumbnailVersion$description
#> [1] "The thumbnail version for use in thumbnail cache invalidation."
#> 
#> 
#> $drive.files.create.media$parameters$trashed
#> $drive.files.create.media$parameters$trashed$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$trashed$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$trashed$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$trashed$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$trashed$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$trashed$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$trashed$description
#> [1] "Whether the file has been trashed, either explicitly or from a trashed parent folder. Only the owner may trash a file, and other users cannot see files in the owner's trash."
#> 
#> 
#> $drive.files.create.media$parameters$trashedTime
#> $drive.files.create.media$parameters$trashedTime$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$trashedTime$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$trashedTime$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$trashedTime$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$trashedTime$format
#> [1] "date-time"
#> 
#> $drive.files.create.media$parameters$trashedTime$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$trashedTime$description
#> [1] "The time that the item was trashed (RFC 3339 date-time). Only populated for Team Drive files."
#> 
#> 
#> $drive.files.create.media$parameters$trashingUser
#> $drive.files.create.media$parameters$trashingUser$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$trashingUser$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$trashingUser$type
#> [1] NA
#> 
#> $drive.files.create.media$parameters$trashingUser$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$trashingUser$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$trashingUser$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$trashingUser$description
#> [1] "If the file has been explicitly trashed, the user who trashed it. Only populated for Team Drive files."
#> 
#> 
#> $drive.files.create.media$parameters$version
#> $drive.files.create.media$parameters$version$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$version$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$version$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$version$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$version$format
#> [1] "int64"
#> 
#> $drive.files.create.media$parameters$version$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$version$description
#> [1] "A monotonically increasing version number for the file. This reflects every change made to the file on the server, even those not visible to the user."
#> 
#> 
#> $drive.files.create.media$parameters$videoMediaMetadata
#> $drive.files.create.media$parameters$videoMediaMetadata$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$videoMediaMetadata$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$videoMediaMetadata$type
#> [1] "object"
#> 
#> $drive.files.create.media$parameters$videoMediaMetadata$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$videoMediaMetadata$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$videoMediaMetadata$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$videoMediaMetadata$description
#> [1] "Additional metadata about video media. This may not be available immediately upon upload."
#> 
#> 
#> $drive.files.create.media$parameters$viewedByMe
#> $drive.files.create.media$parameters$viewedByMe$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$viewedByMe$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$viewedByMe$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$viewedByMe$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$viewedByMe$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$viewedByMe$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$viewedByMe$description
#> [1] "Whether the file has been viewed by this user."
#> 
#> 
#> $drive.files.create.media$parameters$viewedByMeTime
#> $drive.files.create.media$parameters$viewedByMeTime$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$viewedByMeTime$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$viewedByMeTime$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$viewedByMeTime$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$viewedByMeTime$format
#> [1] "date-time"
#> 
#> $drive.files.create.media$parameters$viewedByMeTime$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$viewedByMeTime$description
#> [1] "The last time the file was viewed by the user (RFC 3339 date-time)."
#> 
#> 
#> $drive.files.create.media$parameters$viewersCanCopyContent
#> $drive.files.create.media$parameters$viewersCanCopyContent$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$viewersCanCopyContent$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$viewersCanCopyContent$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$viewersCanCopyContent$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$viewersCanCopyContent$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$viewersCanCopyContent$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$viewersCanCopyContent$description
#> [1] "Whether users with only reader or commenter permission can copy the file's content. This affects copy, download, and print operations."
#> 
#> 
#> $drive.files.create.media$parameters$webContentLink
#> $drive.files.create.media$parameters$webContentLink$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$webContentLink$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$webContentLink$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$webContentLink$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$webContentLink$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$webContentLink$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$webContentLink$description
#> [1] "A link for downloading the content of the file in a browser. This is only available for files with binary content in Drive."
#> 
#> 
#> $drive.files.create.media$parameters$webViewLink
#> $drive.files.create.media$parameters$webViewLink$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$webViewLink$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$webViewLink$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$webViewLink$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$webViewLink$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$webViewLink$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$webViewLink$description
#> [1] "A link for opening the file in a relevant Google editor or viewer in a browser."
#> 
#> 
#> $drive.files.create.media$parameters$writersCanShare
#> $drive.files.create.media$parameters$writersCanShare$location
#> [1] "body"
#> 
#> $drive.files.create.media$parameters$writersCanShare$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$writersCanShare$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$writersCanShare$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$writersCanShare$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$writersCanShare$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$writersCanShare$description
#> [1] "Whether users with only writer permission can modify the file's permissions. Not populated for Team Drive files."
#> 
#> 
#> $drive.files.create.media$parameters$alt
#> $drive.files.create.media$parameters$alt$location
#> [1] "query"
#> 
#> $drive.files.create.media$parameters$alt$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$alt$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$alt$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$alt$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$alt$enum
#> [1] "json"
#> 
#> $drive.files.create.media$parameters$alt$description
#> [1] "Data format for the response."
#> 
#> 
#> $drive.files.create.media$parameters$fields
#> $drive.files.create.media$parameters$fields$location
#> [1] "query"
#> 
#> $drive.files.create.media$parameters$fields$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$fields$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$fields$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$fields$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$fields$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$fields$description
#> [1] "Selector specifying which fields to include in a partial response."
#> 
#> 
#> $drive.files.create.media$parameters$key
#> $drive.files.create.media$parameters$key$location
#> [1] "query"
#> 
#> $drive.files.create.media$parameters$key$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$key$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$key$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$key$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$key$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$key$description
#> [1] "API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token."
#> 
#> 
#> $drive.files.create.media$parameters$oauth_token
#> $drive.files.create.media$parameters$oauth_token$location
#> [1] "query"
#> 
#> $drive.files.create.media$parameters$oauth_token$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$oauth_token$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$oauth_token$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$oauth_token$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$oauth_token$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$oauth_token$description
#> [1] "OAuth 2.0 token for the current user."
#> 
#> 
#> $drive.files.create.media$parameters$prettyPrint
#> $drive.files.create.media$parameters$prettyPrint$location
#> [1] "query"
#> 
#> $drive.files.create.media$parameters$prettyPrint$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$prettyPrint$type
#> [1] "boolean"
#> 
#> $drive.files.create.media$parameters$prettyPrint$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$prettyPrint$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$prettyPrint$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$prettyPrint$description
#> [1] "Returns response with indentations and line breaks."
#> 
#> 
#> $drive.files.create.media$parameters$quotaUser
#> $drive.files.create.media$parameters$quotaUser$location
#> [1] "query"
#> 
#> $drive.files.create.media$parameters$quotaUser$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$quotaUser$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$quotaUser$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$quotaUser$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$quotaUser$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$quotaUser$description
#> [1] "Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided."
#> 
#> 
#> $drive.files.create.media$parameters$userIp
#> $drive.files.create.media$parameters$userIp$location
#> [1] "query"
#> 
#> $drive.files.create.media$parameters$userIp$required
#> [1] NA
#> 
#> $drive.files.create.media$parameters$userIp$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$userIp$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$userIp$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$userIp$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$userIp$description
#> [1] "IP address of the site where the request originates. Use this if you want to enforce per-user limits."
#> 
#> 
#> $drive.files.create.media$parameters$uploadType
#> $drive.files.create.media$parameters$uploadType$location
#> [1] "query"
#> 
#> $drive.files.create.media$parameters$uploadType$required
#> [1] TRUE
#> 
#> $drive.files.create.media$parameters$uploadType$type
#> [1] "string"
#> 
#> $drive.files.create.media$parameters$uploadType$repeated
#> [1] NA
#> 
#> $drive.files.create.media$parameters$uploadType$format
#> [1] NA
#> 
#> $drive.files.create.media$parameters$uploadType$enum
#> [1] NA
#> 
#> $drive.files.create.media$parameters$uploadType$description
#> [1] NA
#> 
#> 
#> 
#>