List Drive endpoints

Returns a list of selected Drive API v3 endpoints, as stored inside the googledrive package. The names of this list (or the id sub-elements) are the nicknames that can be used to specify an endpoint in request_generate(). For each endpoint, we store its nickname or id, the associated HTTP verb, the path, and details about the parameters. This list is derived programmatically from the Drive API v3 Discovery Document.

drive_endpoints(i = NULL)

Arguments

i
The name(s) or integer index(ices) of the endpoints to return. Optional. By default, the entire list is returned.

Value

A list containing some or all of the subset of the Drive API v3 endpoints that are used internally by googledrive.

Examples

str(drive_endpoints(), max.level = 2)

#> List of 48
#>  $ drive.about.get                :List of 7
#>   ..$ id         : chr "drive.about.get"
#>   ..$ httpMethod : chr "GET"
#>   ..$ path       : 'fs_path' chr "drive/v3/about"
#>   ..$ scopes     : chr "drive, drive.appdata, drive.file, drive.metadata, drive.metadata.readonly, drive.photos.readonly, drive.readonly"
#>   ..$ description: chr "Gets information about the user, the user's Drive, and system capabilities."
#>   ..$ response   : chr "About"
#>   ..$ parameters :List of 7
#>  $ drive.changes.getStartPageToken:List of 7
#>   ..$ id         : chr "drive.changes.getStartPageToken"
#>   ..$ httpMethod : chr "GET"
#>   ..$ path       : 'fs_path' chr "drive/v3/changes/startPageToken"
#>   ..$ parameters :List of 11
#>   ..$ scopes     : chr "drive, drive.appdata, drive.file, drive.metadata, drive.metadata.readonly, drive.photos.readonly, drive.readonly"
#>   ..$ description: chr "Gets the starting pageToken for listing future changes."
#>   ..$ response   : chr "StartPageToken"
#>  $ drive.changes.list             :List of 9
#>   ..$ id                  : chr "drive.changes.list"
#>   ..$ httpMethod          : chr "GET"
#>   ..$ path                : 'fs_path' chr "drive/v3/changes"
#>   ..$ parameters          :List of 19
#>   ..$ scopes              : chr "drive, drive.appdata, drive.file, drive.metadata, drive.metadata.readonly, drive.photos.readonly, drive.readonly"
#>   ..$ description         : chr "Lists the changes for a user or shared drive."
#>   ..$ response            : chr "ChangeList"
#>   ..$ parameterOrder      : chr "pageToken"
#>   ..$ supportsSubscription: logi TRUE
#>  $ drive.changes.watch            :List of 10
#>   ..$ id                  : chr "drive.changes.watch"
#>   ..$ httpMethod          : chr "POST"
#>   ..$ path                : 'fs_path' chr "drive/v3/changes/watch"
#>   ..$ parameters          :List of 29
#>   ..$ scopes              : chr "drive, drive.appdata, drive.file, drive.metadata, drive.metadata.readonly, drive.photos.readonly, drive.readonly"
#>   ..$ description         : chr "Subscribes to changes for a user."
#>   ..$ request             : chr "Channel"
#>   ..$ response            : chr "Channel"
#>   ..$ parameterOrder      : chr "pageToken"
#>   ..$ supportsSubscription: logi TRUE
#>  $ drive.channels.stop            :List of 7
#>   ..$ id         : chr "drive.channels.stop"
#>   ..$ httpMethod : chr "POST"
#>   ..$ path       : 'fs_path' chr "drive/v3/channels/stop"
#>   ..$ scopes     : chr "drive, drive.appdata, drive.file, drive.metadata, drive.metadata.readonly, drive.photos.readonly, drive.readonly"
#>   ..$ description: chr "Stop watching resources through this channel"
#>   ..$ request    : chr "Channel"
#>   ..$ parameters :List of 17
#>  $ drive.comments.create          :List of 9
#>   ..$ id            : chr "drive.comments.create"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/comments"
#>   ..$ parameters    :List of 20
#>   ..$ scopes        : chr "drive, drive.file"
#>   ..$ description   : chr "Creates a new comment on a file."
#>   ..$ request       : chr "Comment"
#>   ..$ response      : chr "Comment"
#>   ..$ parameterOrder: chr "fileId"
#>  $ drive.comments.delete          :List of 7
#>   ..$ id            : chr "drive.comments.delete"
#>   ..$ httpMethod    : chr "DELETE"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/comments/{commentId}"
#>   ..$ parameters    :List of 9
#>   ..$ scopes        : chr "drive, drive.file"
#>   ..$ description   : chr "Deletes a comment."
#>   ..$ parameterOrder: chr [1:2] "fileId" "commentId"
#>  $ drive.comments.get             :List of 8
#>   ..$ id            : chr "drive.comments.get"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/comments/{commentId}"
#>   ..$ parameters    :List of 10
#>   ..$ scopes        : chr "drive, drive.file, drive.readonly"
#>   ..$ description   : chr "Gets a comment by ID."
#>   ..$ response      : chr "Comment"
#>   ..$ parameterOrder: chr [1:2] "fileId" "commentId"
#>  $ drive.comments.list            :List of 8
#>   ..$ id            : chr "drive.comments.list"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/comments"
#>   ..$ parameters    :List of 12
#>   ..$ scopes        : chr "drive, drive.file, drive.readonly"
#>   ..$ description   : chr "Lists a file's comments."
#>   ..$ response      : chr "CommentList"
#>   ..$ parameterOrder: chr "fileId"
#>  $ drive.comments.update          :List of 9
#>   ..$ id            : chr "drive.comments.update"
#>   ..$ httpMethod    : chr "PATCH"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/comments/{commentId}"
#>   ..$ parameters    :List of 21
#>   ..$ scopes        : chr "drive, drive.file"
#>   ..$ description   : chr "Updates a comment with patch semantics."
#>   ..$ request       : chr "Comment"
#>   ..$ response      : chr "Comment"
#>   ..$ parameterOrder: chr [1:2] "fileId" "commentId"
#>  $ drive.drives.create            :List of 9
#>   ..$ id            : chr "drive.drives.create"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "drive/v3/drives"
#>   ..$ parameters    :List of 19
#>   ..$ scopes        : chr "drive"
#>   ..$ description   : chr "Creates a new shared drive."
#>   ..$ request       : chr "Drive"
#>   ..$ response      : chr "Drive"
#>   ..$ parameterOrder: chr "requestId"
#>  $ drive.drives.delete            :List of 7
#>   ..$ id            : chr "drive.drives.delete"
#>   ..$ httpMethod    : chr "DELETE"
#>   ..$ path          : 'fs_path' chr "drive/v3/drives/{driveId}"
#>   ..$ parameters    :List of 8
#>   ..$ scopes        : chr "drive"
#>   ..$ description   : chr "Permanently deletes a shared drive for which the user is an organizer. The shared drive cannot contain any untrashed items."
#>   ..$ parameterOrder: chr "driveId"
#>  $ drive.drives.get               :List of 8
#>   ..$ id            : chr "drive.drives.get"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "drive/v3/drives/{driveId}"
#>   ..$ parameters    :List of 9
#>   ..$ scopes        : chr "drive, drive.readonly"
#>   ..$ description   : chr "Gets a shared drive's metadata by ID."
#>   ..$ response      : chr "Drive"
#>   ..$ parameterOrder: chr "driveId"
#>  $ drive.drives.hide              :List of 8
#>   ..$ id            : chr "drive.drives.hide"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "drive/v3/drives/{driveId}/hide"
#>   ..$ parameters    :List of 8
#>   ..$ scopes        : chr "drive"
#>   ..$ description   : chr "Hides a shared drive from the default view."
#>   ..$ response      : chr "Drive"
#>   ..$ parameterOrder: chr "driveId"
#>  $ drive.drives.list              :List of 7
#>   ..$ id         : chr "drive.drives.list"
#>   ..$ httpMethod : chr "GET"
#>   ..$ path       : 'fs_path' chr "drive/v3/drives"
#>   ..$ parameters :List of 11
#>   ..$ scopes     : chr "drive, drive.readonly"
#>   ..$ description: chr "Lists the user's shared drives."
#>   ..$ response   : chr "DriveList"
#>  $ drive.drives.unhide            :List of 8
#>   ..$ id            : chr "drive.drives.unhide"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "drive/v3/drives/{driveId}/unhide"
#>   ..$ parameters    :List of 8
#>   ..$ scopes        : chr "drive"
#>   ..$ description   : chr "Restores a shared drive to the default view."
#>   ..$ response      : chr "Drive"
#>   ..$ parameterOrder: chr "driveId"
#>  $ drive.drives.update            :List of 9
#>   ..$ id            : chr "drive.drives.update"
#>   ..$ httpMethod    : chr "PATCH"
#>   ..$ path          : 'fs_path' chr "drive/v3/drives/{driveId}"
#>   ..$ parameters    :List of 20
#>   ..$ scopes        : chr "drive"
#>   ..$ description   : chr "Updates the metadate for a shared drive."
#>   ..$ request       : chr "Drive"
#>   ..$ response      : chr "Drive"
#>   ..$ parameterOrder: chr "driveId"
#>  $ drive.files.copy               :List of 9
#>   ..$ id            : chr "drive.files.copy"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/copy"
#>   ..$ parameters    :List of 68
#>   ..$ scopes        : chr "drive, drive.appdata, drive.file, drive.photos.readonly"
#>   ..$ description   : chr "Creates a copy of a file and applies any requested updates with patch semantics."
#>   ..$ request       : chr "File"
#>   ..$ response      : chr "File"
#>   ..$ parameterOrder: chr "fileId"
#>  $ drive.files.create             :List of 11
#>   ..$ id                  : chr "drive.files.create"
#>   ..$ httpMethod          : chr "POST"
#>   ..$ path                : 'fs_path' chr "drive/v3/files"
#>   ..$ parameters          :List of 68
#>   ..$ scopes              : chr "drive, drive.appdata, drive.file"
#>   ..$ description         : chr "Creates a new file."
#>   ..$ request             : chr "File"
#>   ..$ response            : chr "File"
#>   ..$ mediaUpload         :List of 3
#>   ..$ supportsMediaUpload : logi TRUE
#>   ..$ supportsSubscription: logi TRUE
#>  $ drive.files.delete             :List of 7
#>   ..$ id            : chr "drive.files.delete"
#>   ..$ httpMethod    : chr "DELETE"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}"
#>   ..$ parameters    :List of 10
#>   ..$ scopes        : chr "drive, drive.appdata, drive.file"
#>   ..$ description   : chr "Permanently deletes a file owned by the user without moving it to the trash. If the file belongs to a shared dr"| __truncated__
#>   ..$ parameterOrder: chr "fileId"
#>  $ drive.files.emptyTrash         :List of 6
#>   ..$ id         : chr "drive.files.emptyTrash"
#>   ..$ httpMethod : chr "DELETE"
#>   ..$ path       : 'fs_path' chr "drive/v3/files/trash"
#>   ..$ scopes     : chr "drive"
#>   ..$ description: chr "Permanently deletes all of the user's trashed files."
#>   ..$ parameters :List of 7
#>  $ drive.files.export             :List of 8
#>   ..$ id                   : chr "drive.files.export"
#>   ..$ httpMethod           : chr "GET"
#>   ..$ path                 : 'fs_path' chr "drive/v3/files/{fileId}/export"
#>   ..$ parameters           :List of 9
#>   ..$ scopes               : chr "drive, drive.file, drive.readonly"
#>   ..$ description          : chr "Exports a Google Doc to the requested MIME type and returns the exported content. Please note that the exported"| __truncated__
#>   ..$ supportsMediaDownload: logi TRUE
#>   ..$ parameterOrder       : chr [1:2] "fileId" "mimeType"
#>  $ drive.files.generateIds        :List of 7
#>   ..$ id         : chr "drive.files.generateIds"
#>   ..$ httpMethod : chr "GET"
#>   ..$ path       : 'fs_path' chr "drive/v3/files/generateIds"
#>   ..$ parameters :List of 9
#>   ..$ scopes     : chr "drive, drive.appdata, drive.file"
#>   ..$ description: chr "Generates a set of file IDs which can be provided in create or copy requests."
#>   ..$ response   : chr "GeneratedIds"
#>  $ drive.files.get                :List of 11
#>   ..$ id                     : chr "drive.files.get"
#>   ..$ httpMethod             : chr "GET"
#>   ..$ path                   : 'fs_path' chr "drive/v3/files/{fileId}"
#>   ..$ parameters             :List of 11
#>   ..$ scopes                 : chr "drive, drive.appdata, drive.file, drive.metadata, drive.metadata.readonly, drive.photos.readonly, drive.readonly"
#>   ..$ description            : chr "Gets a file's metadata or content by ID."
#>   ..$ response               : chr "File"
#>   ..$ supportsMediaDownload  : logi TRUE
#>   ..$ useMediaDownloadService: logi TRUE
#>   ..$ parameterOrder         : chr "fileId"
#>   ..$ supportsSubscription   : logi TRUE
#>  $ drive.files.list               :List of 7
#>   ..$ id         : chr "drive.files.list"
#>   ..$ httpMethod : chr "GET"
#>   ..$ path       : 'fs_path' chr "drive/v3/files"
#>   ..$ parameters :List of 20
#>   ..$ scopes     : chr "drive, drive.appdata, drive.file, drive.metadata, drive.metadata.readonly, drive.photos.readonly, drive.readonly"
#>   ..$ description: chr "Lists or searches files."
#>   ..$ response   : chr "FileList"
#>  $ drive.files.update             :List of 11
#>   ..$ id                 : chr "drive.files.update"
#>   ..$ httpMethod         : chr "PATCH"
#>   ..$ path               : 'fs_path' chr "drive/v3/files/{fileId}"
#>   ..$ parameters         :List of 70
#>   ..$ scopes             : chr "drive, drive.appdata, drive.file, drive.metadata, drive.scripts"
#>   ..$ description        : chr "Updates a file's metadata and/or content with patch semantics."
#>   ..$ request            : chr "File"
#>   ..$ response           : chr "File"
#>   ..$ mediaUpload        :List of 3
#>   ..$ supportsMediaUpload: logi TRUE
#>   ..$ parameterOrder     : chr "fileId"
#>  $ drive.files.watch              :List of 12
#>   ..$ id                     : chr "drive.files.watch"
#>   ..$ httpMethod             : chr "POST"
#>   ..$ path                   : 'fs_path' chr "drive/v3/files/{fileId}/watch"
#>   ..$ parameters             :List of 21
#>   ..$ scopes                 : chr "drive, drive.appdata, drive.file, drive.metadata, drive.metadata.readonly, drive.photos.readonly, drive.readonly"
#>   ..$ description            : chr "Subscribes to changes to a file"
#>   ..$ request                : chr "Channel"
#>   ..$ response               : chr "Channel"
#>   ..$ supportsMediaDownload  : logi TRUE
#>   ..$ useMediaDownloadService: logi TRUE
#>   ..$ parameterOrder         : chr "fileId"
#>   ..$ supportsSubscription   : logi TRUE
#>  $ drive.permissions.create       :List of 9
#>   ..$ id            : chr "drive.permissions.create"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/permissions"
#>   ..$ parameters    :List of 27
#>   ..$ scopes        : chr "drive, drive.file"
#>   ..$ description   : chr "Creates a permission for a file or shared drive."
#>   ..$ request       : chr "Permission"
#>   ..$ response      : chr "Permission"
#>   ..$ parameterOrder: chr "fileId"
#>  $ drive.permissions.delete       :List of 7
#>   ..$ id            : chr "drive.permissions.delete"
#>   ..$ httpMethod    : chr "DELETE"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/permissions/{permissionId}"
#>   ..$ parameters    :List of 12
#>   ..$ scopes        : chr "drive, drive.file"
#>   ..$ description   : chr "Deletes a permission."
#>   ..$ parameterOrder: chr [1:2] "fileId" "permissionId"
#>  $ drive.permissions.get          :List of 8
#>   ..$ id            : chr "drive.permissions.get"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/permissions/{permissionId}"
#>   ..$ parameters    :List of 12
#>   ..$ scopes        : chr "drive, drive.file, drive.metadata, drive.metadata.readonly, drive.photos.readonly, drive.readonly"
#>   ..$ description   : chr "Gets a permission by ID."
#>   ..$ response      : chr "Permission"
#>   ..$ parameterOrder: chr [1:2] "fileId" "permissionId"
#>  $ drive.permissions.list         :List of 8
#>   ..$ id            : chr "drive.permissions.list"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/permissions"
#>   ..$ parameters    :List of 13
#>   ..$ scopes        : chr "drive, drive.file, drive.metadata, drive.metadata.readonly, drive.photos.readonly, drive.readonly"
#>   ..$ description   : chr "Lists a file's or shared drive's permissions."
#>   ..$ response      : chr "PermissionList"
#>   ..$ parameterOrder: chr "fileId"
#>  $ drive.permissions.update       :List of 9
#>   ..$ id            : chr "drive.permissions.update"
#>   ..$ httpMethod    : chr "PATCH"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/permissions/{permissionId}"
#>   ..$ parameters    :List of 27
#>   ..$ scopes        : chr "drive, drive.file"
#>   ..$ description   : chr "Updates a permission with patch semantics."
#>   ..$ request       : chr "Permission"
#>   ..$ response      : chr "Permission"
#>   ..$ parameterOrder: chr [1:2] "fileId" "permissionId"
#>  $ drive.replies.create           :List of 9
#>   ..$ id            : chr "drive.replies.create"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/comments/{commentId}/replies"
#>   ..$ parameters    :List of 18
#>   ..$ scopes        : chr "drive, drive.file"
#>   ..$ description   : chr "Creates a new reply to a comment."
#>   ..$ request       : chr "Reply"
#>   ..$ response      : chr "Reply"
#>   ..$ parameterOrder: chr [1:2] "fileId" "commentId"
#>  $ drive.replies.delete           :List of 7
#>   ..$ id            : chr "drive.replies.delete"
#>   ..$ httpMethod    : chr "DELETE"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/comments/{commentId}/replies/{replyId}"
#>   ..$ parameters    :List of 10
#>   ..$ scopes        : chr "drive, drive.file"
#>   ..$ description   : chr "Deletes a reply."
#>   ..$ parameterOrder: chr [1:3] "fileId" "commentId" "replyId"
#>  $ drive.replies.get              :List of 8
#>   ..$ id            : chr "drive.replies.get"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/comments/{commentId}/replies/{replyId}"
#>   ..$ parameters    :List of 11
#>   ..$ scopes        : chr "drive, drive.file, drive.readonly"
#>   ..$ description   : chr "Gets a reply by ID."
#>   ..$ response      : chr "Reply"
#>   ..$ parameterOrder: chr [1:3] "fileId" "commentId" "replyId"
#>  $ drive.replies.list             :List of 8
#>   ..$ id            : chr "drive.replies.list"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/comments/{commentId}/replies"
#>   ..$ parameters    :List of 12
#>   ..$ scopes        : chr "drive, drive.file, drive.readonly"
#>   ..$ description   : chr "Lists a comment's replies."
#>   ..$ response      : chr "ReplyList"
#>   ..$ parameterOrder: chr [1:2] "fileId" "commentId"
#>  $ drive.replies.update           :List of 9
#>   ..$ id            : chr "drive.replies.update"
#>   ..$ httpMethod    : chr "PATCH"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/comments/{commentId}/replies/{replyId}"
#>   ..$ parameters    :List of 19
#>   ..$ scopes        : chr "drive, drive.file"
#>   ..$ description   : chr "Updates a reply with patch semantics."
#>   ..$ request       : chr "Reply"
#>   ..$ response      : chr "Reply"
#>   ..$ parameterOrder: chr [1:3] "fileId" "commentId" "replyId"
#>  $ drive.revisions.delete         :List of 7
#>   ..$ id            : chr "drive.revisions.delete"
#>   ..$ httpMethod    : chr "DELETE"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/revisions/{revisionId}"
#>   ..$ parameters    :List of 9
#>   ..$ scopes        : chr "drive, drive.appdata, drive.file"
#>   ..$ description   : chr "Permanently deletes a file version. You can only delete revisions for files with binary content in Google Drive"| __truncated__
#>   ..$ parameterOrder: chr [1:2] "fileId" "revisionId"
#>  $ drive.revisions.get            :List of 10
#>   ..$ id                     : chr "drive.revisions.get"
#>   ..$ httpMethod             : chr "GET"
#>   ..$ path                   : 'fs_path' chr "drive/v3/files/{fileId}/revisions/{revisionId}"
#>   ..$ parameters             :List of 10
#>   ..$ scopes                 : chr "drive, drive.appdata, drive.file, drive.metadata, drive.metadata.readonly, drive.photos.readonly, drive.readonly"
#>   ..$ description            : chr "Gets a revision's metadata or content by ID."
#>   ..$ response               : chr "Revision"
#>   ..$ supportsMediaDownload  : logi TRUE
#>   ..$ useMediaDownloadService: logi TRUE
#>   ..$ parameterOrder         : chr [1:2] "fileId" "revisionId"
#>  $ drive.revisions.list           :List of 8
#>   ..$ id            : chr "drive.revisions.list"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/revisions"
#>   ..$ parameters    :List of 10
#>   ..$ scopes        : chr "drive, drive.appdata, drive.file, drive.metadata, drive.metadata.readonly, drive.photos.readonly, drive.readonly"
#>   ..$ description   : chr "Lists a file's revisions."
#>   ..$ response      : chr "RevisionList"
#>   ..$ parameterOrder: chr "fileId"
#>  $ drive.revisions.update         :List of 9
#>   ..$ id            : chr "drive.revisions.update"
#>   ..$ httpMethod    : chr "PATCH"
#>   ..$ path          : 'fs_path' chr "drive/v3/files/{fileId}/revisions/{revisionId}"
#>   ..$ parameters    :List of 22
#>   ..$ scopes        : chr "drive, drive.appdata, drive.file"
#>   ..$ description   : chr "Updates a revision with patch semantics."
#>   ..$ request       : chr "Revision"
#>   ..$ response      : chr "Revision"
#>   ..$ parameterOrder: chr [1:2] "fileId" "revisionId"
#>  $ drive.teamdrives.create        :List of 9
#>   ..$ id            : chr "drive.teamdrives.create"
#>   ..$ httpMethod    : chr "POST"
#>   ..$ path          : 'fs_path' chr "drive/v3/teamdrives"
#>   ..$ parameters    :List of 18
#>   ..$ scopes        : chr "drive"
#>   ..$ description   : chr "Deprecated use drives.create instead."
#>   ..$ request       : chr "TeamDrive"
#>   ..$ response      : chr "TeamDrive"
#>   ..$ parameterOrder: chr "requestId"
#>  $ drive.teamdrives.delete        :List of 7
#>   ..$ id            : chr "drive.teamdrives.delete"
#>   ..$ httpMethod    : chr "DELETE"
#>   ..$ path          : 'fs_path' chr "drive/v3/teamdrives/{teamDriveId}"
#>   ..$ parameters    :List of 8
#>   ..$ scopes        : chr "drive"
#>   ..$ description   : chr "Deprecated use drives.delete instead."
#>   ..$ parameterOrder: chr "teamDriveId"
#>  $ drive.teamdrives.get           :List of 8
#>   ..$ id            : chr "drive.teamdrives.get"
#>   ..$ httpMethod    : chr "GET"
#>   ..$ path          : 'fs_path' chr "drive/v3/teamdrives/{teamDriveId}"
#>   ..$ parameters    :List of 9
#>   ..$ scopes        : chr "drive, drive.readonly"
#>   ..$ description   : chr "Deprecated use drives.get instead."
#>   ..$ response      : chr "TeamDrive"
#>   ..$ parameterOrder: chr "teamDriveId"
#>  $ drive.teamdrives.list          :List of 7
#>   ..$ id         : chr "drive.teamdrives.list"
#>   ..$ httpMethod : chr "GET"
#>   ..$ path       : 'fs_path' chr "drive/v3/teamdrives"
#>   ..$ parameters :List of 11
#>   ..$ scopes     : chr "drive, drive.readonly"
#>   ..$ description: chr "Deprecated use drives.list instead."
#>   ..$ response   : chr "TeamDriveList"
#>  $ drive.teamdrives.update        :List of 9
#>   ..$ id            : chr "drive.teamdrives.update"
#>   ..$ httpMethod    : chr "PATCH"
#>   ..$ path          : 'fs_path' chr "drive/v3/teamdrives/{teamDriveId}"
#>   ..$ parameters    :List of 19
#>   ..$ scopes        : chr "drive"
#>   ..$ description   : chr "Deprecated use drives.update instead"
#>   ..$ request       : chr "TeamDrive"
#>   ..$ response      : chr "TeamDrive"
#>   ..$ parameterOrder: chr "teamDriveId"
#>  $ drive.files.update.media       :List of 11
#>   ..$ id                 : chr "drive.files.update.media"
#>   ..$ httpMethod         : chr "PATCH"
#>   ..$ path               : chr "/upload/drive/v3/files/{fileId}"
#>   ..$ parameters         :List of 71
#>   ..$ scopes             : chr "drive, drive.appdata, drive.file, drive.metadata, drive.scripts"
#>   ..$ description        : chr "Updates a file's metadata and/or content with patch semantics."
#>   ..$ request            : chr "File"
#>   ..$ response           : chr "File"
#>   ..$ mediaUpload        :List of 3
#>   ..$ supportsMediaUpload: logi TRUE
#>   ..$ parameterOrder     : chr "fileId"
#>  $ drive.files.create.media       :List of 11
#>   ..$ id                  : chr "drive.files.create.media"
#>   ..$ httpMethod          : chr "POST"
#>   ..$ path                : chr "/upload/drive/v3/files"
#>   ..$ parameters          :List of 69
#>   ..$ scopes              : chr "drive, drive.appdata, drive.file"
#>   ..$ description         : chr "Creates a new file."
#>   ..$ request             : chr "File"
#>   ..$ response            : chr "File"
#>   ..$ mediaUpload         :List of 3
#>   ..$ supportsMediaUpload : logi TRUE
#>   ..$ supportsSubscription: logi TRUE

drive_endpoints("drive.files.delete")

#> $drive.files.delete
#> $drive.files.delete$id
#> [1] "drive.files.delete"
#> 
#> $drive.files.delete$httpMethod
#> [1] "DELETE"
#> 
#> $drive.files.delete$path
#> drive/v3/files/{fileId}
#> 
#> $drive.files.delete$parameters
#> $drive.files.delete$parameters$fileId
#> $drive.files.delete$parameters$fileId$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$fileId$description
#> [1] "The ID of the file."
#> 
#> $drive.files.delete$parameters$fileId$required
#> [1] TRUE
#> 
#> $drive.files.delete$parameters$fileId$location
#> [1] "path"
#> 
#> 
#> $drive.files.delete$parameters$supportsAllDrives
#> $drive.files.delete$parameters$supportsAllDrives$type
#> [1] "boolean"
#> 
#> $drive.files.delete$parameters$supportsAllDrives$description
#> [1] "Deprecated - Whether the requesting application supports both My Drives and shared drives. This parameter will only be effective until June 1, 2020. Afterwards all applications are assumed to support shared drives."
#> 
#> $drive.files.delete$parameters$supportsAllDrives$default
#> [1] "false"
#> 
#> $drive.files.delete$parameters$supportsAllDrives$location
#> [1] "query"
#> 
#> 
#> $drive.files.delete$parameters$supportsTeamDrives
#> $drive.files.delete$parameters$supportsTeamDrives$type
#> [1] "boolean"
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$description
#> [1] "Deprecated use supportsAllDrives instead."
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$default
#> [1] "false"
#> 
#> $drive.files.delete$parameters$supportsTeamDrives$location
#> [1] "query"
#> 
#> 
#> $drive.files.delete$parameters$alt
#> $drive.files.delete$parameters$alt$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$alt$description
#> [1] "Data format for the response."
#> 
#> $drive.files.delete$parameters$alt$default
#> [1] "json"
#> 
#> $drive.files.delete$parameters$alt$enum
#> [1] "json"
#> 
#> $drive.files.delete$parameters$alt$enumDescriptions
#> [1] "Responses with Content-Type of application/json"
#> 
#> $drive.files.delete$parameters$alt$location
#> [1] "query"
#> 
#> 
#> $drive.files.delete$parameters$fields
#> $drive.files.delete$parameters$fields$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$fields$description
#> [1] "Selector specifying which fields to include in a partial response."
#> 
#> $drive.files.delete$parameters$fields$location
#> [1] "query"
#> 
#> 
#> $drive.files.delete$parameters$key
#> $drive.files.delete$parameters$key$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$key$description
#> [1] "API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token."
#> 
#> $drive.files.delete$parameters$key$location
#> [1] "query"
#> 
#> 
#> $drive.files.delete$parameters$oauth_token
#> $drive.files.delete$parameters$oauth_token$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$oauth_token$description
#> [1] "OAuth 2.0 token for the current user."
#> 
#> $drive.files.delete$parameters$oauth_token$location
#> [1] "query"
#> 
#> 
#> $drive.files.delete$parameters$prettyPrint
#> $drive.files.delete$parameters$prettyPrint$type
#> [1] "boolean"
#> 
#> $drive.files.delete$parameters$prettyPrint$description
#> [1] "Returns response with indentations and line breaks."
#> 
#> $drive.files.delete$parameters$prettyPrint$default
#> [1] "true"
#> 
#> $drive.files.delete$parameters$prettyPrint$location
#> [1] "query"
#> 
#> 
#> $drive.files.delete$parameters$quotaUser
#> $drive.files.delete$parameters$quotaUser$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$quotaUser$description
#> [1] "An opaque string that represents a user for quota purposes. Must not exceed 40 characters."
#> 
#> $drive.files.delete$parameters$quotaUser$location
#> [1] "query"
#> 
#> 
#> $drive.files.delete$parameters$userIp
#> $drive.files.delete$parameters$userIp$type
#> [1] "string"
#> 
#> $drive.files.delete$parameters$userIp$description
#> [1] "Deprecated. Please use quotaUser instead."
#> 
#> $drive.files.delete$parameters$userIp$location
#> [1] "query"
#> 
#> 
#> 
#> $drive.files.delete$scopes
#> [1] "drive, drive.appdata, drive.file"
#> 
#> $drive.files.delete$description
#> [1] "Permanently deletes a file owned by the user without moving it to the trash. If the file belongs to a shared drive the user must be an organizer on the parent. If the target is a folder, all descendants owned by the user are also deleted."
#> 
#> $drive.files.delete$parameterOrder
#> [1] "fileId"
#> 
#> 

drive_endpoints(4)

#> $drive.changes.watch
#> $drive.changes.watch$id
#> [1] "drive.changes.watch"
#> 
#> $drive.changes.watch$httpMethod
#> [1] "POST"
#> 
#> $drive.changes.watch$path
#> drive/v3/changes/watch
#> 
#> $drive.changes.watch$parameters
#> $drive.changes.watch$parameters$driveId
#> $drive.changes.watch$parameters$driveId$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$driveId$description
#> [1] "The shared drive from which changes will be returned. If specified the change IDs will be reflective of the shared drive; use the combined drive ID and change ID as an identifier."
#> 
#> $drive.changes.watch$parameters$driveId$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$includeCorpusRemovals
#> $drive.changes.watch$parameters$includeCorpusRemovals$type
#> [1] "boolean"
#> 
#> $drive.changes.watch$parameters$includeCorpusRemovals$description
#> [1] "Whether changes should include the file resource if the file is still accessible by the user at the time of the request, even when a file was removed from the list of changes and there will be no further change entries for this file."
#> 
#> $drive.changes.watch$parameters$includeCorpusRemovals$default
#> [1] "false"
#> 
#> $drive.changes.watch$parameters$includeCorpusRemovals$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$includeItemsFromAllDrives
#> $drive.changes.watch$parameters$includeItemsFromAllDrives$type
#> [1] "boolean"
#> 
#> $drive.changes.watch$parameters$includeItemsFromAllDrives$description
#> [1] "Deprecated - Whether both My Drive and shared drive items should be included in results. This parameter will only be effective until June 1, 2020. Afterwards shared drive items will be included in the results."
#> 
#> $drive.changes.watch$parameters$includeItemsFromAllDrives$default
#> [1] "false"
#> 
#> $drive.changes.watch$parameters$includeItemsFromAllDrives$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$includeRemoved
#> $drive.changes.watch$parameters$includeRemoved$type
#> [1] "boolean"
#> 
#> $drive.changes.watch$parameters$includeRemoved$description
#> [1] "Whether to include changes indicating that items have been removed from the list of changes, for example by deletion or loss of access."
#> 
#> $drive.changes.watch$parameters$includeRemoved$default
#> [1] "true"
#> 
#> $drive.changes.watch$parameters$includeRemoved$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$includeTeamDriveItems
#> $drive.changes.watch$parameters$includeTeamDriveItems$type
#> [1] "boolean"
#> 
#> $drive.changes.watch$parameters$includeTeamDriveItems$description
#> [1] "Deprecated use includeItemsFromAllDrives instead."
#> 
#> $drive.changes.watch$parameters$includeTeamDriveItems$default
#> [1] "false"
#> 
#> $drive.changes.watch$parameters$includeTeamDriveItems$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$pageSize
#> $drive.changes.watch$parameters$pageSize$type
#> [1] "integer"
#> 
#> $drive.changes.watch$parameters$pageSize$description
#> [1] "The maximum number of changes to return per page."
#> 
#> $drive.changes.watch$parameters$pageSize$default
#> [1] "100"
#> 
#> $drive.changes.watch$parameters$pageSize$format
#> [1] "int32"
#> 
#> $drive.changes.watch$parameters$pageSize$minimum
#> [1] "1"
#> 
#> $drive.changes.watch$parameters$pageSize$maximum
#> [1] "1000"
#> 
#> $drive.changes.watch$parameters$pageSize$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$pageToken
#> $drive.changes.watch$parameters$pageToken$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$pageToken$description
#> [1] "The token for continuing a previous list request on the next page. This should be set to the value of 'nextPageToken' from the previous response or to the response from the getStartPageToken method."
#> 
#> $drive.changes.watch$parameters$pageToken$required
#> [1] TRUE
#> 
#> $drive.changes.watch$parameters$pageToken$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$restrictToMyDrive
#> $drive.changes.watch$parameters$restrictToMyDrive$type
#> [1] "boolean"
#> 
#> $drive.changes.watch$parameters$restrictToMyDrive$description
#> [1] "Whether to restrict the results to changes inside the My Drive hierarchy. This omits changes to files such as those in the Application Data folder or shared files which have not been added to My Drive."
#> 
#> $drive.changes.watch$parameters$restrictToMyDrive$default
#> [1] "false"
#> 
#> $drive.changes.watch$parameters$restrictToMyDrive$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$spaces
#> $drive.changes.watch$parameters$spaces$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$spaces$description
#> [1] "A comma-separated list of spaces to query within the user corpus. Supported values are 'drive', 'appDataFolder' and 'photos'."
#> 
#> $drive.changes.watch$parameters$spaces$default
#> [1] "drive"
#> 
#> $drive.changes.watch$parameters$spaces$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$supportsAllDrives
#> $drive.changes.watch$parameters$supportsAllDrives$type
#> [1] "boolean"
#> 
#> $drive.changes.watch$parameters$supportsAllDrives$description
#> [1] "Deprecated - Whether the requesting application supports both My Drives and shared drives. This parameter will only be effective until June 1, 2020. Afterwards all applications are assumed to support shared drives."
#> 
#> $drive.changes.watch$parameters$supportsAllDrives$default
#> [1] "false"
#> 
#> $drive.changes.watch$parameters$supportsAllDrives$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$supportsTeamDrives
#> $drive.changes.watch$parameters$supportsTeamDrives$type
#> [1] "boolean"
#> 
#> $drive.changes.watch$parameters$supportsTeamDrives$description
#> [1] "Deprecated use supportsAllDrives instead."
#> 
#> $drive.changes.watch$parameters$supportsTeamDrives$default
#> [1] "false"
#> 
#> $drive.changes.watch$parameters$supportsTeamDrives$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$teamDriveId
#> $drive.changes.watch$parameters$teamDriveId$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$teamDriveId$description
#> [1] "Deprecated use driveId instead."
#> 
#> $drive.changes.watch$parameters$teamDriveId$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$address
#> $drive.changes.watch$parameters$address$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$address$description
#> [1] "The address where notifications are delivered for this channel."
#> 
#> $drive.changes.watch$parameters$address$location
#> [1] "body"
#> 
#> 
#> $drive.changes.watch$parameters$expiration
#> $drive.changes.watch$parameters$expiration$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$expiration$description
#> [1] "Date and time of notification channel expiration, expressed as a Unix timestamp, in milliseconds. Optional."
#> 
#> $drive.changes.watch$parameters$expiration$format
#> [1] "int64"
#> 
#> $drive.changes.watch$parameters$expiration$location
#> [1] "body"
#> 
#> 
#> $drive.changes.watch$parameters$id
#> $drive.changes.watch$parameters$id$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$id$description
#> [1] "A UUID or similar unique string that identifies this channel."
#> 
#> $drive.changes.watch$parameters$id$location
#> [1] "body"
#> 
#> 
#> $drive.changes.watch$parameters$kind
#> $drive.changes.watch$parameters$kind$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$kind$description
#> [1] "Identifies this as a notification channel used to watch for changes to a resource, which is \"api#channel\"."
#> 
#> $drive.changes.watch$parameters$kind$default
#> [1] "api#channel"
#> 
#> $drive.changes.watch$parameters$kind$location
#> [1] "body"
#> 
#> 
#> $drive.changes.watch$parameters$params
#> $drive.changes.watch$parameters$params$type
#> [1] "object"
#> 
#> $drive.changes.watch$parameters$params$description
#> [1] "Additional parameters controlling delivery channel behavior. Optional."
#> 
#> $drive.changes.watch$parameters$params$additionalProperties
#> $drive.changes.watch$parameters$params$additionalProperties$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$params$additionalProperties$description
#> [1] "Declares a new parameter by name."
#> 
#> 
#> $drive.changes.watch$parameters$params$location
#> [1] "body"
#> 
#> 
#> $drive.changes.watch$parameters$payload
#> $drive.changes.watch$parameters$payload$type
#> [1] "boolean"
#> 
#> $drive.changes.watch$parameters$payload$description
#> [1] "A Boolean value to indicate whether payload is wanted. Optional."
#> 
#> $drive.changes.watch$parameters$payload$location
#> [1] "body"
#> 
#> 
#> $drive.changes.watch$parameters$resourceId
#> $drive.changes.watch$parameters$resourceId$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$resourceId$description
#> [1] "An opaque ID that identifies the resource being watched on this channel. Stable across different API versions."
#> 
#> $drive.changes.watch$parameters$resourceId$location
#> [1] "body"
#> 
#> 
#> $drive.changes.watch$parameters$resourceUri
#> $drive.changes.watch$parameters$resourceUri$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$resourceUri$description
#> [1] "A version-specific identifier for the watched resource."
#> 
#> $drive.changes.watch$parameters$resourceUri$location
#> [1] "body"
#> 
#> 
#> $drive.changes.watch$parameters$token
#> $drive.changes.watch$parameters$token$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$token$description
#> [1] "An arbitrary string delivered to the target address with each notification delivered over this channel. Optional."
#> 
#> $drive.changes.watch$parameters$token$location
#> [1] "body"
#> 
#> 
#> $drive.changes.watch$parameters$type
#> $drive.changes.watch$parameters$type$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$type$description
#> [1] "The type of delivery mechanism used for this channel."
#> 
#> $drive.changes.watch$parameters$type$location
#> [1] "body"
#> 
#> 
#> $drive.changes.watch$parameters$alt
#> $drive.changes.watch$parameters$alt$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$alt$description
#> [1] "Data format for the response."
#> 
#> $drive.changes.watch$parameters$alt$default
#> [1] "json"
#> 
#> $drive.changes.watch$parameters$alt$enum
#> [1] "json"
#> 
#> $drive.changes.watch$parameters$alt$enumDescriptions
#> [1] "Responses with Content-Type of application/json"
#> 
#> $drive.changes.watch$parameters$alt$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$fields
#> $drive.changes.watch$parameters$fields$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$fields$description
#> [1] "Selector specifying which fields to include in a partial response."
#> 
#> $drive.changes.watch$parameters$fields$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$key
#> $drive.changes.watch$parameters$key$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$key$description
#> [1] "API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token."
#> 
#> $drive.changes.watch$parameters$key$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$oauth_token
#> $drive.changes.watch$parameters$oauth_token$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$oauth_token$description
#> [1] "OAuth 2.0 token for the current user."
#> 
#> $drive.changes.watch$parameters$oauth_token$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$prettyPrint
#> $drive.changes.watch$parameters$prettyPrint$type
#> [1] "boolean"
#> 
#> $drive.changes.watch$parameters$prettyPrint$description
#> [1] "Returns response with indentations and line breaks."
#> 
#> $drive.changes.watch$parameters$prettyPrint$default
#> [1] "true"
#> 
#> $drive.changes.watch$parameters$prettyPrint$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$quotaUser
#> $drive.changes.watch$parameters$quotaUser$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$quotaUser$description
#> [1] "An opaque string that represents a user for quota purposes. Must not exceed 40 characters."
#> 
#> $drive.changes.watch$parameters$quotaUser$location
#> [1] "query"
#> 
#> 
#> $drive.changes.watch$parameters$userIp
#> $drive.changes.watch$parameters$userIp$type
#> [1] "string"
#> 
#> $drive.changes.watch$parameters$userIp$description
#> [1] "Deprecated. Please use quotaUser instead."
#> 
#> $drive.changes.watch$parameters$userIp$location
#> [1] "query"
#> 
#> 
#> 
#> $drive.changes.watch$scopes
#> [1] "drive, drive.appdata, drive.file, drive.metadata, drive.metadata.readonly, drive.photos.readonly, drive.readonly"
#> 
#> $drive.changes.watch$description
#> [1] "Subscribes to changes for a user."
#> 
#> $drive.changes.watch$request
#> [1] "Channel"
#> 
#> $drive.changes.watch$response
#> [1] "Channel"
#> 
#> $drive.changes.watch$parameterOrder
#> [1] "pageToken"
#> 
#> $drive.changes.watch$supportsSubscription
#> [1] TRUE
#> 
#>