drive_examples {googledrive}	R Documentation

Example files

Description

googledrive makes a variety of example files – both local and remote – available for use in examples and reprexes. These functions help you access the example files. See vignette("example-files", package = "googledrive") for more.

Usage

drive_examples_local(matches)

drive_examples_remote(matches)

drive_example_local(matches)

drive_example_remote(matches)

Arguments

matches
A regular expression that matches the name of the desired example file(s). This argument is optional for the plural forms (drive_examples_local() and drive_examples_remote() and, if provided, multiple matches are allowed. The single forms (drive_example_local() and drive_example_reomote() require this argument and require that there is exactly one match.

Value

For drive_example_local() and drive_examples_local(), one or more local filepaths.

For drive_example_remote() and drive_examples_remote(), a dribble.

Examples

drive_examples_remote()
drive_examples_remote("chicken")
drive_example_remote("chicken_doc")

drive_examples_local() %>% basename()
drive_examples_local("chicken") %>% basename()
drive_example_local("imdb")

[Package googledrive version 2.0.0 Index]