Get Drive files by path or id

Retrieve metadata for files specified via path or via file id. Note that Google Drive does NOT behave like your local file system:

You can get zero, one, or more files back for each file name or path! On Google Drive, file and folder names need not be unique, even at a given level of the hierarchy.

A file or folder can also have multiple direct parents, so a single Drive file can potentially be represented by multiple paths.

In contrast, a file id will always identify at most one Drive file. Finally, note also that a folder is just a specific type of file on Drive.

If the files are specified via path, versus id, the returned dribble will include a path variable. To add path information to any dribble that lacks it, use drive_add_path(). If you want to list the contents of a folder, use drive_ls(). For general searching, use drive_find().

Note: Team Drives are only available to users of certain enhanced Google services, such as G Suite Enterprise, G Suite Business, or G Suite for Education.

drive_get(path = NULL, id = NULL, team_drive = NULL, corpora = NULL,
  verbose = TRUE)

Arguments

path
Character vector of path(s) to get. Use a trailing slash to indicate explicitly that a path is a folder, which can disambiguate if there is a file of the same name (yes this is possible on Drive!). A character vector marked with as_id() is treated as if it was provided via the id argument.

id
Character vector of Drive file ids or URLs (it is first processed with as_id()). If both path and id are non-NULL, id is silently ignored.

team_drive
Anything that identifies one specific Team Drive: its name, its id or URL marked with as_id(), or a dribble. Is pre-processed with as_teamdrive() and then drive_corpus().

corpora
Character, specifying the search collection. Only relevant in the Team Drives context. Is pre-processed together with team_drive by drive_corpus().

verbose
Logical, indicating whether to print informative messages (default TRUE).

Value

An object of class dribble, a tibble with one row per item.

See also

Wraps the files.get endpoint and, if you specify files by name or path, also calls files.list:

https://developers.google.com/drive/v3/reference/files/update

https://developers.google.com/drive/v3/reference/files/list

Examples

# NOT RUN {
## get info about your "My Drive" root folder
drive_get("~/")
## the API reserves the file id "root" for your root folder
drive_get(id = "root")
drive_get(id = "root") %>% drive_add_path()

## multiple names
drive_get(c("abc", "def"))

## multiple names, one of which must be a folder
drive_get(c("abc", "def/"))

## query by file id(s)
drive_get(id = "abcdefgeh123456789")
drive_get(as_id("abcdefgeh123456789"))
drive_get(id = c("abcdefgh123456789", "jklmnopq123456789"))

## access the Team Drive named "foo"
foo <- teamdrive_get("foo")
drive_get(c("this.jpg", "that-file"), team_drive = foo)
drive_get(as_id("123456789"), team_drive = foo)

## search all Team Drives and other files user has accessed
drive_get(c("this.jpg", "that-file"), corpora = "user,allTeamDrives")
# }