List contents of a folder.

List the contents of a folder on Google Drive, nonrecursively. Optionally, filter for a regex in the file names and/or on MIME type. This is a thin wrapper around drive_search(). drive_list() is an alias for drive_ls().

drive_ls(path = "~/", pattern = NULL, type = NULL, ...)

drive_list(path = "~/", pattern = NULL, type = NULL, ...)

Arguments

path
Specifies a single folder on Google Drive whose contents you want to list. Can be an actual path (character), a file id marked with as_id(), or a dribble.

pattern
Character. If provided, only the files whose names match this regular expression are returned.

type
Character. If provided, only files of this type will be returned. Can be anything that drive_mime_type() knows how to handle. This is processed by googledrive and sent as a query parameter.

...
Query parameters to pass along to the API query.

Value

An object of class dribble, a tibble with one row per file.

Examples

# NOT RUN {
## get contents of the folder 'abc' (non-recursive)
drive_ls("abc")

## get contents of folder 'abc' that contain the
## letters 'def'
drive_ls(path = "abc", pattern = "def")

## get all Google spreadsheets in folder 'abc'
## that contain the letters 'def'
drive_ls(path = "abc", pattern = "def", type = "spreadsheet")
# }