Lookup MIME type

This is a helper to determine which MIME type should be used for a file. Three types of input are acceptable:

Native Google Drive file types. Important examples:

"document" for Google Docs

"folder" for folders

"presentation" for Google Slides

"spreadsheet" for Google Sheets

File extensions, such as "pdf", "csv", etc.

MIME types accepted by Google Drive (these are simply passed through).

drive_mime_type(type = NULL)

Arguments

type
Character. Google Drive file type, file extension, or MIME type. Pass the sentinel expose() if you want to get the full table used for validation and lookup, i.e. all MIME types known to be relevant to the Drive API.

Value

Character. MIME type.

Examples

## get the mime type for Google Spreadsheets
drive_mime_type("spreadsheet")

#> [1] "application/vnd.google-apps.spreadsheet"

## get the mime type for jpegs
drive_mime_type("jpeg")

#> [1] "image/jpeg"

## it's vectorized
drive_mime_type(c("presentation", "pdf", "image/gif"))

#> [1] "application/vnd.google-apps.presentation"
#> [2] "application/pdf"                         
#> [3] "image/gif"                               

## see the internal tibble of MIME types known to the Drive API
drive_mime_type(expose())

#> [90m# A tibble: 77 x 4[39m
#>    mime_type                                           ext   human_type default
#>    [3m[90m<chr>[39m[23m                                               [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m      [3m[90m<lgl>[39m[23m  
#> [90m 1[39m application/x-vnd.oasis.opendocument.presentation   [31mNA[39m    [31mNA[39m         [31mNA[39m     
#> [90m 2[39m text/tab-separated-values                           tsv   tsv        TRUE   
#> [90m 3[39m image/jpeg                                          jpeg  jpeg       TRUE   
#> [90m 4[39m image/jpeg                                          jpg   jpg        FALSE  
#> [90m 5[39m image/jpeg                                          jpe   jpe        FALSE  
#> [90m 6[39m image/bmp                                           [31mNA[39m    [31mNA[39m         [31mNA[39m     
#> [90m 7[39m image/gif                                           gif   gif        TRUE   
#> [90m 8[39m application/vnd.ms-excel.sheet.macroenabled.12      [31mNA[39m    [31mNA[39m         [31mNA[39m     
#> [90m 9[39m application/vnd.openxmlformats-officedocument.word… dotx  dotx       TRUE   
#> [90m10[39m application/vnd.ms-powerpoint.presentation.macroen… [31mNA[39m    [31mNA[39m         [31mNA[39m     
#> [90m# … with 67 more rows[39m