Create a folder on Drive.

Create a folder on Drive.

drive_mkdir(path = NULL, name = NULL, verbose = TRUE)

Arguments

path
Specifies target destination for the new folder on Google Drive. Can be an actual path (character), a file id marked with as_id(), or a dribble. If specified as an actual path, it is best to explicitly indicate if it's a folder by including a trailing slash, since it cannot always be worked out from the context of the call.

name
Character, new folder name if not specified as part of path. This will force path to be treated as a folder, even if it is character and lacks a trailing slash.

verbose
Logical, indicating whether to print informative messages (default TRUE).

Value

An object of class dribble, a tibble with one row per file.

Examples

# NOT RUN {
## Create folder named "def" in existing folder "abc".
drive_mkdir("abc/def")

## This will also create a folder named "def" in folder "abc".
drive_mkdir(path = "abc", name = "def")

## Another way to create a folder named "def" in folder "abc",
## this time with parent folder stored in a dribble.
abc <- as_dribble("abc")
drive_mkdir(path = abc, name = "def")

## Yet another way to do this,
## this time with parent folder provide via id.
drive_mkdir(path = as_id(abc$id), name = "def")
# }