Move a Drive file

Move a Drive file to a different folder, give it a different name, or both.

Usage

drive_mv(
  file,
  path = NULL,
  name = NULL,
  overwrite = NA,
  verbose = deprecated()
)

Arguments

file

Something that identifies the file of interest on your Google Drive. Can be a name or path, a file id or URL marked with as_id(), or a dribble.

path

Specifies target destination for the file on Google Drive. Can be an actual path (character), a file id marked with as_id(), or a dribble.

If path is a shortcut to a folder, it is automatically resolved to its target folder.

If path is given as a path (as opposed to a dribble or an id), it is best to explicitly indicate if it's a folder by including a trailing slash, since it cannot always be worked out from the context of the call. By default, the file stays in its current folder.

name

Character, new file name if not specified as part of path. This will force path to be interpreted as a folder, even if it is character and lacks a trailing slash. By default, the file keeps its current name.

overwrite

Logical, indicating whether to check for a pre-existing file at the targetted "filepath". The quotes around "filepath" refer to the fact that Drive does not impose a 1-to-1 relationship between filepaths and files, like a typical file system; read more about that in drive_get().

NA (default): Just do the operation, even if it results in multiple files with the same filepath.

TRUE: Check for a pre-existing file at the filepath. If there is zero or one, move a pre-existing file to the trash, then carry on. Note that the new file does not inherit any properties from the old one, such as sharing or publishing settings. It will have a new file ID. An error is thrown if two or more pre-existing files are found.

FALSE: Error if there is any pre-existing file at the filepath.

Note that existence checks, based on filepath, are expensive operations, i.e. they require additional API calls.

verbose

This logical argument to individual googledrive functions is deprecated. To globally suppress googledrive messaging, use options(googledrive_quiet = TRUE) (the default behaviour is to emit informational messages). To suppress messaging in a more limited way, use the helpers local_drive_quiet() or with_drive_quiet().

Value

An object of class dribble, a tibble with one row per file.

See also

Examples

# create a file to move
file <- drive_example_remote("chicken.txt") %>%
  drive_cp("chicken-mv.txt")
#> Original file:
#> • chicken.txt <id: 1wOLeWVRkTb6lDmLRiOhg9iKM7DlN762Y>
#> Copied to file:
#> • chicken-mv.txt <id: 1jnc7M6ui44Jqo8KT4Hvl__03-O40qmpH>

# rename it, but leave in current folder (root folder, in this case)
file <- drive_mv(file, "chicken-mv-renamed.txt")
#> Original file:
#> • chicken-mv.txt <id: 1jnc7M6ui44Jqo8KT4Hvl__03-O40qmpH>
#> Has been renamed:
#> • chicken-mv-renamed.txt <id: 1jnc7M6ui44Jqo8KT4Hvl__03-O40qmpH>

# create a folder to move the file into
folder <- drive_mkdir("mv-folder")
#> Error in confirm_clear_path(path, name): Unclear if `path` specifies parent folder or full path to the new
#> file, including its name.
#> See `?as_dribble()` for advice on how to make this clear.

# move the file and rename it again,
# specify destination as a dribble
file <- drive_mv(file, path = folder, name = "chicken-mv-re-renamed.txt")
#> Error: object 'folder' not found

# verify renamed file is now in the folder
drive_ls(folder)
#> Error: object 'folder' not found

# move the file back to root folder
file <- drive_mv(file, "~/")
#> ! Nothing to be done.

# move it again
# specify destination as path with trailing slash
# to ensure we get a move vs. renaming it to "mv-folder"
file <- drive_mv(file, "mv-folder/")
#> Original file:
#> • chicken-mv-renamed.txt <id: 1jnc7M6ui44Jqo8KT4Hvl__03-O40qmpH>
#> Has been moved:
#> • mv-folder/chicken-mv-renamed.txt
#>   <id: 1jnc7M6ui44Jqo8KT4Hvl__03-O40qmpH>

# `overwrite = FALSE` errors if something already exists at target filepath
# THIS WILL ERROR!
drive_create("name-squatter-mv", path = "~/")
#> Created Drive file:
#> • name-squatter-mv <id: 1FJphxfT8bY1vZ-Rw34ik3m0IEwfkaTh2>
#> With MIME type:
#> • application/octet-stream
drive_mv(file, path = "~/", name = "name-squatter-mv", overwrite = FALSE)
#> Error in check_for_overwrite(parent = params[["addParents"]] %||% parent_before,     name = params[["name"]] %||% file$name, overwrite = overwrite): 149 items already exist at the target filepath and `overwrite =
#> FALSE`:
#> • name-squatter-mv <id: 1FJphxfT8bY1vZ-Rw34ik3m0IEwfkaTh2>
#> • name-squatter-mv <id: 1RVhR_pm9vmMuXPRG-vqybp8BQtqp0vh->
#> • name-squatter-mv <id: 19_OnShH50oZanyEyZmDx-nA6IP5ItfsY>
#> • name-squatter-mv <id: 1j-zIE-h98-pTZJHlqWA6OoN4XLATaa6x>
#> • name-squatter-mv <id: 1MMIW037tKPcmGEqsbakS0EtOVy9hXddf>
#>   … and 144 more

# `overwrite = TRUE` moves the existing item to trash, then proceeds
drive_mv(file, path = "~/", name = "name-squatter-mv", overwrite = TRUE)
#> Error in check_for_overwrite(parent = params[["addParents"]] %||% parent_before,     name = params[["name"]] %||% file$name, overwrite = overwrite): Multiple items already exist at the target filepath.
#> • name-squatter-mv <id: 1FJphxfT8bY1vZ-Rw34ik3m0IEwfkaTh2>
#> • name-squatter-mv <id: 1RVhR_pm9vmMuXPRG-vqybp8BQtqp0vh->
#> • name-squatter-mv <id: 19_OnShH50oZanyEyZmDx-nA6IP5ItfsY>
#> • name-squatter-mv <id: 1j-zIE-h98-pTZJHlqWA6OoN4XLATaa6x>
#> • name-squatter-mv <id: 1MMIW037tKPcmGEqsbakS0EtOVy9hXddf>
#>   … and 144 more
#> Although `overwrite = TRUE`, it's not clear which item to overwrite.
#> Use `overwrite = NA` to suppress this check. Exiting.

# Clean up
drive_rm(file, folder)
#> Error: object 'folder' not found