Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file
Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what
Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via what appears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 x 4
#>    name          path                id                          drive_resource 
#>  * <chr>         <chr>               <chr>                       <list>         
#>  1 name-squatter ~/name-squatter     1xTfbXGc9Jgp79C1Fi8rgMT0Z_… <named list [3…
#>  2 name-squatter ~/name-squatter     1lyITYj1J_8oGSZ4-glOjeAyYE… <named list [3…
#>  3 name-squatter ~/name-squatter     1wfClch9-GsrQ6dyL1aDifyjcQ… <named list [3…
#>  4 name-squatter ~/name-squatter     19kwAjx00T_K86we2yI4Zhb-X1… <named list [3…
#>  5 name-squatter ~/name-squatter     1vDnIumOzpcpWqLLVpehm3_5aa… <named list [3…
#>  6 chicken_sheet ~/chicken_sheet     1s0kEHcqG2PyciERoGq52L_Qwz… <named list [3…
#>  7 r_logo.jpg    ~/r_logo.jpg        1wFAZdmBiSRu4GShsqurxD7wID… <named list [4…
#>  8 THANKS        ~/THANKS            19URV7BT0_E1KhYdfDODszK5ai… <named list [4…
#>  9 googledrive-… ~/abc/def/googledr… 1h1lhFfQrDZevE2OEX10-rbi2B… <named list [3…
#> 10 def           ~/abc/def/          1ALSW_Nqs7FsPOcrJ6MqyBoRm0… <named list [3…
drive_reveal(files, "permissions")
#> # A dribble: 10 x 5
#>    name         shared id                      drive_resource  permissions_reso…
#>  * <chr>        <lgl>  <chr>                   <list>          <list>           
#>  1 name-squatt… FALSE  1xTfbXGc9Jgp79C1Fi8rgM… <named list [3… <named list [2]> 
#>  2 name-squatt… FALSE  1lyITYj1J_8oGSZ4-glOje… <named list [3… <named list [2]> 
#>  3 name-squatt… FALSE  1wfClch9-GsrQ6dyL1aDif… <named list [3… <named list [2]> 
#>  4 name-squatt… FALSE  19kwAjx00T_K86we2yI4Zh… <named list [3… <named list [2]> 
#>  5 name-squatt… FALSE  1vDnIumOzpcpWqLLVpehm3… <named list [3… <named list [2]> 
#>  6 chicken_she… FALSE  1s0kEHcqG2PyciERoGq52L… <named list [3… <named list [2]> 
#>  7 r_logo.jpg   TRUE   1wFAZdmBiSRu4GShsqurxD… <named list [4… <named list [2]> 
#>  8 THANKS       TRUE   19URV7BT0_E1KhYdfDODsz… <named list [4… <named list [2]> 
#>  9 googledrive… FALSE  1h1lhFfQrDZevE2OEX10-r… <named list [3… <named list [2]> 
#> 10 def          FALSE  1ALSW_Nqs7FsPOcrJ6MqyB… <named list [3… <named list [2]> 
drive_reveal(files, "published")
#> # A dribble: 10 x 5
#>    name        published id                     drive_resource  revision_resour…
#>  * <chr>       <lgl>     <chr>                  <list>          <list>          
#>  1 name-squat… FALSE     1xTfbXGc9Jgp79C1Fi8rg… <named list [3… <named list [9]>
#>  2 name-squat… FALSE     1lyITYj1J_8oGSZ4-glOj… <named list [3… <named list [9]>
#>  3 name-squat… FALSE     1wfClch9-GsrQ6dyL1aDi… <named list [3… <named list [9]>
#>  4 name-squat… FALSE     19kwAjx00T_K86we2yI4Z… <named list [3… <named list [9]>
#>  5 name-squat… FALSE     1vDnIumOzpcpWqLLVpehm… <named list [3… <named list [9]>
#>  6 chicken_sh… FALSE     1s0kEHcqG2PyciERoGq52… <named list [3… <named list [7]>
#>  7 r_logo.jpg  FALSE     1wFAZdmBiSRu4GShsqurx… <named list [4… <named list [10…
#>  8 THANKS      FALSE     19URV7BT0_E1KhYdfDODs… <named list [4… <named list [10…
#>  9 googledriv… FALSE     1h1lhFfQrDZevE2OEX10-… <named list [3… <named list [10…
#> 10 def         FALSE     1ALSW_Nqs7FsPOcrJ6Mqy… <named list [3… <NULL>          

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 x 4
#>    name          id_parent              id                       drive_resource 
#>    <chr>         <chr>                  <chr>                    <list>         
#>  1 name-squatter 0AO_RMaBzcP63Uk9PVA    1xTfbXGc9Jgp79C1Fi8rgMT… <named list [3…
#>  2 name-squatter 0AO_RMaBzcP63Uk9PVA    1lyITYj1J_8oGSZ4-glOjeA… <named list [3…
#>  3 name-squatter 0AO_RMaBzcP63Uk9PVA    1wfClch9-GsrQ6dyL1aDify… <named list [3…
#>  4 name-squatter 0AO_RMaBzcP63Uk9PVA    19kwAjx00T_K86we2yI4Zhb… <named list [3…
#>  5 name-squatter 0AO_RMaBzcP63Uk9PVA    1vDnIumOzpcpWqLLVpehm3_… <named list [3…
#>  6 chicken_sheet 0AO_RMaBzcP63Uk9PVA    1s0kEHcqG2PyciERoGq52L_… <named list [3…
#>  7 r_logo.jpg    0AO_RMaBzcP63Uk9PVA    1wFAZdmBiSRu4GShsqurxD7… <named list [4…
#>  8 THANKS        0AO_RMaBzcP63Uk9PVA    19URV7BT0_E1KhYdfDODszK… <named list [4…
#>  9 googledrive-… 1ALSW_Nqs7FsPOcrJ6Mqy… 1h1lhFfQrDZevE2OEX10-rb… <named list [3…
#> 10 def           1o89YN5n4325GbUA86Wp6… 1ALSW_Nqs7FsPOcrJ6MqyBo… <named list [3…

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 x 4
#>    name             trashed id                                  drive_resource  
#>  * <chr>            <lgl>   <chr>                               <list>          
#>  1 name-squatter    TRUE    1xTfbXGc9Jgp79C1Fi8rgMT0Z_esWHgGz   <named list [39…
#>  2 name-squatter    TRUE    1lyITYj1J_8oGSZ4-glOjeAyYE65O2tok   <named list [39…
#>  3 name-squatter    TRUE    1wfClch9-GsrQ6dyL1aDifyjcQ1bCbOCk   <named list [39…
#>  4 name-squatter    TRUE    19kwAjx00T_K86we2yI4Zhb-X1OUieg6c   <named list [39…
#>  5 name-squatter    TRUE    1vDnIumOzpcpWqLLVpehm3_5aaupwZVll   <named list [39…
#>  6 chicken_sheet    FALSE   1s0kEHcqG2PyciERoGq52L_Qwzp4y3__rB… <named list [35…
#>  7 r_logo.jpg       FALSE   1wFAZdmBiSRu4GShsqurxD7wIDSCZvPud   <named list [41…
#>  8 THANKS           FALSE   19URV7BT0_E1KhYdfDODszK5aiELOwTSz   <named list [40…
#>  9 googledrive-NEW… FALSE   1h1lhFfQrDZevE2OEX10-rbi2BfvGogFm   <named list [39…
#> 10 def              FALSE   1ALSW_Nqs7FsPOcrJ6MqyBoRm03gansmn   <named list [33…
drive_reveal(files, "mime_type")
#> # A dribble: 10 x 4
#>    name         mime_type                id                      drive_resource 
#>  * <chr>        <chr>                    <chr>                   <list>         
#>  1 name-squatt… application/octet-stream 1xTfbXGc9Jgp79C1Fi8rgM… <named list [3…
#>  2 name-squatt… application/octet-stream 1lyITYj1J_8oGSZ4-glOje… <named list [3…
#>  3 name-squatt… application/octet-stream 1wfClch9-GsrQ6dyL1aDif… <named list [3…
#>  4 name-squatt… application/octet-stream 19kwAjx00T_K86we2yI4Zh… <named list [3…
#>  5 name-squatt… application/octet-stream 1vDnIumOzpcpWqLLVpehm3… <named list [3…
#>  6 chicken_she… application/vnd.google-… 1s0kEHcqG2PyciERoGq52L… <named list [3…
#>  7 r_logo.jpg   image/jpeg               1wFAZdmBiSRu4GShsqurxD… <named list [4…
#>  8 THANKS       text/plain               19URV7BT0_E1KhYdfDODsz… <named list [4…
#>  9 googledrive… text/markdown            1h1lhFfQrDZevE2OEX10-r… <named list [3…
#> 10 def          application/vnd.google-… 1ALSW_Nqs7FsPOcrJ6MqyB… <named list [3…
drive_reveal(files, "starred")
#> # A dribble: 10 x 4
#>    name             starred id                                  drive_resource  
#>  * <chr>            <lgl>   <chr>                               <list>          
#>  1 name-squatter    FALSE   1xTfbXGc9Jgp79C1Fi8rgMT0Z_esWHgGz   <named list [39…
#>  2 name-squatter    FALSE   1lyITYj1J_8oGSZ4-glOjeAyYE65O2tok   <named list [39…
#>  3 name-squatter    FALSE   1wfClch9-GsrQ6dyL1aDifyjcQ1bCbOCk   <named list [39…
#>  4 name-squatter    FALSE   19kwAjx00T_K86we2yI4Zhb-X1OUieg6c   <named list [39…
#>  5 name-squatter    FALSE   1vDnIumOzpcpWqLLVpehm3_5aaupwZVll   <named list [39…
#>  6 chicken_sheet    FALSE   1s0kEHcqG2PyciERoGq52L_Qwzp4y3__rB… <named list [35…
#>  7 r_logo.jpg       TRUE    1wFAZdmBiSRu4GShsqurxD7wIDSCZvPud   <named list [41…
#>  8 THANKS           TRUE    19URV7BT0_E1KhYdfDODszK5aiELOwTSz   <named list [40…
#>  9 googledrive-NEW… FALSE   1h1lhFfQrDZevE2OEX10-rbi2BfvGogFm   <named list [39…
#> 10 def              FALSE   1ALSW_Nqs7FsPOcrJ6MqyBoRm03gansmn   <named list [33…
drive_reveal(files, "description")
#> # A dribble: 10 x 4
#>    name            description id                               drive_resource  
#>  * <chr>           <list>      <chr>                            <list>          
#>  1 name-squatter   <NULL>      1xTfbXGc9Jgp79C1Fi8rgMT0Z_esWHg… <named list [39…
#>  2 name-squatter   <NULL>      1lyITYj1J_8oGSZ4-glOjeAyYE65O2t… <named list [39…
#>  3 name-squatter   <NULL>      1wfClch9-GsrQ6dyL1aDifyjcQ1bCbO… <named list [39…
#>  4 name-squatter   <NULL>      19kwAjx00T_K86we2yI4Zhb-X1OUieg… <named list [39…
#>  5 name-squatter   <NULL>      1vDnIumOzpcpWqLLVpehm3_5aaupwZV… <named list [39…
#>  6 chicken_sheet   <NULL>      1s0kEHcqG2PyciERoGq52L_Qwzp4y3_… <named list [35…
#>  7 r_logo.jpg      <NULL>      1wFAZdmBiSRu4GShsqurxD7wIDSCZvP… <named list [41…
#>  8 THANKS          <NULL>      19URV7BT0_E1KhYdfDODszK5aiELOwT… <named list [40…
#>  9 googledrive-NE… <NULL>      1h1lhFfQrDZevE2OEX10-rbi2BfvGog… <named list [39…
#> 10 def             <NULL>      1ALSW_Nqs7FsPOcrJ6MqyBoRm03gans… <named list [33…
drive_reveal(files, "version")
#> # A dribble: 10 x 4
#>    name             version id                                  drive_resource  
#>  * <chr>            <chr>   <chr>                               <list>          
#>  1 name-squatter    2       1xTfbXGc9Jgp79C1Fi8rgMT0Z_esWHgGz   <named list [39…
#>  2 name-squatter    2       1lyITYj1J_8oGSZ4-glOjeAyYE65O2tok   <named list [39…
#>  3 name-squatter    2       1wfClch9-GsrQ6dyL1aDifyjcQ1bCbOCk   <named list [39…
#>  4 name-squatter    2       19kwAjx00T_K86we2yI4Zhb-X1OUieg6c   <named list [39…
#>  5 name-squatter    2       1vDnIumOzpcpWqLLVpehm3_5aaupwZVll   <named list [39…
#>  6 chicken_sheet    5       1s0kEHcqG2PyciERoGq52L_Qwzp4y3__rB… <named list [35…
#>  7 r_logo.jpg       3       1wFAZdmBiSRu4GShsqurxD7wIDSCZvPud   <named list [41…
#>  8 THANKS           4       19URV7BT0_E1KhYdfDODszK5aiELOwTSz   <named list [40…
#>  9 googledrive-NEW… 1       1h1lhFfQrDZevE2OEX10-rbi2BfvGogFm   <named list [39…
#> 10 def              1       1ALSW_Nqs7FsPOcrJ6MqyBoRm03gansmn   <named list [33…
drive_reveal(files, "web_view_link")
#> # A dribble: 10 x 4
#>    name       web_view_link                     id                drive_resource
#>  * <chr>      <chr>                             <chr>             <list>        
#>  1 name-squa… https://drive.google.com/file/d/… 1xTfbXGc9Jgp79C1… <named list […
#>  2 name-squa… https://drive.google.com/file/d/… 1lyITYj1J_8oGSZ4… <named list […
#>  3 name-squa… https://drive.google.com/file/d/… 1wfClch9-GsrQ6dy… <named list […
#>  4 name-squa… https://drive.google.com/file/d/… 19kwAjx00T_K86we… <named list […
#>  5 name-squa… https://drive.google.com/file/d/… 1vDnIumOzpcpWqLL… <named list […
#>  6 chicken_s… https://docs.google.com/spreadsh… 1s0kEHcqG2PyciER… <named list […
#>  7 r_logo.jpg https://drive.google.com/file/d/… 1wFAZdmBiSRu4GSh… <named list […
#>  8 THANKS     https://drive.google.com/file/d/… 19URV7BT0_E1KhYd… <named list […
#>  9 googledri… https://drive.google.com/file/d/… 1h1lhFfQrDZevE2O… <named list […
#> 10 def        https://drive.google.com/drive/f… 1ALSW_Nqs7FsPOcr… <named list […
drive_reveal(files, "modified_time")
#> # A dribble: 10 x 4
#>    name          modified_time       id                          drive_resource 
#>    <chr>         <dttm>              <chr>                       <list>         
#>  1 name-squatter 2021-07-01 05:26:57 1xTfbXGc9Jgp79C1Fi8rgMT0Z_… <named list [3…
#>  2 name-squatter 2021-07-01 05:26:41 1lyITYj1J_8oGSZ4-glOjeAyYE… <named list [3…
#>  3 name-squatter 2021-07-01 05:16:07 1wfClch9-GsrQ6dyL1aDifyjcQ… <named list [3…
#>  4 name-squatter 2021-07-01 05:15:37 19kwAjx00T_K86we2yI4Zhb-X1… <named list [3…
#>  5 name-squatter 2021-07-01 05:15:22 1vDnIumOzpcpWqLLVpehm3_5aa… <named list [3…
#>  6 chicken_sheet 2021-06-30 22:41:06 1s0kEHcqG2PyciERoGq52L_Qwz… <named list [3…
#>  7 r_logo.jpg    2021-06-30 22:43:02 1wFAZdmBiSRu4GShsqurxD7wID… <named list [4…
#>  8 THANKS        2021-06-30 22:43:01 19URV7BT0_E1KhYdfDODszK5ai… <named list [4…
#>  9 googledrive-… 2021-06-30 22:41:50 1h1lhFfQrDZevE2OEX10-rbi2B… <named list [3…
#> 10 def           2021-06-30 22:41:43 1ALSW_Nqs7FsPOcrJ6MqyBoRm0… <named list [3…
drive_reveal(files, "created_time")
#> # A dribble: 10 x 4
#>    name          created_time        id                          drive_resource 
#>    <chr>         <dttm>              <chr>                       <list>         
#>  1 name-squatter 2021-07-01 05:26:57 1xTfbXGc9Jgp79C1Fi8rgMT0Z_… <named list [3…
#>  2 name-squatter 2021-07-01 05:26:41 1lyITYj1J_8oGSZ4-glOjeAyYE… <named list [3…
#>  3 name-squatter 2021-07-01 05:16:07 1wfClch9-GsrQ6dyL1aDifyjcQ… <named list [3…
#>  4 name-squatter 2021-07-01 05:15:37 19kwAjx00T_K86we2yI4Zhb-X1… <named list [3…
#>  5 name-squatter 2021-07-01 05:15:22 1vDnIumOzpcpWqLLVpehm3_5aa… <named list [3…
#>  6 chicken_sheet 2021-06-30 22:41:05 1s0kEHcqG2PyciERoGq52L_Qwz… <named list [3…
#>  7 r_logo.jpg    2021-06-30 22:41:13 1wFAZdmBiSRu4GShsqurxD7wID… <named list [4…
#>  8 THANKS        2021-06-30 22:41:35 19URV7BT0_E1KhYdfDODszK5ai… <named list [4…
#>  9 googledrive-… 2021-06-30 22:41:50 1h1lhFfQrDZevE2OEX10-rbi2B… <named list [3…
#> 10 def           2021-06-30 22:41:43 1ALSW_Nqs7FsPOcrJ6MqyBoRm0… <named list [3…
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 x 4
#>    name            owned_by_me id                               drive_resource  
#>  * <chr>           <lgl>       <chr>                            <list>          
#>  1 name-squatter   TRUE        1xTfbXGc9Jgp79C1Fi8rgMT0Z_esWHg… <named list [39…
#>  2 name-squatter   TRUE        1lyITYj1J_8oGSZ4-glOjeAyYE65O2t… <named list [39…
#>  3 name-squatter   TRUE        1wfClch9-GsrQ6dyL1aDifyjcQ1bCbO… <named list [39…
#>  4 name-squatter   TRUE        19kwAjx00T_K86we2yI4Zhb-X1OUieg… <named list [39…
#>  5 name-squatter   TRUE        1vDnIumOzpcpWqLLVpehm3_5aaupwZV… <named list [39…
#>  6 chicken_sheet   TRUE        1s0kEHcqG2PyciERoGq52L_Qwzp4y3_… <named list [35…
#>  7 r_logo.jpg      TRUE        1wFAZdmBiSRu4GShsqurxD7wIDSCZvP… <named list [41…
#>  8 THANKS          TRUE        19URV7BT0_E1KhYdfDODszK5aiELOwT… <named list [40…
#>  9 googledrive-NE… TRUE        1h1lhFfQrDZevE2OEX10-rbi2BfvGog… <named list [39…
#> 10 def             TRUE        1ALSW_Nqs7FsPOcrJ6MqyBoRm03gans… <named list [33…
drive_reveal(files, "size")
#> # A dribble: 10 x 4
#>    name              size  id                                   drive_resource  
#>  * <chr>             <chr> <chr>                                <list>          
#>  1 name-squatter     0     1xTfbXGc9Jgp79C1Fi8rgMT0Z_esWHgGz    <named list [39…
#>  2 name-squatter     0     1lyITYj1J_8oGSZ4-glOjeAyYE65O2tok    <named list [39…
#>  3 name-squatter     0     1wfClch9-GsrQ6dyL1aDifyjcQ1bCbOCk    <named list [39…
#>  4 name-squatter     0     19kwAjx00T_K86we2yI4Zhb-X1OUieg6c    <named list [39…
#>  5 name-squatter     0     1vDnIumOzpcpWqLLVpehm3_5aaupwZVll    <named list [39…
#>  6 chicken_sheet     NA    1s0kEHcqG2PyciERoGq52L_Qwzp4y3__rBV… <named list [35…
#>  7 r_logo.jpg        15985 1wFAZdmBiSRu4GShsqurxD7wIDSCZvPud    <named list [41…
#>  8 THANKS            3776  19URV7BT0_E1KhYdfDODszK5aiELOwTSz    <named list [40…
#>  9 googledrive-NEWS… 17552 1h1lhFfQrDZevE2OEX10-rbi2BfvGogFm    <named list [39…
#> 10 def               NA    1ALSW_Nqs7FsPOcrJ6MqyBoRm03gansmn    <named list [33…
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 x 4
#>    name           quota_bytes_used id                            drive_resource 
#>  * <chr>          <chr>            <chr>                         <list>         
#>  1 name-squatter  0                1xTfbXGc9Jgp79C1Fi8rgMT0Z_es… <named list [3…
#>  2 name-squatter  0                1lyITYj1J_8oGSZ4-glOjeAyYE65… <named list [3…
#>  3 name-squatter  0                1wfClch9-GsrQ6dyL1aDifyjcQ1b… <named list [3…
#>  4 name-squatter  0                19kwAjx00T_K86we2yI4Zhb-X1OU… <named list [3…
#>  5 name-squatter  0                1vDnIumOzpcpWqLLVpehm3_5aaup… <named list [3…
#>  6 chicken_sheet  0                1s0kEHcqG2PyciERoGq52L_Qwzp4… <named list [3…
#>  7 r_logo.jpg     15985            1wFAZdmBiSRu4GShsqurxD7wIDSC… <named list [4…
#>  8 THANKS         3776             19URV7BT0_E1KhYdfDODszK5aiEL… <named list [4…
#>  9 googledrive-N… 17552            1h1lhFfQrDZevE2OEX10-rbi2Bfv… <named list [3…
#> 10 def            0                1ALSW_Nqs7FsPOcrJ6MqyBoRm03g… <named list [3…

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 x 4
#>   name     path  id                  drive_resource   
#> * <chr>    <chr> <chr>               <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>