Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what

Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via whatappears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name                 path                   id                drive_resource 
#>    <chr>                <chr>                  <drv_id>          <list>         
#>  1 name-squatter-rename ~/name-squatter-rename 11htG4nGA3fxtb6T… <named list [3…
#>  2 name-squatter-mv     ~/name-squatter-mv     1fgkSXStWJz5ErfJ… <named list [3…
#>  3 name-squatter-upload ~/name-squatter-upload 1edy6nFgMFufoInG… <named list [3…
#>  4 name-squatter-rename ~/name-squatter-rename 1y3_0FWsiHnKNMz1… <named list [3…
#>  5 name-squatter-mv     ~/name-squatter-mv     1eFXjbC54DlT3615… <named list [3…
#>  6 name-squatter-upload ~/name-squatter-upload 1-oFj3HpLd6U5_Du… <named list [3…
#>  7 name-squatter-rename ~/name-squatter-rename 1phfAXGRR9JFkKEE… <named list [3…
#>  8 name-squatter-mv     ~/name-squatter-mv     1ptV13HLCrkkMOPA… <named list [3…
#>  9 name-squatter-upload ~/name-squatter-upload 15Ptb85suZQ1yfdA… <named list [3…
#> 10 name-squatter-rename ~/name-squatter-rename 1o1qMT8TUy5DG04g… <named list [3…
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name                 shared id              drive_resource  permissions_reso…
#>    <chr>                <lgl>  <drv_id>        <list>          <list>           
#>  1 name-squatter-rename FALSE  11htG4nGA3fxtb… <named list [3… <named list [2]> 
#>  2 name-squatter-mv     FALSE  1fgkSXStWJz5Er… <named list [3… <named list [2]> 
#>  3 name-squatter-upload FALSE  1edy6nFgMFufoI… <named list [3… <named list [2]> 
#>  4 name-squatter-rename FALSE  1y3_0FWsiHnKNM… <named list [3… <named list [2]> 
#>  5 name-squatter-mv     FALSE  1eFXjbC54DlT36… <named list [3… <named list [2]> 
#>  6 name-squatter-upload FALSE  1-oFj3HpLd6U5_… <named list [3… <named list [2]> 
#>  7 name-squatter-rename FALSE  1phfAXGRR9JFkK… <named list [3… <named list [2]> 
#>  8 name-squatter-mv     FALSE  1ptV13HLCrkkMO… <named list [3… <named list [2]> 
#>  9 name-squatter-upload FALSE  15Ptb85suZQ1yf… <named list [3… <named list [2]> 
#> 10 name-squatter-rename FALSE  1o1qMT8TUy5DG0… <named list [3… <named list [2]> 
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name                 published id            drive_resource  revision_resour…
#>    <chr>                <lgl>     <drv_id>      <list>          <list>          
#>  1 name-squatter-rename FALSE     11htG4nGA3fx… <named list [3… <named list [9]>
#>  2 name-squatter-mv     FALSE     1fgkSXStWJz5… <named list [3… <named list [9]>
#>  3 name-squatter-upload FALSE     1edy6nFgMFuf… <named list [3… <named list [9]>
#>  4 name-squatter-rename FALSE     1y3_0FWsiHnK… <named list [3… <named list [9]>
#>  5 name-squatter-mv     FALSE     1eFXjbC54DlT… <named list [3… <named list [9]>
#>  6 name-squatter-upload FALSE     1-oFj3HpLd6U… <named list [3… <named list [9]>
#>  7 name-squatter-rename FALSE     1phfAXGRR9JF… <named list [3… <named list [9]>
#>  8 name-squatter-mv     FALSE     1ptV13HLCrkk… <named list [3… <named list [9]>
#>  9 name-squatter-upload FALSE     15Ptb85suZQ1… <named list [3… <named list [9]>
#> 10 name-squatter-rename FALSE     1o1qMT8TUy5D… <named list [3… <named list [9]>

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name                 id_parent           id                   drive_resource 
#>    <chr>                <drv_id>            <drv_id>             <list>         
#>  1 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 11htG4nGA3fxtb6Tv-n… <named list [3…
#>  2 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1fgkSXStWJz5ErfJrP5… <named list [3…
#>  3 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1edy6nFgMFufoInG1xX… <named list [3…
#>  4 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1y3_0FWsiHnKNMz1DL4… <named list [3…
#>  5 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1eFXjbC54DlT3615rrf… <named list [3…
#>  6 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1-oFj3HpLd6U5_Duumm… <named list [3…
#>  7 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1phfAXGRR9JFkKEE3JS… <named list [3…
#>  8 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1ptV13HLCrkkMOPA2dh… <named list [3…
#>  9 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 15Ptb85suZQ1yfdAKh8… <named list [3…
#> 10 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1o1qMT8TUy5DG04gxcn… <named list [3…

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name                 trashed id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename TRUE    11htG4nGA3fxtb6Tv-nGTNdAtn3oC1wMn <named list […
#>  2 name-squatter-mv     TRUE    1fgkSXStWJz5ErfJrP5M4oQvAEZwc-Gds <named list […
#>  3 name-squatter-upload TRUE    1edy6nFgMFufoInG1xX-NtfI4xcFXtgwR <named list […
#>  4 name-squatter-rename TRUE    1y3_0FWsiHnKNMz1DL4C8k1rqxRcCq2KE <named list […
#>  5 name-squatter-mv     TRUE    1eFXjbC54DlT3615rrf7ayJrMoPBBj7Jd <named list […
#>  6 name-squatter-upload TRUE    1-oFj3HpLd6U5_DuummQVZbiAf4tCctH4 <named list […
#>  7 name-squatter-rename TRUE    1phfAXGRR9JFkKEE3JS42IJOm3MAKHLAw <named list […
#>  8 name-squatter-mv     TRUE    1ptV13HLCrkkMOPA2dh4zztYEg_DWAG6K <named list […
#>  9 name-squatter-upload TRUE    15Ptb85suZQ1yfdAKh86z94II4Tb93sWh <named list […
#> 10 name-squatter-rename TRUE    1o1qMT8TUy5DG04gxcn70eLQ5wcFIp9Go <named list […
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name                 mime_type                id              drive_resource 
#>    <chr>                <chr>                    <drv_id>        <list>         
#>  1 name-squatter-rename application/octet-stream 11htG4nGA3fxtb… <named list [3…
#>  2 name-squatter-mv     application/octet-stream 1fgkSXStWJz5Er… <named list [3…
#>  3 name-squatter-upload application/octet-stream 1edy6nFgMFufoI… <named list [3…
#>  4 name-squatter-rename application/octet-stream 1y3_0FWsiHnKNM… <named list [3…
#>  5 name-squatter-mv     application/octet-stream 1eFXjbC54DlT36… <named list [3…
#>  6 name-squatter-upload application/octet-stream 1-oFj3HpLd6U5_… <named list [3…
#>  7 name-squatter-rename application/octet-stream 1phfAXGRR9JFkK… <named list [3…
#>  8 name-squatter-mv     application/octet-stream 1ptV13HLCrkkMO… <named list [3…
#>  9 name-squatter-upload application/octet-stream 15Ptb85suZQ1yf… <named list [3…
#> 10 name-squatter-rename application/octet-stream 1o1qMT8TUy5DG0… <named list [3…
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name                 starred id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename FALSE   11htG4nGA3fxtb6Tv-nGTNdAtn3oC1wMn <named list […
#>  2 name-squatter-mv     FALSE   1fgkSXStWJz5ErfJrP5M4oQvAEZwc-Gds <named list […
#>  3 name-squatter-upload FALSE   1edy6nFgMFufoInG1xX-NtfI4xcFXtgwR <named list […
#>  4 name-squatter-rename FALSE   1y3_0FWsiHnKNMz1DL4C8k1rqxRcCq2KE <named list […
#>  5 name-squatter-mv     FALSE   1eFXjbC54DlT3615rrf7ayJrMoPBBj7Jd <named list […
#>  6 name-squatter-upload FALSE   1-oFj3HpLd6U5_DuummQVZbiAf4tCctH4 <named list […
#>  7 name-squatter-rename FALSE   1phfAXGRR9JFkKEE3JS42IJOm3MAKHLAw <named list […
#>  8 name-squatter-mv     FALSE   1ptV13HLCrkkMOPA2dh4zztYEg_DWAG6K <named list […
#>  9 name-squatter-upload FALSE   15Ptb85suZQ1yfdAKh86z94II4Tb93sWh <named list […
#> 10 name-squatter-rename FALSE   1o1qMT8TUy5DG04gxcn70eLQ5wcFIp9Go <named list […
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name                 description id                          drive_resource  
#>    <chr>                <list>      <drv_id>                    <list>          
#>  1 name-squatter-rename <NULL>      11htG4nGA3fxtb6Tv-nGTNdAtn… <named list [39…
#>  2 name-squatter-mv     <NULL>      1fgkSXStWJz5ErfJrP5M4oQvAE… <named list [39…
#>  3 name-squatter-upload <NULL>      1edy6nFgMFufoInG1xX-NtfI4x… <named list [39…
#>  4 name-squatter-rename <NULL>      1y3_0FWsiHnKNMz1DL4C8k1rqx… <named list [39…
#>  5 name-squatter-mv     <NULL>      1eFXjbC54DlT3615rrf7ayJrMo… <named list [39…
#>  6 name-squatter-upload <NULL>      1-oFj3HpLd6U5_DuummQVZbiAf… <named list [39…
#>  7 name-squatter-rename <NULL>      1phfAXGRR9JFkKEE3JS42IJOm3… <named list [39…
#>  8 name-squatter-mv     <NULL>      1ptV13HLCrkkMOPA2dh4zztYEg… <named list [39…
#>  9 name-squatter-upload <NULL>      15Ptb85suZQ1yfdAKh86z94II4… <named list [39…
#> 10 name-squatter-rename <NULL>      1o1qMT8TUy5DG04gxcn70eLQ5w… <named list [39…
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name                 version id                                drive_resource
#>    <chr>                <chr>   <drv_id>                          <list>        
#>  1 name-squatter-rename 2       11htG4nGA3fxtb6Tv-nGTNdAtn3oC1wMn <named list […
#>  2 name-squatter-mv     2       1fgkSXStWJz5ErfJrP5M4oQvAEZwc-Gds <named list […
#>  3 name-squatter-upload 2       1edy6nFgMFufoInG1xX-NtfI4xcFXtgwR <named list […
#>  4 name-squatter-rename 2       1y3_0FWsiHnKNMz1DL4C8k1rqxRcCq2KE <named list […
#>  5 name-squatter-mv     2       1eFXjbC54DlT3615rrf7ayJrMoPBBj7Jd <named list […
#>  6 name-squatter-upload 2       1-oFj3HpLd6U5_DuummQVZbiAf4tCctH4 <named list […
#>  7 name-squatter-rename 2       1phfAXGRR9JFkKEE3JS42IJOm3MAKHLAw <named list […
#>  8 name-squatter-mv     2       1ptV13HLCrkkMOPA2dh4zztYEg_DWAG6K <named list […
#>  9 name-squatter-upload 2       15Ptb85suZQ1yfdAKh86z94II4Tb93sWh <named list […
#> 10 name-squatter-rename 2       1o1qMT8TUy5DG04gxcn70eLQ5wcFIp9Go <named list […
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name                 web_view_link                id           drive_resource
#>    <chr>                <chr>                        <drv_id>     <list>        
#>  1 name-squatter-rename https://drive.google.com/fi… 11htG4nGA3f… <named list […
#>  2 name-squatter-mv     https://drive.google.com/fi… 1fgkSXStWJz… <named list […
#>  3 name-squatter-upload https://drive.google.com/fi… 1edy6nFgMFu… <named list […
#>  4 name-squatter-rename https://drive.google.com/fi… 1y3_0FWsiHn… <named list […
#>  5 name-squatter-mv     https://drive.google.com/fi… 1eFXjbC54Dl… <named list […
#>  6 name-squatter-upload https://drive.google.com/fi… 1-oFj3HpLd6… <named list […
#>  7 name-squatter-rename https://drive.google.com/fi… 1phfAXGRR9J… <named list […
#>  8 name-squatter-mv     https://drive.google.com/fi… 1ptV13HLCrk… <named list […
#>  9 name-squatter-upload https://drive.google.com/fi… 15Ptb85suZQ… <named list […
#> 10 name-squatter-rename https://drive.google.com/fi… 1o1qMT8TUy5… <named list […
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name                 modified_time       id                   drive_resource 
#>    <chr>                <dttm>              <drv_id>             <list>         
#>  1 name-squatter-rename 2022-01-09 13:46:00 11htG4nGA3fxtb6Tv-n… <named list [3…
#>  2 name-squatter-mv     2022-01-09 13:45:47 1fgkSXStWJz5ErfJrP5… <named list [3…
#>  3 name-squatter-upload 2022-01-08 13:47:07 1edy6nFgMFufoInG1xX… <named list [3…
#>  4 name-squatter-rename 2022-01-08 13:46:34 1y3_0FWsiHnKNMz1DL4… <named list [3…
#>  5 name-squatter-mv     2022-01-08 13:46:19 1eFXjbC54DlT3615rrf… <named list [3…
#>  6 name-squatter-upload 2022-01-07 13:48:29 1-oFj3HpLd6U5_Duumm… <named list [3…
#>  7 name-squatter-rename 2022-01-07 13:47:56 1phfAXGRR9JFkKEE3JS… <named list [3…
#>  8 name-squatter-mv     2022-01-07 13:47:41 1ptV13HLCrkkMOPA2dh… <named list [3…
#>  9 name-squatter-upload 2022-01-06 13:49:10 15Ptb85suZQ1yfdAKh8… <named list [3…
#> 10 name-squatter-rename 2022-01-06 13:48:37 1o1qMT8TUy5DG04gxcn… <named list [3…
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name                 created_time        id                   drive_resource 
#>    <chr>                <dttm>              <drv_id>             <list>         
#>  1 name-squatter-rename 2022-01-09 13:46:00 11htG4nGA3fxtb6Tv-n… <named list [3…
#>  2 name-squatter-mv     2022-01-09 13:45:47 1fgkSXStWJz5ErfJrP5… <named list [3…
#>  3 name-squatter-upload 2022-01-08 13:47:07 1edy6nFgMFufoInG1xX… <named list [3…
#>  4 name-squatter-rename 2022-01-08 13:46:34 1y3_0FWsiHnKNMz1DL4… <named list [3…
#>  5 name-squatter-mv     2022-01-08 13:46:19 1eFXjbC54DlT3615rrf… <named list [3…
#>  6 name-squatter-upload 2022-01-07 13:48:29 1-oFj3HpLd6U5_Duumm… <named list [3…
#>  7 name-squatter-rename 2022-01-07 13:47:56 1phfAXGRR9JFkKEE3JS… <named list [3…
#>  8 name-squatter-mv     2022-01-07 13:47:41 1ptV13HLCrkkMOPA2dh… <named list [3…
#>  9 name-squatter-upload 2022-01-06 13:49:10 15Ptb85suZQ1yfdAKh8… <named list [3…
#> 10 name-squatter-rename 2022-01-06 13:48:37 1o1qMT8TUy5DG04gxcn… <named list [3…
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name                 owned_by_me id                          drive_resource  
#>    <chr>                <lgl>       <drv_id>                    <list>          
#>  1 name-squatter-rename TRUE        11htG4nGA3fxtb6Tv-nGTNdAtn… <named list [39…
#>  2 name-squatter-mv     TRUE        1fgkSXStWJz5ErfJrP5M4oQvAE… <named list [39…
#>  3 name-squatter-upload TRUE        1edy6nFgMFufoInG1xX-NtfI4x… <named list [39…
#>  4 name-squatter-rename TRUE        1y3_0FWsiHnKNMz1DL4C8k1rqx… <named list [39…
#>  5 name-squatter-mv     TRUE        1eFXjbC54DlT3615rrf7ayJrMo… <named list [39…
#>  6 name-squatter-upload TRUE        1-oFj3HpLd6U5_DuummQVZbiAf… <named list [39…
#>  7 name-squatter-rename TRUE        1phfAXGRR9JFkKEE3JS42IJOm3… <named list [39…
#>  8 name-squatter-mv     TRUE        1ptV13HLCrkkMOPA2dh4zztYEg… <named list [39…
#>  9 name-squatter-upload TRUE        15Ptb85suZQ1yfdAKh86z94II4… <named list [39…
#> 10 name-squatter-rename TRUE        1o1qMT8TUy5DG04gxcn70eLQ5w… <named list [39…
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name                 size  id                                drive_resource  
#>    <chr>                <chr> <drv_id>                          <list>          
#>  1 name-squatter-rename 0     11htG4nGA3fxtb6Tv-nGTNdAtn3oC1wMn <named list [39…
#>  2 name-squatter-mv     0     1fgkSXStWJz5ErfJrP5M4oQvAEZwc-Gds <named list [39…
#>  3 name-squatter-upload 0     1edy6nFgMFufoInG1xX-NtfI4xcFXtgwR <named list [39…
#>  4 name-squatter-rename 0     1y3_0FWsiHnKNMz1DL4C8k1rqxRcCq2KE <named list [39…
#>  5 name-squatter-mv     0     1eFXjbC54DlT3615rrf7ayJrMoPBBj7Jd <named list [39…
#>  6 name-squatter-upload 0     1-oFj3HpLd6U5_DuummQVZbiAf4tCctH4 <named list [39…
#>  7 name-squatter-rename 0     1phfAXGRR9JFkKEE3JS42IJOm3MAKHLAw <named list [39…
#>  8 name-squatter-mv     0     1ptV13HLCrkkMOPA2dh4zztYEg_DWAG6K <named list [39…
#>  9 name-squatter-upload 0     15Ptb85suZQ1yfdAKh86z94II4Tb93sWh <named list [39…
#> 10 name-squatter-rename 0     1o1qMT8TUy5DG04gxcn70eLQ5wcFIp9Go <named list [39…
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name                 quota_bytes_used id                      drive_resource 
#>    <chr>                <chr>            <drv_id>                <list>         
#>  1 name-squatter-rename 0                11htG4nGA3fxtb6Tv-nGTN… <named list [3…
#>  2 name-squatter-mv     0                1fgkSXStWJz5ErfJrP5M4o… <named list [3…
#>  3 name-squatter-upload 0                1edy6nFgMFufoInG1xX-Nt… <named list [3…
#>  4 name-squatter-rename 0                1y3_0FWsiHnKNMz1DL4C8k… <named list [3…
#>  5 name-squatter-mv     0                1eFXjbC54DlT3615rrf7ay… <named list [3…
#>  6 name-squatter-upload 0                1-oFj3HpLd6U5_DuummQVZ… <named list [3…
#>  7 name-squatter-rename 0                1phfAXGRR9JFkKEE3JS42I… <named list [3…
#>  8 name-squatter-mv     0                1ptV13HLCrkkMOPA2dh4zz… <named list [3…
#>  9 name-squatter-upload 0                15Ptb85suZQ1yfdAKh86z9… <named list [3…
#> 10 name-squatter-rename 0                1o1qMT8TUy5DG04gxcn70e… <named list [3…

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>