Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what

Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via whatappears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name                 path                   id                 drive_resource
#>    <chr>                <chr>                  <drv_id>           <list>        
#>  1 name-squatter-rename ~/name-squatter-rename 1FlZtCAUMzkP6Yr61… <named list>  
#>  2 name-squatter-mv     ~/name-squatter-mv     1fXUcrM9Cj2hHejtC… <named list>  
#>  3 name-squatter-upload ~/name-squatter-upload 1yKcoa6LSDS--tW-t… <named list>  
#>  4 name-squatter-rename ~/name-squatter-rename 10Ez9J-9jBrG159bH… <named list>  
#>  5 name-squatter-mv     ~/name-squatter-mv     1eOSCqNace-6DKvBg… <named list>  
#>  6 name-squatter-upload ~/name-squatter-upload 1M_u3fm8EmeH73yTi… <named list>  
#>  7 name-squatter-rename ~/name-squatter-rename 17j3lXOQiV0XBe8nn… <named list>  
#>  8 name-squatter-mv     ~/name-squatter-mv     1E2b0FiO-s5EEh90Q… <named list>  
#>  9 name-squatter-upload ~/name-squatter-upload 1K8FqSHHEcoQq0xue… <named list>  
#> 10 name-squatter-rename ~/name-squatter-rename 1u8eaTU9pnofrMjxl… <named list>  
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name                 shared id                drive_resource permissions_res…
#>    <chr>                <lgl>  <drv_id>          <list>         <list>          
#>  1 name-squatter-rename FALSE  1FlZtCAUMzkP6Yr6… <named list>   <named list [2]>
#>  2 name-squatter-mv     FALSE  1fXUcrM9Cj2hHejt… <named list>   <named list [2]>
#>  3 name-squatter-upload FALSE  1yKcoa6LSDS--tW-… <named list>   <named list [2]>
#>  4 name-squatter-rename FALSE  10Ez9J-9jBrG159b… <named list>   <named list [2]>
#>  5 name-squatter-mv     FALSE  1eOSCqNace-6DKvB… <named list>   <named list [2]>
#>  6 name-squatter-upload FALSE  1M_u3fm8EmeH73yT… <named list>   <named list [2]>
#>  7 name-squatter-rename FALSE  17j3lXOQiV0XBe8n… <named list>   <named list [2]>
#>  8 name-squatter-mv     FALSE  1E2b0FiO-s5EEh90… <named list>   <named list [2]>
#>  9 name-squatter-upload FALSE  1K8FqSHHEcoQq0xu… <named list>   <named list [2]>
#> 10 name-squatter-rename FALSE  1u8eaTU9pnofrMjx… <named list>   <named list [2]>
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name                 published id             drive_resource revision_resour…
#>    <chr>                <lgl>     <drv_id>       <list>         <list>          
#>  1 name-squatter-rename FALSE     1FlZtCAUMzkP6… <named list>   <named list [9]>
#>  2 name-squatter-mv     FALSE     1fXUcrM9Cj2hH… <named list>   <named list [9]>
#>  3 name-squatter-upload FALSE     1yKcoa6LSDS--… <named list>   <named list [9]>
#>  4 name-squatter-rename FALSE     10Ez9J-9jBrG1… <named list>   <named list [9]>
#>  5 name-squatter-mv     FALSE     1eOSCqNace-6D… <named list>   <named list [9]>
#>  6 name-squatter-upload FALSE     1M_u3fm8EmeH7… <named list>   <named list [9]>
#>  7 name-squatter-rename FALSE     17j3lXOQiV0XB… <named list>   <named list [9]>
#>  8 name-squatter-mv     FALSE     1E2b0FiO-s5EE… <named list>   <named list [9]>
#>  9 name-squatter-upload FALSE     1K8FqSHHEcoQq… <named list>   <named list [9]>
#> 10 name-squatter-rename FALSE     1u8eaTU9pnofr… <named list>   <named list [9]>

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name                 id_parent           id                    drive_resource
#>    <chr>                <drv_id>            <drv_id>              <list>        
#>  1 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1FlZtCAUMzkP6Yr61Rlu… <named list>  
#>  2 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1fXUcrM9Cj2hHejtCByN… <named list>  
#>  3 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1yKcoa6LSDS--tW-ted_… <named list>  
#>  4 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 10Ez9J-9jBrG159bHUqP… <named list>  
#>  5 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1eOSCqNace-6DKvBgJMQ… <named list>  
#>  6 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1M_u3fm8EmeH73yTiWRX… <named list>  
#>  7 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 17j3lXOQiV0XBe8nnwVF… <named list>  
#>  8 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1E2b0FiO-s5EEh90QDaZ… <named list>  
#>  9 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1K8FqSHHEcoQq0xueQnb… <named list>  
#> 10 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1u8eaTU9pnofrMjxlzrz… <named list>  

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name                 trashed id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename TRUE    1FlZtCAUMzkP6Yr61Rlu3qzT-xIN5MEkD <named list>  
#>  2 name-squatter-mv     TRUE    1fXUcrM9Cj2hHejtCByNM2z6_UShZAiEz <named list>  
#>  3 name-squatter-upload TRUE    1yKcoa6LSDS--tW-ted_QtG3DkaNHTK4c <named list>  
#>  4 name-squatter-rename TRUE    10Ez9J-9jBrG159bHUqPMnCJqEnYFx_M1 <named list>  
#>  5 name-squatter-mv     TRUE    1eOSCqNace-6DKvBgJMQLN8I1dDXoP9rc <named list>  
#>  6 name-squatter-upload TRUE    1M_u3fm8EmeH73yTiWRXajujzQ0o0nsdZ <named list>  
#>  7 name-squatter-rename TRUE    17j3lXOQiV0XBe8nnwVFEBKay6JfDg6Am <named list>  
#>  8 name-squatter-mv     TRUE    1E2b0FiO-s5EEh90QDaZU868UaMYo1ZJd <named list>  
#>  9 name-squatter-upload TRUE    1K8FqSHHEcoQq0xueQnb8mmDskvwRAB39 <named list>  
#> 10 name-squatter-rename TRUE    1u8eaTU9pnofrMjxlzrzXT3yrRkymH2ZV <named list>  
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name                 mime_type                id               drive_resource
#>    <chr>                <chr>                    <drv_id>         <list>        
#>  1 name-squatter-rename application/octet-stream 1FlZtCAUMzkP6Yr… <named list>  
#>  2 name-squatter-mv     application/octet-stream 1fXUcrM9Cj2hHej… <named list>  
#>  3 name-squatter-upload application/octet-stream 1yKcoa6LSDS--tW… <named list>  
#>  4 name-squatter-rename application/octet-stream 10Ez9J-9jBrG159… <named list>  
#>  5 name-squatter-mv     application/octet-stream 1eOSCqNace-6DKv… <named list>  
#>  6 name-squatter-upload application/octet-stream 1M_u3fm8EmeH73y… <named list>  
#>  7 name-squatter-rename application/octet-stream 17j3lXOQiV0XBe8… <named list>  
#>  8 name-squatter-mv     application/octet-stream 1E2b0FiO-s5EEh9… <named list>  
#>  9 name-squatter-upload application/octet-stream 1K8FqSHHEcoQq0x… <named list>  
#> 10 name-squatter-rename application/octet-stream 1u8eaTU9pnofrMj… <named list>  
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name                 starred id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename FALSE   1FlZtCAUMzkP6Yr61Rlu3qzT-xIN5MEkD <named list>  
#>  2 name-squatter-mv     FALSE   1fXUcrM9Cj2hHejtCByNM2z6_UShZAiEz <named list>  
#>  3 name-squatter-upload FALSE   1yKcoa6LSDS--tW-ted_QtG3DkaNHTK4c <named list>  
#>  4 name-squatter-rename FALSE   10Ez9J-9jBrG159bHUqPMnCJqEnYFx_M1 <named list>  
#>  5 name-squatter-mv     FALSE   1eOSCqNace-6DKvBgJMQLN8I1dDXoP9rc <named list>  
#>  6 name-squatter-upload FALSE   1M_u3fm8EmeH73yTiWRXajujzQ0o0nsdZ <named list>  
#>  7 name-squatter-rename FALSE   17j3lXOQiV0XBe8nnwVFEBKay6JfDg6Am <named list>  
#>  8 name-squatter-mv     FALSE   1E2b0FiO-s5EEh90QDaZU868UaMYo1ZJd <named list>  
#>  9 name-squatter-upload FALSE   1K8FqSHHEcoQq0xueQnb8mmDskvwRAB39 <named list>  
#> 10 name-squatter-rename FALSE   1u8eaTU9pnofrMjxlzrzXT3yrRkymH2ZV <named list>  
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name                 description id                            drive_resource
#>    <chr>                <list>      <drv_id>                      <list>        
#>  1 name-squatter-rename <NULL>      1FlZtCAUMzkP6Yr61Rlu3qzT-xIN… <named list>  
#>  2 name-squatter-mv     <NULL>      1fXUcrM9Cj2hHejtCByNM2z6_USh… <named list>  
#>  3 name-squatter-upload <NULL>      1yKcoa6LSDS--tW-ted_QtG3DkaN… <named list>  
#>  4 name-squatter-rename <NULL>      10Ez9J-9jBrG159bHUqPMnCJqEnY… <named list>  
#>  5 name-squatter-mv     <NULL>      1eOSCqNace-6DKvBgJMQLN8I1dDX… <named list>  
#>  6 name-squatter-upload <NULL>      1M_u3fm8EmeH73yTiWRXajujzQ0o… <named list>  
#>  7 name-squatter-rename <NULL>      17j3lXOQiV0XBe8nnwVFEBKay6Jf… <named list>  
#>  8 name-squatter-mv     <NULL>      1E2b0FiO-s5EEh90QDaZU868UaMY… <named list>  
#>  9 name-squatter-upload <NULL>      1K8FqSHHEcoQq0xueQnb8mmDskvw… <named list>  
#> 10 name-squatter-rename <NULL>      1u8eaTU9pnofrMjxlzrzXT3yrRky… <named list>  
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name                 version id                                drive_resource
#>    <chr>                <chr>   <drv_id>                          <list>        
#>  1 name-squatter-rename 2       1FlZtCAUMzkP6Yr61Rlu3qzT-xIN5MEkD <named list>  
#>  2 name-squatter-mv     2       1fXUcrM9Cj2hHejtCByNM2z6_UShZAiEz <named list>  
#>  3 name-squatter-upload 2       1yKcoa6LSDS--tW-ted_QtG3DkaNHTK4c <named list>  
#>  4 name-squatter-rename 2       10Ez9J-9jBrG159bHUqPMnCJqEnYFx_M1 <named list>  
#>  5 name-squatter-mv     2       1eOSCqNace-6DKvBgJMQLN8I1dDXoP9rc <named list>  
#>  6 name-squatter-upload 2       1M_u3fm8EmeH73yTiWRXajujzQ0o0nsdZ <named list>  
#>  7 name-squatter-rename 2       17j3lXOQiV0XBe8nnwVFEBKay6JfDg6Am <named list>  
#>  8 name-squatter-mv     2       1E2b0FiO-s5EEh90QDaZU868UaMYo1ZJd <named list>  
#>  9 name-squatter-upload 2       1K8FqSHHEcoQq0xueQnb8mmDskvwRAB39 <named list>  
#> 10 name-squatter-rename 2       1u8eaTU9pnofrMjxlzrzXT3yrRkymH2ZV <named list>  
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name                 web_view_link                       id    drive_resource
#>    <chr>                <chr>                               <drv> <list>        
#>  1 name-squatter-rename https://drive.google.com/file/d/1F… 1FlZ… <named list>  
#>  2 name-squatter-mv     https://drive.google.com/file/d/1f… 1fXU… <named list>  
#>  3 name-squatter-upload https://drive.google.com/file/d/1y… 1yKc… <named list>  
#>  4 name-squatter-rename https://drive.google.com/file/d/10… 10Ez… <named list>  
#>  5 name-squatter-mv     https://drive.google.com/file/d/1e… 1eOS… <named list>  
#>  6 name-squatter-upload https://drive.google.com/file/d/1M… 1M_u… <named list>  
#>  7 name-squatter-rename https://drive.google.com/file/d/17… 17j3… <named list>  
#>  8 name-squatter-mv     https://drive.google.com/file/d/1E… 1E2b… <named list>  
#>  9 name-squatter-upload https://drive.google.com/file/d/1K… 1K8F… <named list>  
#> 10 name-squatter-rename https://drive.google.com/file/d/1u… 1u8e… <named list>  
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name                 modified_time       id                    drive_resource
#>    <chr>                <dttm>              <drv_id>              <list>        
#>  1 name-squatter-rename 2022-02-28 13:43:14 1FlZtCAUMzkP6Yr61Rlu… <named list>  
#>  2 name-squatter-mv     2022-02-28 13:42:58 1fXUcrM9Cj2hHejtCByN… <named list>  
#>  3 name-squatter-upload 2022-02-27 13:42:20 1yKcoa6LSDS--tW-ted_… <named list>  
#>  4 name-squatter-rename 2022-02-27 13:41:48 10Ez9J-9jBrG159bHUqP… <named list>  
#>  5 name-squatter-mv     2022-02-27 13:41:33 1eOSCqNace-6DKvBgJMQ… <named list>  
#>  6 name-squatter-upload 2022-02-26 13:42:02 1M_u3fm8EmeH73yTiWRX… <named list>  
#>  7 name-squatter-rename 2022-02-26 13:41:26 17j3lXOQiV0XBe8nnwVF… <named list>  
#>  8 name-squatter-mv     2022-02-26 13:41:09 1E2b0FiO-s5EEh90QDaZ… <named list>  
#>  9 name-squatter-upload 2022-02-25 13:43:14 1K8FqSHHEcoQq0xueQnb… <named list>  
#> 10 name-squatter-rename 2022-02-25 13:42:37 1u8eaTU9pnofrMjxlzrz… <named list>  
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name                 created_time        id                    drive_resource
#>    <chr>                <dttm>              <drv_id>              <list>        
#>  1 name-squatter-rename 2022-02-28 13:43:14 1FlZtCAUMzkP6Yr61Rlu… <named list>  
#>  2 name-squatter-mv     2022-02-28 13:42:58 1fXUcrM9Cj2hHejtCByN… <named list>  
#>  3 name-squatter-upload 2022-02-27 13:42:20 1yKcoa6LSDS--tW-ted_… <named list>  
#>  4 name-squatter-rename 2022-02-27 13:41:48 10Ez9J-9jBrG159bHUqP… <named list>  
#>  5 name-squatter-mv     2022-02-27 13:41:33 1eOSCqNace-6DKvBgJMQ… <named list>  
#>  6 name-squatter-upload 2022-02-26 13:42:02 1M_u3fm8EmeH73yTiWRX… <named list>  
#>  7 name-squatter-rename 2022-02-26 13:41:26 17j3lXOQiV0XBe8nnwVF… <named list>  
#>  8 name-squatter-mv     2022-02-26 13:41:09 1E2b0FiO-s5EEh90QDaZ… <named list>  
#>  9 name-squatter-upload 2022-02-25 13:43:14 1K8FqSHHEcoQq0xueQnb… <named list>  
#> 10 name-squatter-rename 2022-02-25 13:42:37 1u8eaTU9pnofrMjxlzrz… <named list>  
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name                 owned_by_me id                            drive_resource
#>    <chr>                <lgl>       <drv_id>                      <list>        
#>  1 name-squatter-rename TRUE        1FlZtCAUMzkP6Yr61Rlu3qzT-xIN… <named list>  
#>  2 name-squatter-mv     TRUE        1fXUcrM9Cj2hHejtCByNM2z6_USh… <named list>  
#>  3 name-squatter-upload TRUE        1yKcoa6LSDS--tW-ted_QtG3DkaN… <named list>  
#>  4 name-squatter-rename TRUE        10Ez9J-9jBrG159bHUqPMnCJqEnY… <named list>  
#>  5 name-squatter-mv     TRUE        1eOSCqNace-6DKvBgJMQLN8I1dDX… <named list>  
#>  6 name-squatter-upload TRUE        1M_u3fm8EmeH73yTiWRXajujzQ0o… <named list>  
#>  7 name-squatter-rename TRUE        17j3lXOQiV0XBe8nnwVFEBKay6Jf… <named list>  
#>  8 name-squatter-mv     TRUE        1E2b0FiO-s5EEh90QDaZU868UaMY… <named list>  
#>  9 name-squatter-upload TRUE        1K8FqSHHEcoQq0xueQnb8mmDskvw… <named list>  
#> 10 name-squatter-rename TRUE        1u8eaTU9pnofrMjxlzrzXT3yrRky… <named list>  
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name                 size  id                                drive_resource
#>    <chr>                <chr> <drv_id>                          <list>        
#>  1 name-squatter-rename 0     1FlZtCAUMzkP6Yr61Rlu3qzT-xIN5MEkD <named list>  
#>  2 name-squatter-mv     0     1fXUcrM9Cj2hHejtCByNM2z6_UShZAiEz <named list>  
#>  3 name-squatter-upload 0     1yKcoa6LSDS--tW-ted_QtG3DkaNHTK4c <named list>  
#>  4 name-squatter-rename 0     10Ez9J-9jBrG159bHUqPMnCJqEnYFx_M1 <named list>  
#>  5 name-squatter-mv     0     1eOSCqNace-6DKvBgJMQLN8I1dDXoP9rc <named list>  
#>  6 name-squatter-upload 0     1M_u3fm8EmeH73yTiWRXajujzQ0o0nsdZ <named list>  
#>  7 name-squatter-rename 0     17j3lXOQiV0XBe8nnwVFEBKay6JfDg6Am <named list>  
#>  8 name-squatter-mv     0     1E2b0FiO-s5EEh90QDaZU868UaMYo1ZJd <named list>  
#>  9 name-squatter-upload 0     1K8FqSHHEcoQq0xueQnb8mmDskvwRAB39 <named list>  
#> 10 name-squatter-rename 0     1u8eaTU9pnofrMjxlzrzXT3yrRkymH2ZV <named list>  
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name                 quota_bytes_used id                       drive_resource
#>    <chr>                <chr>            <drv_id>                 <list>        
#>  1 name-squatter-rename 0                1FlZtCAUMzkP6Yr61Rlu3qz… <named list>  
#>  2 name-squatter-mv     0                1fXUcrM9Cj2hHejtCByNM2z… <named list>  
#>  3 name-squatter-upload 0                1yKcoa6LSDS--tW-ted_QtG… <named list>  
#>  4 name-squatter-rename 0                10Ez9J-9jBrG159bHUqPMnC… <named list>  
#>  5 name-squatter-mv     0                1eOSCqNace-6DKvBgJMQLN8… <named list>  
#>  6 name-squatter-upload 0                1M_u3fm8EmeH73yTiWRXaju… <named list>  
#>  7 name-squatter-rename 0                17j3lXOQiV0XBe8nnwVFEBK… <named list>  
#>  8 name-squatter-mv     0                1E2b0FiO-s5EEh90QDaZU86… <named list>  
#>  9 name-squatter-upload 0                1K8FqSHHEcoQq0xueQnb8mm… <named list>  
#> 10 name-squatter-rename 0                1u8eaTU9pnofrMjxlzrzXT3… <named list>  

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>