Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what

Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via whatappears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name                 path                   id                 drive_resource
#>    <chr>                <chr>                  <drv_id>           <list>        
#>  1 name-squatter-rename ~/name-squatter-rename 1aouSIeSgFjCWNGkf… <named list>  
#>  2 name-squatter-mv     ~/name-squatter-mv     1GqIESl9brnuZIVH3… <named list>  
#>  3 name-squatter-upload ~/name-squatter-upload 1Dct61kDLxwYg9XcI… <named list>  
#>  4 name-squatter-rename ~/name-squatter-rename 1ApTJOCRd6vG7avJ4… <named list>  
#>  5 name-squatter-mv     ~/name-squatter-mv     1mZ0BynI1flDnLpLD… <named list>  
#>  6 name-squatter-upload ~/name-squatter-upload 1l5FlasbAX0i39x6I… <named list>  
#>  7 name-squatter-rename ~/name-squatter-rename 1xc3fj6HuGwTt8nSs… <named list>  
#>  8 name-squatter-mv     ~/name-squatter-mv     10Mfk4GI_x3_vWgmC… <named list>  
#>  9 name-squatter-upload ~/name-squatter-upload 1XBeoFlrysmYq7JsB… <named list>  
#> 10 name-squatter-rename ~/name-squatter-rename 1cte5Q0aLmF1aPumO… <named list>  
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name                 shared id                drive_resource permissions_res…
#>    <chr>                <lgl>  <drv_id>          <list>         <list>          
#>  1 name-squatter-rename FALSE  1aouSIeSgFjCWNGk… <named list>   <named list [2]>
#>  2 name-squatter-mv     FALSE  1GqIESl9brnuZIVH… <named list>   <named list [2]>
#>  3 name-squatter-upload FALSE  1Dct61kDLxwYg9Xc… <named list>   <named list [2]>
#>  4 name-squatter-rename FALSE  1ApTJOCRd6vG7avJ… <named list>   <named list [2]>
#>  5 name-squatter-mv     FALSE  1mZ0BynI1flDnLpL… <named list>   <named list [2]>
#>  6 name-squatter-upload FALSE  1l5FlasbAX0i39x6… <named list>   <named list [2]>
#>  7 name-squatter-rename FALSE  1xc3fj6HuGwTt8nS… <named list>   <named list [2]>
#>  8 name-squatter-mv     FALSE  10Mfk4GI_x3_vWgm… <named list>   <named list [2]>
#>  9 name-squatter-upload FALSE  1XBeoFlrysmYq7Js… <named list>   <named list [2]>
#> 10 name-squatter-rename FALSE  1cte5Q0aLmF1aPum… <named list>   <named list [2]>
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name                 published id             drive_resource revision_resour…
#>    <chr>                <lgl>     <drv_id>       <list>         <list>          
#>  1 name-squatter-rename FALSE     1aouSIeSgFjCW… <named list>   <named list [9]>
#>  2 name-squatter-mv     FALSE     1GqIESl9brnuZ… <named list>   <named list [9]>
#>  3 name-squatter-upload FALSE     1Dct61kDLxwYg… <named list>   <named list [9]>
#>  4 name-squatter-rename FALSE     1ApTJOCRd6vG7… <named list>   <named list [9]>
#>  5 name-squatter-mv     FALSE     1mZ0BynI1flDn… <named list>   <named list [9]>
#>  6 name-squatter-upload FALSE     1l5FlasbAX0i3… <named list>   <named list [9]>
#>  7 name-squatter-rename FALSE     1xc3fj6HuGwTt… <named list>   <named list [9]>
#>  8 name-squatter-mv     FALSE     10Mfk4GI_x3_v… <named list>   <named list [9]>
#>  9 name-squatter-upload FALSE     1XBeoFlrysmYq… <named list>   <named list [9]>
#> 10 name-squatter-rename FALSE     1cte5Q0aLmF1a… <named list>   <named list [9]>

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name                 id_parent           id                    drive_resource
#>    <chr>                <drv_id>            <drv_id>              <list>        
#>  1 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1aouSIeSgFjCWNGkfe4r… <named list>  
#>  2 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1GqIESl9brnuZIVH3YtJ… <named list>  
#>  3 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1Dct61kDLxwYg9XcI7ma… <named list>  
#>  4 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1ApTJOCRd6vG7avJ4B1q… <named list>  
#>  5 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1mZ0BynI1flDnLpLDynZ… <named list>  
#>  6 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1l5FlasbAX0i39x6IhAT… <named list>  
#>  7 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1xc3fj6HuGwTt8nSsLeO… <named list>  
#>  8 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 10Mfk4GI_x3_vWgmCJyS… <named list>  
#>  9 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1XBeoFlrysmYq7JsBBKP… <named list>  
#> 10 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1cte5Q0aLmF1aPumOjvB… <named list>  

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name                 trashed id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename TRUE    1aouSIeSgFjCWNGkfe4rYXMzsO48uOtHG <named list>  
#>  2 name-squatter-mv     TRUE    1GqIESl9brnuZIVH3YtJmUEvyUJDSWsO- <named list>  
#>  3 name-squatter-upload TRUE    1Dct61kDLxwYg9XcI7maevPi6o-m114vY <named list>  
#>  4 name-squatter-rename TRUE    1ApTJOCRd6vG7avJ4B1qFqN1GHXRXU-UX <named list>  
#>  5 name-squatter-mv     TRUE    1mZ0BynI1flDnLpLDynZRvJRP1UXDw2kQ <named list>  
#>  6 name-squatter-upload TRUE    1l5FlasbAX0i39x6IhATPsrsPKQZ0DplB <named list>  
#>  7 name-squatter-rename TRUE    1xc3fj6HuGwTt8nSsLeOt45udtECoNm_A <named list>  
#>  8 name-squatter-mv     TRUE    10Mfk4GI_x3_vWgmCJySDeIDbNb4cPD5g <named list>  
#>  9 name-squatter-upload TRUE    1XBeoFlrysmYq7JsBBKPWZC9gC6Cy2KGH <named list>  
#> 10 name-squatter-rename TRUE    1cte5Q0aLmF1aPumOjvBFCLyJ3hR1Z1oZ <named list>  
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name                 mime_type                id               drive_resource
#>    <chr>                <chr>                    <drv_id>         <list>        
#>  1 name-squatter-rename application/octet-stream 1aouSIeSgFjCWNG… <named list>  
#>  2 name-squatter-mv     application/octet-stream 1GqIESl9brnuZIV… <named list>  
#>  3 name-squatter-upload application/octet-stream 1Dct61kDLxwYg9X… <named list>  
#>  4 name-squatter-rename application/octet-stream 1ApTJOCRd6vG7av… <named list>  
#>  5 name-squatter-mv     application/octet-stream 1mZ0BynI1flDnLp… <named list>  
#>  6 name-squatter-upload application/octet-stream 1l5FlasbAX0i39x… <named list>  
#>  7 name-squatter-rename application/octet-stream 1xc3fj6HuGwTt8n… <named list>  
#>  8 name-squatter-mv     application/octet-stream 10Mfk4GI_x3_vWg… <named list>  
#>  9 name-squatter-upload application/octet-stream 1XBeoFlrysmYq7J… <named list>  
#> 10 name-squatter-rename application/octet-stream 1cte5Q0aLmF1aPu… <named list>  
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name                 starred id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename FALSE   1aouSIeSgFjCWNGkfe4rYXMzsO48uOtHG <named list>  
#>  2 name-squatter-mv     FALSE   1GqIESl9brnuZIVH3YtJmUEvyUJDSWsO- <named list>  
#>  3 name-squatter-upload FALSE   1Dct61kDLxwYg9XcI7maevPi6o-m114vY <named list>  
#>  4 name-squatter-rename FALSE   1ApTJOCRd6vG7avJ4B1qFqN1GHXRXU-UX <named list>  
#>  5 name-squatter-mv     FALSE   1mZ0BynI1flDnLpLDynZRvJRP1UXDw2kQ <named list>  
#>  6 name-squatter-upload FALSE   1l5FlasbAX0i39x6IhATPsrsPKQZ0DplB <named list>  
#>  7 name-squatter-rename FALSE   1xc3fj6HuGwTt8nSsLeOt45udtECoNm_A <named list>  
#>  8 name-squatter-mv     FALSE   10Mfk4GI_x3_vWgmCJySDeIDbNb4cPD5g <named list>  
#>  9 name-squatter-upload FALSE   1XBeoFlrysmYq7JsBBKPWZC9gC6Cy2KGH <named list>  
#> 10 name-squatter-rename FALSE   1cte5Q0aLmF1aPumOjvBFCLyJ3hR1Z1oZ <named list>  
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name                 description id                            drive_resource
#>    <chr>                <list>      <drv_id>                      <list>        
#>  1 name-squatter-rename <NULL>      1aouSIeSgFjCWNGkfe4rYXMzsO48… <named list>  
#>  2 name-squatter-mv     <NULL>      1GqIESl9brnuZIVH3YtJmUEvyUJD… <named list>  
#>  3 name-squatter-upload <NULL>      1Dct61kDLxwYg9XcI7maevPi6o-m… <named list>  
#>  4 name-squatter-rename <NULL>      1ApTJOCRd6vG7avJ4B1qFqN1GHXR… <named list>  
#>  5 name-squatter-mv     <NULL>      1mZ0BynI1flDnLpLDynZRvJRP1UX… <named list>  
#>  6 name-squatter-upload <NULL>      1l5FlasbAX0i39x6IhATPsrsPKQZ… <named list>  
#>  7 name-squatter-rename <NULL>      1xc3fj6HuGwTt8nSsLeOt45udtEC… <named list>  
#>  8 name-squatter-mv     <NULL>      10Mfk4GI_x3_vWgmCJySDeIDbNb4… <named list>  
#>  9 name-squatter-upload <NULL>      1XBeoFlrysmYq7JsBBKPWZC9gC6C… <named list>  
#> 10 name-squatter-rename <NULL>      1cte5Q0aLmF1aPumOjvBFCLyJ3hR… <named list>  
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name                 version id                                drive_resource
#>    <chr>                <chr>   <drv_id>                          <list>        
#>  1 name-squatter-rename 2       1aouSIeSgFjCWNGkfe4rYXMzsO48uOtHG <named list>  
#>  2 name-squatter-mv     2       1GqIESl9brnuZIVH3YtJmUEvyUJDSWsO- <named list>  
#>  3 name-squatter-upload 2       1Dct61kDLxwYg9XcI7maevPi6o-m114vY <named list>  
#>  4 name-squatter-rename 2       1ApTJOCRd6vG7avJ4B1qFqN1GHXRXU-UX <named list>  
#>  5 name-squatter-mv     2       1mZ0BynI1flDnLpLDynZRvJRP1UXDw2kQ <named list>  
#>  6 name-squatter-upload 2       1l5FlasbAX0i39x6IhATPsrsPKQZ0DplB <named list>  
#>  7 name-squatter-rename 2       1xc3fj6HuGwTt8nSsLeOt45udtECoNm_A <named list>  
#>  8 name-squatter-mv     2       10Mfk4GI_x3_vWgmCJySDeIDbNb4cPD5g <named list>  
#>  9 name-squatter-upload 2       1XBeoFlrysmYq7JsBBKPWZC9gC6Cy2KGH <named list>  
#> 10 name-squatter-rename 2       1cte5Q0aLmF1aPumOjvBFCLyJ3hR1Z1oZ <named list>  
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name                 web_view_link                       id    drive_resource
#>    <chr>                <chr>                               <drv> <list>        
#>  1 name-squatter-rename https://drive.google.com/file/d/1a… 1aou… <named list>  
#>  2 name-squatter-mv     https://drive.google.com/file/d/1G… 1GqI… <named list>  
#>  3 name-squatter-upload https://drive.google.com/file/d/1D… 1Dct… <named list>  
#>  4 name-squatter-rename https://drive.google.com/file/d/1A… 1ApT… <named list>  
#>  5 name-squatter-mv     https://drive.google.com/file/d/1m… 1mZ0… <named list>  
#>  6 name-squatter-upload https://drive.google.com/file/d/1l… 1l5F… <named list>  
#>  7 name-squatter-rename https://drive.google.com/file/d/1x… 1xc3… <named list>  
#>  8 name-squatter-mv     https://drive.google.com/file/d/10… 10Mf… <named list>  
#>  9 name-squatter-upload https://drive.google.com/file/d/1X… 1XBe… <named list>  
#> 10 name-squatter-rename https://drive.google.com/file/d/1c… 1cte… <named list>  
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name                 modified_time       id                    drive_resource
#>    <chr>                <dttm>              <drv_id>              <list>        
#>  1 name-squatter-rename 2022-02-14 13:43:09 1aouSIeSgFjCWNGkfe4r… <named list>  
#>  2 name-squatter-mv     2022-02-14 13:42:54 1GqIESl9brnuZIVH3YtJ… <named list>  
#>  3 name-squatter-upload 2022-02-13 13:41:11 1Dct61kDLxwYg9XcI7ma… <named list>  
#>  4 name-squatter-rename 2022-02-13 13:40:40 1ApTJOCRd6vG7avJ4B1q… <named list>  
#>  5 name-squatter-mv     2022-02-13 13:40:24 1mZ0BynI1flDnLpLDynZ… <named list>  
#>  6 name-squatter-upload 2022-02-12 13:41:39 1l5FlasbAX0i39x6IhAT… <named list>  
#>  7 name-squatter-rename 2022-02-12 13:41:05 1xc3fj6HuGwTt8nSsLeO… <named list>  
#>  8 name-squatter-mv     2022-02-12 13:40:49 10Mfk4GI_x3_vWgmCJyS… <named list>  
#>  9 name-squatter-upload 2022-02-11 13:41:43 1XBeoFlrysmYq7JsBBKP… <named list>  
#> 10 name-squatter-rename 2022-02-11 13:41:06 1cte5Q0aLmF1aPumOjvB… <named list>  
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name                 created_time        id                    drive_resource
#>    <chr>                <dttm>              <drv_id>              <list>        
#>  1 name-squatter-rename 2022-02-14 13:43:09 1aouSIeSgFjCWNGkfe4r… <named list>  
#>  2 name-squatter-mv     2022-02-14 13:42:54 1GqIESl9brnuZIVH3YtJ… <named list>  
#>  3 name-squatter-upload 2022-02-13 13:41:11 1Dct61kDLxwYg9XcI7ma… <named list>  
#>  4 name-squatter-rename 2022-02-13 13:40:40 1ApTJOCRd6vG7avJ4B1q… <named list>  
#>  5 name-squatter-mv     2022-02-13 13:40:24 1mZ0BynI1flDnLpLDynZ… <named list>  
#>  6 name-squatter-upload 2022-02-12 13:41:39 1l5FlasbAX0i39x6IhAT… <named list>  
#>  7 name-squatter-rename 2022-02-12 13:41:05 1xc3fj6HuGwTt8nSsLeO… <named list>  
#>  8 name-squatter-mv     2022-02-12 13:40:49 10Mfk4GI_x3_vWgmCJyS… <named list>  
#>  9 name-squatter-upload 2022-02-11 13:41:43 1XBeoFlrysmYq7JsBBKP… <named list>  
#> 10 name-squatter-rename 2022-02-11 13:41:06 1cte5Q0aLmF1aPumOjvB… <named list>  
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name                 owned_by_me id                            drive_resource
#>    <chr>                <lgl>       <drv_id>                      <list>        
#>  1 name-squatter-rename TRUE        1aouSIeSgFjCWNGkfe4rYXMzsO48… <named list>  
#>  2 name-squatter-mv     TRUE        1GqIESl9brnuZIVH3YtJmUEvyUJD… <named list>  
#>  3 name-squatter-upload TRUE        1Dct61kDLxwYg9XcI7maevPi6o-m… <named list>  
#>  4 name-squatter-rename TRUE        1ApTJOCRd6vG7avJ4B1qFqN1GHXR… <named list>  
#>  5 name-squatter-mv     TRUE        1mZ0BynI1flDnLpLDynZRvJRP1UX… <named list>  
#>  6 name-squatter-upload TRUE        1l5FlasbAX0i39x6IhATPsrsPKQZ… <named list>  
#>  7 name-squatter-rename TRUE        1xc3fj6HuGwTt8nSsLeOt45udtEC… <named list>  
#>  8 name-squatter-mv     TRUE        10Mfk4GI_x3_vWgmCJySDeIDbNb4… <named list>  
#>  9 name-squatter-upload TRUE        1XBeoFlrysmYq7JsBBKPWZC9gC6C… <named list>  
#> 10 name-squatter-rename TRUE        1cte5Q0aLmF1aPumOjvBFCLyJ3hR… <named list>  
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name                 size  id                                drive_resource
#>    <chr>                <chr> <drv_id>                          <list>        
#>  1 name-squatter-rename 0     1aouSIeSgFjCWNGkfe4rYXMzsO48uOtHG <named list>  
#>  2 name-squatter-mv     0     1GqIESl9brnuZIVH3YtJmUEvyUJDSWsO- <named list>  
#>  3 name-squatter-upload 0     1Dct61kDLxwYg9XcI7maevPi6o-m114vY <named list>  
#>  4 name-squatter-rename 0     1ApTJOCRd6vG7avJ4B1qFqN1GHXRXU-UX <named list>  
#>  5 name-squatter-mv     0     1mZ0BynI1flDnLpLDynZRvJRP1UXDw2kQ <named list>  
#>  6 name-squatter-upload 0     1l5FlasbAX0i39x6IhATPsrsPKQZ0DplB <named list>  
#>  7 name-squatter-rename 0     1xc3fj6HuGwTt8nSsLeOt45udtECoNm_A <named list>  
#>  8 name-squatter-mv     0     10Mfk4GI_x3_vWgmCJySDeIDbNb4cPD5g <named list>  
#>  9 name-squatter-upload 0     1XBeoFlrysmYq7JsBBKPWZC9gC6Cy2KGH <named list>  
#> 10 name-squatter-rename 0     1cte5Q0aLmF1aPumOjvBFCLyJ3hR1Z1oZ <named list>  
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name                 quota_bytes_used id                       drive_resource
#>    <chr>                <chr>            <drv_id>                 <list>        
#>  1 name-squatter-rename 0                1aouSIeSgFjCWNGkfe4rYXM… <named list>  
#>  2 name-squatter-mv     0                1GqIESl9brnuZIVH3YtJmUE… <named list>  
#>  3 name-squatter-upload 0                1Dct61kDLxwYg9XcI7maevP… <named list>  
#>  4 name-squatter-rename 0                1ApTJOCRd6vG7avJ4B1qFqN… <named list>  
#>  5 name-squatter-mv     0                1mZ0BynI1flDnLpLDynZRvJ… <named list>  
#>  6 name-squatter-upload 0                1l5FlasbAX0i39x6IhATPsr… <named list>  
#>  7 name-squatter-rename 0                1xc3fj6HuGwTt8nSsLeOt45… <named list>  
#>  8 name-squatter-mv     0                10Mfk4GI_x3_vWgmCJySDeI… <named list>  
#>  9 name-squatter-upload 0                1XBeoFlrysmYq7JsBBKPWZC… <named list>  
#> 10 name-squatter-rename 0                1cte5Q0aLmF1aPumOjvBFCL… <named list>  

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>