Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file
Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what
Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via what appears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 x 4
#>    name            path                  id                      drive_resource 
#>  * <chr>           <chr>                 <chr>                   <list>         
#>  1 name-squatter   ~/name-squatter       1nYRdy9YJ-b129egrXPCiy… <named list [3…
#>  2 name-squatter   ~/name-squatter       1CeZIjUb_WFC5CL_7jG0E0… <named list [3…
#>  3 chicken-cp.txt  ~/chicken-cp.txt      1ardlyQOXHY8Rku86rVUcq… <named list [4…
#>  4 name-squatter   ~/name-squatter       1vuKodf7AhBA97ed1Ia4Ad… <named list [3…
#>  5 name-squatter   ~/name-squatter       10wYsaVA4ArhJHjNvmTgFs… <named list [3…
#>  6 Rlogo.pdf       ~/Rlogo.pdf           1opPZyo4WUiue56qlF8SO5… <named list [4…
#>  7 THANKS          ~/THANKS              1JCQ06Wj6AjntiyjnHsXmm… <named list [4…
#>  8 googledrive-NE… ~/abc/def/googledriv… 1wMnFwFSIG0eQ4UUUEbn6-… <named list [3…
#>  9 def             ~/abc/def/            12xjsE3eIf84VSFF5hNhSE… <named list [3…
#> 10 abc             ~/abc/                1qkS667DEnvDzTuVBSyj5N… <named list [3…
drive_reveal(files, "permissions")
#> # A dribble: 10 x 5
#>    name          shared id                    drive_resource  permissions_resou…
#>  * <chr>         <lgl>  <chr>                 <list>          <list>            
#>  1 name-squatter FALSE  1nYRdy9YJ-b129egrXPC… <named list [3… <named list [2]>  
#>  2 name-squatter FALSE  1CeZIjUb_WFC5CL_7jG0… <named list [3… <named list [2]>  
#>  3 chicken-cp.t… FALSE  1ardlyQOXHY8Rku86rVU… <named list [4… <named list [2]>  
#>  4 name-squatter FALSE  1vuKodf7AhBA97ed1Ia4… <named list [3… <named list [2]>  
#>  5 name-squatter FALSE  10wYsaVA4ArhJHjNvmTg… <named list [3… <named list [2]>  
#>  6 Rlogo.pdf     TRUE   1opPZyo4WUiue56qlF8S… <named list [4… <named list [2]>  
#>  7 THANKS        TRUE   1JCQ06Wj6AjntiyjnHsX… <named list [4… <named list [2]>  
#>  8 googledrive-… FALSE  1wMnFwFSIG0eQ4UUUEbn… <named list [3… <named list [2]>  
#>  9 def           FALSE  12xjsE3eIf84VSFF5hNh… <named list [3… <named list [2]>  
#> 10 abc           FALSE  1qkS667DEnvDzTuVBSyj… <named list [3… <named list [2]>  
drive_reveal(files, "published")
#> # A dribble: 10 x 5
#>    name          published id                   drive_resource  revision_resour…
#>  * <chr>         <lgl>     <chr>                <list>          <list>          
#>  1 name-squatter FALSE     1nYRdy9YJ-b129egrXP… <named list [3… <named list [9]>
#>  2 name-squatter FALSE     1CeZIjUb_WFC5CL_7jG… <named list [3… <named list [9]>
#>  3 chicken-cp.t… FALSE     1ardlyQOXHY8Rku86rV… <named list [4… <named list [10…
#>  4 name-squatter FALSE     1vuKodf7AhBA97ed1Ia… <named list [3… <named list [9]>
#>  5 name-squatter FALSE     10wYsaVA4ArhJHjNvmT… <named list [3… <named list [9]>
#>  6 Rlogo.pdf     FALSE     1opPZyo4WUiue56qlF8… <named list [4… <named list [10…
#>  7 THANKS        FALSE     1JCQ06Wj6AjntiyjnHs… <named list [4… <named list [10…
#>  8 googledrive-… FALSE     1wMnFwFSIG0eQ4UUUEb… <named list [3… <named list [10…
#>  9 def           FALSE     12xjsE3eIf84VSFF5hN… <named list [3… <NULL>          
#> 10 abc           FALSE     1qkS667DEnvDzTuVBSy… <named list [3… <NULL>          

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 x 4
#>    name           id_parent                id                    drive_resource 
#>    <chr>          <chr>                    <chr>                 <list>         
#>  1 name-squatter  0AO_RMaBzcP63Uk9PVA      1nYRdy9YJ-b129egrXPC… <named list [3…
#>  2 name-squatter  0AO_RMaBzcP63Uk9PVA      1CeZIjUb_WFC5CL_7jG0… <named list [3…
#>  3 chicken-cp.txt 0AO_RMaBzcP63Uk9PVA      1ardlyQOXHY8Rku86rVU… <named list [4…
#>  4 name-squatter  0AO_RMaBzcP63Uk9PVA      1vuKodf7AhBA97ed1Ia4… <named list [3…
#>  5 name-squatter  0AO_RMaBzcP63Uk9PVA      10wYsaVA4ArhJHjNvmTg… <named list [3…
#>  6 Rlogo.pdf      0AO_RMaBzcP63Uk9PVA      1opPZyo4WUiue56qlF8S… <named list [4…
#>  7 THANKS         0AO_RMaBzcP63Uk9PVA      1JCQ06Wj6AjntiyjnHsX… <named list [4…
#>  8 googledrive-N… 12xjsE3eIf84VSFF5hNhSEO… 1wMnFwFSIG0eQ4UUUEbn… <named list [3…
#>  9 def            1qkS667DEnvDzTuVBSyj5Ne… 12xjsE3eIf84VSFF5hNh… <named list [3…
#> 10 abc            0AO_RMaBzcP63Uk9PVA      1qkS667DEnvDzTuVBSyj… <named list [3…

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 x 4
#>    name                trashed id                               drive_resource  
#>  * <chr>               <lgl>   <chr>                            <list>          
#>  1 name-squatter       TRUE    1nYRdy9YJ-b129egrXPCiyDFbtRz2Z-… <named list [39…
#>  2 name-squatter       TRUE    1CeZIjUb_WFC5CL_7jG0E0nPSrqy1Ps… <named list [39…
#>  3 chicken-cp.txt      TRUE    1ardlyQOXHY8Rku86rVUcqb4pgfy2TG… <named list [40…
#>  4 name-squatter       TRUE    1vuKodf7AhBA97ed1Ia4AdfQJ5jdOzl… <named list [39…
#>  5 name-squatter       TRUE    10wYsaVA4ArhJHjNvmTgFsWrWKbU4UE… <named list [39…
#>  6 Rlogo.pdf           FALSE   1opPZyo4WUiue56qlF8SO5ipaTJwPLz… <named list [40…
#>  7 THANKS              FALSE   1JCQ06Wj6AjntiyjnHsXmmcHZWCkBNg… <named list [40…
#>  8 googledrive-NEWS.md FALSE   1wMnFwFSIG0eQ4UUUEbn6-udF3GoKY3… <named list [39…
#>  9 def                 FALSE   12xjsE3eIf84VSFF5hNhSEOnQF8DDSH… <named list [33…
#> 10 abc                 FALSE   1qkS667DEnvDzTuVBSyj5NeGzIGbrAy… <named list [33…
drive_reveal(files, "mime_type")
#> # A dribble: 10 x 4
#>    name           mime_type                id                    drive_resource 
#>  * <chr>          <chr>                    <chr>                 <list>         
#>  1 name-squatter  application/octet-stream 1nYRdy9YJ-b129egrXPC… <named list [3…
#>  2 name-squatter  application/octet-stream 1CeZIjUb_WFC5CL_7jG0… <named list [3…
#>  3 chicken-cp.txt text/plain               1ardlyQOXHY8Rku86rVU… <named list [4…
#>  4 name-squatter  application/octet-stream 1vuKodf7AhBA97ed1Ia4… <named list [3…
#>  5 name-squatter  application/octet-stream 10wYsaVA4ArhJHjNvmTg… <named list [3…
#>  6 Rlogo.pdf      application/pdf          1opPZyo4WUiue56qlF8S… <named list [4…
#>  7 THANKS         text/plain               1JCQ06Wj6AjntiyjnHsX… <named list [4…
#>  8 googledrive-N… text/markdown            1wMnFwFSIG0eQ4UUUEbn… <named list [3…
#>  9 def            application/vnd.google-… 12xjsE3eIf84VSFF5hNh… <named list [3…
#> 10 abc            application/vnd.google-… 1qkS667DEnvDzTuVBSyj… <named list [3…
drive_reveal(files, "starred")
#> # A dribble: 10 x 4
#>    name                starred id                               drive_resource  
#>  * <chr>               <lgl>   <chr>                            <list>          
#>  1 name-squatter       FALSE   1nYRdy9YJ-b129egrXPCiyDFbtRz2Z-… <named list [39…
#>  2 name-squatter       FALSE   1CeZIjUb_WFC5CL_7jG0E0nPSrqy1Ps… <named list [39…
#>  3 chicken-cp.txt      FALSE   1ardlyQOXHY8Rku86rVUcqb4pgfy2TG… <named list [40…
#>  4 name-squatter       FALSE   1vuKodf7AhBA97ed1Ia4AdfQJ5jdOzl… <named list [39…
#>  5 name-squatter       FALSE   10wYsaVA4ArhJHjNvmTgFsWrWKbU4UE… <named list [39…
#>  6 Rlogo.pdf           TRUE    1opPZyo4WUiue56qlF8SO5ipaTJwPLz… <named list [40…
#>  7 THANKS              TRUE    1JCQ06Wj6AjntiyjnHsXmmcHZWCkBNg… <named list [40…
#>  8 googledrive-NEWS.md FALSE   1wMnFwFSIG0eQ4UUUEbn6-udF3GoKY3… <named list [39…
#>  9 def                 FALSE   12xjsE3eIf84VSFF5hNhSEOnQF8DDSH… <named list [33…
#> 10 abc                 FALSE   1qkS667DEnvDzTuVBSyj5NeGzIGbrAy… <named list [33…
drive_reveal(files, "description")
#> # A dribble: 10 x 4
#>    name               description id                            drive_resource  
#>  * <chr>              <list>      <chr>                         <list>          
#>  1 name-squatter      <NULL>      1nYRdy9YJ-b129egrXPCiyDFbtRz… <named list [39…
#>  2 name-squatter      <NULL>      1CeZIjUb_WFC5CL_7jG0E0nPSrqy… <named list [39…
#>  3 chicken-cp.txt     <NULL>      1ardlyQOXHY8Rku86rVUcqb4pgfy… <named list [40…
#>  4 name-squatter      <NULL>      1vuKodf7AhBA97ed1Ia4AdfQJ5jd… <named list [39…
#>  5 name-squatter      <NULL>      10wYsaVA4ArhJHjNvmTgFsWrWKbU… <named list [39…
#>  6 Rlogo.pdf          <NULL>      1opPZyo4WUiue56qlF8SO5ipaTJw… <named list [40…
#>  7 THANKS             <NULL>      1JCQ06Wj6AjntiyjnHsXmmcHZWCk… <named list [40…
#>  8 googledrive-NEWS.… <NULL>      1wMnFwFSIG0eQ4UUUEbn6-udF3Go… <named list [39…
#>  9 def                <NULL>      12xjsE3eIf84VSFF5hNhSEOnQF8D… <named list [33…
#> 10 abc                <NULL>      1qkS667DEnvDzTuVBSyj5NeGzIGb… <named list [33…
drive_reveal(files, "version")
#> # A dribble: 10 x 4
#>    name                version id                               drive_resource  
#>  * <chr>               <chr>   <chr>                            <list>          
#>  1 name-squatter       2       1nYRdy9YJ-b129egrXPCiyDFbtRz2Z-… <named list [39…
#>  2 name-squatter       2       1CeZIjUb_WFC5CL_7jG0E0nPSrqy1Ps… <named list [39…
#>  3 chicken-cp.txt      3       1ardlyQOXHY8Rku86rVUcqb4pgfy2TG… <named list [40…
#>  4 name-squatter       2       1vuKodf7AhBA97ed1Ia4AdfQJ5jdOzl… <named list [39…
#>  5 name-squatter       2       10wYsaVA4ArhJHjNvmTgFsWrWKbU4UE… <named list [39…
#>  6 Rlogo.pdf           4       1opPZyo4WUiue56qlF8SO5ipaTJwPLz… <named list [40…
#>  7 THANKS              4       1JCQ06Wj6AjntiyjnHsXmmcHZWCkBNg… <named list [40…
#>  8 googledrive-NEWS.md 1       1wMnFwFSIG0eQ4UUUEbn6-udF3GoKY3… <named list [39…
#>  9 def                 1       12xjsE3eIf84VSFF5hNhSEOnQF8DDSH… <named list [33…
#> 10 abc                 1       1qkS667DEnvDzTuVBSyj5NeGzIGbrAy… <named list [33…
drive_reveal(files, "web_view_link")
#> # A dribble: 10 x 4
#>    name       web_view_link                       id             drive_resource 
#>  * <chr>      <chr>                               <chr>          <list>         
#>  1 name-squa… https://drive.google.com/file/d/1n… 1nYRdy9YJ-b12… <named list [3…
#>  2 name-squa… https://drive.google.com/file/d/1C… 1CeZIjUb_WFC5… <named list [3…
#>  3 chicken-c… https://drive.google.com/file/d/1a… 1ardlyQOXHY8R… <named list [4…
#>  4 name-squa… https://drive.google.com/file/d/1v… 1vuKodf7AhBA9… <named list [3…
#>  5 name-squa… https://drive.google.com/file/d/10… 10wYsaVA4ArhJ… <named list [3…
#>  6 Rlogo.pdf  https://drive.google.com/file/d/1o… 1opPZyo4WUiue… <named list [4…
#>  7 THANKS     https://drive.google.com/file/d/1J… 1JCQ06Wj6Ajnt… <named list [4…
#>  8 googledri… https://drive.google.com/file/d/1w… 1wMnFwFSIG0eQ… <named list [3…
#>  9 def        https://drive.google.com/drive/fol… 12xjsE3eIf84V… <named list [3…
#> 10 abc        https://drive.google.com/drive/fol… 1qkS667DEnvDz… <named list [3…
drive_reveal(files, "modified_time")
#> # A dribble: 10 x 4
#>    name            modified_time       id                       drive_resource  
#>    <chr>           <dttm>              <chr>                    <list>          
#>  1 name-squatter   2021-06-30 13:45:14 1nYRdy9YJ-b129egrXPCiyD… <named list [39…
#>  2 name-squatter   2021-06-30 13:44:55 1CeZIjUb_WFC5CL_7jG0E0n… <named list [39…
#>  3 chicken-cp.txt  2021-06-30 13:44:07 1ardlyQOXHY8Rku86rVUcqb… <named list [40…
#>  4 name-squatter   2021-06-30 05:35:35 1vuKodf7AhBA97ed1Ia4Adf… <named list [39…
#>  5 name-squatter   2021-06-30 05:29:49 10wYsaVA4ArhJHjNvmTgFsW… <named list [39…
#>  6 Rlogo.pdf       2021-04-28 16:17:31 1opPZyo4WUiue56qlF8SO5i… <named list [40…
#>  7 THANKS          2021-04-28 16:17:29 1JCQ06Wj6AjntiyjnHsXmmc… <named list [40…
#>  8 googledrive-NE… 2021-04-28 16:17:14 1wMnFwFSIG0eQ4UUUEbn6-u… <named list [39…
#>  9 def             2021-04-28 16:17:02 12xjsE3eIf84VSFF5hNhSEO… <named list [33…
#> 10 abc             2021-04-28 16:16:57 1qkS667DEnvDzTuVBSyj5Ne… <named list [33…
drive_reveal(files, "created_time")
#> # A dribble: 10 x 4
#>    name            created_time        id                       drive_resource  
#>    <chr>           <dttm>              <chr>                    <list>          
#>  1 name-squatter   2021-06-30 13:45:14 1nYRdy9YJ-b129egrXPCiyD… <named list [39…
#>  2 name-squatter   2021-06-30 13:44:55 1CeZIjUb_WFC5CL_7jG0E0n… <named list [39…
#>  3 chicken-cp.txt  2021-06-30 13:44:07 1ardlyQOXHY8Rku86rVUcqb… <named list [40…
#>  4 name-squatter   2021-06-30 05:35:35 1vuKodf7AhBA97ed1Ia4Adf… <named list [39…
#>  5 name-squatter   2021-06-30 05:29:49 10wYsaVA4ArhJHjNvmTgFsW… <named list [39…
#>  6 Rlogo.pdf       2021-04-28 16:16:49 1opPZyo4WUiue56qlF8SO5i… <named list [40…
#>  7 THANKS          2021-04-28 16:16:53 1JCQ06Wj6AjntiyjnHsXmmc… <named list [40…
#>  8 googledrive-NE… 2021-04-28 16:17:14 1wMnFwFSIG0eQ4UUUEbn6-u… <named list [39…
#>  9 def             2021-04-28 16:17:02 12xjsE3eIf84VSFF5hNhSEO… <named list [33…
#> 10 abc             2021-04-28 16:16:57 1qkS667DEnvDzTuVBSyj5Ne… <named list [33…
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 x 4
#>    name               owned_by_me id                            drive_resource  
#>  * <chr>              <lgl>       <chr>                         <list>          
#>  1 name-squatter      TRUE        1nYRdy9YJ-b129egrXPCiyDFbtRz… <named list [39…
#>  2 name-squatter      TRUE        1CeZIjUb_WFC5CL_7jG0E0nPSrqy… <named list [39…
#>  3 chicken-cp.txt     TRUE        1ardlyQOXHY8Rku86rVUcqb4pgfy… <named list [40…
#>  4 name-squatter      TRUE        1vuKodf7AhBA97ed1Ia4AdfQJ5jd… <named list [39…
#>  5 name-squatter      TRUE        10wYsaVA4ArhJHjNvmTgFsWrWKbU… <named list [39…
#>  6 Rlogo.pdf          TRUE        1opPZyo4WUiue56qlF8SO5ipaTJw… <named list [40…
#>  7 THANKS             TRUE        1JCQ06Wj6AjntiyjnHsXmmcHZWCk… <named list [40…
#>  8 googledrive-NEWS.… TRUE        1wMnFwFSIG0eQ4UUUEbn6-udF3Go… <named list [39…
#>  9 def                TRUE        12xjsE3eIf84VSFF5hNhSEOnQF8D… <named list [33…
#> 10 abc                TRUE        1qkS667DEnvDzTuVBSyj5NeGzIGb… <named list [33…
drive_reveal(files, "size")
#> # A dribble: 10 x 4
#>    name                size  id                                drive_resource   
#>  * <chr>               <chr> <chr>                             <list>           
#>  1 name-squatter       0     1nYRdy9YJ-b129egrXPCiyDFbtRz2Z--4 <named list [39]>
#>  2 name-squatter       0     1CeZIjUb_WFC5CL_7jG0E0nPSrqy1PsHv <named list [39]>
#>  3 chicken-cp.txt      165   1ardlyQOXHY8Rku86rVUcqb4pgfy2TGwH <named list [40]>
#>  4 name-squatter       0     1vuKodf7AhBA97ed1Ia4AdfQJ5jdOzlHW <named list [39]>
#>  5 name-squatter       0     10wYsaVA4ArhJHjNvmTgFsWrWKbU4UE_p <named list [39]>
#>  6 Rlogo.pdf           6755  1opPZyo4WUiue56qlF8SO5ipaTJwPLzZg <named list [40]>
#>  7 THANKS              3776  1JCQ06Wj6AjntiyjnHsXmmcHZWCkBNgBJ <named list [40]>
#>  8 googledrive-NEWS.md 12728 1wMnFwFSIG0eQ4UUUEbn6-udF3GoKY3xN <named list [39]>
#>  9 def                 NA    12xjsE3eIf84VSFF5hNhSEOnQF8DDSHkc <named list [33]>
#> 10 abc                 NA    1qkS667DEnvDzTuVBSyj5NeGzIGbrAy9Q <named list [33]>
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 x 4
#>    name             quota_bytes_used id                         drive_resource  
#>  * <chr>            <chr>            <chr>                      <list>          
#>  1 name-squatter    0                1nYRdy9YJ-b129egrXPCiyDFb… <named list [39…
#>  2 name-squatter    0                1CeZIjUb_WFC5CL_7jG0E0nPS… <named list [39…
#>  3 chicken-cp.txt   165              1ardlyQOXHY8Rku86rVUcqb4p… <named list [40…
#>  4 name-squatter    0                1vuKodf7AhBA97ed1Ia4AdfQJ… <named list [39…
#>  5 name-squatter    0                10wYsaVA4ArhJHjNvmTgFsWrW… <named list [39…
#>  6 Rlogo.pdf        6755             1opPZyo4WUiue56qlF8SO5ipa… <named list [40…
#>  7 THANKS           3776             1JCQ06Wj6AjntiyjnHsXmmcHZ… <named list [40…
#>  8 googledrive-NEW… 12728            1wMnFwFSIG0eQ4UUUEbn6-udF… <named list [39…
#>  9 def              0                12xjsE3eIf84VSFF5hNhSEOnQ… <named list [33…
#> 10 abc              0                1qkS667DEnvDzTuVBSyj5NeGz… <named list [33…

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 x 4
#>   name     path  id                  drive_resource   
#> * <chr>    <chr> <chr>               <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>