Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what

Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via whatappears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name                 path                   id                drive_resource 
#>    <chr>                <chr>                  <drv_id>          <list>         
#>  1 name-squatter-rename ~/name-squatter-rename 1bwM9TnZECaxLbrG… <named list [3…
#>  2 name-squatter-mv     ~/name-squatter-mv     14wnpbp7ukusIN_j… <named list [3…
#>  3 name-squatter-upload ~/name-squatter-upload 1nk3haWe6kk-UdxK… <named list [3…
#>  4 name-squatter-rename ~/name-squatter-rename 1qibxE4EI5yJDohE… <named list [3…
#>  5 name-squatter-mv     ~/name-squatter-mv     1Boc8WQVNe76BI0D… <named list [3…
#>  6 name-squatter-upload ~/name-squatter-upload 1D6sQD1tAGhrgGeX… <named list [3…
#>  7 name-squatter-rename ~/name-squatter-rename 1mKilMWRTKAlbPzY… <named list [3…
#>  8 name-squatter-mv     ~/name-squatter-mv     1hnclHtn1NOWFNga… <named list [3…
#>  9 name-squatter-upload ~/name-squatter-upload 1QoSbLqqDHnW6h3Q… <named list [3…
#> 10 name-squatter-rename ~/name-squatter-rename 16TUtxbnj3QQFQht… <named list [3…
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name                 shared id              drive_resource  permissions_reso…
#>    <chr>                <lgl>  <drv_id>        <list>          <list>           
#>  1 name-squatter-rename FALSE  1bwM9TnZECaxLb… <named list [3… <named list [2]> 
#>  2 name-squatter-mv     FALSE  14wnpbp7ukusIN… <named list [3… <named list [2]> 
#>  3 name-squatter-upload FALSE  1nk3haWe6kk-Ud… <named list [3… <named list [2]> 
#>  4 name-squatter-rename FALSE  1qibxE4EI5yJDo… <named list [3… <named list [2]> 
#>  5 name-squatter-mv     FALSE  1Boc8WQVNe76BI… <named list [3… <named list [2]> 
#>  6 name-squatter-upload FALSE  1D6sQD1tAGhrgG… <named list [3… <named list [2]> 
#>  7 name-squatter-rename FALSE  1mKilMWRTKAlbP… <named list [3… <named list [2]> 
#>  8 name-squatter-mv     FALSE  1hnclHtn1NOWFN… <named list [3… <named list [2]> 
#>  9 name-squatter-upload FALSE  1QoSbLqqDHnW6h… <named list [3… <named list [2]> 
#> 10 name-squatter-rename FALSE  16TUtxbnj3QQFQ… <named list [3… <named list [2]> 
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name                 published id            drive_resource  revision_resour…
#>    <chr>                <lgl>     <drv_id>      <list>          <list>          
#>  1 name-squatter-rename FALSE     1bwM9TnZECax… <named list [3… <named list [9]>
#>  2 name-squatter-mv     FALSE     14wnpbp7ukus… <named list [3… <named list [9]>
#>  3 name-squatter-upload FALSE     1nk3haWe6kk-… <named list [3… <named list [9]>
#>  4 name-squatter-rename FALSE     1qibxE4EI5yJ… <named list [3… <named list [9]>
#>  5 name-squatter-mv     FALSE     1Boc8WQVNe76… <named list [3… <named list [9]>
#>  6 name-squatter-upload FALSE     1D6sQD1tAGhr… <named list [3… <named list [9]>
#>  7 name-squatter-rename FALSE     1mKilMWRTKAl… <named list [3… <named list [9]>
#>  8 name-squatter-mv     FALSE     1hnclHtn1NOW… <named list [3… <named list [9]>
#>  9 name-squatter-upload FALSE     1QoSbLqqDHnW… <named list [3… <named list [9]>
#> 10 name-squatter-rename FALSE     16TUtxbnj3QQ… <named list [3… <named list [9]>

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name                 id_parent           id                   drive_resource 
#>    <chr>                <drv_id>            <drv_id>             <list>         
#>  1 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1bwM9TnZECaxLbrGYD0… <named list [3…
#>  2 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 14wnpbp7ukusIN_jkkC… <named list [3…
#>  3 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1nk3haWe6kk-UdxKlP4… <named list [3…
#>  4 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1qibxE4EI5yJDohEPkA… <named list [3…
#>  5 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1Boc8WQVNe76BI0DcAn… <named list [3…
#>  6 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1D6sQD1tAGhrgGeXAZW… <named list [3…
#>  7 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1mKilMWRTKAlbPzYrIZ… <named list [3…
#>  8 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1hnclHtn1NOWFNgakvZ… <named list [3…
#>  9 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1QoSbLqqDHnW6h3QRdg… <named list [3…
#> 10 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 16TUtxbnj3QQFQht7ow… <named list [3…

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name                 trashed id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename TRUE    1bwM9TnZECaxLbrGYD0uOXvoA6P1Tyulj <named list […
#>  2 name-squatter-mv     TRUE    14wnpbp7ukusIN_jkkC72lb_d6JS2cfqP <named list […
#>  3 name-squatter-upload TRUE    1nk3haWe6kk-UdxKlP47xy_pwXqoKbBB8 <named list […
#>  4 name-squatter-rename TRUE    1qibxE4EI5yJDohEPkAbf3pscpdfXmJpG <named list […
#>  5 name-squatter-mv     TRUE    1Boc8WQVNe76BI0DcAn972BD1CQ7w-LZM <named list […
#>  6 name-squatter-upload TRUE    1D6sQD1tAGhrgGeXAZWg0h7En3824GONL <named list […
#>  7 name-squatter-rename TRUE    1mKilMWRTKAlbPzYrIZCyVG-3_pklNQPB <named list […
#>  8 name-squatter-mv     TRUE    1hnclHtn1NOWFNgakvZVG-8Js2G1KHsOs <named list […
#>  9 name-squatter-upload TRUE    1QoSbLqqDHnW6h3QRdgeXgzozugvAn0OM <named list […
#> 10 name-squatter-rename TRUE    16TUtxbnj3QQFQht7owdouGurK-v-Q_vZ <named list […
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name                 mime_type                id              drive_resource 
#>    <chr>                <chr>                    <drv_id>        <list>         
#>  1 name-squatter-rename application/octet-stream 1bwM9TnZECaxLb… <named list [3…
#>  2 name-squatter-mv     application/octet-stream 14wnpbp7ukusIN… <named list [3…
#>  3 name-squatter-upload application/octet-stream 1nk3haWe6kk-Ud… <named list [3…
#>  4 name-squatter-rename application/octet-stream 1qibxE4EI5yJDo… <named list [3…
#>  5 name-squatter-mv     application/octet-stream 1Boc8WQVNe76BI… <named list [3…
#>  6 name-squatter-upload application/octet-stream 1D6sQD1tAGhrgG… <named list [3…
#>  7 name-squatter-rename application/octet-stream 1mKilMWRTKAlbP… <named list [3…
#>  8 name-squatter-mv     application/octet-stream 1hnclHtn1NOWFN… <named list [3…
#>  9 name-squatter-upload application/octet-stream 1QoSbLqqDHnW6h… <named list [3…
#> 10 name-squatter-rename application/octet-stream 16TUtxbnj3QQFQ… <named list [3…
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name                 starred id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename FALSE   1bwM9TnZECaxLbrGYD0uOXvoA6P1Tyulj <named list […
#>  2 name-squatter-mv     FALSE   14wnpbp7ukusIN_jkkC72lb_d6JS2cfqP <named list […
#>  3 name-squatter-upload FALSE   1nk3haWe6kk-UdxKlP47xy_pwXqoKbBB8 <named list […
#>  4 name-squatter-rename FALSE   1qibxE4EI5yJDohEPkAbf3pscpdfXmJpG <named list […
#>  5 name-squatter-mv     FALSE   1Boc8WQVNe76BI0DcAn972BD1CQ7w-LZM <named list […
#>  6 name-squatter-upload FALSE   1D6sQD1tAGhrgGeXAZWg0h7En3824GONL <named list […
#>  7 name-squatter-rename FALSE   1mKilMWRTKAlbPzYrIZCyVG-3_pklNQPB <named list […
#>  8 name-squatter-mv     FALSE   1hnclHtn1NOWFNgakvZVG-8Js2G1KHsOs <named list […
#>  9 name-squatter-upload FALSE   1QoSbLqqDHnW6h3QRdgeXgzozugvAn0OM <named list […
#> 10 name-squatter-rename FALSE   16TUtxbnj3QQFQht7owdouGurK-v-Q_vZ <named list […
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name                 description id                          drive_resource  
#>    <chr>                <list>      <drv_id>                    <list>          
#>  1 name-squatter-rename <NULL>      1bwM9TnZECaxLbrGYD0uOXvoA6… <named list [39…
#>  2 name-squatter-mv     <NULL>      14wnpbp7ukusIN_jkkC72lb_d6… <named list [39…
#>  3 name-squatter-upload <NULL>      1nk3haWe6kk-UdxKlP47xy_pwX… <named list [39…
#>  4 name-squatter-rename <NULL>      1qibxE4EI5yJDohEPkAbf3pscp… <named list [39…
#>  5 name-squatter-mv     <NULL>      1Boc8WQVNe76BI0DcAn972BD1C… <named list [39…
#>  6 name-squatter-upload <NULL>      1D6sQD1tAGhrgGeXAZWg0h7En3… <named list [39…
#>  7 name-squatter-rename <NULL>      1mKilMWRTKAlbPzYrIZCyVG-3_… <named list [39…
#>  8 name-squatter-mv     <NULL>      1hnclHtn1NOWFNgakvZVG-8Js2… <named list [39…
#>  9 name-squatter-upload <NULL>      1QoSbLqqDHnW6h3QRdgeXgzozu… <named list [39…
#> 10 name-squatter-rename <NULL>      16TUtxbnj3QQFQht7owdouGurK… <named list [39…
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name                 version id                                drive_resource
#>    <chr>                <chr>   <drv_id>                          <list>        
#>  1 name-squatter-rename 2       1bwM9TnZECaxLbrGYD0uOXvoA6P1Tyulj <named list […
#>  2 name-squatter-mv     2       14wnpbp7ukusIN_jkkC72lb_d6JS2cfqP <named list […
#>  3 name-squatter-upload 2       1nk3haWe6kk-UdxKlP47xy_pwXqoKbBB8 <named list […
#>  4 name-squatter-rename 2       1qibxE4EI5yJDohEPkAbf3pscpdfXmJpG <named list […
#>  5 name-squatter-mv     2       1Boc8WQVNe76BI0DcAn972BD1CQ7w-LZM <named list […
#>  6 name-squatter-upload 2       1D6sQD1tAGhrgGeXAZWg0h7En3824GONL <named list […
#>  7 name-squatter-rename 2       1mKilMWRTKAlbPzYrIZCyVG-3_pklNQPB <named list […
#>  8 name-squatter-mv     2       1hnclHtn1NOWFNgakvZVG-8Js2G1KHsOs <named list […
#>  9 name-squatter-upload 2       1QoSbLqqDHnW6h3QRdgeXgzozugvAn0OM <named list […
#> 10 name-squatter-rename 2       16TUtxbnj3QQFQht7owdouGurK-v-Q_vZ <named list […
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name                 web_view_link                id           drive_resource
#>    <chr>                <chr>                        <drv_id>     <list>        
#>  1 name-squatter-rename https://drive.google.com/fi… 1bwM9TnZECa… <named list […
#>  2 name-squatter-mv     https://drive.google.com/fi… 14wnpbp7uku… <named list […
#>  3 name-squatter-upload https://drive.google.com/fi… 1nk3haWe6kk… <named list […
#>  4 name-squatter-rename https://drive.google.com/fi… 1qibxE4EI5y… <named list […
#>  5 name-squatter-mv     https://drive.google.com/fi… 1Boc8WQVNe7… <named list […
#>  6 name-squatter-upload https://drive.google.com/fi… 1D6sQD1tAGh… <named list […
#>  7 name-squatter-rename https://drive.google.com/fi… 1mKilMWRTKA… <named list […
#>  8 name-squatter-mv     https://drive.google.com/fi… 1hnclHtn1NO… <named list […
#>  9 name-squatter-upload https://drive.google.com/fi… 1QoSbLqqDHn… <named list […
#> 10 name-squatter-rename https://drive.google.com/fi… 16TUtxbnj3Q… <named list […
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name                 modified_time       id                   drive_resource 
#>    <chr>                <dttm>              <drv_id>             <list>         
#>  1 name-squatter-rename 2022-01-15 13:46:40 1bwM9TnZECaxLbrGYD0… <named list [3…
#>  2 name-squatter-mv     2022-01-15 13:46:25 14wnpbp7ukusIN_jkkC… <named list [3…
#>  3 name-squatter-upload 2022-01-14 13:49:08 1nk3haWe6kk-UdxKlP4… <named list [3…
#>  4 name-squatter-rename 2022-01-14 13:48:36 1qibxE4EI5yJDohEPkA… <named list [3…
#>  5 name-squatter-mv     2022-01-14 13:48:21 1Boc8WQVNe76BI0DcAn… <named list [3…
#>  6 name-squatter-upload 2022-01-13 13:48:55 1D6sQD1tAGhrgGeXAZW… <named list [3…
#>  7 name-squatter-rename 2022-01-13 13:48:21 1mKilMWRTKAlbPzYrIZ… <named list [3…
#>  8 name-squatter-mv     2022-01-13 13:48:04 1hnclHtn1NOWFNgakvZ… <named list [3…
#>  9 name-squatter-upload 2022-01-12 13:49:58 1QoSbLqqDHnW6h3QRdg… <named list [3…
#> 10 name-squatter-rename 2022-01-12 13:49:21 16TUtxbnj3QQFQht7ow… <named list [3…
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name                 created_time        id                   drive_resource 
#>    <chr>                <dttm>              <drv_id>             <list>         
#>  1 name-squatter-rename 2022-01-15 13:46:40 1bwM9TnZECaxLbrGYD0… <named list [3…
#>  2 name-squatter-mv     2022-01-15 13:46:25 14wnpbp7ukusIN_jkkC… <named list [3…
#>  3 name-squatter-upload 2022-01-14 13:49:08 1nk3haWe6kk-UdxKlP4… <named list [3…
#>  4 name-squatter-rename 2022-01-14 13:48:36 1qibxE4EI5yJDohEPkA… <named list [3…
#>  5 name-squatter-mv     2022-01-14 13:48:21 1Boc8WQVNe76BI0DcAn… <named list [3…
#>  6 name-squatter-upload 2022-01-13 13:48:55 1D6sQD1tAGhrgGeXAZW… <named list [3…
#>  7 name-squatter-rename 2022-01-13 13:48:21 1mKilMWRTKAlbPzYrIZ… <named list [3…
#>  8 name-squatter-mv     2022-01-13 13:48:04 1hnclHtn1NOWFNgakvZ… <named list [3…
#>  9 name-squatter-upload 2022-01-12 13:49:58 1QoSbLqqDHnW6h3QRdg… <named list [3…
#> 10 name-squatter-rename 2022-01-12 13:49:21 16TUtxbnj3QQFQht7ow… <named list [3…
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name                 owned_by_me id                          drive_resource  
#>    <chr>                <lgl>       <drv_id>                    <list>          
#>  1 name-squatter-rename TRUE        1bwM9TnZECaxLbrGYD0uOXvoA6… <named list [39…
#>  2 name-squatter-mv     TRUE        14wnpbp7ukusIN_jkkC72lb_d6… <named list [39…
#>  3 name-squatter-upload TRUE        1nk3haWe6kk-UdxKlP47xy_pwX… <named list [39…
#>  4 name-squatter-rename TRUE        1qibxE4EI5yJDohEPkAbf3pscp… <named list [39…
#>  5 name-squatter-mv     TRUE        1Boc8WQVNe76BI0DcAn972BD1C… <named list [39…
#>  6 name-squatter-upload TRUE        1D6sQD1tAGhrgGeXAZWg0h7En3… <named list [39…
#>  7 name-squatter-rename TRUE        1mKilMWRTKAlbPzYrIZCyVG-3_… <named list [39…
#>  8 name-squatter-mv     TRUE        1hnclHtn1NOWFNgakvZVG-8Js2… <named list [39…
#>  9 name-squatter-upload TRUE        1QoSbLqqDHnW6h3QRdgeXgzozu… <named list [39…
#> 10 name-squatter-rename TRUE        16TUtxbnj3QQFQht7owdouGurK… <named list [39…
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name                 size  id                                drive_resource  
#>    <chr>                <chr> <drv_id>                          <list>          
#>  1 name-squatter-rename 0     1bwM9TnZECaxLbrGYD0uOXvoA6P1Tyulj <named list [39…
#>  2 name-squatter-mv     0     14wnpbp7ukusIN_jkkC72lb_d6JS2cfqP <named list [39…
#>  3 name-squatter-upload 0     1nk3haWe6kk-UdxKlP47xy_pwXqoKbBB8 <named list [39…
#>  4 name-squatter-rename 0     1qibxE4EI5yJDohEPkAbf3pscpdfXmJpG <named list [39…
#>  5 name-squatter-mv     0     1Boc8WQVNe76BI0DcAn972BD1CQ7w-LZM <named list [39…
#>  6 name-squatter-upload 0     1D6sQD1tAGhrgGeXAZWg0h7En3824GONL <named list [39…
#>  7 name-squatter-rename 0     1mKilMWRTKAlbPzYrIZCyVG-3_pklNQPB <named list [39…
#>  8 name-squatter-mv     0     1hnclHtn1NOWFNgakvZVG-8Js2G1KHsOs <named list [39…
#>  9 name-squatter-upload 0     1QoSbLqqDHnW6h3QRdgeXgzozugvAn0OM <named list [39…
#> 10 name-squatter-rename 0     16TUtxbnj3QQFQht7owdouGurK-v-Q_vZ <named list [39…
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name                 quota_bytes_used id                      drive_resource 
#>    <chr>                <chr>            <drv_id>                <list>         
#>  1 name-squatter-rename 0                1bwM9TnZECaxLbrGYD0uOX… <named list [3…
#>  2 name-squatter-mv     0                14wnpbp7ukusIN_jkkC72l… <named list [3…
#>  3 name-squatter-upload 0                1nk3haWe6kk-UdxKlP47xy… <named list [3…
#>  4 name-squatter-rename 0                1qibxE4EI5yJDohEPkAbf3… <named list [3…
#>  5 name-squatter-mv     0                1Boc8WQVNe76BI0DcAn972… <named list [3…
#>  6 name-squatter-upload 0                1D6sQD1tAGhrgGeXAZWg0h… <named list [3…
#>  7 name-squatter-rename 0                1mKilMWRTKAlbPzYrIZCyV… <named list [3…
#>  8 name-squatter-mv     0                1hnclHtn1NOWFNgakvZVG-… <named list [3…
#>  9 name-squatter-upload 0                1QoSbLqqDHnW6h3QRdgeXg… <named list [3…
#> 10 name-squatter-rename 0                16TUtxbnj3QQFQht7owdou… <named list [3…

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>