Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what

Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via whatappears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name                 path                   id                drive_resource 
#>    <chr>                <chr>                  <drv_id>          <list>         
#>  1 name-squatter-rename ~/name-squatter-rename 1sGK5Mp0YRBpncn0… <named list [3…
#>  2 name-squatter-mv     ~/name-squatter-mv     1p8hXhHAfTQNmUgd… <named list [3…
#>  3 name-squatter-upload ~/name-squatter-upload 1D6KTaJEMb_PeuLW… <named list [3…
#>  4 name-squatter-rename ~/name-squatter-rename 1N9IBPAzZJ4THNNF… <named list [3…
#>  5 name-squatter-mv     ~/name-squatter-mv     1Wyf-rEOPG7OcSQR… <named list [3…
#>  6 name-squatter-upload ~/name-squatter-upload 1nb78WMgFWQab9Q8… <named list [3…
#>  7 name-squatter-rename ~/name-squatter-rename 1sGHpOgnhWBQ1Esr… <named list [3…
#>  8 name-squatter-mv     ~/name-squatter-mv     16RaPJC6N2SEk24i… <named list [3…
#>  9 name-squatter-upload ~/name-squatter-upload 1nrOCSmFGl0Jb3wE… <named list [3…
#> 10 name-squatter-rename ~/name-squatter-rename 1svdMpQLgqPK7x2Q… <named list [3…
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name                 shared id              drive_resource  permissions_reso…
#>    <chr>                <lgl>  <drv_id>        <list>          <list>           
#>  1 name-squatter-rename FALSE  1sGK5Mp0YRBpnc… <named list [3… <named list [2]> 
#>  2 name-squatter-mv     FALSE  1p8hXhHAfTQNmU… <named list [3… <named list [2]> 
#>  3 name-squatter-upload FALSE  1D6KTaJEMb_Peu… <named list [3… <named list [2]> 
#>  4 name-squatter-rename FALSE  1N9IBPAzZJ4THN… <named list [3… <named list [2]> 
#>  5 name-squatter-mv     FALSE  1Wyf-rEOPG7OcS… <named list [3… <named list [2]> 
#>  6 name-squatter-upload FALSE  1nb78WMgFWQab9… <named list [3… <named list [2]> 
#>  7 name-squatter-rename FALSE  1sGHpOgnhWBQ1E… <named list [3… <named list [2]> 
#>  8 name-squatter-mv     FALSE  16RaPJC6N2SEk2… <named list [3… <named list [2]> 
#>  9 name-squatter-upload FALSE  1nrOCSmFGl0Jb3… <named list [3… <named list [2]> 
#> 10 name-squatter-rename FALSE  1svdMpQLgqPK7x… <named list [3… <named list [2]> 
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name                 published id            drive_resource  revision_resour…
#>    <chr>                <lgl>     <drv_id>      <list>          <list>          
#>  1 name-squatter-rename FALSE     1sGK5Mp0YRBp… <named list [3… <named list [9]>
#>  2 name-squatter-mv     FALSE     1p8hXhHAfTQN… <named list [3… <named list [9]>
#>  3 name-squatter-upload FALSE     1D6KTaJEMb_P… <named list [3… <named list [9]>
#>  4 name-squatter-rename FALSE     1N9IBPAzZJ4T… <named list [3… <named list [9]>
#>  5 name-squatter-mv     FALSE     1Wyf-rEOPG7O… <named list [3… <named list [9]>
#>  6 name-squatter-upload FALSE     1nb78WMgFWQa… <named list [3… <named list [9]>
#>  7 name-squatter-rename FALSE     1sGHpOgnhWBQ… <named list [3… <named list [9]>
#>  8 name-squatter-mv     FALSE     16RaPJC6N2SE… <named list [3… <named list [9]>
#>  9 name-squatter-upload FALSE     1nrOCSmFGl0J… <named list [3… <named list [9]>
#> 10 name-squatter-rename FALSE     1svdMpQLgqPK… <named list [3… <named list [9]>

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name                 id_parent           id                   drive_resource 
#>    <chr>                <drv_id>            <drv_id>             <list>         
#>  1 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1sGK5Mp0YRBpncn0YUt… <named list [3…
#>  2 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1p8hXhHAfTQNmUgdJAL… <named list [3…
#>  3 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1D6KTaJEMb_PeuLW4wL… <named list [3…
#>  4 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1N9IBPAzZJ4THNNFeYY… <named list [3…
#>  5 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1Wyf-rEOPG7OcSQRqDt… <named list [3…
#>  6 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1nb78WMgFWQab9Q8jYd… <named list [3…
#>  7 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1sGHpOgnhWBQ1Esrk1K… <named list [3…
#>  8 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 16RaPJC6N2SEk24ivW_… <named list [3…
#>  9 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1nrOCSmFGl0Jb3wEgTO… <named list [3…
#> 10 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1svdMpQLgqPK7x2Q_3-… <named list [3…

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name                 trashed id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename TRUE    1sGK5Mp0YRBpncn0YUt_Ml7mRyw_SLYaU <named list […
#>  2 name-squatter-mv     TRUE    1p8hXhHAfTQNmUgdJALB8jOGXBBLg1whM <named list […
#>  3 name-squatter-upload TRUE    1D6KTaJEMb_PeuLW4wLYLOwstmDHXAtEq <named list […
#>  4 name-squatter-rename TRUE    1N9IBPAzZJ4THNNFeYYGQduEeEFhzVgxW <named list […
#>  5 name-squatter-mv     TRUE    1Wyf-rEOPG7OcSQRqDtF__KwiXm3vmCiy <named list […
#>  6 name-squatter-upload TRUE    1nb78WMgFWQab9Q8jYdA927lWHFgLOveI <named list […
#>  7 name-squatter-rename TRUE    1sGHpOgnhWBQ1Esrk1KWNNtADsfSfUbHU <named list […
#>  8 name-squatter-mv     TRUE    16RaPJC6N2SEk24ivW_iWe6w5Z-2Z4_H1 <named list […
#>  9 name-squatter-upload TRUE    1nrOCSmFGl0Jb3wEgTONj22vTkQSxXoDU <named list […
#> 10 name-squatter-rename TRUE    1svdMpQLgqPK7x2Q_3-4XOSIn3CiZuc2w <named list […
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name                 mime_type                id              drive_resource 
#>    <chr>                <chr>                    <drv_id>        <list>         
#>  1 name-squatter-rename application/octet-stream 1sGK5Mp0YRBpnc… <named list [3…
#>  2 name-squatter-mv     application/octet-stream 1p8hXhHAfTQNmU… <named list [3…
#>  3 name-squatter-upload application/octet-stream 1D6KTaJEMb_Peu… <named list [3…
#>  4 name-squatter-rename application/octet-stream 1N9IBPAzZJ4THN… <named list [3…
#>  5 name-squatter-mv     application/octet-stream 1Wyf-rEOPG7OcS… <named list [3…
#>  6 name-squatter-upload application/octet-stream 1nb78WMgFWQab9… <named list [3…
#>  7 name-squatter-rename application/octet-stream 1sGHpOgnhWBQ1E… <named list [3…
#>  8 name-squatter-mv     application/octet-stream 16RaPJC6N2SEk2… <named list [3…
#>  9 name-squatter-upload application/octet-stream 1nrOCSmFGl0Jb3… <named list [3…
#> 10 name-squatter-rename application/octet-stream 1svdMpQLgqPK7x… <named list [3…
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name                 starred id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename FALSE   1sGK5Mp0YRBpncn0YUt_Ml7mRyw_SLYaU <named list […
#>  2 name-squatter-mv     FALSE   1p8hXhHAfTQNmUgdJALB8jOGXBBLg1whM <named list […
#>  3 name-squatter-upload FALSE   1D6KTaJEMb_PeuLW4wLYLOwstmDHXAtEq <named list […
#>  4 name-squatter-rename FALSE   1N9IBPAzZJ4THNNFeYYGQduEeEFhzVgxW <named list […
#>  5 name-squatter-mv     FALSE   1Wyf-rEOPG7OcSQRqDtF__KwiXm3vmCiy <named list […
#>  6 name-squatter-upload FALSE   1nb78WMgFWQab9Q8jYdA927lWHFgLOveI <named list […
#>  7 name-squatter-rename FALSE   1sGHpOgnhWBQ1Esrk1KWNNtADsfSfUbHU <named list […
#>  8 name-squatter-mv     FALSE   16RaPJC6N2SEk24ivW_iWe6w5Z-2Z4_H1 <named list […
#>  9 name-squatter-upload FALSE   1nrOCSmFGl0Jb3wEgTONj22vTkQSxXoDU <named list […
#> 10 name-squatter-rename FALSE   1svdMpQLgqPK7x2Q_3-4XOSIn3CiZuc2w <named list […
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name                 description id                          drive_resource  
#>    <chr>                <list>      <drv_id>                    <list>          
#>  1 name-squatter-rename <NULL>      1sGK5Mp0YRBpncn0YUt_Ml7mRy… <named list [39…
#>  2 name-squatter-mv     <NULL>      1p8hXhHAfTQNmUgdJALB8jOGXB… <named list [39…
#>  3 name-squatter-upload <NULL>      1D6KTaJEMb_PeuLW4wLYLOwstm… <named list [39…
#>  4 name-squatter-rename <NULL>      1N9IBPAzZJ4THNNFeYYGQduEeE… <named list [39…
#>  5 name-squatter-mv     <NULL>      1Wyf-rEOPG7OcSQRqDtF__KwiX… <named list [39…
#>  6 name-squatter-upload <NULL>      1nb78WMgFWQab9Q8jYdA927lWH… <named list [39…
#>  7 name-squatter-rename <NULL>      1sGHpOgnhWBQ1Esrk1KWNNtADs… <named list [39…
#>  8 name-squatter-mv     <NULL>      16RaPJC6N2SEk24ivW_iWe6w5Z… <named list [39…
#>  9 name-squatter-upload <NULL>      1nrOCSmFGl0Jb3wEgTONj22vTk… <named list [39…
#> 10 name-squatter-rename <NULL>      1svdMpQLgqPK7x2Q_3-4XOSIn3… <named list [39…
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name                 version id                                drive_resource
#>    <chr>                <chr>   <drv_id>                          <list>        
#>  1 name-squatter-rename 2       1sGK5Mp0YRBpncn0YUt_Ml7mRyw_SLYaU <named list […
#>  2 name-squatter-mv     2       1p8hXhHAfTQNmUgdJALB8jOGXBBLg1whM <named list […
#>  3 name-squatter-upload 2       1D6KTaJEMb_PeuLW4wLYLOwstmDHXAtEq <named list […
#>  4 name-squatter-rename 2       1N9IBPAzZJ4THNNFeYYGQduEeEFhzVgxW <named list […
#>  5 name-squatter-mv     2       1Wyf-rEOPG7OcSQRqDtF__KwiXm3vmCiy <named list […
#>  6 name-squatter-upload 2       1nb78WMgFWQab9Q8jYdA927lWHFgLOveI <named list […
#>  7 name-squatter-rename 2       1sGHpOgnhWBQ1Esrk1KWNNtADsfSfUbHU <named list […
#>  8 name-squatter-mv     2       16RaPJC6N2SEk24ivW_iWe6w5Z-2Z4_H1 <named list […
#>  9 name-squatter-upload 2       1nrOCSmFGl0Jb3wEgTONj22vTkQSxXoDU <named list […
#> 10 name-squatter-rename 2       1svdMpQLgqPK7x2Q_3-4XOSIn3CiZuc2w <named list […
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name                 web_view_link                id           drive_resource
#>    <chr>                <chr>                        <drv_id>     <list>        
#>  1 name-squatter-rename https://drive.google.com/fi… 1sGK5Mp0YRB… <named list […
#>  2 name-squatter-mv     https://drive.google.com/fi… 1p8hXhHAfTQ… <named list […
#>  3 name-squatter-upload https://drive.google.com/fi… 1D6KTaJEMb_… <named list […
#>  4 name-squatter-rename https://drive.google.com/fi… 1N9IBPAzZJ4… <named list […
#>  5 name-squatter-mv     https://drive.google.com/fi… 1Wyf-rEOPG7… <named list […
#>  6 name-squatter-upload https://drive.google.com/fi… 1nb78WMgFWQ… <named list […
#>  7 name-squatter-rename https://drive.google.com/fi… 1sGHpOgnhWB… <named list […
#>  8 name-squatter-mv     https://drive.google.com/fi… 16RaPJC6N2S… <named list […
#>  9 name-squatter-upload https://drive.google.com/fi… 1nrOCSmFGl0… <named list […
#> 10 name-squatter-rename https://drive.google.com/fi… 1svdMpQLgqP… <named list […
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name                 modified_time       id                   drive_resource 
#>    <chr>                <dttm>              <drv_id>             <list>         
#>  1 name-squatter-rename 2021-12-25 13:45:40 1sGK5Mp0YRBpncn0YUt… <named list [3…
#>  2 name-squatter-mv     2021-12-25 13:45:23 1p8hXhHAfTQNmUgdJAL… <named list [3…
#>  3 name-squatter-upload 2021-12-24 13:46:50 1D6KTaJEMb_PeuLW4wL… <named list [3…
#>  4 name-squatter-rename 2021-12-24 13:46:18 1N9IBPAzZJ4THNNFeYY… <named list [3…
#>  5 name-squatter-mv     2021-12-24 13:46:03 1Wyf-rEOPG7OcSQRqDt… <named list [3…
#>  6 name-squatter-upload 2021-12-23 13:46:55 1nb78WMgFWQab9Q8jYd… <named list [3…
#>  7 name-squatter-rename 2021-12-23 13:46:23 1sGHpOgnhWBQ1Esrk1K… <named list [3…
#>  8 name-squatter-mv     2021-12-23 13:46:09 16RaPJC6N2SEk24ivW_… <named list [3…
#>  9 name-squatter-upload 2021-12-21 13:47:29 1nrOCSmFGl0Jb3wEgTO… <named list [3…
#> 10 name-squatter-rename 2021-12-21 13:46:58 1svdMpQLgqPK7x2Q_3-… <named list [3…
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name                 created_time        id                   drive_resource 
#>    <chr>                <dttm>              <drv_id>             <list>         
#>  1 name-squatter-rename 2021-12-25 13:45:40 1sGK5Mp0YRBpncn0YUt… <named list [3…
#>  2 name-squatter-mv     2021-12-25 13:45:23 1p8hXhHAfTQNmUgdJAL… <named list [3…
#>  3 name-squatter-upload 2021-12-24 13:46:50 1D6KTaJEMb_PeuLW4wL… <named list [3…
#>  4 name-squatter-rename 2021-12-24 13:46:18 1N9IBPAzZJ4THNNFeYY… <named list [3…
#>  5 name-squatter-mv     2021-12-24 13:46:03 1Wyf-rEOPG7OcSQRqDt… <named list [3…
#>  6 name-squatter-upload 2021-12-23 13:46:55 1nb78WMgFWQab9Q8jYd… <named list [3…
#>  7 name-squatter-rename 2021-12-23 13:46:23 1sGHpOgnhWBQ1Esrk1K… <named list [3…
#>  8 name-squatter-mv     2021-12-23 13:46:09 16RaPJC6N2SEk24ivW_… <named list [3…
#>  9 name-squatter-upload 2021-12-21 13:47:29 1nrOCSmFGl0Jb3wEgTO… <named list [3…
#> 10 name-squatter-rename 2021-12-21 13:46:58 1svdMpQLgqPK7x2Q_3-… <named list [3…
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name                 owned_by_me id                          drive_resource  
#>    <chr>                <lgl>       <drv_id>                    <list>          
#>  1 name-squatter-rename TRUE        1sGK5Mp0YRBpncn0YUt_Ml7mRy… <named list [39…
#>  2 name-squatter-mv     TRUE        1p8hXhHAfTQNmUgdJALB8jOGXB… <named list [39…
#>  3 name-squatter-upload TRUE        1D6KTaJEMb_PeuLW4wLYLOwstm… <named list [39…
#>  4 name-squatter-rename TRUE        1N9IBPAzZJ4THNNFeYYGQduEeE… <named list [39…
#>  5 name-squatter-mv     TRUE        1Wyf-rEOPG7OcSQRqDtF__KwiX… <named list [39…
#>  6 name-squatter-upload TRUE        1nb78WMgFWQab9Q8jYdA927lWH… <named list [39…
#>  7 name-squatter-rename TRUE        1sGHpOgnhWBQ1Esrk1KWNNtADs… <named list [39…
#>  8 name-squatter-mv     TRUE        16RaPJC6N2SEk24ivW_iWe6w5Z… <named list [39…
#>  9 name-squatter-upload TRUE        1nrOCSmFGl0Jb3wEgTONj22vTk… <named list [39…
#> 10 name-squatter-rename TRUE        1svdMpQLgqPK7x2Q_3-4XOSIn3… <named list [39…
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name                 size  id                                drive_resource  
#>    <chr>                <chr> <drv_id>                          <list>          
#>  1 name-squatter-rename 0     1sGK5Mp0YRBpncn0YUt_Ml7mRyw_SLYaU <named list [39…
#>  2 name-squatter-mv     0     1p8hXhHAfTQNmUgdJALB8jOGXBBLg1whM <named list [39…
#>  3 name-squatter-upload 0     1D6KTaJEMb_PeuLW4wLYLOwstmDHXAtEq <named list [39…
#>  4 name-squatter-rename 0     1N9IBPAzZJ4THNNFeYYGQduEeEFhzVgxW <named list [39…
#>  5 name-squatter-mv     0     1Wyf-rEOPG7OcSQRqDtF__KwiXm3vmCiy <named list [39…
#>  6 name-squatter-upload 0     1nb78WMgFWQab9Q8jYdA927lWHFgLOveI <named list [39…
#>  7 name-squatter-rename 0     1sGHpOgnhWBQ1Esrk1KWNNtADsfSfUbHU <named list [39…
#>  8 name-squatter-mv     0     16RaPJC6N2SEk24ivW_iWe6w5Z-2Z4_H1 <named list [39…
#>  9 name-squatter-upload 0     1nrOCSmFGl0Jb3wEgTONj22vTkQSxXoDU <named list [39…
#> 10 name-squatter-rename 0     1svdMpQLgqPK7x2Q_3-4XOSIn3CiZuc2w <named list [39…
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name                 quota_bytes_used id                      drive_resource 
#>    <chr>                <chr>            <drv_id>                <list>         
#>  1 name-squatter-rename 0                1sGK5Mp0YRBpncn0YUt_Ml… <named list [3…
#>  2 name-squatter-mv     0                1p8hXhHAfTQNmUgdJALB8j… <named list [3…
#>  3 name-squatter-upload 0                1D6KTaJEMb_PeuLW4wLYLO… <named list [3…
#>  4 name-squatter-rename 0                1N9IBPAzZJ4THNNFeYYGQd… <named list [3…
#>  5 name-squatter-mv     0                1Wyf-rEOPG7OcSQRqDtF__… <named list [3…
#>  6 name-squatter-upload 0                1nb78WMgFWQab9Q8jYdA92… <named list [3…
#>  7 name-squatter-rename 0                1sGHpOgnhWBQ1Esrk1KWNN… <named list [3…
#>  8 name-squatter-mv     0                16RaPJC6N2SEk24ivW_iWe… <named list [3…
#>  9 name-squatter-upload 0                1nrOCSmFGl0Jb3wEgTONj2… <named list [3…
#> 10 name-squatter-rename 0                1svdMpQLgqPK7x2Q_3-4XO… <named list [3…

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>