Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what

Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via whatappears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name                 path                   id                 drive_resource
#>    <chr>                <chr>                  <drv_id>           <list>        
#>  1 name-squatter-rename ~/name-squatter-rename 18lS9WkXVt32akbgy… <named list>  
#>  2 name-squatter-mv     ~/name-squatter-mv     1mQDwdIafCalWPDFl… <named list>  
#>  3 name-squatter-upload ~/name-squatter-upload 1iLzlXosyC3YEAe4Z… <named list>  
#>  4 name-squatter-rename ~/name-squatter-rename 1GmLsADT05v-EnFqB… <named list>  
#>  5 name-squatter-mv     ~/name-squatter-mv     1Q51YVITRWPEBtjur… <named list>  
#>  6 name-squatter-upload ~/name-squatter-upload 1e03fT5Pwjtybws6K… <named list>  
#>  7 name-squatter-rename ~/name-squatter-rename 1yyxLosuliwRZxwRb… <named list>  
#>  8 name-squatter-mv     ~/name-squatter-mv     1s7I8K2EyhibP1Ttz… <named list>  
#>  9 name-squatter-upload ~/name-squatter-upload 13x7aGDDwPTECjdXZ… <named list>  
#> 10 name-squatter-rename ~/name-squatter-rename 1oy1nEjavHKW7LInQ… <named list>  
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name                 shared id                drive_resource permissions_res…
#>    <chr>                <lgl>  <drv_id>          <list>         <list>          
#>  1 name-squatter-rename FALSE  18lS9WkXVt32akbg… <named list>   <named list [2]>
#>  2 name-squatter-mv     FALSE  1mQDwdIafCalWPDF… <named list>   <named list [2]>
#>  3 name-squatter-upload FALSE  1iLzlXosyC3YEAe4… <named list>   <named list [2]>
#>  4 name-squatter-rename FALSE  1GmLsADT05v-EnFq… <named list>   <named list [2]>
#>  5 name-squatter-mv     FALSE  1Q51YVITRWPEBtju… <named list>   <named list [2]>
#>  6 name-squatter-upload FALSE  1e03fT5Pwjtybws6… <named list>   <named list [2]>
#>  7 name-squatter-rename FALSE  1yyxLosuliwRZxwR… <named list>   <named list [2]>
#>  8 name-squatter-mv     FALSE  1s7I8K2EyhibP1Tt… <named list>   <named list [2]>
#>  9 name-squatter-upload FALSE  13x7aGDDwPTECjdX… <named list>   <named list [2]>
#> 10 name-squatter-rename FALSE  1oy1nEjavHKW7LIn… <named list>   <named list [2]>
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name                 published id             drive_resource revision_resour…
#>    <chr>                <lgl>     <drv_id>       <list>         <list>          
#>  1 name-squatter-rename FALSE     18lS9WkXVt32a… <named list>   <named list [9]>
#>  2 name-squatter-mv     FALSE     1mQDwdIafCalW… <named list>   <named list [9]>
#>  3 name-squatter-upload FALSE     1iLzlXosyC3YE… <named list>   <named list [9]>
#>  4 name-squatter-rename FALSE     1GmLsADT05v-E… <named list>   <named list [9]>
#>  5 name-squatter-mv     FALSE     1Q51YVITRWPEB… <named list>   <named list [9]>
#>  6 name-squatter-upload FALSE     1e03fT5Pwjtyb… <named list>   <named list [9]>
#>  7 name-squatter-rename FALSE     1yyxLosuliwRZ… <named list>   <named list [9]>
#>  8 name-squatter-mv     FALSE     1s7I8K2EyhibP… <named list>   <named list [9]>
#>  9 name-squatter-upload FALSE     13x7aGDDwPTEC… <named list>   <named list [9]>
#> 10 name-squatter-rename FALSE     1oy1nEjavHKW7… <named list>   <named list [9]>

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name                 id_parent           id                    drive_resource
#>    <chr>                <drv_id>            <drv_id>              <list>        
#>  1 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 18lS9WkXVt32akbgypBa… <named list>  
#>  2 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1mQDwdIafCalWPDFli7d… <named list>  
#>  3 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1iLzlXosyC3YEAe4Z3P-… <named list>  
#>  4 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1GmLsADT05v-EnFqBxt-… <named list>  
#>  5 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1Q51YVITRWPEBtjurWf2… <named list>  
#>  6 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1e03fT5Pwjtybws6KT0a… <named list>  
#>  7 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1yyxLosuliwRZxwRbTbF… <named list>  
#>  8 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1s7I8K2EyhibP1TtzcfQ… <named list>  
#>  9 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 13x7aGDDwPTECjdXZU6o… <named list>  
#> 10 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1oy1nEjavHKW7LInQ6xq… <named list>  

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name                 trashed id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename TRUE    18lS9WkXVt32akbgypBao0Uk89Nb_VZAd <named list>  
#>  2 name-squatter-mv     TRUE    1mQDwdIafCalWPDFli7dhhc8Ap7WeAIG8 <named list>  
#>  3 name-squatter-upload TRUE    1iLzlXosyC3YEAe4Z3P-HHGHMcyOI2cTk <named list>  
#>  4 name-squatter-rename TRUE    1GmLsADT05v-EnFqBxt-qe4FK19J7m8_s <named list>  
#>  5 name-squatter-mv     TRUE    1Q51YVITRWPEBtjurWf2gy5p1_RgbVt7t <named list>  
#>  6 name-squatter-upload TRUE    1e03fT5Pwjtybws6KT0aIgbx73Q01Zo3S <named list>  
#>  7 name-squatter-rename TRUE    1yyxLosuliwRZxwRbTbFLNpNs1ekdXjR_ <named list>  
#>  8 name-squatter-mv     TRUE    1s7I8K2EyhibP1TtzcfQoPgADM2r381UY <named list>  
#>  9 name-squatter-upload TRUE    13x7aGDDwPTECjdXZU6oocQw7aoGKCIIL <named list>  
#> 10 name-squatter-rename TRUE    1oy1nEjavHKW7LInQ6xq77k4ZAVTmFPfB <named list>  
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name                 mime_type                id               drive_resource
#>    <chr>                <chr>                    <drv_id>         <list>        
#>  1 name-squatter-rename application/octet-stream 18lS9WkXVt32akb… <named list>  
#>  2 name-squatter-mv     application/octet-stream 1mQDwdIafCalWPD… <named list>  
#>  3 name-squatter-upload application/octet-stream 1iLzlXosyC3YEAe… <named list>  
#>  4 name-squatter-rename application/octet-stream 1GmLsADT05v-EnF… <named list>  
#>  5 name-squatter-mv     application/octet-stream 1Q51YVITRWPEBtj… <named list>  
#>  6 name-squatter-upload application/octet-stream 1e03fT5Pwjtybws… <named list>  
#>  7 name-squatter-rename application/octet-stream 1yyxLosuliwRZxw… <named list>  
#>  8 name-squatter-mv     application/octet-stream 1s7I8K2EyhibP1T… <named list>  
#>  9 name-squatter-upload application/octet-stream 13x7aGDDwPTECjd… <named list>  
#> 10 name-squatter-rename application/octet-stream 1oy1nEjavHKW7LI… <named list>  
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name                 starred id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename FALSE   18lS9WkXVt32akbgypBao0Uk89Nb_VZAd <named list>  
#>  2 name-squatter-mv     FALSE   1mQDwdIafCalWPDFli7dhhc8Ap7WeAIG8 <named list>  
#>  3 name-squatter-upload FALSE   1iLzlXosyC3YEAe4Z3P-HHGHMcyOI2cTk <named list>  
#>  4 name-squatter-rename FALSE   1GmLsADT05v-EnFqBxt-qe4FK19J7m8_s <named list>  
#>  5 name-squatter-mv     FALSE   1Q51YVITRWPEBtjurWf2gy5p1_RgbVt7t <named list>  
#>  6 name-squatter-upload FALSE   1e03fT5Pwjtybws6KT0aIgbx73Q01Zo3S <named list>  
#>  7 name-squatter-rename FALSE   1yyxLosuliwRZxwRbTbFLNpNs1ekdXjR_ <named list>  
#>  8 name-squatter-mv     FALSE   1s7I8K2EyhibP1TtzcfQoPgADM2r381UY <named list>  
#>  9 name-squatter-upload FALSE   13x7aGDDwPTECjdXZU6oocQw7aoGKCIIL <named list>  
#> 10 name-squatter-rename FALSE   1oy1nEjavHKW7LInQ6xq77k4ZAVTmFPfB <named list>  
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name                 description id                            drive_resource
#>    <chr>                <list>      <drv_id>                      <list>        
#>  1 name-squatter-rename <NULL>      18lS9WkXVt32akbgypBao0Uk89Nb… <named list>  
#>  2 name-squatter-mv     <NULL>      1mQDwdIafCalWPDFli7dhhc8Ap7W… <named list>  
#>  3 name-squatter-upload <NULL>      1iLzlXosyC3YEAe4Z3P-HHGHMcyO… <named list>  
#>  4 name-squatter-rename <NULL>      1GmLsADT05v-EnFqBxt-qe4FK19J… <named list>  
#>  5 name-squatter-mv     <NULL>      1Q51YVITRWPEBtjurWf2gy5p1_Rg… <named list>  
#>  6 name-squatter-upload <NULL>      1e03fT5Pwjtybws6KT0aIgbx73Q0… <named list>  
#>  7 name-squatter-rename <NULL>      1yyxLosuliwRZxwRbTbFLNpNs1ek… <named list>  
#>  8 name-squatter-mv     <NULL>      1s7I8K2EyhibP1TtzcfQoPgADM2r… <named list>  
#>  9 name-squatter-upload <NULL>      13x7aGDDwPTECjdXZU6oocQw7aoG… <named list>  
#> 10 name-squatter-rename <NULL>      1oy1nEjavHKW7LInQ6xq77k4ZAVT… <named list>  
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name                 version id                                drive_resource
#>    <chr>                <chr>   <drv_id>                          <list>        
#>  1 name-squatter-rename 2       18lS9WkXVt32akbgypBao0Uk89Nb_VZAd <named list>  
#>  2 name-squatter-mv     2       1mQDwdIafCalWPDFli7dhhc8Ap7WeAIG8 <named list>  
#>  3 name-squatter-upload 2       1iLzlXosyC3YEAe4Z3P-HHGHMcyOI2cTk <named list>  
#>  4 name-squatter-rename 2       1GmLsADT05v-EnFqBxt-qe4FK19J7m8_s <named list>  
#>  5 name-squatter-mv     2       1Q51YVITRWPEBtjurWf2gy5p1_RgbVt7t <named list>  
#>  6 name-squatter-upload 2       1e03fT5Pwjtybws6KT0aIgbx73Q01Zo3S <named list>  
#>  7 name-squatter-rename 2       1yyxLosuliwRZxwRbTbFLNpNs1ekdXjR_ <named list>  
#>  8 name-squatter-mv     2       1s7I8K2EyhibP1TtzcfQoPgADM2r381UY <named list>  
#>  9 name-squatter-upload 2       13x7aGDDwPTECjdXZU6oocQw7aoGKCIIL <named list>  
#> 10 name-squatter-rename 2       1oy1nEjavHKW7LInQ6xq77k4ZAVTmFPfB <named list>  
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name                 web_view_link                       id    drive_resource
#>    <chr>                <chr>                               <drv> <list>        
#>  1 name-squatter-rename https://drive.google.com/file/d/18… 18lS… <named list>  
#>  2 name-squatter-mv     https://drive.google.com/file/d/1m… 1mQD… <named list>  
#>  3 name-squatter-upload https://drive.google.com/file/d/1i… 1iLz… <named list>  
#>  4 name-squatter-rename https://drive.google.com/file/d/1G… 1GmL… <named list>  
#>  5 name-squatter-mv     https://drive.google.com/file/d/1Q… 1Q51… <named list>  
#>  6 name-squatter-upload https://drive.google.com/file/d/1e… 1e03… <named list>  
#>  7 name-squatter-rename https://drive.google.com/file/d/1y… 1yyx… <named list>  
#>  8 name-squatter-mv     https://drive.google.com/file/d/1s… 1s7I… <named list>  
#>  9 name-squatter-upload https://drive.google.com/file/d/13… 13x7… <named list>  
#> 10 name-squatter-rename https://drive.google.com/file/d/1o… 1oy1… <named list>  
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name                 modified_time       id                    drive_resource
#>    <chr>                <dttm>              <drv_id>              <list>        
#>  1 name-squatter-rename 2022-02-18 13:42:15 18lS9WkXVt32akbgypBa… <named list>  
#>  2 name-squatter-mv     2022-02-18 13:42:00 1mQDwdIafCalWPDFli7d… <named list>  
#>  3 name-squatter-upload 2022-02-17 13:42:12 1iLzlXosyC3YEAe4Z3P-… <named list>  
#>  4 name-squatter-rename 2022-02-17 13:41:37 1GmLsADT05v-EnFqBxt-… <named list>  
#>  5 name-squatter-mv     2022-02-17 13:41:20 1Q51YVITRWPEBtjurWf2… <named list>  
#>  6 name-squatter-upload 2022-02-16 13:42:24 1e03fT5Pwjtybws6KT0a… <named list>  
#>  7 name-squatter-rename 2022-02-16 13:41:47 1yyxLosuliwRZxwRbTbF… <named list>  
#>  8 name-squatter-mv     2022-02-16 13:41:33 1s7I8K2EyhibP1TtzcfQ… <named list>  
#>  9 name-squatter-upload 2022-02-15 13:41:50 13x7aGDDwPTECjdXZU6o… <named list>  
#> 10 name-squatter-rename 2022-02-15 13:41:18 1oy1nEjavHKW7LInQ6xq… <named list>  
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name                 created_time        id                    drive_resource
#>    <chr>                <dttm>              <drv_id>              <list>        
#>  1 name-squatter-rename 2022-02-18 13:42:15 18lS9WkXVt32akbgypBa… <named list>  
#>  2 name-squatter-mv     2022-02-18 13:42:00 1mQDwdIafCalWPDFli7d… <named list>  
#>  3 name-squatter-upload 2022-02-17 13:42:12 1iLzlXosyC3YEAe4Z3P-… <named list>  
#>  4 name-squatter-rename 2022-02-17 13:41:37 1GmLsADT05v-EnFqBxt-… <named list>  
#>  5 name-squatter-mv     2022-02-17 13:41:20 1Q51YVITRWPEBtjurWf2… <named list>  
#>  6 name-squatter-upload 2022-02-16 13:42:24 1e03fT5Pwjtybws6KT0a… <named list>  
#>  7 name-squatter-rename 2022-02-16 13:41:47 1yyxLosuliwRZxwRbTbF… <named list>  
#>  8 name-squatter-mv     2022-02-16 13:41:33 1s7I8K2EyhibP1TtzcfQ… <named list>  
#>  9 name-squatter-upload 2022-02-15 13:41:50 13x7aGDDwPTECjdXZU6o… <named list>  
#> 10 name-squatter-rename 2022-02-15 13:41:18 1oy1nEjavHKW7LInQ6xq… <named list>  
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name                 owned_by_me id                            drive_resource
#>    <chr>                <lgl>       <drv_id>                      <list>        
#>  1 name-squatter-rename TRUE        18lS9WkXVt32akbgypBao0Uk89Nb… <named list>  
#>  2 name-squatter-mv     TRUE        1mQDwdIafCalWPDFli7dhhc8Ap7W… <named list>  
#>  3 name-squatter-upload TRUE        1iLzlXosyC3YEAe4Z3P-HHGHMcyO… <named list>  
#>  4 name-squatter-rename TRUE        1GmLsADT05v-EnFqBxt-qe4FK19J… <named list>  
#>  5 name-squatter-mv     TRUE        1Q51YVITRWPEBtjurWf2gy5p1_Rg… <named list>  
#>  6 name-squatter-upload TRUE        1e03fT5Pwjtybws6KT0aIgbx73Q0… <named list>  
#>  7 name-squatter-rename TRUE        1yyxLosuliwRZxwRbTbFLNpNs1ek… <named list>  
#>  8 name-squatter-mv     TRUE        1s7I8K2EyhibP1TtzcfQoPgADM2r… <named list>  
#>  9 name-squatter-upload TRUE        13x7aGDDwPTECjdXZU6oocQw7aoG… <named list>  
#> 10 name-squatter-rename TRUE        1oy1nEjavHKW7LInQ6xq77k4ZAVT… <named list>  
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name                 size  id                                drive_resource
#>    <chr>                <chr> <drv_id>                          <list>        
#>  1 name-squatter-rename 0     18lS9WkXVt32akbgypBao0Uk89Nb_VZAd <named list>  
#>  2 name-squatter-mv     0     1mQDwdIafCalWPDFli7dhhc8Ap7WeAIG8 <named list>  
#>  3 name-squatter-upload 0     1iLzlXosyC3YEAe4Z3P-HHGHMcyOI2cTk <named list>  
#>  4 name-squatter-rename 0     1GmLsADT05v-EnFqBxt-qe4FK19J7m8_s <named list>  
#>  5 name-squatter-mv     0     1Q51YVITRWPEBtjurWf2gy5p1_RgbVt7t <named list>  
#>  6 name-squatter-upload 0     1e03fT5Pwjtybws6KT0aIgbx73Q01Zo3S <named list>  
#>  7 name-squatter-rename 0     1yyxLosuliwRZxwRbTbFLNpNs1ekdXjR_ <named list>  
#>  8 name-squatter-mv     0     1s7I8K2EyhibP1TtzcfQoPgADM2r381UY <named list>  
#>  9 name-squatter-upload 0     13x7aGDDwPTECjdXZU6oocQw7aoGKCIIL <named list>  
#> 10 name-squatter-rename 0     1oy1nEjavHKW7LInQ6xq77k4ZAVTmFPfB <named list>  
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name                 quota_bytes_used id                       drive_resource
#>    <chr>                <chr>            <drv_id>                 <list>        
#>  1 name-squatter-rename 0                18lS9WkXVt32akbgypBao0U… <named list>  
#>  2 name-squatter-mv     0                1mQDwdIafCalWPDFli7dhhc… <named list>  
#>  3 name-squatter-upload 0                1iLzlXosyC3YEAe4Z3P-HHG… <named list>  
#>  4 name-squatter-rename 0                1GmLsADT05v-EnFqBxt-qe4… <named list>  
#>  5 name-squatter-mv     0                1Q51YVITRWPEBtjurWf2gy5… <named list>  
#>  6 name-squatter-upload 0                1e03fT5Pwjtybws6KT0aIgb… <named list>  
#>  7 name-squatter-rename 0                1yyxLosuliwRZxwRbTbFLNp… <named list>  
#>  8 name-squatter-mv     0                1s7I8K2EyhibP1TtzcfQoPg… <named list>  
#>  9 name-squatter-upload 0                13x7aGDDwPTECjdXZU6oocQ… <named list>  
#> 10 name-squatter-rename 0                1oy1nEjavHKW7LInQ6xq77k… <named list>  

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>