Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file
Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what
Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via what appears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 x 4
#>    name          path            id                             drive_resource  
#>    <chr>         <chr>           <drv_id>                       <list>          
#>  1 name-squatter ~/name-squatter 1VB_v7sc9GL6Ip3hxGmdhqmhNzRLM… <named list [39…
#>  2 name-squatter ~/name-squatter 10mh7n8qg3X3htSZMv4LIz7zDgyTp… <named list [39…
#>  3 name-squatter ~/name-squatter 15hBQs6kdDEziSIAyXScvFU_vhS0D… <named list [39…
#>  4 name-squatter ~/name-squatter 1Mr_LUlT_xOs-RbVWZrDehhv6Hjg2… <named list [39…
#>  5 name-squatter ~/name-squatter 155yQeUGetb_7G5er81ZfZDI-xBHS… <named list [39…
#>  6 name-squatter ~/name-squatter 1q2YQ8bHZ246wUndsl4Md1bVmsy5F… <named list [39…
#>  7 name-squatter ~/name-squatter 1VaA88IWJNie-QAJ4tcJiypDCPHnF… <named list [39…
#>  8 name-squatter ~/name-squatter 17DJVtTTv-ztodHJN55OzEpc58xXo… <named list [39…
#>  9 name-squatter ~/name-squatter 17eb5zC7I-cN8xE7yJGm006TU9b6-… <named list [39…
#> 10 name-squatter ~/name-squatter 1yrb9uqylJsX2tXfH4KBBNHSvExBS… <named list [39…
drive_reveal(files, "permissions")
#> # A dribble: 10 x 5
#>    name        shared id                      drive_resource  permissions_resou…
#>    <chr>       <lgl>  <drv_id>                <list>          <list>            
#>  1 name-squat… FALSE  1VB_v7sc9GL6Ip3hxGmdhq… <named list [3… <named list [2]>  
#>  2 name-squat… FALSE  10mh7n8qg3X3htSZMv4LIz… <named list [3… <named list [2]>  
#>  3 name-squat… FALSE  15hBQs6kdDEziSIAyXScvF… <named list [3… <named list [2]>  
#>  4 name-squat… FALSE  1Mr_LUlT_xOs-RbVWZrDeh… <named list [3… <named list [2]>  
#>  5 name-squat… FALSE  155yQeUGetb_7G5er81ZfZ… <named list [3… <named list [2]>  
#>  6 name-squat… FALSE  1q2YQ8bHZ246wUndsl4Md1… <named list [3… <named list [2]>  
#>  7 name-squat… FALSE  1VaA88IWJNie-QAJ4tcJiy… <named list [3… <named list [2]>  
#>  8 name-squat… FALSE  17DJVtTTv-ztodHJN55OzE… <named list [3… <named list [2]>  
#>  9 name-squat… FALSE  17eb5zC7I-cN8xE7yJGm00… <named list [3… <named list [2]>  
#> 10 name-squat… FALSE  1yrb9uqylJsX2tXfH4KBBN… <named list [3… <named list [2]>  
drive_reveal(files, "published")
#> # A dribble: 10 x 5
#>    name       published id                     drive_resource   revision_resour…
#>    <chr>      <lgl>     <drv_id>               <list>           <list>          
#>  1 name-squa… FALSE     1VB_v7sc9GL6Ip3hxGmdh… <named list [39… <named list [9]>
#>  2 name-squa… FALSE     10mh7n8qg3X3htSZMv4LI… <named list [39… <named list [9]>
#>  3 name-squa… FALSE     15hBQs6kdDEziSIAyXScv… <named list [39… <named list [9]>
#>  4 name-squa… FALSE     1Mr_LUlT_xOs-RbVWZrDe… <named list [39… <named list [9]>
#>  5 name-squa… FALSE     155yQeUGetb_7G5er81Zf… <named list [39… <named list [9]>
#>  6 name-squa… FALSE     1q2YQ8bHZ246wUndsl4Md… <named list [39… <named list [9]>
#>  7 name-squa… FALSE     1VaA88IWJNie-QAJ4tcJi… <named list [39… <named list [9]>
#>  8 name-squa… FALSE     17DJVtTTv-ztodHJN55Oz… <named list [39… <named list [9]>
#>  9 name-squa… FALSE     17eb5zC7I-cN8xE7yJGm0… <named list [39… <named list [9]>
#> 10 name-squa… FALSE     1yrb9uqylJsX2tXfH4KBB… <named list [39… <named list [9]>

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 x 4
#>    name         id_parent          id                           drive_resource  
#>    <chr>        <drv_id>           <drv_id>                     <list>          
#>  1 name-squatt… 0AO_RMaBzcP63Uk9P… 1VB_v7sc9GL6Ip3hxGmdhqmhNzR… <named list [39…
#>  2 name-squatt… 0AO_RMaBzcP63Uk9P… 10mh7n8qg3X3htSZMv4LIz7zDgy… <named list [39…
#>  3 name-squatt… 0AO_RMaBzcP63Uk9P… 15hBQs6kdDEziSIAyXScvFU_vhS… <named list [39…
#>  4 name-squatt… 0AO_RMaBzcP63Uk9P… 1Mr_LUlT_xOs-RbVWZrDehhv6Hj… <named list [39…
#>  5 name-squatt… 0AO_RMaBzcP63Uk9P… 155yQeUGetb_7G5er81ZfZDI-xB… <named list [39…
#>  6 name-squatt… 0AO_RMaBzcP63Uk9P… 1q2YQ8bHZ246wUndsl4Md1bVmsy… <named list [39…
#>  7 name-squatt… 0AO_RMaBzcP63Uk9P… 1VaA88IWJNie-QAJ4tcJiypDCPH… <named list [39…
#>  8 name-squatt… 0AO_RMaBzcP63Uk9P… 17DJVtTTv-ztodHJN55OzEpc58x… <named list [39…
#>  9 name-squatt… 0AO_RMaBzcP63Uk9P… 17eb5zC7I-cN8xE7yJGm006TU9b… <named list [39…
#> 10 name-squatt… 0AO_RMaBzcP63Uk9P… 1yrb9uqylJsX2tXfH4KBBNHSvEx… <named list [39…

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 x 4
#>    name          trashed id                                drive_resource   
#>    <chr>         <lgl>   <drv_id>                          <list>           
#>  1 name-squatter TRUE    1VB_v7sc9GL6Ip3hxGmdhqmhNzRLMNmeJ <named list [39]>
#>  2 name-squatter TRUE    10mh7n8qg3X3htSZMv4LIz7zDgyTpgW3G <named list [39]>
#>  3 name-squatter TRUE    15hBQs6kdDEziSIAyXScvFU_vhS0DO2sc <named list [39]>
#>  4 name-squatter TRUE    1Mr_LUlT_xOs-RbVWZrDehhv6Hjg2_Ta2 <named list [39]>
#>  5 name-squatter TRUE    155yQeUGetb_7G5er81ZfZDI-xBHS0o9D <named list [39]>
#>  6 name-squatter TRUE    1q2YQ8bHZ246wUndsl4Md1bVmsy5F8OGb <named list [39]>
#>  7 name-squatter TRUE    1VaA88IWJNie-QAJ4tcJiypDCPHnF-PTW <named list [39]>
#>  8 name-squatter TRUE    17DJVtTTv-ztodHJN55OzEpc58xXoV16I <named list [39]>
#>  9 name-squatter TRUE    17eb5zC7I-cN8xE7yJGm006TU9b6-bGya <named list [39]>
#> 10 name-squatter TRUE    1yrb9uqylJsX2tXfH4KBBNHSvExBSh5f2 <named list [39]>
drive_reveal(files, "mime_type")
#> # A dribble: 10 x 4
#>    name         mime_type            id                         drive_resource  
#>    <chr>        <chr>                <drv_id>                   <list>          
#>  1 name-squatt… application/octet-s… 1VB_v7sc9GL6Ip3hxGmdhqmhN… <named list [39…
#>  2 name-squatt… application/octet-s… 10mh7n8qg3X3htSZMv4LIz7zD… <named list [39…
#>  3 name-squatt… application/octet-s… 15hBQs6kdDEziSIAyXScvFU_v… <named list [39…
#>  4 name-squatt… application/octet-s… 1Mr_LUlT_xOs-RbVWZrDehhv6… <named list [39…
#>  5 name-squatt… application/octet-s… 155yQeUGetb_7G5er81ZfZDI-… <named list [39…
#>  6 name-squatt… application/octet-s… 1q2YQ8bHZ246wUndsl4Md1bVm… <named list [39…
#>  7 name-squatt… application/octet-s… 1VaA88IWJNie-QAJ4tcJiypDC… <named list [39…
#>  8 name-squatt… application/octet-s… 17DJVtTTv-ztodHJN55OzEpc5… <named list [39…
#>  9 name-squatt… application/octet-s… 17eb5zC7I-cN8xE7yJGm006TU… <named list [39…
#> 10 name-squatt… application/octet-s… 1yrb9uqylJsX2tXfH4KBBNHSv… <named list [39…
drive_reveal(files, "starred")
#> # A dribble: 10 x 4
#>    name          starred id                                drive_resource   
#>    <chr>         <lgl>   <drv_id>                          <list>           
#>  1 name-squatter FALSE   1VB_v7sc9GL6Ip3hxGmdhqmhNzRLMNmeJ <named list [39]>
#>  2 name-squatter FALSE   10mh7n8qg3X3htSZMv4LIz7zDgyTpgW3G <named list [39]>
#>  3 name-squatter FALSE   15hBQs6kdDEziSIAyXScvFU_vhS0DO2sc <named list [39]>
#>  4 name-squatter FALSE   1Mr_LUlT_xOs-RbVWZrDehhv6Hjg2_Ta2 <named list [39]>
#>  5 name-squatter FALSE   155yQeUGetb_7G5er81ZfZDI-xBHS0o9D <named list [39]>
#>  6 name-squatter FALSE   1q2YQ8bHZ246wUndsl4Md1bVmsy5F8OGb <named list [39]>
#>  7 name-squatter FALSE   1VaA88IWJNie-QAJ4tcJiypDCPHnF-PTW <named list [39]>
#>  8 name-squatter FALSE   17DJVtTTv-ztodHJN55OzEpc58xXoV16I <named list [39]>
#>  9 name-squatter FALSE   17eb5zC7I-cN8xE7yJGm006TU9b6-bGya <named list [39]>
#> 10 name-squatter FALSE   1yrb9uqylJsX2tXfH4KBBNHSvExBSh5f2 <named list [39]>
drive_reveal(files, "description")
#> # A dribble: 10 x 4
#>    name          description id                                drive_resource   
#>    <chr>         <list>      <drv_id>                          <list>           
#>  1 name-squatter <NULL>      1VB_v7sc9GL6Ip3hxGmdhqmhNzRLMNmeJ <named list [39]>
#>  2 name-squatter <NULL>      10mh7n8qg3X3htSZMv4LIz7zDgyTpgW3G <named list [39]>
#>  3 name-squatter <NULL>      15hBQs6kdDEziSIAyXScvFU_vhS0DO2sc <named list [39]>
#>  4 name-squatter <NULL>      1Mr_LUlT_xOs-RbVWZrDehhv6Hjg2_Ta2 <named list [39]>
#>  5 name-squatter <NULL>      155yQeUGetb_7G5er81ZfZDI-xBHS0o9D <named list [39]>
#>  6 name-squatter <NULL>      1q2YQ8bHZ246wUndsl4Md1bVmsy5F8OGb <named list [39]>
#>  7 name-squatter <NULL>      1VaA88IWJNie-QAJ4tcJiypDCPHnF-PTW <named list [39]>
#>  8 name-squatter <NULL>      17DJVtTTv-ztodHJN55OzEpc58xXoV16I <named list [39]>
#>  9 name-squatter <NULL>      17eb5zC7I-cN8xE7yJGm006TU9b6-bGya <named list [39]>
#> 10 name-squatter <NULL>      1yrb9uqylJsX2tXfH4KBBNHSvExBSh5f2 <named list [39]>
drive_reveal(files, "version")
#> # A dribble: 10 x 4
#>    name          version id                                drive_resource   
#>    <chr>         <chr>   <drv_id>                          <list>           
#>  1 name-squatter 2       1VB_v7sc9GL6Ip3hxGmdhqmhNzRLMNmeJ <named list [39]>
#>  2 name-squatter 2       10mh7n8qg3X3htSZMv4LIz7zDgyTpgW3G <named list [39]>
#>  3 name-squatter 2       15hBQs6kdDEziSIAyXScvFU_vhS0DO2sc <named list [39]>
#>  4 name-squatter 2       1Mr_LUlT_xOs-RbVWZrDehhv6Hjg2_Ta2 <named list [39]>
#>  5 name-squatter 2       155yQeUGetb_7G5er81ZfZDI-xBHS0o9D <named list [39]>
#>  6 name-squatter 2       1q2YQ8bHZ246wUndsl4Md1bVmsy5F8OGb <named list [39]>
#>  7 name-squatter 2       1VaA88IWJNie-QAJ4tcJiypDCPHnF-PTW <named list [39]>
#>  8 name-squatter 2       17DJVtTTv-ztodHJN55OzEpc58xXoV16I <named list [39]>
#>  9 name-squatter 2       17eb5zC7I-cN8xE7yJGm006TU9b6-bGya <named list [39]>
#> 10 name-squatter 2       1yrb9uqylJsX2tXfH4KBBNHSvExBSh5f2 <named list [39]>
drive_reveal(files, "web_view_link")
#> # A dribble: 10 x 4
#>    name     web_view_link                        id              drive_resource 
#>    <chr>    <chr>                                <drv_id>        <list>         
#>  1 name-sq… https://drive.google.com/file/d/1VB… 1VB_v7sc9GL6Ip… <named list [3…
#>  2 name-sq… https://drive.google.com/file/d/10m… 10mh7n8qg3X3ht… <named list [3…
#>  3 name-sq… https://drive.google.com/file/d/15h… 15hBQs6kdDEziS… <named list [3…
#>  4 name-sq… https://drive.google.com/file/d/1Mr… 1Mr_LUlT_xOs-R… <named list [3…
#>  5 name-sq… https://drive.google.com/file/d/155… 155yQeUGetb_7G… <named list [3…
#>  6 name-sq… https://drive.google.com/file/d/1q2… 1q2YQ8bHZ246wU… <named list [3…
#>  7 name-sq… https://drive.google.com/file/d/1Va… 1VaA88IWJNie-Q… <named list [3…
#>  8 name-sq… https://drive.google.com/file/d/17D… 17DJVtTTv-ztod… <named list [3…
#>  9 name-sq… https://drive.google.com/file/d/17e… 17eb5zC7I-cN8x… <named list [3…
#> 10 name-sq… https://drive.google.com/file/d/1yr… 1yrb9uqylJsX2t… <named list [3…
drive_reveal(files, "modified_time")
#> # A dribble: 10 x 4
#>    name         modified_time       id                          drive_resource  
#>    <chr>        <dttm>              <drv_id>                    <list>          
#>  1 name-squatt… 2021-07-08 17:28:36 1VB_v7sc9GL6Ip3hxGmdhqmhNz… <named list [39…
#>  2 name-squatt… 2021-07-08 17:28:22 10mh7n8qg3X3htSZMv4LIz7zDg… <named list [39…
#>  3 name-squatt… 2021-07-08 13:49:06 15hBQs6kdDEziSIAyXScvFU_vh… <named list [39…
#>  4 name-squatt… 2021-07-08 13:48:32 1Mr_LUlT_xOs-RbVWZrDehhv6H… <named list [39…
#>  5 name-squatt… 2021-07-08 13:48:17 155yQeUGetb_7G5er81ZfZDI-x… <named list [39…
#>  6 name-squatt… 2021-07-07 22:54:27 1q2YQ8bHZ246wUndsl4Md1bVms… <named list [39…
#>  7 name-squatt… 2021-07-07 22:53:45 1VaA88IWJNie-QAJ4tcJiypDCP… <named list [39…
#>  8 name-squatt… 2021-07-07 22:53:24 17DJVtTTv-ztodHJN55OzEpc58… <named list [39…
#>  9 name-squatt… 2021-07-07 21:56:23 17eb5zC7I-cN8xE7yJGm006TU9… <named list [39…
#> 10 name-squatt… 2021-07-07 21:55:52 1yrb9uqylJsX2tXfH4KBBNHSvE… <named list [39…
drive_reveal(files, "created_time")
#> # A dribble: 10 x 4
#>    name         created_time        id                          drive_resource  
#>    <chr>        <dttm>              <drv_id>                    <list>          
#>  1 name-squatt… 2021-07-08 17:28:36 1VB_v7sc9GL6Ip3hxGmdhqmhNz… <named list [39…
#>  2 name-squatt… 2021-07-08 17:28:22 10mh7n8qg3X3htSZMv4LIz7zDg… <named list [39…
#>  3 name-squatt… 2021-07-08 13:49:06 15hBQs6kdDEziSIAyXScvFU_vh… <named list [39…
#>  4 name-squatt… 2021-07-08 13:48:32 1Mr_LUlT_xOs-RbVWZrDehhv6H… <named list [39…
#>  5 name-squatt… 2021-07-08 13:48:17 155yQeUGetb_7G5er81ZfZDI-x… <named list [39…
#>  6 name-squatt… 2021-07-07 22:54:27 1q2YQ8bHZ246wUndsl4Md1bVms… <named list [39…
#>  7 name-squatt… 2021-07-07 22:53:45 1VaA88IWJNie-QAJ4tcJiypDCP… <named list [39…
#>  8 name-squatt… 2021-07-07 22:53:24 17DJVtTTv-ztodHJN55OzEpc58… <named list [39…
#>  9 name-squatt… 2021-07-07 21:56:23 17eb5zC7I-cN8xE7yJGm006TU9… <named list [39…
#> 10 name-squatt… 2021-07-07 21:55:52 1yrb9uqylJsX2tXfH4KBBNHSvE… <named list [39…
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 x 4
#>    name          owned_by_me id                                drive_resource   
#>    <chr>         <lgl>       <drv_id>                          <list>           
#>  1 name-squatter TRUE        1VB_v7sc9GL6Ip3hxGmdhqmhNzRLMNmeJ <named list [39]>
#>  2 name-squatter TRUE        10mh7n8qg3X3htSZMv4LIz7zDgyTpgW3G <named list [39]>
#>  3 name-squatter TRUE        15hBQs6kdDEziSIAyXScvFU_vhS0DO2sc <named list [39]>
#>  4 name-squatter TRUE        1Mr_LUlT_xOs-RbVWZrDehhv6Hjg2_Ta2 <named list [39]>
#>  5 name-squatter TRUE        155yQeUGetb_7G5er81ZfZDI-xBHS0o9D <named list [39]>
#>  6 name-squatter TRUE        1q2YQ8bHZ246wUndsl4Md1bVmsy5F8OGb <named list [39]>
#>  7 name-squatter TRUE        1VaA88IWJNie-QAJ4tcJiypDCPHnF-PTW <named list [39]>
#>  8 name-squatter TRUE        17DJVtTTv-ztodHJN55OzEpc58xXoV16I <named list [39]>
#>  9 name-squatter TRUE        17eb5zC7I-cN8xE7yJGm006TU9b6-bGya <named list [39]>
#> 10 name-squatter TRUE        1yrb9uqylJsX2tXfH4KBBNHSvExBSh5f2 <named list [39]>
drive_reveal(files, "size")
#> # A dribble: 10 x 4
#>    name          size  id                                drive_resource   
#>    <chr>         <chr> <drv_id>                          <list>           
#>  1 name-squatter 0     1VB_v7sc9GL6Ip3hxGmdhqmhNzRLMNmeJ <named list [39]>
#>  2 name-squatter 0     10mh7n8qg3X3htSZMv4LIz7zDgyTpgW3G <named list [39]>
#>  3 name-squatter 0     15hBQs6kdDEziSIAyXScvFU_vhS0DO2sc <named list [39]>
#>  4 name-squatter 0     1Mr_LUlT_xOs-RbVWZrDehhv6Hjg2_Ta2 <named list [39]>
#>  5 name-squatter 0     155yQeUGetb_7G5er81ZfZDI-xBHS0o9D <named list [39]>
#>  6 name-squatter 0     1q2YQ8bHZ246wUndsl4Md1bVmsy5F8OGb <named list [39]>
#>  7 name-squatter 0     1VaA88IWJNie-QAJ4tcJiypDCPHnF-PTW <named list [39]>
#>  8 name-squatter 0     17DJVtTTv-ztodHJN55OzEpc58xXoV16I <named list [39]>
#>  9 name-squatter 0     17eb5zC7I-cN8xE7yJGm006TU9b6-bGya <named list [39]>
#> 10 name-squatter 0     1yrb9uqylJsX2tXfH4KBBNHSvExBSh5f2 <named list [39]>
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 x 4
#>    name         quota_bytes_used id                             drive_resource  
#>    <chr>        <chr>            <drv_id>                       <list>          
#>  1 name-squatt… 0                1VB_v7sc9GL6Ip3hxGmdhqmhNzRLM… <named list [39…
#>  2 name-squatt… 0                10mh7n8qg3X3htSZMv4LIz7zDgyTp… <named list [39…
#>  3 name-squatt… 0                15hBQs6kdDEziSIAyXScvFU_vhS0D… <named list [39…
#>  4 name-squatt… 0                1Mr_LUlT_xOs-RbVWZrDehhv6Hjg2… <named list [39…
#>  5 name-squatt… 0                155yQeUGetb_7G5er81ZfZDI-xBHS… <named list [39…
#>  6 name-squatt… 0                1q2YQ8bHZ246wUndsl4Md1bVmsy5F… <named list [39…
#>  7 name-squatt… 0                1VaA88IWJNie-QAJ4tcJiypDCPHnF… <named list [39…
#>  8 name-squatt… 0                17DJVtTTv-ztodHJN55OzEpc58xXo… <named list [39…
#>  9 name-squatt… 0                17eb5zC7I-cN8xE7yJGm006TU9b6-… <named list [39…
#> 10 name-squatt… 0                1yrb9uqylJsX2tXfH4KBBNHSvExBS… <named list [39…

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 x 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>