Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what

Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via whatappears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name                 path                   id                drive_resource 
#>    <chr>                <chr>                  <drv_id>          <list>         
#>  1 name-squatter-rename ~/name-squatter-rename 1Rnimmb5oeedCCNZ… <named list [3…
#>  2 name-squatter-mv     ~/name-squatter-mv     1dMS3qLQqULcwb9S… <named list [3…
#>  3 name-squatter-upload ~/name-squatter-upload 1WpwhD7CeUq8Rvxz… <named list [3…
#>  4 name-squatter-rename ~/name-squatter-rename 1Vl4IKYYEds4UuGN… <named list [3…
#>  5 name-squatter-mv     ~/name-squatter-mv     19ZCvNMVA5-E4bxg… <named list [3…
#>  6 name-squatter-upload ~/name-squatter-upload 1Q4ceCNl3RL3q4AL… <named list [3…
#>  7 name-squatter-rename ~/name-squatter-rename 1sQTQr8oFbQQGVhT… <named list [3…
#>  8 name-squatter-mv     ~/name-squatter-mv     1XKdD4614QrZlQQZ… <named list [3…
#>  9 name-squatter-upload ~/name-squatter-upload 1P6w9aBeubbB9MR3… <named list [3…
#> 10 name-squatter-rename ~/name-squatter-rename 1LIuCYBtX308z36f… <named list [3…
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name                 shared id              drive_resource  permissions_reso…
#>    <chr>                <lgl>  <drv_id>        <list>          <list>           
#>  1 name-squatter-rename FALSE  1Rnimmb5oeedCC… <named list [3… <named list [2]> 
#>  2 name-squatter-mv     FALSE  1dMS3qLQqULcwb… <named list [3… <named list [2]> 
#>  3 name-squatter-upload FALSE  1WpwhD7CeUq8Rv… <named list [3… <named list [2]> 
#>  4 name-squatter-rename FALSE  1Vl4IKYYEds4Uu… <named list [3… <named list [2]> 
#>  5 name-squatter-mv     FALSE  19ZCvNMVA5-E4b… <named list [3… <named list [2]> 
#>  6 name-squatter-upload FALSE  1Q4ceCNl3RL3q4… <named list [3… <named list [2]> 
#>  7 name-squatter-rename FALSE  1sQTQr8oFbQQGV… <named list [3… <named list [2]> 
#>  8 name-squatter-mv     FALSE  1XKdD4614QrZlQ… <named list [3… <named list [2]> 
#>  9 name-squatter-upload FALSE  1P6w9aBeubbB9M… <named list [3… <named list [2]> 
#> 10 name-squatter-rename FALSE  1LIuCYBtX308z3… <named list [3… <named list [2]> 
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name                 published id            drive_resource  revision_resour…
#>    <chr>                <lgl>     <drv_id>      <list>          <list>          
#>  1 name-squatter-rename FALSE     1Rnimmb5oeed… <named list [3… <named list [9]>
#>  2 name-squatter-mv     FALSE     1dMS3qLQqULc… <named list [3… <named list [9]>
#>  3 name-squatter-upload FALSE     1WpwhD7CeUq8… <named list [3… <named list [9]>
#>  4 name-squatter-rename FALSE     1Vl4IKYYEds4… <named list [3… <named list [9]>
#>  5 name-squatter-mv     FALSE     19ZCvNMVA5-E… <named list [3… <named list [9]>
#>  6 name-squatter-upload FALSE     1Q4ceCNl3RL3… <named list [3… <named list [9]>
#>  7 name-squatter-rename FALSE     1sQTQr8oFbQQ… <named list [3… <named list [9]>
#>  8 name-squatter-mv     FALSE     1XKdD4614QrZ… <named list [3… <named list [9]>
#>  9 name-squatter-upload FALSE     1P6w9aBeubbB… <named list [3… <named list [9]>
#> 10 name-squatter-rename FALSE     1LIuCYBtX308… <named list [3… <named list [9]>

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name                 id_parent           id                   drive_resource 
#>    <chr>                <drv_id>            <drv_id>             <list>         
#>  1 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1Rnimmb5oeedCCNZ9ho… <named list [3…
#>  2 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1dMS3qLQqULcwb9SKBL… <named list [3…
#>  3 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1WpwhD7CeUq8RvxzVOo… <named list [3…
#>  4 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1Vl4IKYYEds4UuGNNDg… <named list [3…
#>  5 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 19ZCvNMVA5-E4bxgJu8… <named list [3…
#>  6 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1Q4ceCNl3RL3q4ALOhD… <named list [3…
#>  7 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1sQTQr8oFbQQGVhTCQk… <named list [3…
#>  8 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1XKdD4614QrZlQQZNkz… <named list [3…
#>  9 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1P6w9aBeubbB9MR3FCF… <named list [3…
#> 10 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1LIuCYBtX308z36faj7… <named list [3…

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name                 trashed id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename TRUE    1Rnimmb5oeedCCNZ9ho7mS9NLrpOMpo1P <named list […
#>  2 name-squatter-mv     TRUE    1dMS3qLQqULcwb9SKBLGWcU4CAJqiaSPT <named list […
#>  3 name-squatter-upload TRUE    1WpwhD7CeUq8RvxzVOoKjwMcFUNbnciQ- <named list […
#>  4 name-squatter-rename TRUE    1Vl4IKYYEds4UuGNNDgd19Y02gpc3EP_U <named list […
#>  5 name-squatter-mv     TRUE    19ZCvNMVA5-E4bxgJu8JaEHWIyIkTFT0g <named list […
#>  6 name-squatter-upload TRUE    1Q4ceCNl3RL3q4ALOhDRrkRJ5M_5BCahE <named list […
#>  7 name-squatter-rename TRUE    1sQTQr8oFbQQGVhTCQkol40wDV6Zg8a1y <named list […
#>  8 name-squatter-mv     TRUE    1XKdD4614QrZlQQZNkzx29Z_DhldMv5ZI <named list […
#>  9 name-squatter-upload TRUE    1P6w9aBeubbB9MR3FCFU1yx8tYWq-I87w <named list […
#> 10 name-squatter-rename TRUE    1LIuCYBtX308z36faj7i4hB2NQNQQYPfB <named list […
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name                 mime_type                id              drive_resource 
#>    <chr>                <chr>                    <drv_id>        <list>         
#>  1 name-squatter-rename application/octet-stream 1Rnimmb5oeedCC… <named list [3…
#>  2 name-squatter-mv     application/octet-stream 1dMS3qLQqULcwb… <named list [3…
#>  3 name-squatter-upload application/octet-stream 1WpwhD7CeUq8Rv… <named list [3…
#>  4 name-squatter-rename application/octet-stream 1Vl4IKYYEds4Uu… <named list [3…
#>  5 name-squatter-mv     application/octet-stream 19ZCvNMVA5-E4b… <named list [3…
#>  6 name-squatter-upload application/octet-stream 1Q4ceCNl3RL3q4… <named list [3…
#>  7 name-squatter-rename application/octet-stream 1sQTQr8oFbQQGV… <named list [3…
#>  8 name-squatter-mv     application/octet-stream 1XKdD4614QrZlQ… <named list [3…
#>  9 name-squatter-upload application/octet-stream 1P6w9aBeubbB9M… <named list [3…
#> 10 name-squatter-rename application/octet-stream 1LIuCYBtX308z3… <named list [3…
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name                 starred id                                drive_resource
#>    <chr>                <lgl>   <drv_id>                          <list>        
#>  1 name-squatter-rename FALSE   1Rnimmb5oeedCCNZ9ho7mS9NLrpOMpo1P <named list […
#>  2 name-squatter-mv     FALSE   1dMS3qLQqULcwb9SKBLGWcU4CAJqiaSPT <named list […
#>  3 name-squatter-upload FALSE   1WpwhD7CeUq8RvxzVOoKjwMcFUNbnciQ- <named list […
#>  4 name-squatter-rename FALSE   1Vl4IKYYEds4UuGNNDgd19Y02gpc3EP_U <named list […
#>  5 name-squatter-mv     FALSE   19ZCvNMVA5-E4bxgJu8JaEHWIyIkTFT0g <named list […
#>  6 name-squatter-upload FALSE   1Q4ceCNl3RL3q4ALOhDRrkRJ5M_5BCahE <named list […
#>  7 name-squatter-rename FALSE   1sQTQr8oFbQQGVhTCQkol40wDV6Zg8a1y <named list […
#>  8 name-squatter-mv     FALSE   1XKdD4614QrZlQQZNkzx29Z_DhldMv5ZI <named list […
#>  9 name-squatter-upload FALSE   1P6w9aBeubbB9MR3FCFU1yx8tYWq-I87w <named list […
#> 10 name-squatter-rename FALSE   1LIuCYBtX308z36faj7i4hB2NQNQQYPfB <named list […
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name                 description id                          drive_resource  
#>    <chr>                <list>      <drv_id>                    <list>          
#>  1 name-squatter-rename <NULL>      1Rnimmb5oeedCCNZ9ho7mS9NLr… <named list [39…
#>  2 name-squatter-mv     <NULL>      1dMS3qLQqULcwb9SKBLGWcU4CA… <named list [39…
#>  3 name-squatter-upload <NULL>      1WpwhD7CeUq8RvxzVOoKjwMcFU… <named list [39…
#>  4 name-squatter-rename <NULL>      1Vl4IKYYEds4UuGNNDgd19Y02g… <named list [39…
#>  5 name-squatter-mv     <NULL>      19ZCvNMVA5-E4bxgJu8JaEHWIy… <named list [39…
#>  6 name-squatter-upload <NULL>      1Q4ceCNl3RL3q4ALOhDRrkRJ5M… <named list [39…
#>  7 name-squatter-rename <NULL>      1sQTQr8oFbQQGVhTCQkol40wDV… <named list [39…
#>  8 name-squatter-mv     <NULL>      1XKdD4614QrZlQQZNkzx29Z_Dh… <named list [39…
#>  9 name-squatter-upload <NULL>      1P6w9aBeubbB9MR3FCFU1yx8tY… <named list [39…
#> 10 name-squatter-rename <NULL>      1LIuCYBtX308z36faj7i4hB2NQ… <named list [39…
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name                 version id                                drive_resource
#>    <chr>                <chr>   <drv_id>                          <list>        
#>  1 name-squatter-rename 2       1Rnimmb5oeedCCNZ9ho7mS9NLrpOMpo1P <named list […
#>  2 name-squatter-mv     2       1dMS3qLQqULcwb9SKBLGWcU4CAJqiaSPT <named list […
#>  3 name-squatter-upload 2       1WpwhD7CeUq8RvxzVOoKjwMcFUNbnciQ- <named list […
#>  4 name-squatter-rename 2       1Vl4IKYYEds4UuGNNDgd19Y02gpc3EP_U <named list […
#>  5 name-squatter-mv     2       19ZCvNMVA5-E4bxgJu8JaEHWIyIkTFT0g <named list […
#>  6 name-squatter-upload 2       1Q4ceCNl3RL3q4ALOhDRrkRJ5M_5BCahE <named list […
#>  7 name-squatter-rename 2       1sQTQr8oFbQQGVhTCQkol40wDV6Zg8a1y <named list […
#>  8 name-squatter-mv     2       1XKdD4614QrZlQQZNkzx29Z_DhldMv5ZI <named list […
#>  9 name-squatter-upload 2       1P6w9aBeubbB9MR3FCFU1yx8tYWq-I87w <named list […
#> 10 name-squatter-rename 2       1LIuCYBtX308z36faj7i4hB2NQNQQYPfB <named list […
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name                 web_view_link                id           drive_resource
#>    <chr>                <chr>                        <drv_id>     <list>        
#>  1 name-squatter-rename https://drive.google.com/fi… 1Rnimmb5oee… <named list […
#>  2 name-squatter-mv     https://drive.google.com/fi… 1dMS3qLQqUL… <named list […
#>  3 name-squatter-upload https://drive.google.com/fi… 1WpwhD7CeUq… <named list […
#>  4 name-squatter-rename https://drive.google.com/fi… 1Vl4IKYYEds… <named list […
#>  5 name-squatter-mv     https://drive.google.com/fi… 19ZCvNMVA5-… <named list […
#>  6 name-squatter-upload https://drive.google.com/fi… 1Q4ceCNl3RL… <named list […
#>  7 name-squatter-rename https://drive.google.com/fi… 1sQTQr8oFbQ… <named list […
#>  8 name-squatter-mv     https://drive.google.com/fi… 1XKdD4614Qr… <named list […
#>  9 name-squatter-upload https://drive.google.com/fi… 1P6w9aBeubb… <named list […
#> 10 name-squatter-rename https://drive.google.com/fi… 1LIuCYBtX30… <named list […
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name                 modified_time       id                   drive_resource 
#>    <chr>                <dttm>              <drv_id>             <list>         
#>  1 name-squatter-rename 2022-01-04 13:46:59 1Rnimmb5oeedCCNZ9ho… <named list [3…
#>  2 name-squatter-mv     2022-01-04 13:46:44 1dMS3qLQqULcwb9SKBL… <named list [3…
#>  3 name-squatter-upload 2022-01-03 13:47:44 1WpwhD7CeUq8RvxzVOo… <named list [3…
#>  4 name-squatter-rename 2022-01-03 13:47:12 1Vl4IKYYEds4UuGNNDg… <named list [3…
#>  5 name-squatter-mv     2022-01-03 13:46:57 19ZCvNMVA5-E4bxgJu8… <named list [3…
#>  6 name-squatter-upload 2022-01-02 13:46:50 1Q4ceCNl3RL3q4ALOhD… <named list [3…
#>  7 name-squatter-rename 2022-01-02 13:46:15 1sQTQr8oFbQQGVhTCQk… <named list [3…
#>  8 name-squatter-mv     2022-01-02 13:45:58 1XKdD4614QrZlQQZNkz… <named list [3…
#>  9 name-squatter-upload 2022-01-01 13:46:20 1P6w9aBeubbB9MR3FCF… <named list [3…
#> 10 name-squatter-rename 2022-01-01 13:45:46 1LIuCYBtX308z36faj7… <named list [3…
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name                 created_time        id                   drive_resource 
#>    <chr>                <dttm>              <drv_id>             <list>         
#>  1 name-squatter-rename 2022-01-04 13:46:59 1Rnimmb5oeedCCNZ9ho… <named list [3…
#>  2 name-squatter-mv     2022-01-04 13:46:44 1dMS3qLQqULcwb9SKBL… <named list [3…
#>  3 name-squatter-upload 2022-01-03 13:47:44 1WpwhD7CeUq8RvxzVOo… <named list [3…
#>  4 name-squatter-rename 2022-01-03 13:47:12 1Vl4IKYYEds4UuGNNDg… <named list [3…
#>  5 name-squatter-mv     2022-01-03 13:46:57 19ZCvNMVA5-E4bxgJu8… <named list [3…
#>  6 name-squatter-upload 2022-01-02 13:46:50 1Q4ceCNl3RL3q4ALOhD… <named list [3…
#>  7 name-squatter-rename 2022-01-02 13:46:15 1sQTQr8oFbQQGVhTCQk… <named list [3…
#>  8 name-squatter-mv     2022-01-02 13:45:58 1XKdD4614QrZlQQZNkz… <named list [3…
#>  9 name-squatter-upload 2022-01-01 13:46:20 1P6w9aBeubbB9MR3FCF… <named list [3…
#> 10 name-squatter-rename 2022-01-01 13:45:46 1LIuCYBtX308z36faj7… <named list [3…
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name                 owned_by_me id                          drive_resource  
#>    <chr>                <lgl>       <drv_id>                    <list>          
#>  1 name-squatter-rename TRUE        1Rnimmb5oeedCCNZ9ho7mS9NLr… <named list [39…
#>  2 name-squatter-mv     TRUE        1dMS3qLQqULcwb9SKBLGWcU4CA… <named list [39…
#>  3 name-squatter-upload TRUE        1WpwhD7CeUq8RvxzVOoKjwMcFU… <named list [39…
#>  4 name-squatter-rename TRUE        1Vl4IKYYEds4UuGNNDgd19Y02g… <named list [39…
#>  5 name-squatter-mv     TRUE        19ZCvNMVA5-E4bxgJu8JaEHWIy… <named list [39…
#>  6 name-squatter-upload TRUE        1Q4ceCNl3RL3q4ALOhDRrkRJ5M… <named list [39…
#>  7 name-squatter-rename TRUE        1sQTQr8oFbQQGVhTCQkol40wDV… <named list [39…
#>  8 name-squatter-mv     TRUE        1XKdD4614QrZlQQZNkzx29Z_Dh… <named list [39…
#>  9 name-squatter-upload TRUE        1P6w9aBeubbB9MR3FCFU1yx8tY… <named list [39…
#> 10 name-squatter-rename TRUE        1LIuCYBtX308z36faj7i4hB2NQ… <named list [39…
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name                 size  id                                drive_resource  
#>    <chr>                <chr> <drv_id>                          <list>          
#>  1 name-squatter-rename 0     1Rnimmb5oeedCCNZ9ho7mS9NLrpOMpo1P <named list [39…
#>  2 name-squatter-mv     0     1dMS3qLQqULcwb9SKBLGWcU4CAJqiaSPT <named list [39…
#>  3 name-squatter-upload 0     1WpwhD7CeUq8RvxzVOoKjwMcFUNbnciQ- <named list [39…
#>  4 name-squatter-rename 0     1Vl4IKYYEds4UuGNNDgd19Y02gpc3EP_U <named list [39…
#>  5 name-squatter-mv     0     19ZCvNMVA5-E4bxgJu8JaEHWIyIkTFT0g <named list [39…
#>  6 name-squatter-upload 0     1Q4ceCNl3RL3q4ALOhDRrkRJ5M_5BCahE <named list [39…
#>  7 name-squatter-rename 0     1sQTQr8oFbQQGVhTCQkol40wDV6Zg8a1y <named list [39…
#>  8 name-squatter-mv     0     1XKdD4614QrZlQQZNkzx29Z_DhldMv5ZI <named list [39…
#>  9 name-squatter-upload 0     1P6w9aBeubbB9MR3FCFU1yx8tYWq-I87w <named list [39…
#> 10 name-squatter-rename 0     1LIuCYBtX308z36faj7i4hB2NQNQQYPfB <named list [39…
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name                 quota_bytes_used id                      drive_resource 
#>    <chr>                <chr>            <drv_id>                <list>         
#>  1 name-squatter-rename 0                1Rnimmb5oeedCCNZ9ho7mS… <named list [3…
#>  2 name-squatter-mv     0                1dMS3qLQqULcwb9SKBLGWc… <named list [3…
#>  3 name-squatter-upload 0                1WpwhD7CeUq8RvxzVOoKjw… <named list [3…
#>  4 name-squatter-rename 0                1Vl4IKYYEds4UuGNNDgd19… <named list [3…
#>  5 name-squatter-mv     0                19ZCvNMVA5-E4bxgJu8JaE… <named list [3…
#>  6 name-squatter-upload 0                1Q4ceCNl3RL3q4ALOhDRrk… <named list [3…
#>  7 name-squatter-rename 0                1sQTQr8oFbQQGVhTCQkol4… <named list [3…
#>  8 name-squatter-mv     0                1XKdD4614QrZlQQZNkzx29… <named list [3…
#>  9 name-squatter-upload 0                1P6w9aBeubbB9MR3FCFU1y… <named list [3…
#> 10 name-squatter-rename 0                1LIuCYBtX308z36faj7i4h… <named list [3…

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>