Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file
Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what
Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via what appears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name          path            id                                drive_resource
#>    <chr>         <chr>           <drv_id>                          <list>        
#>  1 name-squatter ~/name-squatter 1W0LpT7f9_AZd8X4zi6ixnjv4MmtsuCjz <named list […
#>  2 name-squatter ~/name-squatter 1agVDGb64Y6ssGW2mBhUKyLAB2Diwqfpg <named list […
#>  3 name-squatter ~/name-squatter 1PBzxw_Ij-XMwnDM4moQ1SHJe7IpLZYzu <named list […
#>  4 name-squatter ~/name-squatter 1xbsMhPsrsOn-GLv4YDgmFJTfB0moPEGA <named list […
#>  5 name-squatter ~/name-squatter 1BtAxk5ZF99CEhH84vq2VxeE66Mgg8k6k <named list […
#>  6 name-squatter ~/name-squatter 1Zmvd7_zUP8d17F8sAYCprmWOT2oBoYGN <named list […
#>  7 name-squatter ~/name-squatter 19Y7bQtDs43oc96Vhaj6w9ou7dfaBVKHw <named list […
#>  8 name-squatter ~/name-squatter 1w2FUce0a1Esv2K9sEn4s8gPZE7cHTj4v <named list […
#>  9 name-squatter ~/name-squatter 18hzoJiJruyXqtE4zrd25MAe5Y5-Vi2xH <named list […
#> 10 name-squatter ~/name-squatter 1MX82LqNHapLLsfDx0UjlVIgvhIqZe8sr <named list […
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name          shared id                                drive_resource permissions_res…
#>    <chr>         <lgl>  <drv_id>                          <list>         <list>          
#>  1 name-squatter FALSE  1W0LpT7f9_AZd8X4zi6ixnjv4MmtsuCjz <named list [… <named list [2]>
#>  2 name-squatter FALSE  1agVDGb64Y6ssGW2mBhUKyLAB2Diwqfpg <named list [… <named list [2]>
#>  3 name-squatter FALSE  1PBzxw_Ij-XMwnDM4moQ1SHJe7IpLZYzu <named list [… <named list [2]>
#>  4 name-squatter FALSE  1xbsMhPsrsOn-GLv4YDgmFJTfB0moPEGA <named list [… <named list [2]>
#>  5 name-squatter FALSE  1BtAxk5ZF99CEhH84vq2VxeE66Mgg8k6k <named list [… <named list [2]>
#>  6 name-squatter FALSE  1Zmvd7_zUP8d17F8sAYCprmWOT2oBoYGN <named list [… <named list [2]>
#>  7 name-squatter FALSE  19Y7bQtDs43oc96Vhaj6w9ou7dfaBVKHw <named list [… <named list [2]>
#>  8 name-squatter FALSE  1w2FUce0a1Esv2K9sEn4s8gPZE7cHTj4v <named list [… <named list [2]>
#>  9 name-squatter FALSE  18hzoJiJruyXqtE4zrd25MAe5Y5-Vi2xH <named list [… <named list [2]>
#> 10 name-squatter FALSE  1MX82LqNHapLLsfDx0UjlVIgvhIqZe8sr <named list [… <named list [2]>
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name          published id                  drive_resource   revision_resour…
#>    <chr>         <lgl>     <drv_id>            <list>           <list>          
#>  1 name-squatter FALSE     1W0LpT7f9_AZd8X4zi… <named list [39… <named list [9]>
#>  2 name-squatter FALSE     1agVDGb64Y6ssGW2mB… <named list [39… <named list [9]>
#>  3 name-squatter FALSE     1PBzxw_Ij-XMwnDM4m… <named list [39… <named list [9]>
#>  4 name-squatter FALSE     1xbsMhPsrsOn-GLv4Y… <named list [39… <named list [9]>
#>  5 name-squatter FALSE     1BtAxk5ZF99CEhH84v… <named list [39… <named list [9]>
#>  6 name-squatter FALSE     1Zmvd7_zUP8d17F8sA… <named list [39… <named list [9]>
#>  7 name-squatter FALSE     19Y7bQtDs43oc96Vha… <named list [39… <named list [9]>
#>  8 name-squatter FALSE     1w2FUce0a1Esv2K9sE… <named list [39… <named list [9]>
#>  9 name-squatter FALSE     18hzoJiJruyXqtE4zr… <named list [39… <named list [9]>
#> 10 name-squatter FALSE     1MX82LqNHapLLsfDx0… <named list [39… <named list [9]>

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name          id_parent           id                                drive_resource
#>    <chr>         <drv_id>            <drv_id>                          <list>        
#>  1 name-squatter 0AO_RMaBzcP63Uk9PVA 1W0LpT7f9_AZd8X4zi6ixnjv4MmtsuCjz <named list […
#>  2 name-squatter 0AO_RMaBzcP63Uk9PVA 1agVDGb64Y6ssGW2mBhUKyLAB2Diwqfpg <named list […
#>  3 name-squatter 0AO_RMaBzcP63Uk9PVA 1PBzxw_Ij-XMwnDM4moQ1SHJe7IpLZYzu <named list […
#>  4 name-squatter 0AO_RMaBzcP63Uk9PVA 1xbsMhPsrsOn-GLv4YDgmFJTfB0moPEGA <named list […
#>  5 name-squatter 0AO_RMaBzcP63Uk9PVA 1BtAxk5ZF99CEhH84vq2VxeE66Mgg8k6k <named list […
#>  6 name-squatter 0AO_RMaBzcP63Uk9PVA 1Zmvd7_zUP8d17F8sAYCprmWOT2oBoYGN <named list […
#>  7 name-squatter 0AO_RMaBzcP63Uk9PVA 19Y7bQtDs43oc96Vhaj6w9ou7dfaBVKHw <named list […
#>  8 name-squatter 0AO_RMaBzcP63Uk9PVA 1w2FUce0a1Esv2K9sEn4s8gPZE7cHTj4v <named list […
#>  9 name-squatter 0AO_RMaBzcP63Uk9PVA 18hzoJiJruyXqtE4zrd25MAe5Y5-Vi2xH <named list […
#> 10 name-squatter 0AO_RMaBzcP63Uk9PVA 1MX82LqNHapLLsfDx0UjlVIgvhIqZe8sr <named list […

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name          trashed id                                drive_resource   
#>    <chr>         <lgl>   <drv_id>                          <list>           
#>  1 name-squatter FALSE   1W0LpT7f9_AZd8X4zi6ixnjv4MmtsuCjz <named list [39]>
#>  2 name-squatter FALSE   1agVDGb64Y6ssGW2mBhUKyLAB2Diwqfpg <named list [39]>
#>  3 name-squatter FALSE   1PBzxw_Ij-XMwnDM4moQ1SHJe7IpLZYzu <named list [39]>
#>  4 name-squatter FALSE   1xbsMhPsrsOn-GLv4YDgmFJTfB0moPEGA <named list [39]>
#>  5 name-squatter FALSE   1BtAxk5ZF99CEhH84vq2VxeE66Mgg8k6k <named list [39]>
#>  6 name-squatter FALSE   1Zmvd7_zUP8d17F8sAYCprmWOT2oBoYGN <named list [39]>
#>  7 name-squatter FALSE   19Y7bQtDs43oc96Vhaj6w9ou7dfaBVKHw <named list [39]>
#>  8 name-squatter FALSE   1w2FUce0a1Esv2K9sEn4s8gPZE7cHTj4v <named list [39]>
#>  9 name-squatter FALSE   18hzoJiJruyXqtE4zrd25MAe5Y5-Vi2xH <named list [39]>
#> 10 name-squatter FALSE   1MX82LqNHapLLsfDx0UjlVIgvhIqZe8sr <named list [39]>
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name          mime_type                id                     drive_resource 
#>    <chr>         <chr>                    <drv_id>               <list>         
#>  1 name-squatter application/octet-stream 1W0LpT7f9_AZd8X4zi6ix… <named list [3…
#>  2 name-squatter application/octet-stream 1agVDGb64Y6ssGW2mBhUK… <named list [3…
#>  3 name-squatter application/octet-stream 1PBzxw_Ij-XMwnDM4moQ1… <named list [3…
#>  4 name-squatter application/octet-stream 1xbsMhPsrsOn-GLv4YDgm… <named list [3…
#>  5 name-squatter application/octet-stream 1BtAxk5ZF99CEhH84vq2V… <named list [3…
#>  6 name-squatter application/octet-stream 1Zmvd7_zUP8d17F8sAYCp… <named list [3…
#>  7 name-squatter application/octet-stream 19Y7bQtDs43oc96Vhaj6w… <named list [3…
#>  8 name-squatter application/octet-stream 1w2FUce0a1Esv2K9sEn4s… <named list [3…
#>  9 name-squatter application/octet-stream 18hzoJiJruyXqtE4zrd25… <named list [3…
#> 10 name-squatter application/octet-stream 1MX82LqNHapLLsfDx0Ujl… <named list [3…
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name          starred id                                drive_resource   
#>    <chr>         <lgl>   <drv_id>                          <list>           
#>  1 name-squatter FALSE   1W0LpT7f9_AZd8X4zi6ixnjv4MmtsuCjz <named list [39]>
#>  2 name-squatter FALSE   1agVDGb64Y6ssGW2mBhUKyLAB2Diwqfpg <named list [39]>
#>  3 name-squatter FALSE   1PBzxw_Ij-XMwnDM4moQ1SHJe7IpLZYzu <named list [39]>
#>  4 name-squatter FALSE   1xbsMhPsrsOn-GLv4YDgmFJTfB0moPEGA <named list [39]>
#>  5 name-squatter FALSE   1BtAxk5ZF99CEhH84vq2VxeE66Mgg8k6k <named list [39]>
#>  6 name-squatter FALSE   1Zmvd7_zUP8d17F8sAYCprmWOT2oBoYGN <named list [39]>
#>  7 name-squatter FALSE   19Y7bQtDs43oc96Vhaj6w9ou7dfaBVKHw <named list [39]>
#>  8 name-squatter FALSE   1w2FUce0a1Esv2K9sEn4s8gPZE7cHTj4v <named list [39]>
#>  9 name-squatter FALSE   18hzoJiJruyXqtE4zrd25MAe5Y5-Vi2xH <named list [39]>
#> 10 name-squatter FALSE   1MX82LqNHapLLsfDx0UjlVIgvhIqZe8sr <named list [39]>
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name          description id                                drive_resource   
#>    <chr>         <list>      <drv_id>                          <list>           
#>  1 name-squatter <NULL>      1W0LpT7f9_AZd8X4zi6ixnjv4MmtsuCjz <named list [39]>
#>  2 name-squatter <NULL>      1agVDGb64Y6ssGW2mBhUKyLAB2Diwqfpg <named list [39]>
#>  3 name-squatter <NULL>      1PBzxw_Ij-XMwnDM4moQ1SHJe7IpLZYzu <named list [39]>
#>  4 name-squatter <NULL>      1xbsMhPsrsOn-GLv4YDgmFJTfB0moPEGA <named list [39]>
#>  5 name-squatter <NULL>      1BtAxk5ZF99CEhH84vq2VxeE66Mgg8k6k <named list [39]>
#>  6 name-squatter <NULL>      1Zmvd7_zUP8d17F8sAYCprmWOT2oBoYGN <named list [39]>
#>  7 name-squatter <NULL>      19Y7bQtDs43oc96Vhaj6w9ou7dfaBVKHw <named list [39]>
#>  8 name-squatter <NULL>      1w2FUce0a1Esv2K9sEn4s8gPZE7cHTj4v <named list [39]>
#>  9 name-squatter <NULL>      18hzoJiJruyXqtE4zrd25MAe5Y5-Vi2xH <named list [39]>
#> 10 name-squatter <NULL>      1MX82LqNHapLLsfDx0UjlVIgvhIqZe8sr <named list [39]>
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name          version id                                drive_resource   
#>    <chr>         <chr>   <drv_id>                          <list>           
#>  1 name-squatter 1       1W0LpT7f9_AZd8X4zi6ixnjv4MmtsuCjz <named list [39]>
#>  2 name-squatter 1       1agVDGb64Y6ssGW2mBhUKyLAB2Diwqfpg <named list [39]>
#>  3 name-squatter 1       1PBzxw_Ij-XMwnDM4moQ1SHJe7IpLZYzu <named list [39]>
#>  4 name-squatter 1       1xbsMhPsrsOn-GLv4YDgmFJTfB0moPEGA <named list [39]>
#>  5 name-squatter 1       1BtAxk5ZF99CEhH84vq2VxeE66Mgg8k6k <named list [39]>
#>  6 name-squatter 1       1Zmvd7_zUP8d17F8sAYCprmWOT2oBoYGN <named list [39]>
#>  7 name-squatter 1       19Y7bQtDs43oc96Vhaj6w9ou7dfaBVKHw <named list [39]>
#>  8 name-squatter 1       1w2FUce0a1Esv2K9sEn4s8gPZE7cHTj4v <named list [39]>
#>  9 name-squatter 1       18hzoJiJruyXqtE4zrd25MAe5Y5-Vi2xH <named list [39]>
#> 10 name-squatter 1       1MX82LqNHapLLsfDx0UjlVIgvhIqZe8sr <named list [39]>
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name          web_view_link                     id             drive_resource
#>    <chr>         <chr>                             <drv_id>       <list>        
#>  1 name-squatter https://drive.google.com/file/d/… 1W0LpT7f9_AZd… <named list […
#>  2 name-squatter https://drive.google.com/file/d/… 1agVDGb64Y6ss… <named list […
#>  3 name-squatter https://drive.google.com/file/d/… 1PBzxw_Ij-XMw… <named list […
#>  4 name-squatter https://drive.google.com/file/d/… 1xbsMhPsrsOn-… <named list […
#>  5 name-squatter https://drive.google.com/file/d/… 1BtAxk5ZF99CE… <named list […
#>  6 name-squatter https://drive.google.com/file/d/… 1Zmvd7_zUP8d1… <named list […
#>  7 name-squatter https://drive.google.com/file/d/… 19Y7bQtDs43oc… <named list […
#>  8 name-squatter https://drive.google.com/file/d/… 1w2FUce0a1Esv… <named list […
#>  9 name-squatter https://drive.google.com/file/d/… 18hzoJiJruyXq… <named list […
#> 10 name-squatter https://drive.google.com/file/d/… 1MX82LqNHapLL… <named list […
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name          modified_time       id                                drive_resource
#>    <chr>         <dttm>              <drv_id>                          <list>        
#>  1 name-squatter 2021-10-18 13:43:46 1W0LpT7f9_AZd8X4zi6ixnjv4MmtsuCjz <named list […
#>  2 name-squatter 2021-10-17 13:42:08 1agVDGb64Y6ssGW2mBhUKyLAB2Diwqfpg <named list […
#>  3 name-squatter 2021-10-16 13:42:05 1PBzxw_Ij-XMwnDM4moQ1SHJe7IpLZYzu <named list […
#>  4 name-squatter 2021-10-15 13:43:20 1xbsMhPsrsOn-GLv4YDgmFJTfB0moPEGA <named list […
#>  5 name-squatter 2021-10-14 13:44:19 1BtAxk5ZF99CEhH84vq2VxeE66Mgg8k6k <named list […
#>  6 name-squatter 2021-10-13 13:48:57 1Zmvd7_zUP8d17F8sAYCprmWOT2oBoYGN <named list […
#>  7 name-squatter 2021-10-12 13:47:06 19Y7bQtDs43oc96Vhaj6w9ou7dfaBVKHw <named list […
#>  8 name-squatter 2021-10-11 13:45:19 1w2FUce0a1Esv2K9sEn4s8gPZE7cHTj4v <named list […
#>  9 name-squatter 2021-10-10 13:42:11 18hzoJiJruyXqtE4zrd25MAe5Y5-Vi2xH <named list […
#> 10 name-squatter 2021-10-09 13:42:00 1MX82LqNHapLLsfDx0UjlVIgvhIqZe8sr <named list […
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name          created_time        id                                drive_resource
#>    <chr>         <dttm>              <drv_id>                          <list>        
#>  1 name-squatter 2021-10-18 13:43:46 1W0LpT7f9_AZd8X4zi6ixnjv4MmtsuCjz <named list […
#>  2 name-squatter 2021-10-17 13:42:08 1agVDGb64Y6ssGW2mBhUKyLAB2Diwqfpg <named list […
#>  3 name-squatter 2021-10-16 13:42:05 1PBzxw_Ij-XMwnDM4moQ1SHJe7IpLZYzu <named list […
#>  4 name-squatter 2021-10-15 13:43:20 1xbsMhPsrsOn-GLv4YDgmFJTfB0moPEGA <named list […
#>  5 name-squatter 2021-10-14 13:44:19 1BtAxk5ZF99CEhH84vq2VxeE66Mgg8k6k <named list […
#>  6 name-squatter 2021-10-13 13:48:57 1Zmvd7_zUP8d17F8sAYCprmWOT2oBoYGN <named list […
#>  7 name-squatter 2021-10-12 13:47:06 19Y7bQtDs43oc96Vhaj6w9ou7dfaBVKHw <named list […
#>  8 name-squatter 2021-10-11 13:45:19 1w2FUce0a1Esv2K9sEn4s8gPZE7cHTj4v <named list […
#>  9 name-squatter 2021-10-10 13:42:11 18hzoJiJruyXqtE4zrd25MAe5Y5-Vi2xH <named list […
#> 10 name-squatter 2021-10-09 13:42:00 1MX82LqNHapLLsfDx0UjlVIgvhIqZe8sr <named list […
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name          owned_by_me id                                drive_resource   
#>    <chr>         <lgl>       <drv_id>                          <list>           
#>  1 name-squatter TRUE        1W0LpT7f9_AZd8X4zi6ixnjv4MmtsuCjz <named list [39]>
#>  2 name-squatter TRUE        1agVDGb64Y6ssGW2mBhUKyLAB2Diwqfpg <named list [39]>
#>  3 name-squatter TRUE        1PBzxw_Ij-XMwnDM4moQ1SHJe7IpLZYzu <named list [39]>
#>  4 name-squatter TRUE        1xbsMhPsrsOn-GLv4YDgmFJTfB0moPEGA <named list [39]>
#>  5 name-squatter TRUE        1BtAxk5ZF99CEhH84vq2VxeE66Mgg8k6k <named list [39]>
#>  6 name-squatter TRUE        1Zmvd7_zUP8d17F8sAYCprmWOT2oBoYGN <named list [39]>
#>  7 name-squatter TRUE        19Y7bQtDs43oc96Vhaj6w9ou7dfaBVKHw <named list [39]>
#>  8 name-squatter TRUE        1w2FUce0a1Esv2K9sEn4s8gPZE7cHTj4v <named list [39]>
#>  9 name-squatter TRUE        18hzoJiJruyXqtE4zrd25MAe5Y5-Vi2xH <named list [39]>
#> 10 name-squatter TRUE        1MX82LqNHapLLsfDx0UjlVIgvhIqZe8sr <named list [39]>
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name          size  id                                drive_resource   
#>    <chr>         <chr> <drv_id>                          <list>           
#>  1 name-squatter 0     1W0LpT7f9_AZd8X4zi6ixnjv4MmtsuCjz <named list [39]>
#>  2 name-squatter 0     1agVDGb64Y6ssGW2mBhUKyLAB2Diwqfpg <named list [39]>
#>  3 name-squatter 0     1PBzxw_Ij-XMwnDM4moQ1SHJe7IpLZYzu <named list [39]>
#>  4 name-squatter 0     1xbsMhPsrsOn-GLv4YDgmFJTfB0moPEGA <named list [39]>
#>  5 name-squatter 0     1BtAxk5ZF99CEhH84vq2VxeE66Mgg8k6k <named list [39]>
#>  6 name-squatter 0     1Zmvd7_zUP8d17F8sAYCprmWOT2oBoYGN <named list [39]>
#>  7 name-squatter 0     19Y7bQtDs43oc96Vhaj6w9ou7dfaBVKHw <named list [39]>
#>  8 name-squatter 0     1w2FUce0a1Esv2K9sEn4s8gPZE7cHTj4v <named list [39]>
#>  9 name-squatter 0     18hzoJiJruyXqtE4zrd25MAe5Y5-Vi2xH <named list [39]>
#> 10 name-squatter 0     1MX82LqNHapLLsfDx0UjlVIgvhIqZe8sr <named list [39]>
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name          quota_bytes_used id                                drive_resource
#>    <chr>         <chr>            <drv_id>                          <list>        
#>  1 name-squatter 0                1W0LpT7f9_AZd8X4zi6ixnjv4MmtsuCjz <named list […
#>  2 name-squatter 0                1agVDGb64Y6ssGW2mBhUKyLAB2Diwqfpg <named list […
#>  3 name-squatter 0                1PBzxw_Ij-XMwnDM4moQ1SHJe7IpLZYzu <named list […
#>  4 name-squatter 0                1xbsMhPsrsOn-GLv4YDgmFJTfB0moPEGA <named list […
#>  5 name-squatter 0                1BtAxk5ZF99CEhH84vq2VxeE66Mgg8k6k <named list […
#>  6 name-squatter 0                1Zmvd7_zUP8d17F8sAYCprmWOT2oBoYGN <named list […
#>  7 name-squatter 0                19Y7bQtDs43oc96Vhaj6w9ou7dfaBVKHw <named list […
#>  8 name-squatter 0                1w2FUce0a1Esv2K9sEn4s8gPZE7cHTj4v <named list […
#>  9 name-squatter 0                18hzoJiJruyXqtE4zrd25MAe5Y5-Vi2xH <named list […
#> 10 name-squatter 0                1MX82LqNHapLLsfDx0UjlVIgvhIqZe8sr <named list […

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>