Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

Usage

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what

Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via whatappears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name                 path                   id           drive_resource
#>    <chr>                <chr>                  <drv_id>     <list>        
#>  1 name-squatter-rename ~/name-squatter-rename 194V4rMcHO7… <named list>  
#>  2 name-squatter-mv     ~/name-squatter-mv     1RCaKYb_YNm… <named list>  
#>  3 name-squatter-upload ~/name-squatter-upload 1sNZYENlS7K… <named list>  
#>  4 name-squatter-rename ~/name-squatter-rename 1kK656KyPDW… <named list>  
#>  5 name-squatter-mv     ~/name-squatter-mv     1cZSoU06EPk… <named list>  
#>  6 name-squatter-upload ~/name-squatter-upload 1XaPwd1J3wJ… <named list>  
#>  7 name-squatter-rename ~/name-squatter-rename 1bJO9Vb1Lw2… <named list>  
#>  8 name-squatter-mv     ~/name-squatter-mv     1eg67RpFsMF… <named list>  
#>  9 name-squatter-upload ~/name-squatter-upload 1ENrBCnmeGP… <named list>  
#> 10 name-squatter-rename ~/name-squatter-rename 1YgZvlRvw_P… <named list>  
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name                 shared id          drive_resource permissions_res…
#>    <chr>                <lgl>  <drv_id>    <list>         <list>          
#>  1 name-squatter-rename FALSE  194V4rMcHO… <named list>   <named list [2]>
#>  2 name-squatter-mv     FALSE  1RCaKYb_YN… <named list>   <named list [2]>
#>  3 name-squatter-upload FALSE  1sNZYENlS7… <named list>   <named list [2]>
#>  4 name-squatter-rename FALSE  1kK656KyPD… <named list>   <named list [2]>
#>  5 name-squatter-mv     FALSE  1cZSoU06EP… <named list>   <named list [2]>
#>  6 name-squatter-upload FALSE  1XaPwd1J3w… <named list>   <named list [2]>
#>  7 name-squatter-rename FALSE  1bJO9Vb1Lw… <named list>   <named list [2]>
#>  8 name-squatter-mv     FALSE  1eg67RpFsM… <named list>   <named list [2]>
#>  9 name-squatter-upload FALSE  1ENrBCnmeG… <named list>   <named list [2]>
#> 10 name-squatter-rename FALSE  1YgZvlRvw_… <named list>   <named list [2]>
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name                 published id       drive_resource revision_resour…
#>    <chr>                <lgl>     <drv_id> <list>         <list>          
#>  1 name-squatter-rename FALSE     194V4rM… <named list>   <named list [9]>
#>  2 name-squatter-mv     FALSE     1RCaKYb… <named list>   <named list [9]>
#>  3 name-squatter-upload FALSE     1sNZYEN… <named list>   <named list [9]>
#>  4 name-squatter-rename FALSE     1kK656K… <named list>   <named list [9]>
#>  5 name-squatter-mv     FALSE     1cZSoU0… <named list>   <named list [9]>
#>  6 name-squatter-upload FALSE     1XaPwd1… <named list>   <named list [9]>
#>  7 name-squatter-rename FALSE     1bJO9Vb… <named list>   <named list [9]>
#>  8 name-squatter-mv     FALSE     1eg67Rp… <named list>   <named list [9]>
#>  9 name-squatter-upload FALSE     1ENrBCn… <named list>   <named list [9]>
#> 10 name-squatter-rename FALSE     1YgZvlR… <named list>   <named list [9]>

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name                 id_parent           id              drive_resource
#>    <chr>                <drv_id>            <drv_id>        <list>        
#>  1 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 194V4rMcHO7vAh… <named list>  
#>  2 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1RCaKYb_YNmILK… <named list>  
#>  3 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1sNZYENlS7Km-N… <named list>  
#>  4 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1kK656KyPDWfb7… <named list>  
#>  5 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1cZSoU06EPkULW… <named list>  
#>  6 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1XaPwd1J3wJpYC… <named list>  
#>  7 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1bJO9Vb1Lw2toJ… <named list>  
#>  8 name-squatter-mv     0AO_RMaBzcP63Uk9PVA 1eg67RpFsMFJn1… <named list>  
#>  9 name-squatter-upload 0AO_RMaBzcP63Uk9PVA 1ENrBCnmeGP-zb… <named list>  
#> 10 name-squatter-rename 0AO_RMaBzcP63Uk9PVA 1YgZvlRvw_POs-… <named list>  

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name                 trashed id                          drive_resource
#>    <chr>                <lgl>   <drv_id>                    <list>        
#>  1 name-squatter-rename TRUE    194V4rMcHO7vAhpk2Q1vHTVJBR… <named list>  
#>  2 name-squatter-mv     TRUE    1RCaKYb_YNmILKY6VBzb4E7mwW… <named list>  
#>  3 name-squatter-upload TRUE    1sNZYENlS7Km-NyVNt4wRcojzb… <named list>  
#>  4 name-squatter-rename TRUE    1kK656KyPDWfb7-zfPpTlRVx6b… <named list>  
#>  5 name-squatter-mv     TRUE    1cZSoU06EPkULWTOYjVcaWiy_F… <named list>  
#>  6 name-squatter-upload TRUE    1XaPwd1J3wJpYC3dcQEHjiHhte… <named list>  
#>  7 name-squatter-rename TRUE    1bJO9Vb1Lw2toJCrEiqy4zpp7F… <named list>  
#>  8 name-squatter-mv     TRUE    1eg67RpFsMFJn15qvnr5PuVAzU… <named list>  
#>  9 name-squatter-upload TRUE    1ENrBCnmeGP-zb6PFtCW3hX_FV… <named list>  
#> 10 name-squatter-rename TRUE    1YgZvlRvw_POs-8-jtFrFALK_t… <named list>  
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name                 mime_type                id         drive_resource
#>    <chr>                <chr>                    <drv_id>   <list>        
#>  1 name-squatter-rename application/octet-stream 194V4rMcH… <named list>  
#>  2 name-squatter-mv     application/octet-stream 1RCaKYb_Y… <named list>  
#>  3 name-squatter-upload application/octet-stream 1sNZYENlS… <named list>  
#>  4 name-squatter-rename application/octet-stream 1kK656KyP… <named list>  
#>  5 name-squatter-mv     application/octet-stream 1cZSoU06E… <named list>  
#>  6 name-squatter-upload application/octet-stream 1XaPwd1J3… <named list>  
#>  7 name-squatter-rename application/octet-stream 1bJO9Vb1L… <named list>  
#>  8 name-squatter-mv     application/octet-stream 1eg67RpFs… <named list>  
#>  9 name-squatter-upload application/octet-stream 1ENrBCnme… <named list>  
#> 10 name-squatter-rename application/octet-stream 1YgZvlRvw… <named list>  
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name                 starred id                          drive_resource
#>    <chr>                <lgl>   <drv_id>                    <list>        
#>  1 name-squatter-rename FALSE   194V4rMcHO7vAhpk2Q1vHTVJBR… <named list>  
#>  2 name-squatter-mv     FALSE   1RCaKYb_YNmILKY6VBzb4E7mwW… <named list>  
#>  3 name-squatter-upload FALSE   1sNZYENlS7Km-NyVNt4wRcojzb… <named list>  
#>  4 name-squatter-rename FALSE   1kK656KyPDWfb7-zfPpTlRVx6b… <named list>  
#>  5 name-squatter-mv     FALSE   1cZSoU06EPkULWTOYjVcaWiy_F… <named list>  
#>  6 name-squatter-upload FALSE   1XaPwd1J3wJpYC3dcQEHjiHhte… <named list>  
#>  7 name-squatter-rename FALSE   1bJO9Vb1Lw2toJCrEiqy4zpp7F… <named list>  
#>  8 name-squatter-mv     FALSE   1eg67RpFsMFJn15qvnr5PuVAzU… <named list>  
#>  9 name-squatter-upload FALSE   1ENrBCnmeGP-zb6PFtCW3hX_FV… <named list>  
#> 10 name-squatter-rename FALSE   1YgZvlRvw_POs-8-jtFrFALK_t… <named list>  
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name                 description id                      drive_resource
#>    <chr>                <list>      <drv_id>                <list>        
#>  1 name-squatter-rename <NULL>      194V4rMcHO7vAhpk2Q1vHT… <named list>  
#>  2 name-squatter-mv     <NULL>      1RCaKYb_YNmILKY6VBzb4E… <named list>  
#>  3 name-squatter-upload <NULL>      1sNZYENlS7Km-NyVNt4wRc… <named list>  
#>  4 name-squatter-rename <NULL>      1kK656KyPDWfb7-zfPpTlR… <named list>  
#>  5 name-squatter-mv     <NULL>      1cZSoU06EPkULWTOYjVcaW… <named list>  
#>  6 name-squatter-upload <NULL>      1XaPwd1J3wJpYC3dcQEHji… <named list>  
#>  7 name-squatter-rename <NULL>      1bJO9Vb1Lw2toJCrEiqy4z… <named list>  
#>  8 name-squatter-mv     <NULL>      1eg67RpFsMFJn15qvnr5Pu… <named list>  
#>  9 name-squatter-upload <NULL>      1ENrBCnmeGP-zb6PFtCW3h… <named list>  
#> 10 name-squatter-rename <NULL>      1YgZvlRvw_POs-8-jtFrFA… <named list>  
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name                 version id                          drive_resource
#>    <chr>                <chr>   <drv_id>                    <list>        
#>  1 name-squatter-rename 2       194V4rMcHO7vAhpk2Q1vHTVJBR… <named list>  
#>  2 name-squatter-mv     2       1RCaKYb_YNmILKY6VBzb4E7mwW… <named list>  
#>  3 name-squatter-upload 2       1sNZYENlS7Km-NyVNt4wRcojzb… <named list>  
#>  4 name-squatter-rename 2       1kK656KyPDWfb7-zfPpTlRVx6b… <named list>  
#>  5 name-squatter-mv     2       1cZSoU06EPkULWTOYjVcaWiy_F… <named list>  
#>  6 name-squatter-upload 2       1XaPwd1J3wJpYC3dcQEHjiHhte… <named list>  
#>  7 name-squatter-rename 2       1bJO9Vb1Lw2toJCrEiqy4zpp7F… <named list>  
#>  8 name-squatter-mv     2       1eg67RpFsMFJn15qvnr5PuVAzU… <named list>  
#>  9 name-squatter-upload 2       1ENrBCnmeGP-zb6PFtCW3hX_FV… <named list>  
#> 10 name-squatter-rename 2       1YgZvlRvw_POs-8-jtFrFALK_t… <named list>  
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name                 web_view_link                 id    drive_resource
#>    <chr>                <chr>                         <drv> <list>        
#>  1 name-squatter-rename https://drive.google.com/fil… 194V… <named list>  
#>  2 name-squatter-mv     https://drive.google.com/fil… 1RCa… <named list>  
#>  3 name-squatter-upload https://drive.google.com/fil… 1sNZ… <named list>  
#>  4 name-squatter-rename https://drive.google.com/fil… 1kK6… <named list>  
#>  5 name-squatter-mv     https://drive.google.com/fil… 1cZS… <named list>  
#>  6 name-squatter-upload https://drive.google.com/fil… 1XaP… <named list>  
#>  7 name-squatter-rename https://drive.google.com/fil… 1bJO… <named list>  
#>  8 name-squatter-mv     https://drive.google.com/fil… 1eg6… <named list>  
#>  9 name-squatter-upload https://drive.google.com/fil… 1ENr… <named list>  
#> 10 name-squatter-rename https://drive.google.com/fil… 1YgZ… <named list>  
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name                 modified_time       id              drive_resource
#>    <chr>                <dttm>              <drv_id>        <list>        
#>  1 name-squatter-rename 2022-05-04 13:58:50 194V4rMcHO7vAh… <named list>  
#>  2 name-squatter-mv     2022-05-04 13:58:33 1RCaKYb_YNmILK… <named list>  
#>  3 name-squatter-upload 2022-05-03 14:03:53 1sNZYENlS7Km-N… <named list>  
#>  4 name-squatter-rename 2022-05-03 14:03:16 1kK656KyPDWfb7… <named list>  
#>  5 name-squatter-mv     2022-05-03 14:02:59 1cZSoU06EPkULW… <named list>  
#>  6 name-squatter-upload 2022-05-02 14:03:28 1XaPwd1J3wJpYC… <named list>  
#>  7 name-squatter-rename 2022-05-02 14:02:55 1bJO9Vb1Lw2toJ… <named list>  
#>  8 name-squatter-mv     2022-05-02 14:02:40 1eg67RpFsMFJn1… <named list>  
#>  9 name-squatter-upload 2022-05-01 13:48:09 1ENrBCnmeGP-zb… <named list>  
#> 10 name-squatter-rename 2022-05-01 13:47:32 1YgZvlRvw_POs-… <named list>  
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name                 created_time        id              drive_resource
#>    <chr>                <dttm>              <drv_id>        <list>        
#>  1 name-squatter-rename 2022-05-04 13:58:50 194V4rMcHO7vAh… <named list>  
#>  2 name-squatter-mv     2022-05-04 13:58:33 1RCaKYb_YNmILK… <named list>  
#>  3 name-squatter-upload 2022-05-03 14:03:53 1sNZYENlS7Km-N… <named list>  
#>  4 name-squatter-rename 2022-05-03 14:03:16 1kK656KyPDWfb7… <named list>  
#>  5 name-squatter-mv     2022-05-03 14:02:59 1cZSoU06EPkULW… <named list>  
#>  6 name-squatter-upload 2022-05-02 14:03:28 1XaPwd1J3wJpYC… <named list>  
#>  7 name-squatter-rename 2022-05-02 14:02:55 1bJO9Vb1Lw2toJ… <named list>  
#>  8 name-squatter-mv     2022-05-02 14:02:40 1eg67RpFsMFJn1… <named list>  
#>  9 name-squatter-upload 2022-05-01 13:48:09 1ENrBCnmeGP-zb… <named list>  
#> 10 name-squatter-rename 2022-05-01 13:47:32 1YgZvlRvw_POs-… <named list>  
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name                 owned_by_me id                      drive_resource
#>    <chr>                <lgl>       <drv_id>                <list>        
#>  1 name-squatter-rename TRUE        194V4rMcHO7vAhpk2Q1vHT… <named list>  
#>  2 name-squatter-mv     TRUE        1RCaKYb_YNmILKY6VBzb4E… <named list>  
#>  3 name-squatter-upload TRUE        1sNZYENlS7Km-NyVNt4wRc… <named list>  
#>  4 name-squatter-rename TRUE        1kK656KyPDWfb7-zfPpTlR… <named list>  
#>  5 name-squatter-mv     TRUE        1cZSoU06EPkULWTOYjVcaW… <named list>  
#>  6 name-squatter-upload TRUE        1XaPwd1J3wJpYC3dcQEHji… <named list>  
#>  7 name-squatter-rename TRUE        1bJO9Vb1Lw2toJCrEiqy4z… <named list>  
#>  8 name-squatter-mv     TRUE        1eg67RpFsMFJn15qvnr5Pu… <named list>  
#>  9 name-squatter-upload TRUE        1ENrBCnmeGP-zb6PFtCW3h… <named list>  
#> 10 name-squatter-rename TRUE        1YgZvlRvw_POs-8-jtFrFA… <named list>  
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name                 size  id                            drive_resource
#>    <chr>                <chr> <drv_id>                      <list>        
#>  1 name-squatter-rename 0     194V4rMcHO7vAhpk2Q1vHTVJBRS5… <named list>  
#>  2 name-squatter-mv     0     1RCaKYb_YNmILKY6VBzb4E7mwWNT… <named list>  
#>  3 name-squatter-upload 0     1sNZYENlS7Km-NyVNt4wRcojzbG4… <named list>  
#>  4 name-squatter-rename 0     1kK656KyPDWfb7-zfPpTlRVx6bE1… <named list>  
#>  5 name-squatter-mv     0     1cZSoU06EPkULWTOYjVcaWiy_FhQ… <named list>  
#>  6 name-squatter-upload 0     1XaPwd1J3wJpYC3dcQEHjiHhtegQ… <named list>  
#>  7 name-squatter-rename 0     1bJO9Vb1Lw2toJCrEiqy4zpp7F-H… <named list>  
#>  8 name-squatter-mv     0     1eg67RpFsMFJn15qvnr5PuVAzUUe… <named list>  
#>  9 name-squatter-upload 0     1ENrBCnmeGP-zb6PFtCW3hX_FVjw… <named list>  
#> 10 name-squatter-rename 0     1YgZvlRvw_POs-8-jtFrFALK_tzm… <named list>  
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name                 quota_bytes_used id                 drive_resource
#>    <chr>                <chr>            <drv_id>           <list>        
#>  1 name-squatter-rename 0                194V4rMcHO7vAhpk2… <named list>  
#>  2 name-squatter-mv     0                1RCaKYb_YNmILKY6V… <named list>  
#>  3 name-squatter-upload 0                1sNZYENlS7Km-NyVN… <named list>  
#>  4 name-squatter-rename 0                1kK656KyPDWfb7-zf… <named list>  
#>  5 name-squatter-mv     0                1cZSoU06EPkULWTOY… <named list>  
#>  6 name-squatter-upload 0                1XaPwd1J3wJpYC3dc… <named list>  
#>  7 name-squatter-rename 0                1bJO9Vb1Lw2toJCrE… <named list>  
#>  8 name-squatter-mv     0                1eg67RpFsMFJn15qv… <named list>  
#>  9 name-squatter-upload 0                1ENrBCnmeGP-zb6PF… <named list>  
#> 10 name-squatter-rename 0                1YgZvlRvw_POs-8-j… <named list>  

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>