Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

Usage

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what

Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via what

appears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name                 path                   id             drive_reso…¹
#>    <chr>                <chr>                  <drv_id>       <list>      
#>  1 name-squatter-rename ~/name-squatter-rename 1jyxeBlygR8dn… <named list>
#>  2 name-squatter-mv     ~/name-squatter-mv     1h6NkhvM7tM9-… <named list>
#>  3 Lotus 1-2-3          ~/b4xl/Lotus 1-2-3     1W5X4MCJBIJxd… <named list>
#>  4 SuperCalc            ~/b4xl/SuperCalc       1ctjoD2HX05ZE… <named list>
#>  5 ExecuVision          ~/ExecuVision          1TjcZYBVDWN9b… <named list>
#>  6 WordStar             ~/WordStar             187ajd08ITKeH… <named list>
#>  7 name-squatter-upload ~/name-squatter-upload 15fJL62lFDPdZ… <named list>
#>  8 name-squatter-rename ~/name-squatter-rename 1ifZxrlH9FLcS… <named list>
#>  9 name-squatter-mv     ~/name-squatter-mv     1RhGFXAin-dYo… <named list>
#> 10 Lotus 1-2-3          ~/b4xl/Lotus 1-2-3     1h6pIFFxuCvlR… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name                 shared id                drive_reso…¹ permission…²
#>    <chr>                <lgl>  <drv_id>          <list>       <list>      
#>  1 name-squatter-rename FALSE  1jyxeBlygR8dnhaV… <named list> <named list>
#>  2 name-squatter-mv     FALSE  1h6NkhvM7tM9-sf7… <named list> <named list>
#>  3 Lotus 1-2-3          FALSE  1W5X4MCJBIJxdvQ-… <named list> <named list>
#>  4 SuperCalc            FALSE  1ctjoD2HX05ZEu0b… <named list> <named list>
#>  5 ExecuVision          FALSE  1TjcZYBVDWN9bIUt… <named list> <named list>
#>  6 WordStar             FALSE  187ajd08ITKeHbZN… <named list> <named list>
#>  7 name-squatter-upload FALSE  15fJL62lFDPdZqra… <named list> <named list>
#>  8 name-squatter-rename FALSE  1ifZxrlH9FLcSXZt… <named list> <named list>
#>  9 name-squatter-mv     FALSE  1RhGFXAin-dYoKh-… <named list> <named list>
#> 10 Lotus 1-2-3          FALSE  1h6pIFFxuCvlRApQ… <named list> <named list>
#> # … with abbreviated variable names ¹​drive_resource,
#> #   ²​permissions_resource
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name                 published id             drive_reso…¹ revision_r…²
#>    <chr>                <lgl>     <drv_id>       <list>       <list>      
#>  1 name-squatter-rename FALSE     1jyxeBlygR8dn… <named list> <named list>
#>  2 name-squatter-mv     FALSE     1h6NkhvM7tM9-… <named list> <named list>
#>  3 Lotus 1-2-3          FALSE     1W5X4MCJBIJxd… <named list> <named list>
#>  4 SuperCalc            FALSE     1ctjoD2HX05ZE… <named list> <named list>
#>  5 ExecuVision          FALSE     1TjcZYBVDWN9b… <named list> <named list>
#>  6 WordStar             FALSE     187ajd08ITKeH… <named list> <named list>
#>  7 name-squatter-upload FALSE     15fJL62lFDPdZ… <named list> <named list>
#>  8 name-squatter-rename FALSE     1ifZxrlH9FLcS… <named list> <named list>
#>  9 name-squatter-mv     FALSE     1RhGFXAin-dYo… <named list> <named list>
#> 10 Lotus 1-2-3          FALSE     1h6pIFFxuCvlR… <named list> <named list>
#> # … with abbreviated variable names ¹​drive_resource, ²​revision_resource

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name                 id_parent                       id    drive_reso…¹
#>    <chr>                <drv_id>                        <drv> <list>      
#>  1 name-squatter-rename 0AO_RMaBzcP63Uk9PVA             1jyx… <named list>
#>  2 name-squatter-mv     0AO_RMaBzcP63Uk9PVA             1h6N… <named list>
#>  3 Lotus 1-2-3          1S0iWntXFsr7UqMZUnzVm1j7QLssl_… 1W5X… <named list>
#>  4 SuperCalc            1S0iWntXFsr7UqMZUnzVm1j7QLssl_… 1ctj… <named list>
#>  5 ExecuVision          0AO_RMaBzcP63Uk9PVA             1Tjc… <named list>
#>  6 WordStar             0AO_RMaBzcP63Uk9PVA             187a… <named list>
#>  7 name-squatter-upload 0AO_RMaBzcP63Uk9PVA             15fJ… <named list>
#>  8 name-squatter-rename 0AO_RMaBzcP63Uk9PVA             1ifZ… <named list>
#>  9 name-squatter-mv     0AO_RMaBzcP63Uk9PVA             1RhG… <named list>
#> 10 Lotus 1-2-3          1S0iWntXFsr7UqMZUnzVm1j7QLssl_… 1h6p… <named list>
#> # … with abbreviated variable name ¹​drive_resource

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name                 trashed id                            drive_reso…¹
#>    <chr>                <lgl>   <drv_id>                      <list>      
#>  1 name-squatter-rename TRUE    1jyxeBlygR8dnhaVSVpTGAOy7yJx… <named list>
#>  2 name-squatter-mv     TRUE    1h6NkhvM7tM9-sf7anooG1TRRIZq… <named list>
#>  3 Lotus 1-2-3          FALSE   1W5X4MCJBIJxdvQ-a-pNuU4khW14… <named list>
#>  4 SuperCalc            FALSE   1ctjoD2HX05ZEu0bkNxvVWMnvhe0… <named list>
#>  5 ExecuVision          FALSE   1TjcZYBVDWN9bIUtSUtP_PMOUqgZ… <named list>
#>  6 WordStar             FALSE   187ajd08ITKeHbZNgGZeSbRcTfjs… <named list>
#>  7 name-squatter-upload TRUE    15fJL62lFDPdZqra6cOAKbPSQk_o… <named list>
#>  8 name-squatter-rename TRUE    1ifZxrlH9FLcSXZtwc2fGScVV4oV… <named list>
#>  9 name-squatter-mv     TRUE    1RhGFXAin-dYoKh-48Yj60RD_K-a… <named list>
#> 10 Lotus 1-2-3          FALSE   1h6pIFFxuCvlRApQ57snAAyrkklp… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name                 mime_type                       id    drive_reso…¹
#>    <chr>                <chr>                           <drv> <list>      
#>  1 name-squatter-rename application/octet-stream        1jyx… <named list>
#>  2 name-squatter-mv     application/octet-stream        1h6N… <named list>
#>  3 Lotus 1-2-3          application/vnd.google-apps.sp… 1W5X… <named list>
#>  4 SuperCalc            application/vnd.google-apps.sp… 1ctj… <named list>
#>  5 ExecuVision          application/vnd.google-apps.pr… 1Tjc… <named list>
#>  6 WordStar             application/vnd.google-apps.do… 187a… <named list>
#>  7 name-squatter-upload application/octet-stream        15fJ… <named list>
#>  8 name-squatter-rename application/octet-stream        1ifZ… <named list>
#>  9 name-squatter-mv     application/octet-stream        1RhG… <named list>
#> 10 Lotus 1-2-3          application/vnd.google-apps.sp… 1h6p… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name                 starred id                            drive_reso…¹
#>    <chr>                <lgl>   <drv_id>                      <list>      
#>  1 name-squatter-rename FALSE   1jyxeBlygR8dnhaVSVpTGAOy7yJx… <named list>
#>  2 name-squatter-mv     FALSE   1h6NkhvM7tM9-sf7anooG1TRRIZq… <named list>
#>  3 Lotus 1-2-3          FALSE   1W5X4MCJBIJxdvQ-a-pNuU4khW14… <named list>
#>  4 SuperCalc            FALSE   1ctjoD2HX05ZEu0bkNxvVWMnvhe0… <named list>
#>  5 ExecuVision          FALSE   1TjcZYBVDWN9bIUtSUtP_PMOUqgZ… <named list>
#>  6 WordStar             TRUE    187ajd08ITKeHbZNgGZeSbRcTfjs… <named list>
#>  7 name-squatter-upload FALSE   15fJL62lFDPdZqra6cOAKbPSQk_o… <named list>
#>  8 name-squatter-rename FALSE   1ifZxrlH9FLcSXZtwc2fGScVV4oV… <named list>
#>  9 name-squatter-mv     FALSE   1RhGFXAin-dYoKh-48Yj60RD_K-a… <named list>
#> 10 Lotus 1-2-3          FALSE   1h6pIFFxuCvlRApQ57snAAyrkklp… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name                 description                    id     drive_reso…¹
#>    <chr>                <chr>                          <drv_> <list>      
#>  1 name-squatter-rename NA                             1jyxe… <named list>
#>  2 name-squatter-mv     NA                             1h6Nk… <named list>
#>  3 Lotus 1-2-3          NA                             1W5X4… <named list>
#>  4 SuperCalc            NA                             1ctjo… <named list>
#>  5 ExecuVision          deeply nested bullet lists FTW 1TjcZ… <named list>
#>  6 WordStar             NA                             187aj… <named list>
#>  7 name-squatter-upload NA                             15fJL… <named list>
#>  8 name-squatter-rename NA                             1ifZx… <named list>
#>  9 name-squatter-mv     NA                             1RhGF… <named list>
#> 10 Lotus 1-2-3          NA                             1h6pI… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name                 version id                            drive_reso…¹
#>    <chr>                <chr>   <drv_id>                      <list>      
#>  1 name-squatter-rename 2       1jyxeBlygR8dnhaVSVpTGAOy7yJx… <named list>
#>  2 name-squatter-mv     2       1h6NkhvM7tM9-sf7anooG1TRRIZq… <named list>
#>  3 Lotus 1-2-3          4       1W5X4MCJBIJxdvQ-a-pNuU4khW14… <named list>
#>  4 SuperCalc            4       1ctjoD2HX05ZEu0bkNxvVWMnvhe0… <named list>
#>  5 ExecuVision          3       1TjcZYBVDWN9bIUtSUtP_PMOUqgZ… <named list>
#>  6 WordStar             3       187ajd08ITKeHbZNgGZeSbRcTfjs… <named list>
#>  7 name-squatter-upload 2       15fJL62lFDPdZqra6cOAKbPSQk_o… <named list>
#>  8 name-squatter-rename 2       1ifZxrlH9FLcSXZtwc2fGScVV4oV… <named list>
#>  9 name-squatter-mv     2       1RhGFXAin-dYoKh-48Yj60RD_K-a… <named list>
#> 10 Lotus 1-2-3          4       1h6pIFFxuCvlRApQ57snAAyrkklp… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name                 web_view_link                   id    drive_reso…¹
#>    <chr>                <chr>                           <drv> <list>      
#>  1 name-squatter-rename https://drive.google.com/file/… 1jyx… <named list>
#>  2 name-squatter-mv     https://drive.google.com/file/… 1h6N… <named list>
#>  3 Lotus 1-2-3          https://docs.google.com/spread… 1W5X… <named list>
#>  4 SuperCalc            https://docs.google.com/spread… 1ctj… <named list>
#>  5 ExecuVision          https://docs.google.com/presen… 1Tjc… <named list>
#>  6 WordStar             https://docs.google.com/docume… 187a… <named list>
#>  7 name-squatter-upload https://drive.google.com/file/… 15fJ… <named list>
#>  8 name-squatter-rename https://drive.google.com/file/… 1ifZ… <named list>
#>  9 name-squatter-mv     https://drive.google.com/file/… 1RhG… <named list>
#> 10 Lotus 1-2-3          https://docs.google.com/spread… 1h6p… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name                 modified_time       id                drive_reso…¹
#>    <chr>                <dttm>              <drv_id>          <list>      
#>  1 name-squatter-rename 2022-08-15 14:03:12 1jyxeBlygR8dnhaV… <named list>
#>  2 name-squatter-mv     2022-08-15 14:02:46 1h6NkhvM7tM9-sf7… <named list>
#>  3 Lotus 1-2-3          2022-08-15 14:02:12 1W5X4MCJBIJxdvQ-… <named list>
#>  4 SuperCalc            2022-08-15 14:02:08 1ctjoD2HX05ZEu0b… <named list>
#>  5 ExecuVision          2022-08-15 14:02:05 1TjcZYBVDWN9bIUt… <named list>
#>  6 WordStar             2022-08-15 14:02:03 187ajd08ITKeHbZN… <named list>
#>  7 name-squatter-upload 2022-08-14 13:50:47 15fJL62lFDPdZqra… <named list>
#>  8 name-squatter-rename 2022-08-14 13:50:10 1ifZxrlH9FLcSXZt… <named list>
#>  9 name-squatter-mv     2022-08-14 13:49:52 1RhGFXAin-dYoKh-… <named list>
#> 10 Lotus 1-2-3          2022-08-14 13:49:25 1h6pIFFxuCvlRApQ… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name                 created_time        id                drive_reso…¹
#>    <chr>                <dttm>              <drv_id>          <list>      
#>  1 name-squatter-rename 2022-08-15 14:03:12 1jyxeBlygR8dnhaV… <named list>
#>  2 name-squatter-mv     2022-08-15 14:02:46 1h6NkhvM7tM9-sf7… <named list>
#>  3 Lotus 1-2-3          2022-08-15 14:02:12 1W5X4MCJBIJxdvQ-… <named list>
#>  4 SuperCalc            2022-08-15 14:02:08 1ctjoD2HX05ZEu0b… <named list>
#>  5 ExecuVision          2022-08-15 14:02:05 1TjcZYBVDWN9bIUt… <named list>
#>  6 WordStar             2022-08-15 14:02:02 187ajd08ITKeHbZN… <named list>
#>  7 name-squatter-upload 2022-08-14 13:50:47 15fJL62lFDPdZqra… <named list>
#>  8 name-squatter-rename 2022-08-14 13:50:10 1ifZxrlH9FLcSXZt… <named list>
#>  9 name-squatter-mv     2022-08-14 13:49:52 1RhGFXAin-dYoKh-… <named list>
#> 10 Lotus 1-2-3          2022-08-14 13:49:25 1h6pIFFxuCvlRApQ… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name                 owned_by_me id                        drive_reso…¹
#>    <chr>                <lgl>       <drv_id>                  <list>      
#>  1 name-squatter-rename TRUE        1jyxeBlygR8dnhaVSVpTGAOy… <named list>
#>  2 name-squatter-mv     TRUE        1h6NkhvM7tM9-sf7anooG1TR… <named list>
#>  3 Lotus 1-2-3          TRUE        1W5X4MCJBIJxdvQ-a-pNuU4k… <named list>
#>  4 SuperCalc            TRUE        1ctjoD2HX05ZEu0bkNxvVWMn… <named list>
#>  5 ExecuVision          TRUE        1TjcZYBVDWN9bIUtSUtP_PMO… <named list>
#>  6 WordStar             TRUE        187ajd08ITKeHbZNgGZeSbRc… <named list>
#>  7 name-squatter-upload TRUE        15fJL62lFDPdZqra6cOAKbPS… <named list>
#>  8 name-squatter-rename TRUE        1ifZxrlH9FLcSXZtwc2fGScV… <named list>
#>  9 name-squatter-mv     TRUE        1RhGFXAin-dYoKh-48Yj60RD… <named list>
#> 10 Lotus 1-2-3          TRUE        1h6pIFFxuCvlRApQ57snAAyr… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name                 size  id                              drive_reso…¹
#>    <chr>                <chr> <drv_id>                        <list>      
#>  1 name-squatter-rename 0     1jyxeBlygR8dnhaVSVpTGAOy7yJxWJ… <named list>
#>  2 name-squatter-mv     0     1h6NkhvM7tM9-sf7anooG1TRRIZqBC… <named list>
#>  3 Lotus 1-2-3          1024  1W5X4MCJBIJxdvQ-a-pNuU4khW14LZ… <named list>
#>  4 SuperCalc            1024  1ctjoD2HX05ZEu0bkNxvVWMnvhe0ty… <named list>
#>  5 ExecuVision          3483  1TjcZYBVDWN9bIUtSUtP_PMOUqgZdo… <named list>
#>  6 WordStar             1024  187ajd08ITKeHbZNgGZeSbRcTfjsn-… <named list>
#>  7 name-squatter-upload 0     15fJL62lFDPdZqra6cOAKbPSQk_oW6… <named list>
#>  8 name-squatter-rename 0     1ifZxrlH9FLcSXZtwc2fGScVV4oVSZ… <named list>
#>  9 name-squatter-mv     0     1RhGFXAin-dYoKh-48Yj60RD_K-aqt… <named list>
#> 10 Lotus 1-2-3          1024  1h6pIFFxuCvlRApQ57snAAyrkklp4n… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name                 quota_bytes_used id                   drive_reso…¹
#>    <chr>                <chr>            <drv_id>             <list>      
#>  1 name-squatter-rename 0                1jyxeBlygR8dnhaVSVp… <named list>
#>  2 name-squatter-mv     0                1h6NkhvM7tM9-sf7ano… <named list>
#>  3 Lotus 1-2-3          1024             1W5X4MCJBIJxdvQ-a-p… <named list>
#>  4 SuperCalc            1024             1ctjoD2HX05ZEu0bkNx… <named list>
#>  5 ExecuVision          3483             1TjcZYBVDWN9bIUtSUt… <named list>
#>  6 WordStar             1024             187ajd08ITKeHbZNgGZ… <named list>
#>  7 name-squatter-upload 0                15fJL62lFDPdZqra6cO… <named list>
#>  8 name-squatter-rename 0                1ifZxrlH9FLcSXZtwc2… <named list>
#>  9 name-squatter-mv     0                1RhGFXAin-dYoKh-48Y… <named list>
#> 10 Lotus 1-2-3          1024             1h6pIFFxuCvlRApQ57s… <named list>
#> # … with abbreviated variable name ¹​drive_resource

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>