Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

Usage

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what

Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via whatappears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name                 path                   id           drive_resource
#>    <chr>                <chr>                  <drv_id>     <list>        
#>  1 name-squatter-rename ~/name-squatter-rename 1hCRqO7Gvby… <named list>  
#>  2 name-squatter-mv     ~/name-squatter-mv     1eg_iF7aBCn… <named list>  
#>  3 Lotus 1-2-3          ~/b4xl/Lotus 1-2-3     1ZJNN6mjTwB… <named list>  
#>  4 SuperCalc            ~/b4xl/SuperCalc       1w-tZf04XG2… <named list>  
#>  5 ExecuVision          ~/ExecuVision          11hUrPnYGtC… <named list>  
#>  6 WordStar             ~/WordStar             1p7j2KkSYmc… <named list>  
#>  7 name-squatter-upload ~/name-squatter-upload 1_AAyhSwzQg… <named list>  
#>  8 name-squatter-rename ~/name-squatter-rename 1aqqz_eoQf5… <named list>  
#>  9 name-squatter-mv     ~/name-squatter-mv     1qswmczbfvr… <named list>  
#> 10 Lotus 1-2-3          ~/b4xl/Lotus 1-2-3     1eE_t43MqrQ… <named list>  
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name                 shared id          drive_resource permissions_res…
#>    <chr>                <lgl>  <drv_id>    <list>         <list>          
#>  1 name-squatter-rename FALSE  1hCRqO7Gvb… <named list>   <named list [2]>
#>  2 name-squatter-mv     FALSE  1eg_iF7aBC… <named list>   <named list [2]>
#>  3 Lotus 1-2-3          FALSE  1ZJNN6mjTw… <named list>   <named list [2]>
#>  4 SuperCalc            FALSE  1w-tZf04XG… <named list>   <named list [2]>
#>  5 ExecuVision          FALSE  11hUrPnYGt… <named list>   <named list [2]>
#>  6 WordStar             FALSE  1p7j2KkSYm… <named list>   <named list [2]>
#>  7 name-squatter-upload FALSE  1_AAyhSwzQ… <named list>   <named list [2]>
#>  8 name-squatter-rename FALSE  1aqqz_eoQf… <named list>   <named list [2]>
#>  9 name-squatter-mv     FALSE  1qswmczbfv… <named list>   <named list [2]>
#> 10 Lotus 1-2-3          FALSE  1eE_t43Mqr… <named list>   <named list [2]>
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name                 published id       drive_resource revision_resour…
#>    <chr>                <lgl>     <drv_id> <list>         <list>          
#>  1 name-squatter-rename FALSE     1hCRqO7… <named list>   <named list [9]>
#>  2 name-squatter-mv     FALSE     1eg_iF7… <named list>   <named list [9]>
#>  3 Lotus 1-2-3          FALSE     1ZJNN6m… <named list>   <named list [7]>
#>  4 SuperCalc            FALSE     1w-tZf0… <named list>   <named list [7]>
#>  5 ExecuVision          FALSE     11hUrPn… <named list>   <named list [7]>
#>  6 WordStar             FALSE     1p7j2Kk… <named list>   <named list [7]>
#>  7 name-squatter-upload FALSE     1_AAyhS… <named list>   <named list [9]>
#>  8 name-squatter-rename FALSE     1aqqz_e… <named list>   <named list [9]>
#>  9 name-squatter-mv     FALSE     1qswmcz… <named list>   <named list [9]>
#> 10 Lotus 1-2-3          FALSE     1eE_t43… <named list>   <named list [7]>

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name                 id_parent                     id    drive_resource
#>    <chr>                <drv_id>                      <drv> <list>        
#>  1 name-squatter-rename 0AO_RMaBzcP63Uk9PVA           1hCR… <named list>  
#>  2 name-squatter-mv     0AO_RMaBzcP63Uk9PVA           1eg_… <named list>  
#>  3 Lotus 1-2-3          1S0iWntXFsr7UqMZUnzVm1j7QLss… 1ZJN… <named list>  
#>  4 SuperCalc            1S0iWntXFsr7UqMZUnzVm1j7QLss… 1w-t… <named list>  
#>  5 ExecuVision          0AO_RMaBzcP63Uk9PVA           11hU… <named list>  
#>  6 WordStar             0AO_RMaBzcP63Uk9PVA           1p7j… <named list>  
#>  7 name-squatter-upload 0AO_RMaBzcP63Uk9PVA           1_AA… <named list>  
#>  8 name-squatter-rename 0AO_RMaBzcP63Uk9PVA           1aqq… <named list>  
#>  9 name-squatter-mv     0AO_RMaBzcP63Uk9PVA           1qsw… <named list>  
#> 10 Lotus 1-2-3          1S0iWntXFsr7UqMZUnzVm1j7QLss… 1eE_… <named list>  

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name                 trashed id                          drive_resource
#>    <chr>                <lgl>   <drv_id>                    <list>        
#>  1 name-squatter-rename TRUE    1hCRqO7GvbyqaQEzxl4PBx43pP… <named list>  
#>  2 name-squatter-mv     TRUE    1eg_iF7aBCncAYgjYGAA1Bld6A… <named list>  
#>  3 Lotus 1-2-3          FALSE   1ZJNN6mjTwBuVjmGRO0IlsiPYv… <named list>  
#>  4 SuperCalc            FALSE   1w-tZf04XG2lWsSrohtsVgexRb… <named list>  
#>  5 ExecuVision          FALSE   11hUrPnYGtCinv6tAYfrPMCBvO… <named list>  
#>  6 WordStar             FALSE   1p7j2KkSYmcOD56-S9EE0lLKoJ… <named list>  
#>  7 name-squatter-upload TRUE    1_AAyhSwzQga-qkq3HMk5VpeDZ… <named list>  
#>  8 name-squatter-rename TRUE    1aqqz_eoQf5_EgY0uoTAcDwwld… <named list>  
#>  9 name-squatter-mv     TRUE    1qswmczbfvrQXWpAIyLUs4YiYG… <named list>  
#> 10 Lotus 1-2-3          FALSE   1eE_t43MqrQ6KCZOYDGVirPlhs… <named list>  
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name                 mime_type                     id    drive_resource
#>    <chr>                <chr>                         <drv> <list>        
#>  1 name-squatter-rename application/octet-stream      1hCR… <named list>  
#>  2 name-squatter-mv     application/octet-stream      1eg_… <named list>  
#>  3 Lotus 1-2-3          application/vnd.google-apps.… 1ZJN… <named list>  
#>  4 SuperCalc            application/vnd.google-apps.… 1w-t… <named list>  
#>  5 ExecuVision          application/vnd.google-apps.… 11hU… <named list>  
#>  6 WordStar             application/vnd.google-apps.… 1p7j… <named list>  
#>  7 name-squatter-upload application/octet-stream      1_AA… <named list>  
#>  8 name-squatter-rename application/octet-stream      1aqq… <named list>  
#>  9 name-squatter-mv     application/octet-stream      1qsw… <named list>  
#> 10 Lotus 1-2-3          application/vnd.google-apps.… 1eE_… <named list>  
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name                 starred id                          drive_resource
#>    <chr>                <lgl>   <drv_id>                    <list>        
#>  1 name-squatter-rename FALSE   1hCRqO7GvbyqaQEzxl4PBx43pP… <named list>  
#>  2 name-squatter-mv     FALSE   1eg_iF7aBCncAYgjYGAA1Bld6A… <named list>  
#>  3 Lotus 1-2-3          FALSE   1ZJNN6mjTwBuVjmGRO0IlsiPYv… <named list>  
#>  4 SuperCalc            FALSE   1w-tZf04XG2lWsSrohtsVgexRb… <named list>  
#>  5 ExecuVision          FALSE   11hUrPnYGtCinv6tAYfrPMCBvO… <named list>  
#>  6 WordStar             TRUE    1p7j2KkSYmcOD56-S9EE0lLKoJ… <named list>  
#>  7 name-squatter-upload FALSE   1_AAyhSwzQga-qkq3HMk5VpeDZ… <named list>  
#>  8 name-squatter-rename FALSE   1aqqz_eoQf5_EgY0uoTAcDwwld… <named list>  
#>  9 name-squatter-mv     FALSE   1qswmczbfvrQXWpAIyLUs4YiYG… <named list>  
#> 10 Lotus 1-2-3          FALSE   1eE_t43MqrQ6KCZOYDGVirPlhs… <named list>  
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name                 description                   id    drive_resource
#>    <chr>                <chr>                         <drv> <list>        
#>  1 name-squatter-rename NA                            1hCR… <named list>  
#>  2 name-squatter-mv     NA                            1eg_… <named list>  
#>  3 Lotus 1-2-3          NA                            1ZJN… <named list>  
#>  4 SuperCalc            NA                            1w-t… <named list>  
#>  5 ExecuVision          deeply nested bullet lists F… 11hU… <named list>  
#>  6 WordStar             NA                            1p7j… <named list>  
#>  7 name-squatter-upload NA                            1_AA… <named list>  
#>  8 name-squatter-rename NA                            1aqq… <named list>  
#>  9 name-squatter-mv     NA                            1qsw… <named list>  
#> 10 Lotus 1-2-3          NA                            1eE_… <named list>  
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name                 version id                          drive_resource
#>    <chr>                <chr>   <drv_id>                    <list>        
#>  1 name-squatter-rename 2       1hCRqO7GvbyqaQEzxl4PBx43pP… <named list>  
#>  2 name-squatter-mv     2       1eg_iF7aBCncAYgjYGAA1Bld6A… <named list>  
#>  3 Lotus 1-2-3          4       1ZJNN6mjTwBuVjmGRO0IlsiPYv… <named list>  
#>  4 SuperCalc            4       1w-tZf04XG2lWsSrohtsVgexRb… <named list>  
#>  5 ExecuVision          3       11hUrPnYGtCinv6tAYfrPMCBvO… <named list>  
#>  6 WordStar             3       1p7j2KkSYmcOD56-S9EE0lLKoJ… <named list>  
#>  7 name-squatter-upload 2       1_AAyhSwzQga-qkq3HMk5VpeDZ… <named list>  
#>  8 name-squatter-rename 2       1aqqz_eoQf5_EgY0uoTAcDwwld… <named list>  
#>  9 name-squatter-mv     2       1qswmczbfvrQXWpAIyLUs4YiYG… <named list>  
#> 10 Lotus 1-2-3          4       1eE_t43MqrQ6KCZOYDGVirPlhs… <named list>  
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name                 web_view_link                 id    drive_resource
#>    <chr>                <chr>                         <drv> <list>        
#>  1 name-squatter-rename https://drive.google.com/fil… 1hCR… <named list>  
#>  2 name-squatter-mv     https://drive.google.com/fil… 1eg_… <named list>  
#>  3 Lotus 1-2-3          https://docs.google.com/spre… 1ZJN… <named list>  
#>  4 SuperCalc            https://docs.google.com/spre… 1w-t… <named list>  
#>  5 ExecuVision          https://docs.google.com/pres… 11hU… <named list>  
#>  6 WordStar             https://docs.google.com/docu… 1p7j… <named list>  
#>  7 name-squatter-upload https://drive.google.com/fil… 1_AA… <named list>  
#>  8 name-squatter-rename https://drive.google.com/fil… 1aqq… <named list>  
#>  9 name-squatter-mv     https://drive.google.com/fil… 1qsw… <named list>  
#> 10 Lotus 1-2-3          https://docs.google.com/spre… 1eE_… <named list>  
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name                 modified_time       id              drive_resource
#>    <chr>                <dttm>              <drv_id>        <list>        
#>  1 name-squatter-rename 2022-05-23 13:59:47 1hCRqO7GvbyqaQ… <named list>  
#>  2 name-squatter-mv     2022-05-23 13:59:29 1eg_iF7aBCncAY… <named list>  
#>  3 Lotus 1-2-3          2022-05-23 13:59:02 1ZJNN6mjTwBuVj… <named list>  
#>  4 SuperCalc            2022-05-23 13:59:00 1w-tZf04XG2lWs… <named list>  
#>  5 ExecuVision          2022-05-23 13:58:58 11hUrPnYGtCinv… <named list>  
#>  6 WordStar             2022-05-23 13:58:56 1p7j2KkSYmcOD5… <named list>  
#>  7 name-squatter-upload 2022-05-22 13:50:04 1_AAyhSwzQga-q… <named list>  
#>  8 name-squatter-rename 2022-05-22 13:49:23 1aqqz_eoQf5_Eg… <named list>  
#>  9 name-squatter-mv     2022-05-22 13:49:06 1qswmczbfvrQXW… <named list>  
#> 10 Lotus 1-2-3          2022-05-22 13:48:37 1eE_t43MqrQ6KC… <named list>  
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name                 created_time        id              drive_resource
#>    <chr>                <dttm>              <drv_id>        <list>        
#>  1 name-squatter-rename 2022-05-23 13:59:47 1hCRqO7GvbyqaQ… <named list>  
#>  2 name-squatter-mv     2022-05-23 13:59:29 1eg_iF7aBCncAY… <named list>  
#>  3 Lotus 1-2-3          2022-05-23 13:59:01 1ZJNN6mjTwBuVj… <named list>  
#>  4 SuperCalc            2022-05-23 13:59:00 1w-tZf04XG2lWs… <named list>  
#>  5 ExecuVision          2022-05-23 13:58:57 11hUrPnYGtCinv… <named list>  
#>  6 WordStar             2022-05-23 13:58:55 1p7j2KkSYmcOD5… <named list>  
#>  7 name-squatter-upload 2022-05-22 13:50:04 1_AAyhSwzQga-q… <named list>  
#>  8 name-squatter-rename 2022-05-22 13:49:23 1aqqz_eoQf5_Eg… <named list>  
#>  9 name-squatter-mv     2022-05-22 13:49:06 1qswmczbfvrQXW… <named list>  
#> 10 Lotus 1-2-3          2022-05-22 13:48:37 1eE_t43MqrQ6KC… <named list>  
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name                 owned_by_me id                      drive_resource
#>    <chr>                <lgl>       <drv_id>                <list>        
#>  1 name-squatter-rename TRUE        1hCRqO7GvbyqaQEzxl4PBx… <named list>  
#>  2 name-squatter-mv     TRUE        1eg_iF7aBCncAYgjYGAA1B… <named list>  
#>  3 Lotus 1-2-3          TRUE        1ZJNN6mjTwBuVjmGRO0Ils… <named list>  
#>  4 SuperCalc            TRUE        1w-tZf04XG2lWsSrohtsVg… <named list>  
#>  5 ExecuVision          TRUE        11hUrPnYGtCinv6tAYfrPM… <named list>  
#>  6 WordStar             TRUE        1p7j2KkSYmcOD56-S9EE0l… <named list>  
#>  7 name-squatter-upload TRUE        1_AAyhSwzQga-qkq3HMk5V… <named list>  
#>  8 name-squatter-rename TRUE        1aqqz_eoQf5_EgY0uoTAcD… <named list>  
#>  9 name-squatter-mv     TRUE        1qswmczbfvrQXWpAIyLUs4… <named list>  
#> 10 Lotus 1-2-3          TRUE        1eE_t43MqrQ6KCZOYDGVir… <named list>  
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name                 size  id                            drive_resource
#>    <chr>                <chr> <drv_id>                      <list>        
#>  1 name-squatter-rename 0     1hCRqO7GvbyqaQEzxl4PBx43pP7Z… <named list>  
#>  2 name-squatter-mv     0     1eg_iF7aBCncAYgjYGAA1Bld6AO9… <named list>  
#>  3 Lotus 1-2-3          1024  1ZJNN6mjTwBuVjmGRO0IlsiPYvrz… <named list>  
#>  4 SuperCalc            1024  1w-tZf04XG2lWsSrohtsVgexRbTo… <named list>  
#>  5 ExecuVision          3483  11hUrPnYGtCinv6tAYfrPMCBvOqP… <named list>  
#>  6 WordStar             1024  1p7j2KkSYmcOD56-S9EE0lLKoJ8k… <named list>  
#>  7 name-squatter-upload 0     1_AAyhSwzQga-qkq3HMk5VpeDZxu… <named list>  
#>  8 name-squatter-rename 0     1aqqz_eoQf5_EgY0uoTAcDwwldMv… <named list>  
#>  9 name-squatter-mv     0     1qswmczbfvrQXWpAIyLUs4YiYGFE… <named list>  
#> 10 Lotus 1-2-3          1024  1eE_t43MqrQ6KCZOYDGVirPlhshH… <named list>  
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name                 quota_bytes_used id                 drive_resource
#>    <chr>                <chr>            <drv_id>           <list>        
#>  1 name-squatter-rename 0                1hCRqO7GvbyqaQEzx… <named list>  
#>  2 name-squatter-mv     0                1eg_iF7aBCncAYgjY… <named list>  
#>  3 Lotus 1-2-3          1024             1ZJNN6mjTwBuVjmGR… <named list>  
#>  4 SuperCalc            1024             1w-tZf04XG2lWsSro… <named list>  
#>  5 ExecuVision          3483             11hUrPnYGtCinv6tA… <named list>  
#>  6 WordStar             1024             1p7j2KkSYmcOD56-S… <named list>  
#>  7 name-squatter-upload 0                1_AAyhSwzQga-qkq3… <named list>  
#>  8 name-squatter-rename 0                1aqqz_eoQf5_EgY0u… <named list>  
#>  9 name-squatter-mv     0                1qswmczbfvrQXWpAI… <named list>  
#> 10 Lotus 1-2-3          1024             1eE_t43MqrQ6KCZOY… <named list>  

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>