Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

Usage

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what

Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via what

appears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name                 path                   id             drive_reso…¹
#>    <chr>                <chr>                  <drv_id>       <list>      
#>  1 name-squatter-rename ~/name-squatter-rename 1-7u6JQMTO5vX… <named list>
#>  2 name-squatter-mv     ~/name-squatter-mv     1v3XjgNoUlY32… <named list>
#>  3 Lotus 1-2-3          ~/b4xl/Lotus 1-2-3     1civaWaJDfD11… <named list>
#>  4 SuperCalc            ~/b4xl/SuperCalc       12kKc3rxy7FA2… <named list>
#>  5 ExecuVision          ~/ExecuVision          1-TI3Tg9eukFv… <named list>
#>  6 WordStar             ~/WordStar             19Lomz_Gn0vPl… <named list>
#>  7 name-squatter-upload ~/name-squatter-upload 1VLn2-099BORJ… <named list>
#>  8 name-squatter-rename ~/name-squatter-rename 1dOt7G3yFc0rA… <named list>
#>  9 name-squatter-mv     ~/name-squatter-mv     1UCMRYnbGFd4v… <named list>
#> 10 Lotus 1-2-3          ~/b4xl/Lotus 1-2-3     10UleHMYWYFZQ… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name                 shared id                drive_reso…¹ permission…²
#>    <chr>                <lgl>  <drv_id>          <list>       <list>      
#>  1 name-squatter-rename FALSE  1-7u6JQMTO5vXKx4… <named list> <named list>
#>  2 name-squatter-mv     FALSE  1v3XjgNoUlY32YHm… <named list> <named list>
#>  3 Lotus 1-2-3          FALSE  1civaWaJDfD11BQ8… <named list> <named list>
#>  4 SuperCalc            FALSE  12kKc3rxy7FA2jue… <named list> <named list>
#>  5 ExecuVision          FALSE  1-TI3Tg9eukFvUOn… <named list> <named list>
#>  6 WordStar             FALSE  19Lomz_Gn0vPl-Lp… <named list> <named list>
#>  7 name-squatter-upload FALSE  1VLn2-099BORJFLt… <named list> <named list>
#>  8 name-squatter-rename FALSE  1dOt7G3yFc0rAJjv… <named list> <named list>
#>  9 name-squatter-mv     FALSE  1UCMRYnbGFd4v_U6… <named list> <named list>
#> 10 Lotus 1-2-3          FALSE  10UleHMYWYFZQHtt… <named list> <named list>
#> # … with abbreviated variable names ¹​drive_resource,
#> #   ²​permissions_resource
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name                 published id             drive_reso…¹ revision_r…²
#>    <chr>                <lgl>     <drv_id>       <list>       <list>      
#>  1 name-squatter-rename FALSE     1-7u6JQMTO5vX… <named list> <named list>
#>  2 name-squatter-mv     FALSE     1v3XjgNoUlY32… <named list> <named list>
#>  3 Lotus 1-2-3          FALSE     1civaWaJDfD11… <named list> <named list>
#>  4 SuperCalc            FALSE     12kKc3rxy7FA2… <named list> <named list>
#>  5 ExecuVision          FALSE     1-TI3Tg9eukFv… <named list> <named list>
#>  6 WordStar             FALSE     19Lomz_Gn0vPl… <named list> <named list>
#>  7 name-squatter-upload FALSE     1VLn2-099BORJ… <named list> <named list>
#>  8 name-squatter-rename FALSE     1dOt7G3yFc0rA… <named list> <named list>
#>  9 name-squatter-mv     FALSE     1UCMRYnbGFd4v… <named list> <named list>
#> 10 Lotus 1-2-3          FALSE     10UleHMYWYFZQ… <named list> <named list>
#> # … with abbreviated variable names ¹​drive_resource, ²​revision_resource

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name                 id_parent                       id    drive_reso…¹
#>    <chr>                <drv_id>                        <drv> <list>      
#>  1 name-squatter-rename 0AO_RMaBzcP63Uk9PVA             1-7u… <named list>
#>  2 name-squatter-mv     0AO_RMaBzcP63Uk9PVA             1v3X… <named list>
#>  3 Lotus 1-2-3          1S0iWntXFsr7UqMZUnzVm1j7QLssl_… 1civ… <named list>
#>  4 SuperCalc            1S0iWntXFsr7UqMZUnzVm1j7QLssl_… 12kK… <named list>
#>  5 ExecuVision          0AO_RMaBzcP63Uk9PVA             1-TI… <named list>
#>  6 WordStar             0AO_RMaBzcP63Uk9PVA             19Lo… <named list>
#>  7 name-squatter-upload 0AO_RMaBzcP63Uk9PVA             1VLn… <named list>
#>  8 name-squatter-rename 0AO_RMaBzcP63Uk9PVA             1dOt… <named list>
#>  9 name-squatter-mv     0AO_RMaBzcP63Uk9PVA             1UCM… <named list>
#> 10 Lotus 1-2-3          1S0iWntXFsr7UqMZUnzVm1j7QLssl_… 10Ul… <named list>
#> # … with abbreviated variable name ¹​drive_resource

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name                 trashed id                            drive_reso…¹
#>    <chr>                <lgl>   <drv_id>                      <list>      
#>  1 name-squatter-rename TRUE    1-7u6JQMTO5vXKx4n8hbsHQXG0V0… <named list>
#>  2 name-squatter-mv     TRUE    1v3XjgNoUlY32YHmjMmrlH5kWOFL… <named list>
#>  3 Lotus 1-2-3          FALSE   1civaWaJDfD11BQ8sMEfbA9ilZdP… <named list>
#>  4 SuperCalc            FALSE   12kKc3rxy7FA2juesp_G7dJXkRpZ… <named list>
#>  5 ExecuVision          FALSE   1-TI3Tg9eukFvUOnp3P2mcaMrGQ6… <named list>
#>  6 WordStar             FALSE   19Lomz_Gn0vPl-Lp6DU67mmpK6kz… <named list>
#>  7 name-squatter-upload TRUE    1VLn2-099BORJFLtVwJO1zvZ_RNo… <named list>
#>  8 name-squatter-rename TRUE    1dOt7G3yFc0rAJjvYh1SHtN-8cKY… <named list>
#>  9 name-squatter-mv     TRUE    1UCMRYnbGFd4v_U6xF3IcqXqxd-E… <named list>
#> 10 Lotus 1-2-3          FALSE   10UleHMYWYFZQHtt7Mlbe6Htv7lN… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name                 mime_type                       id    drive_reso…¹
#>    <chr>                <chr>                           <drv> <list>      
#>  1 name-squatter-rename application/octet-stream        1-7u… <named list>
#>  2 name-squatter-mv     application/octet-stream        1v3X… <named list>
#>  3 Lotus 1-2-3          application/vnd.google-apps.sp… 1civ… <named list>
#>  4 SuperCalc            application/vnd.google-apps.sp… 12kK… <named list>
#>  5 ExecuVision          application/vnd.google-apps.pr… 1-TI… <named list>
#>  6 WordStar             application/vnd.google-apps.do… 19Lo… <named list>
#>  7 name-squatter-upload application/octet-stream        1VLn… <named list>
#>  8 name-squatter-rename application/octet-stream        1dOt… <named list>
#>  9 name-squatter-mv     application/octet-stream        1UCM… <named list>
#> 10 Lotus 1-2-3          application/vnd.google-apps.sp… 10Ul… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name                 starred id                            drive_reso…¹
#>    <chr>                <lgl>   <drv_id>                      <list>      
#>  1 name-squatter-rename FALSE   1-7u6JQMTO5vXKx4n8hbsHQXG0V0… <named list>
#>  2 name-squatter-mv     FALSE   1v3XjgNoUlY32YHmjMmrlH5kWOFL… <named list>
#>  3 Lotus 1-2-3          FALSE   1civaWaJDfD11BQ8sMEfbA9ilZdP… <named list>
#>  4 SuperCalc            FALSE   12kKc3rxy7FA2juesp_G7dJXkRpZ… <named list>
#>  5 ExecuVision          FALSE   1-TI3Tg9eukFvUOnp3P2mcaMrGQ6… <named list>
#>  6 WordStar             TRUE    19Lomz_Gn0vPl-Lp6DU67mmpK6kz… <named list>
#>  7 name-squatter-upload FALSE   1VLn2-099BORJFLtVwJO1zvZ_RNo… <named list>
#>  8 name-squatter-rename FALSE   1dOt7G3yFc0rAJjvYh1SHtN-8cKY… <named list>
#>  9 name-squatter-mv     FALSE   1UCMRYnbGFd4v_U6xF3IcqXqxd-E… <named list>
#> 10 Lotus 1-2-3          FALSE   10UleHMYWYFZQHtt7Mlbe6Htv7lN… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name                 description                    id     drive_reso…¹
#>    <chr>                <chr>                          <drv_> <list>      
#>  1 name-squatter-rename NA                             1-7u6… <named list>
#>  2 name-squatter-mv     NA                             1v3Xj… <named list>
#>  3 Lotus 1-2-3          NA                             1civa… <named list>
#>  4 SuperCalc            NA                             12kKc… <named list>
#>  5 ExecuVision          deeply nested bullet lists FTW 1-TI3… <named list>
#>  6 WordStar             NA                             19Lom… <named list>
#>  7 name-squatter-upload NA                             1VLn2… <named list>
#>  8 name-squatter-rename NA                             1dOt7… <named list>
#>  9 name-squatter-mv     NA                             1UCMR… <named list>
#> 10 Lotus 1-2-3          NA                             10Ule… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name                 version id                            drive_reso…¹
#>    <chr>                <chr>   <drv_id>                      <list>      
#>  1 name-squatter-rename 2       1-7u6JQMTO5vXKx4n8hbsHQXG0V0… <named list>
#>  2 name-squatter-mv     2       1v3XjgNoUlY32YHmjMmrlH5kWOFL… <named list>
#>  3 Lotus 1-2-3          4       1civaWaJDfD11BQ8sMEfbA9ilZdP… <named list>
#>  4 SuperCalc            4       12kKc3rxy7FA2juesp_G7dJXkRpZ… <named list>
#>  5 ExecuVision          3       1-TI3Tg9eukFvUOnp3P2mcaMrGQ6… <named list>
#>  6 WordStar             3       19Lomz_Gn0vPl-Lp6DU67mmpK6kz… <named list>
#>  7 name-squatter-upload 2       1VLn2-099BORJFLtVwJO1zvZ_RNo… <named list>
#>  8 name-squatter-rename 2       1dOt7G3yFc0rAJjvYh1SHtN-8cKY… <named list>
#>  9 name-squatter-mv     2       1UCMRYnbGFd4v_U6xF3IcqXqxd-E… <named list>
#> 10 Lotus 1-2-3          4       10UleHMYWYFZQHtt7Mlbe6Htv7lN… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name                 web_view_link                   id    drive_reso…¹
#>    <chr>                <chr>                           <drv> <list>      
#>  1 name-squatter-rename https://drive.google.com/file/… 1-7u… <named list>
#>  2 name-squatter-mv     https://drive.google.com/file/… 1v3X… <named list>
#>  3 Lotus 1-2-3          https://docs.google.com/spread… 1civ… <named list>
#>  4 SuperCalc            https://docs.google.com/spread… 12kK… <named list>
#>  5 ExecuVision          https://docs.google.com/presen… 1-TI… <named list>
#>  6 WordStar             https://docs.google.com/docume… 19Lo… <named list>
#>  7 name-squatter-upload https://drive.google.com/file/… 1VLn… <named list>
#>  8 name-squatter-rename https://drive.google.com/file/… 1dOt… <named list>
#>  9 name-squatter-mv     https://drive.google.com/file/… 1UCM… <named list>
#> 10 Lotus 1-2-3          https://docs.google.com/spread… 10Ul… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name                 modified_time       id                drive_reso…¹
#>    <chr>                <dttm>              <drv_id>          <list>      
#>  1 name-squatter-rename 2022-12-31 13:41:42 1-7u6JQMTO5vXKx4… <named list>
#>  2 name-squatter-mv     2022-12-31 13:41:18 1v3XjgNoUlY32YHm… <named list>
#>  3 Lotus 1-2-3          2022-12-31 13:40:49 1civaWaJDfD11BQ8… <named list>
#>  4 SuperCalc            2022-12-31 13:40:46 12kKc3rxy7FA2jue… <named list>
#>  5 ExecuVision          2022-12-31 13:40:42 1-TI3Tg9eukFvUOn… <named list>
#>  6 WordStar             2022-12-31 13:40:39 19Lomz_Gn0vPl-Lp… <named list>
#>  7 name-squatter-upload 2022-12-30 13:41:48 1VLn2-099BORJFLt… <named list>
#>  8 name-squatter-rename 2022-12-30 13:41:07 1dOt7G3yFc0rAJjv… <named list>
#>  9 name-squatter-mv     2022-12-30 13:40:46 1UCMRYnbGFd4v_U6… <named list>
#> 10 Lotus 1-2-3          2022-12-30 13:40:20 10UleHMYWYFZQHtt… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name                 created_time        id                drive_reso…¹
#>    <chr>                <dttm>              <drv_id>          <list>      
#>  1 name-squatter-rename 2022-12-31 13:41:42 1-7u6JQMTO5vXKx4… <named list>
#>  2 name-squatter-mv     2022-12-31 13:41:18 1v3XjgNoUlY32YHm… <named list>
#>  3 Lotus 1-2-3          2022-12-31 13:40:48 1civaWaJDfD11BQ8… <named list>
#>  4 SuperCalc            2022-12-31 13:40:46 12kKc3rxy7FA2jue… <named list>
#>  5 ExecuVision          2022-12-31 13:40:41 1-TI3Tg9eukFvUOn… <named list>
#>  6 WordStar             2022-12-31 13:40:38 19Lomz_Gn0vPl-Lp… <named list>
#>  7 name-squatter-upload 2022-12-30 13:41:48 1VLn2-099BORJFLt… <named list>
#>  8 name-squatter-rename 2022-12-30 13:41:07 1dOt7G3yFc0rAJjv… <named list>
#>  9 name-squatter-mv     2022-12-30 13:40:46 1UCMRYnbGFd4v_U6… <named list>
#> 10 Lotus 1-2-3          2022-12-30 13:40:20 10UleHMYWYFZQHtt… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name                 owned_by_me id                        drive_reso…¹
#>    <chr>                <lgl>       <drv_id>                  <list>      
#>  1 name-squatter-rename TRUE        1-7u6JQMTO5vXKx4n8hbsHQX… <named list>
#>  2 name-squatter-mv     TRUE        1v3XjgNoUlY32YHmjMmrlH5k… <named list>
#>  3 Lotus 1-2-3          TRUE        1civaWaJDfD11BQ8sMEfbA9i… <named list>
#>  4 SuperCalc            TRUE        12kKc3rxy7FA2juesp_G7dJX… <named list>
#>  5 ExecuVision          TRUE        1-TI3Tg9eukFvUOnp3P2mcaM… <named list>
#>  6 WordStar             TRUE        19Lomz_Gn0vPl-Lp6DU67mmp… <named list>
#>  7 name-squatter-upload TRUE        1VLn2-099BORJFLtVwJO1zvZ… <named list>
#>  8 name-squatter-rename TRUE        1dOt7G3yFc0rAJjvYh1SHtN-… <named list>
#>  9 name-squatter-mv     TRUE        1UCMRYnbGFd4v_U6xF3IcqXq… <named list>
#> 10 Lotus 1-2-3          TRUE        10UleHMYWYFZQHtt7Mlbe6Ht… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name                 size  id                              drive_reso…¹
#>    <chr>                <chr> <drv_id>                        <list>      
#>  1 name-squatter-rename 0     1-7u6JQMTO5vXKx4n8hbsHQXG0V0N8… <named list>
#>  2 name-squatter-mv     0     1v3XjgNoUlY32YHmjMmrlH5kWOFLZ9… <named list>
#>  3 Lotus 1-2-3          1024  1civaWaJDfD11BQ8sMEfbA9ilZdPN7… <named list>
#>  4 SuperCalc            1024  12kKc3rxy7FA2juesp_G7dJXkRpZ8H… <named list>
#>  5 ExecuVision          3483  1-TI3Tg9eukFvUOnp3P2mcaMrGQ687… <named list>
#>  6 WordStar             1024  19Lomz_Gn0vPl-Lp6DU67mmpK6kz4M… <named list>
#>  7 name-squatter-upload 0     1VLn2-099BORJFLtVwJO1zvZ_RNoDJ… <named list>
#>  8 name-squatter-rename 0     1dOt7G3yFc0rAJjvYh1SHtN-8cKYbA… <named list>
#>  9 name-squatter-mv     0     1UCMRYnbGFd4v_U6xF3IcqXqxd-EDp… <named list>
#> 10 Lotus 1-2-3          1024  10UleHMYWYFZQHtt7Mlbe6Htv7lNIx… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name                 quota_bytes_used id                   drive_reso…¹
#>    <chr>                <chr>            <drv_id>             <list>      
#>  1 name-squatter-rename 0                1-7u6JQMTO5vXKx4n8h… <named list>
#>  2 name-squatter-mv     0                1v3XjgNoUlY32YHmjMm… <named list>
#>  3 Lotus 1-2-3          1024             1civaWaJDfD11BQ8sME… <named list>
#>  4 SuperCalc            1024             12kKc3rxy7FA2juesp_… <named list>
#>  5 ExecuVision          3483             1-TI3Tg9eukFvUOnp3P… <named list>
#>  6 WordStar             1024             19Lomz_Gn0vPl-Lp6DU… <named list>
#>  7 name-squatter-upload 0                1VLn2-099BORJFLtVwJ… <named list>
#>  8 name-squatter-rename 0                1dOt7G3yFc0rAJjvYh1… <named list>
#>  9 name-squatter-mv     0                1UCMRYnbGFd4v_U6xF3… <named list>
#> 10 Lotus 1-2-3          1024             10UleHMYWYFZQHtt7Ml… <named list>
#> # … with abbreviated variable name ¹​drive_resource

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>