Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

Usage

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what

Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via what

appears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name                 path                   id             drive_reso…¹
#>    <chr>                <chr>                  <drv_id>       <list>      
#>  1 name-squatter-rename ~/name-squatter-rename 1syirfRrGzFNx… <named list>
#>  2 name-squatter-mv     ~/name-squatter-mv     1sKe-9u4Hzt37… <named list>
#>  3 Lotus 1-2-3          ~/b4xl/Lotus 1-2-3     13bZ-QVn4ch3y… <named list>
#>  4 SuperCalc            ~/b4xl/SuperCalc       1l3V9N4AUY0HM… <named list>
#>  5 ExecuVision          ~/ExecuVision          1EOBCxaLcI5eA… <named list>
#>  6 WordStar             ~/WordStar             1O_EU60q_P8as… <named list>
#>  7 name-squatter-upload ~/name-squatter-upload 1gyWv3kRCNPAO… <named list>
#>  8 name-squatter-rename ~/name-squatter-rename 1UpVI-r2NAPxw… <named list>
#>  9 name-squatter-mv     ~/name-squatter-mv     1dzixWPGQxr1j… <named list>
#> 10 Lotus 1-2-3          ~/b4xl/Lotus 1-2-3     1U4kumwHyCScE… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name                 shared id                drive_reso…¹ permission…²
#>    <chr>                <lgl>  <drv_id>          <list>       <list>      
#>  1 name-squatter-rename FALSE  1syirfRrGzFNxhSw… <named list> <named list>
#>  2 name-squatter-mv     FALSE  1sKe-9u4Hzt371ZV… <named list> <named list>
#>  3 Lotus 1-2-3          FALSE  13bZ-QVn4ch3yHtm… <named list> <named list>
#>  4 SuperCalc            FALSE  1l3V9N4AUY0HMiX7… <named list> <named list>
#>  5 ExecuVision          FALSE  1EOBCxaLcI5eAWab… <named list> <named list>
#>  6 WordStar             FALSE  1O_EU60q_P8aslrg… <named list> <named list>
#>  7 name-squatter-upload FALSE  1gyWv3kRCNPAO_A2… <named list> <named list>
#>  8 name-squatter-rename FALSE  1UpVI-r2NAPxwyZg… <named list> <named list>
#>  9 name-squatter-mv     FALSE  1dzixWPGQxr1jM2J… <named list> <named list>
#> 10 Lotus 1-2-3          FALSE  1U4kumwHyCScELXa… <named list> <named list>
#> # … with abbreviated variable names ¹​drive_resource,
#> #   ²​permissions_resource
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name                 published id             drive_reso…¹ revision_r…²
#>    <chr>                <lgl>     <drv_id>       <list>       <list>      
#>  1 name-squatter-rename FALSE     1syirfRrGzFNx… <named list> <named list>
#>  2 name-squatter-mv     FALSE     1sKe-9u4Hzt37… <named list> <named list>
#>  3 Lotus 1-2-3          FALSE     13bZ-QVn4ch3y… <named list> <named list>
#>  4 SuperCalc            FALSE     1l3V9N4AUY0HM… <named list> <named list>
#>  5 ExecuVision          FALSE     1EOBCxaLcI5eA… <named list> <named list>
#>  6 WordStar             FALSE     1O_EU60q_P8as… <named list> <named list>
#>  7 name-squatter-upload FALSE     1gyWv3kRCNPAO… <named list> <named list>
#>  8 name-squatter-rename FALSE     1UpVI-r2NAPxw… <named list> <named list>
#>  9 name-squatter-mv     FALSE     1dzixWPGQxr1j… <named list> <named list>
#> 10 Lotus 1-2-3          FALSE     1U4kumwHyCScE… <named list> <named list>
#> # … with abbreviated variable names ¹​drive_resource, ²​revision_resource

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name                 id_parent                       id    drive_reso…¹
#>    <chr>                <drv_id>                        <drv> <list>      
#>  1 name-squatter-rename 0AO_RMaBzcP63Uk9PVA             1syi… <named list>
#>  2 name-squatter-mv     0AO_RMaBzcP63Uk9PVA             1sKe… <named list>
#>  3 Lotus 1-2-3          1S0iWntXFsr7UqMZUnzVm1j7QLssl_… 13bZ… <named list>
#>  4 SuperCalc            1S0iWntXFsr7UqMZUnzVm1j7QLssl_… 1l3V… <named list>
#>  5 ExecuVision          0AO_RMaBzcP63Uk9PVA             1EOB… <named list>
#>  6 WordStar             0AO_RMaBzcP63Uk9PVA             1O_E… <named list>
#>  7 name-squatter-upload 0AO_RMaBzcP63Uk9PVA             1gyW… <named list>
#>  8 name-squatter-rename 0AO_RMaBzcP63Uk9PVA             1UpV… <named list>
#>  9 name-squatter-mv     0AO_RMaBzcP63Uk9PVA             1dzi… <named list>
#> 10 Lotus 1-2-3          1S0iWntXFsr7UqMZUnzVm1j7QLssl_… 1U4k… <named list>
#> # … with abbreviated variable name ¹​drive_resource

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name                 trashed id                            drive_reso…¹
#>    <chr>                <lgl>   <drv_id>                      <list>      
#>  1 name-squatter-rename TRUE    1syirfRrGzFNxhSwP_HDcAriGjNE… <named list>
#>  2 name-squatter-mv     TRUE    1sKe-9u4Hzt371ZV-CeFvmEbvOCk… <named list>
#>  3 Lotus 1-2-3          FALSE   13bZ-QVn4ch3yHtmvMl60MImYRBx… <named list>
#>  4 SuperCalc            FALSE   1l3V9N4AUY0HMiX7GZxgckzWxKvr… <named list>
#>  5 ExecuVision          FALSE   1EOBCxaLcI5eAWab9qV8_-_wTftR… <named list>
#>  6 WordStar             FALSE   1O_EU60q_P8aslrgAZccb53vyImB… <named list>
#>  7 name-squatter-upload TRUE    1gyWv3kRCNPAO_A2gAkrHwu6pto7… <named list>
#>  8 name-squatter-rename TRUE    1UpVI-r2NAPxwyZgG8k9B6Uw2zCL… <named list>
#>  9 name-squatter-mv     TRUE    1dzixWPGQxr1jM2J24JdsUOBqiK-… <named list>
#> 10 Lotus 1-2-3          FALSE   1U4kumwHyCScELXaH6oGnqUbu8W5… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name                 mime_type                       id    drive_reso…¹
#>    <chr>                <chr>                           <drv> <list>      
#>  1 name-squatter-rename application/octet-stream        1syi… <named list>
#>  2 name-squatter-mv     application/octet-stream        1sKe… <named list>
#>  3 Lotus 1-2-3          application/vnd.google-apps.sp… 13bZ… <named list>
#>  4 SuperCalc            application/vnd.google-apps.sp… 1l3V… <named list>
#>  5 ExecuVision          application/vnd.google-apps.pr… 1EOB… <named list>
#>  6 WordStar             application/vnd.google-apps.do… 1O_E… <named list>
#>  7 name-squatter-upload application/octet-stream        1gyW… <named list>
#>  8 name-squatter-rename application/octet-stream        1UpV… <named list>
#>  9 name-squatter-mv     application/octet-stream        1dzi… <named list>
#> 10 Lotus 1-2-3          application/vnd.google-apps.sp… 1U4k… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name                 starred id                            drive_reso…¹
#>    <chr>                <lgl>   <drv_id>                      <list>      
#>  1 name-squatter-rename FALSE   1syirfRrGzFNxhSwP_HDcAriGjNE… <named list>
#>  2 name-squatter-mv     FALSE   1sKe-9u4Hzt371ZV-CeFvmEbvOCk… <named list>
#>  3 Lotus 1-2-3          FALSE   13bZ-QVn4ch3yHtmvMl60MImYRBx… <named list>
#>  4 SuperCalc            FALSE   1l3V9N4AUY0HMiX7GZxgckzWxKvr… <named list>
#>  5 ExecuVision          FALSE   1EOBCxaLcI5eAWab9qV8_-_wTftR… <named list>
#>  6 WordStar             TRUE    1O_EU60q_P8aslrgAZccb53vyImB… <named list>
#>  7 name-squatter-upload FALSE   1gyWv3kRCNPAO_A2gAkrHwu6pto7… <named list>
#>  8 name-squatter-rename FALSE   1UpVI-r2NAPxwyZgG8k9B6Uw2zCL… <named list>
#>  9 name-squatter-mv     FALSE   1dzixWPGQxr1jM2J24JdsUOBqiK-… <named list>
#> 10 Lotus 1-2-3          FALSE   1U4kumwHyCScELXaH6oGnqUbu8W5… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name                 description                    id     drive_reso…¹
#>    <chr>                <chr>                          <drv_> <list>      
#>  1 name-squatter-rename NA                             1syir… <named list>
#>  2 name-squatter-mv     NA                             1sKe-… <named list>
#>  3 Lotus 1-2-3          NA                             13bZ-… <named list>
#>  4 SuperCalc            NA                             1l3V9… <named list>
#>  5 ExecuVision          deeply nested bullet lists FTW 1EOBC… <named list>
#>  6 WordStar             NA                             1O_EU… <named list>
#>  7 name-squatter-upload NA                             1gyWv… <named list>
#>  8 name-squatter-rename NA                             1UpVI… <named list>
#>  9 name-squatter-mv     NA                             1dzix… <named list>
#> 10 Lotus 1-2-3          NA                             1U4ku… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name                 version id                            drive_reso…¹
#>    <chr>                <chr>   <drv_id>                      <list>      
#>  1 name-squatter-rename 2       1syirfRrGzFNxhSwP_HDcAriGjNE… <named list>
#>  2 name-squatter-mv     2       1sKe-9u4Hzt371ZV-CeFvmEbvOCk… <named list>
#>  3 Lotus 1-2-3          4       13bZ-QVn4ch3yHtmvMl60MImYRBx… <named list>
#>  4 SuperCalc            4       1l3V9N4AUY0HMiX7GZxgckzWxKvr… <named list>
#>  5 ExecuVision          3       1EOBCxaLcI5eAWab9qV8_-_wTftR… <named list>
#>  6 WordStar             3       1O_EU60q_P8aslrgAZccb53vyImB… <named list>
#>  7 name-squatter-upload 2       1gyWv3kRCNPAO_A2gAkrHwu6pto7… <named list>
#>  8 name-squatter-rename 2       1UpVI-r2NAPxwyZgG8k9B6Uw2zCL… <named list>
#>  9 name-squatter-mv     2       1dzixWPGQxr1jM2J24JdsUOBqiK-… <named list>
#> 10 Lotus 1-2-3          4       1U4kumwHyCScELXaH6oGnqUbu8W5… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name                 web_view_link                   id    drive_reso…¹
#>    <chr>                <chr>                           <drv> <list>      
#>  1 name-squatter-rename https://drive.google.com/file/… 1syi… <named list>
#>  2 name-squatter-mv     https://drive.google.com/file/… 1sKe… <named list>
#>  3 Lotus 1-2-3          https://docs.google.com/spread… 13bZ… <named list>
#>  4 SuperCalc            https://docs.google.com/spread… 1l3V… <named list>
#>  5 ExecuVision          https://docs.google.com/presen… 1EOB… <named list>
#>  6 WordStar             https://docs.google.com/docume… 1O_E… <named list>
#>  7 name-squatter-upload https://drive.google.com/file/… 1gyW… <named list>
#>  8 name-squatter-rename https://drive.google.com/file/… 1UpV… <named list>
#>  9 name-squatter-mv     https://drive.google.com/file/… 1dzi… <named list>
#> 10 Lotus 1-2-3          https://docs.google.com/spread… 1U4k… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name                 modified_time       id                drive_reso…¹
#>    <chr>                <dttm>              <drv_id>          <list>      
#>  1 name-squatter-rename 2023-03-06 13:45:10 1syirfRrGzFNxhSw… <named list>
#>  2 name-squatter-mv     2023-03-06 13:44:43 1sKe-9u4Hzt371ZV… <named list>
#>  3 Lotus 1-2-3          2023-03-06 13:44:11 13bZ-QVn4ch3yHtm… <named list>
#>  4 SuperCalc            2023-03-06 13:44:10 1l3V9N4AUY0HMiX7… <named list>
#>  5 ExecuVision          2023-03-06 13:44:05 1EOBCxaLcI5eAWab… <named list>
#>  6 WordStar             2023-03-06 13:44:01 1O_EU60q_P8aslrg… <named list>
#>  7 name-squatter-upload 2023-03-05 13:42:43 1gyWv3kRCNPAO_A2… <named list>
#>  8 name-squatter-rename 2023-03-05 13:41:49 1UpVI-r2NAPxwyZg… <named list>
#>  9 name-squatter-mv     2023-03-05 13:41:21 1dzixWPGQxr1jM2J… <named list>
#> 10 Lotus 1-2-3          2023-03-05 13:40:45 1U4kumwHyCScELXa… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name                 created_time        id                drive_reso…¹
#>    <chr>                <dttm>              <drv_id>          <list>      
#>  1 name-squatter-rename 2023-03-06 13:45:10 1syirfRrGzFNxhSw… <named list>
#>  2 name-squatter-mv     2023-03-06 13:44:43 1sKe-9u4Hzt371ZV… <named list>
#>  3 Lotus 1-2-3          2023-03-06 13:44:11 13bZ-QVn4ch3yHtm… <named list>
#>  4 SuperCalc            2023-03-06 13:44:09 1l3V9N4AUY0HMiX7… <named list>
#>  5 ExecuVision          2023-03-06 13:44:04 1EOBCxaLcI5eAWab… <named list>
#>  6 WordStar             2023-03-06 13:44:00 1O_EU60q_P8aslrg… <named list>
#>  7 name-squatter-upload 2023-03-05 13:42:43 1gyWv3kRCNPAO_A2… <named list>
#>  8 name-squatter-rename 2023-03-05 13:41:49 1UpVI-r2NAPxwyZg… <named list>
#>  9 name-squatter-mv     2023-03-05 13:41:21 1dzixWPGQxr1jM2J… <named list>
#> 10 Lotus 1-2-3          2023-03-05 13:40:45 1U4kumwHyCScELXa… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name                 owned_by_me id                        drive_reso…¹
#>    <chr>                <lgl>       <drv_id>                  <list>      
#>  1 name-squatter-rename TRUE        1syirfRrGzFNxhSwP_HDcAri… <named list>
#>  2 name-squatter-mv     TRUE        1sKe-9u4Hzt371ZV-CeFvmEb… <named list>
#>  3 Lotus 1-2-3          TRUE        13bZ-QVn4ch3yHtmvMl60MIm… <named list>
#>  4 SuperCalc            TRUE        1l3V9N4AUY0HMiX7GZxgckzW… <named list>
#>  5 ExecuVision          TRUE        1EOBCxaLcI5eAWab9qV8_-_w… <named list>
#>  6 WordStar             TRUE        1O_EU60q_P8aslrgAZccb53v… <named list>
#>  7 name-squatter-upload TRUE        1gyWv3kRCNPAO_A2gAkrHwu6… <named list>
#>  8 name-squatter-rename TRUE        1UpVI-r2NAPxwyZgG8k9B6Uw… <named list>
#>  9 name-squatter-mv     TRUE        1dzixWPGQxr1jM2J24JdsUOB… <named list>
#> 10 Lotus 1-2-3          TRUE        1U4kumwHyCScELXaH6oGnqUb… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name                 size  id                              drive_reso…¹
#>    <chr>                <chr> <drv_id>                        <list>      
#>  1 name-squatter-rename 0     1syirfRrGzFNxhSwP_HDcAriGjNE5O… <named list>
#>  2 name-squatter-mv     0     1sKe-9u4Hzt371ZV-CeFvmEbvOCkIj… <named list>
#>  3 Lotus 1-2-3          1024  13bZ-QVn4ch3yHtmvMl60MImYRBxbg… <named list>
#>  4 SuperCalc            1024  1l3V9N4AUY0HMiX7GZxgckzWxKvrQd… <named list>
#>  5 ExecuVision          3483  1EOBCxaLcI5eAWab9qV8_-_wTftREm… <named list>
#>  6 WordStar             1024  1O_EU60q_P8aslrgAZccb53vyImBXY… <named list>
#>  7 name-squatter-upload 0     1gyWv3kRCNPAO_A2gAkrHwu6pto7Gd… <named list>
#>  8 name-squatter-rename 0     1UpVI-r2NAPxwyZgG8k9B6Uw2zCLSS… <named list>
#>  9 name-squatter-mv     0     1dzixWPGQxr1jM2J24JdsUOBqiK-Rc… <named list>
#> 10 Lotus 1-2-3          1024  1U4kumwHyCScELXaH6oGnqUbu8W5bY… <named list>
#> # … with abbreviated variable name ¹​drive_resource
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name                 quota_bytes_used id                   drive_reso…¹
#>    <chr>                <chr>            <drv_id>             <list>      
#>  1 name-squatter-rename 0                1syirfRrGzFNxhSwP_H… <named list>
#>  2 name-squatter-mv     0                1sKe-9u4Hzt371ZV-Ce… <named list>
#>  3 Lotus 1-2-3          1024             13bZ-QVn4ch3yHtmvMl… <named list>
#>  4 SuperCalc            1024             1l3V9N4AUY0HMiX7GZx… <named list>
#>  5 ExecuVision          3483             1EOBCxaLcI5eAWab9qV… <named list>
#>  6 WordStar             1024             1O_EU60q_P8aslrgAZc… <named list>
#>  7 name-squatter-upload 0                1gyWv3kRCNPAO_A2gAk… <named list>
#>  8 name-squatter-rename 0                1UpVI-r2NAPxwyZgG8k… <named list>
#>  9 name-squatter-mv     0                1dzixWPGQxr1jM2J24J… <named list>
#> 10 Lotus 1-2-3          1024             1U4kumwHyCScELXaH6o… <named list>
#> # … with abbreviated variable name ¹​drive_resource

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>