Add a new column of Drive file information

drive_reveal() adds extra information about your Drive files that is not readily available in the default dribble produced by googledrive. Why is this info not always included in the default dribble?

You don't always care about it. There is a lot of esoteric information in the drive_resource that has little value for most users.

It might be "expensive" to get this information and put it into a usable form. For example, revealing a file's "path", "permissions", or "published" status all require additional API calls.

drive_reveal() can also hoist any property out of the drive_resource list-column, when the property's name is passed as the what argument. The resulting new column is simplified if it is easy to do so, e.g., if the individual elements are all string or logical. If what extracts a date-time, we return POSIXct. Otherwise, you'll get a list-column. If this makes you sad, consider using tidyr::hoist() instead. It is more powerful due to a richer "plucking specification" and its ptype and transform arguments. Another useful function is tidyr::unnest_wider().

Usage

drive_reveal(file, what = c("path", "permissions", "published", "parent"))

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

what

Character, describing the type of info you want to add. These values get special handling (more details below):

path

permissions

published

parent

You can also request any property in the drive_resource column by name. The request can be in camelCase or snake_case, but the new column name will always be snake_case. Some examples of what:

mime_type (or mimeType)

trashed

starred

description

version

web_view_link (or webViewLink)

modified_time (or modifiedTime)

created_time (or createdTime)

owned_by_me (or ownedByMe)

size

quota_bytes_used (or quotaBytesUsed)

Value

An object of class dribble, a tibble with one row per file. The additional info requested via what appears in one (or more) extra columns.

File path

When what = "path" the dribble gains a character column holding each file's path. This can be very slow, so use with caution.

The example path ~/a/b/ illustrates two conventions used in googledrive:

The leading ~/ means that the folder a is located in the current user's "My Drive" root folder.

The trailing / means that b, located in a, is a folder or a folder shortcut.

Permissions

When what = "permissions" the dribble gains a logical column shared that indicates whether a file is shared and a new list-column permissions_resource containing lists of Permissions resources.

Publishing

When what = "published" the dribble gains a logical column published that indicates whether a file is published and a new list-column revision_resource containing lists of Revisions resources.

Parent

When what = "parent" the dribble gains a character column id_parent that is the file id of this item's parent folder. This information is available in the drive_resource, but can't just be hoisted out:

Google Drive used to allow files to have multiple parents, but this is no longer supported and googledrive now assumes this is impossible. However, we have seen (very old) files that still have >1 parent folder. If we see this we message about it and drop all but the first parent.

The parents property in drive_resource has an "extra" layer of nesting and needs to be flattened.

If you really want the raw parents property, call drive_reveal(what = "parents").

See also

Examples

# Get a few of your files
files <- drive_find(n_max = 10, trashed = NA)

# the "special" cases that require additional API calls and can be slow
drive_reveal(files, "path")
#> # A dribble: 10 × 4
#>    name                   path                     id       drive_resource
#>    <chr>                  <chr>                    <drv_id> <list>        
#>  1 name-squatter-rename   ~/name-squatter-rename   1iejRst… <named list>  
#>  2 name-squatter-mv       ~/name-squatter-mv       1lyFPxt… <named list>  
#>  3 chicken-mv-renamed.txt ~/mv-folder/chicken-mv-… 1y7LlJm… <named list>  
#>  4 name-squatter-upload   ~/name-squatter-upload   1AT6KZa… <named list>  
#>  5 name-squatter-rename   ~/name-squatter-rename   1DRTur5… <named list>  
#>  6 name-squatter-mv       ~/name-squatter-mv       1kqe58I… <named list>  
#>  7 chicken-mv-renamed.txt ~/mv-folder/chicken-mv-… 1WMuZd4… <named list>  
#>  8 name-squatter-upload   ~/name-squatter-upload   1LULuzk… <named list>  
#>  9 name-squatter-rename   ~/name-squatter-rename   1om8oRG… <named list>  
#> 10 name-squatter-mv       ~/name-squatter-mv       1ENeiEn… <named list>  
drive_reveal(files, "permissions")
#> # A dribble: 10 × 5
#>    name                shared id       drive_resource permissions_resource
#>    <chr>               <lgl>  <drv_id> <list>         <list>              
#>  1 name-squatter-rena… FALSE  1iejRst… <named list>   <named list [2]>    
#>  2 name-squatter-mv    FALSE  1lyFPxt… <named list>   <named list [2]>    
#>  3 chicken-mv-renamed… FALSE  1y7LlJm… <named list>   <named list [2]>    
#>  4 name-squatter-uplo… FALSE  1AT6KZa… <named list>   <named list [2]>    
#>  5 name-squatter-rena… FALSE  1DRTur5… <named list>   <named list [2]>    
#>  6 name-squatter-mv    FALSE  1kqe58I… <named list>   <named list [2]>    
#>  7 chicken-mv-renamed… FALSE  1WMuZd4… <named list>   <named list [2]>    
#>  8 name-squatter-uplo… FALSE  1LULuzk… <named list>   <named list [2]>    
#>  9 name-squatter-rena… FALSE  1om8oRG… <named list>   <named list [2]>    
#> 10 name-squatter-mv    FALSE  1ENeiEn… <named list>   <named list [2]>    
drive_reveal(files, "published")
#> # A dribble: 10 × 5
#>    name                published id       drive_resource revision_resource
#>    <chr>               <lgl>     <drv_id> <list>         <list>           
#>  1 name-squatter-rena… FALSE     1iejRst… <named list>   <named list [9]> 
#>  2 name-squatter-mv    FALSE     1lyFPxt… <named list>   <named list [9]> 
#>  3 chicken-mv-renamed… FALSE     1y7LlJm… <named list>   <named list [10]>
#>  4 name-squatter-uplo… FALSE     1AT6KZa… <named list>   <named list [9]> 
#>  5 name-squatter-rena… FALSE     1DRTur5… <named list>   <named list [9]> 
#>  6 name-squatter-mv    FALSE     1kqe58I… <named list>   <named list [9]> 
#>  7 chicken-mv-renamed… FALSE     1WMuZd4… <named list>   <named list [10]>
#>  8 name-squatter-uplo… FALSE     1LULuzk… <named list>   <named list [9]> 
#>  9 name-squatter-rena… FALSE     1om8oRG… <named list>   <named list [9]> 
#> 10 name-squatter-mv    FALSE     1ENeiEn… <named list>   <named list [9]> 

# a "special" case of digging info out of `drive_resource`, then processing
# a bit
drive_reveal(files, "parent")
#> # A dribble: 10 × 4
#>    name                   id_parent id       drive_resource   
#>    <chr>                  <drv_id>  <drv_id> <list>           
#>  1 name-squatter-rename   0AO_RMa…  1iejRst… <named list [39]>
#>  2 name-squatter-mv       0AO_RMa…  1lyFPxt… <named list [39]>
#>  3 chicken-mv-renamed.txt 1ns19LO…  1y7LlJm… <named list [42]>
#>  4 name-squatter-upload   0AO_RMa…  1AT6KZa… <named list [39]>
#>  5 name-squatter-rename   0AO_RMa…  1DRTur5… <named list [39]>
#>  6 name-squatter-mv       0AO_RMa…  1kqe58I… <named list [39]>
#>  7 chicken-mv-renamed.txt 1ns19LO…  1WMuZd4… <named list [42]>
#>  8 name-squatter-upload   0AO_RMa…  1LULuzk… <named list [39]>
#>  9 name-squatter-rename   0AO_RMa…  1om8oRG… <named list [39]>
#> 10 name-squatter-mv       0AO_RMa…  1ENeiEn… <named list [39]>

# the "simple" cases of digging info out of `drive_resource`
drive_reveal(files, "trashed")
#> # A dribble: 10 × 4
#>    name                   trashed id       drive_resource   
#>    <chr>                  <lgl>   <drv_id> <list>           
#>  1 name-squatter-rename   TRUE    1iejRst… <named list [39]>
#>  2 name-squatter-mv       FALSE   1lyFPxt… <named list [39]>
#>  3 chicken-mv-renamed.txt FALSE   1y7LlJm… <named list [42]>
#>  4 name-squatter-upload   TRUE    1AT6KZa… <named list [39]>
#>  5 name-squatter-rename   TRUE    1DRTur5… <named list [39]>
#>  6 name-squatter-mv       FALSE   1kqe58I… <named list [39]>
#>  7 chicken-mv-renamed.txt FALSE   1WMuZd4… <named list [42]>
#>  8 name-squatter-upload   TRUE    1LULuzk… <named list [39]>
#>  9 name-squatter-rename   TRUE    1om8oRG… <named list [39]>
#> 10 name-squatter-mv       FALSE   1ENeiEn… <named list [39]>
drive_reveal(files, "mime_type")
#> # A dribble: 10 × 4
#>    name                   mime_type                id       drive_resource
#>    <chr>                  <chr>                    <drv_id> <list>        
#>  1 name-squatter-rename   application/octet-stream 1iejRst… <named list>  
#>  2 name-squatter-mv       application/octet-stream 1lyFPxt… <named list>  
#>  3 chicken-mv-renamed.txt text/plain               1y7LlJm… <named list>  
#>  4 name-squatter-upload   application/octet-stream 1AT6KZa… <named list>  
#>  5 name-squatter-rename   application/octet-stream 1DRTur5… <named list>  
#>  6 name-squatter-mv       application/octet-stream 1kqe58I… <named list>  
#>  7 chicken-mv-renamed.txt text/plain               1WMuZd4… <named list>  
#>  8 name-squatter-upload   application/octet-stream 1LULuzk… <named list>  
#>  9 name-squatter-rename   application/octet-stream 1om8oRG… <named list>  
#> 10 name-squatter-mv       application/octet-stream 1ENeiEn… <named list>  
drive_reveal(files, "starred")
#> # A dribble: 10 × 4
#>    name                   starred id       drive_resource   
#>    <chr>                  <lgl>   <drv_id> <list>           
#>  1 name-squatter-rename   FALSE   1iejRst… <named list [39]>
#>  2 name-squatter-mv       FALSE   1lyFPxt… <named list [39]>
#>  3 chicken-mv-renamed.txt FALSE   1y7LlJm… <named list [42]>
#>  4 name-squatter-upload   FALSE   1AT6KZa… <named list [39]>
#>  5 name-squatter-rename   FALSE   1DRTur5… <named list [39]>
#>  6 name-squatter-mv       FALSE   1kqe58I… <named list [39]>
#>  7 chicken-mv-renamed.txt FALSE   1WMuZd4… <named list [42]>
#>  8 name-squatter-upload   FALSE   1LULuzk… <named list [39]>
#>  9 name-squatter-rename   FALSE   1om8oRG… <named list [39]>
#> 10 name-squatter-mv       FALSE   1ENeiEn… <named list [39]>
drive_reveal(files, "description")
#> # A dribble: 10 × 4
#>    name                   description id       drive_resource   
#>    <chr>                  <list>      <drv_id> <list>           
#>  1 name-squatter-rename   <NULL>      1iejRst… <named list [39]>
#>  2 name-squatter-mv       <NULL>      1lyFPxt… <named list [39]>
#>  3 chicken-mv-renamed.txt <NULL>      1y7LlJm… <named list [42]>
#>  4 name-squatter-upload   <NULL>      1AT6KZa… <named list [39]>
#>  5 name-squatter-rename   <NULL>      1DRTur5… <named list [39]>
#>  6 name-squatter-mv       <NULL>      1kqe58I… <named list [39]>
#>  7 chicken-mv-renamed.txt <NULL>      1WMuZd4… <named list [42]>
#>  8 name-squatter-upload   <NULL>      1LULuzk… <named list [39]>
#>  9 name-squatter-rename   <NULL>      1om8oRG… <named list [39]>
#> 10 name-squatter-mv       <NULL>      1ENeiEn… <named list [39]>
drive_reveal(files, "version")
#> # A dribble: 10 × 4
#>    name                   version id       drive_resource   
#>    <chr>                  <chr>   <drv_id> <list>           
#>  1 name-squatter-rename   2       1iejRst… <named list [39]>
#>  2 name-squatter-mv       1       1lyFPxt… <named list [39]>
#>  3 chicken-mv-renamed.txt 5       1y7LlJm… <named list [42]>
#>  4 name-squatter-upload   2       1AT6KZa… <named list [39]>
#>  5 name-squatter-rename   2       1DRTur5… <named list [39]>
#>  6 name-squatter-mv       1       1kqe58I… <named list [39]>
#>  7 chicken-mv-renamed.txt 5       1WMuZd4… <named list [42]>
#>  8 name-squatter-upload   2       1LULuzk… <named list [39]>
#>  9 name-squatter-rename   2       1om8oRG… <named list [39]>
#> 10 name-squatter-mv       1       1ENeiEn… <named list [39]>
drive_reveal(files, "web_view_link")
#> # A dribble: 10 × 4
#>    name                   web_view_link            id       drive_resource
#>    <chr>                  <chr>                    <drv_id> <list>        
#>  1 name-squatter-rename   https://drive.google.co… 1iejRst… <named list>  
#>  2 name-squatter-mv       https://drive.google.co… 1lyFPxt… <named list>  
#>  3 chicken-mv-renamed.txt https://drive.google.co… 1y7LlJm… <named list>  
#>  4 name-squatter-upload   https://drive.google.co… 1AT6KZa… <named list>  
#>  5 name-squatter-rename   https://drive.google.co… 1DRTur5… <named list>  
#>  6 name-squatter-mv       https://drive.google.co… 1kqe58I… <named list>  
#>  7 chicken-mv-renamed.txt https://drive.google.co… 1WMuZd4… <named list>  
#>  8 name-squatter-upload   https://drive.google.co… 1LULuzk… <named list>  
#>  9 name-squatter-rename   https://drive.google.co… 1om8oRG… <named list>  
#> 10 name-squatter-mv       https://drive.google.co… 1ENeiEn… <named list>  
drive_reveal(files, "modified_time")
#> # A dribble: 10 × 4
#>    name                   modified_time       id       drive_resource   
#>    <chr>                  <dttm>              <drv_id> <list>           
#>  1 name-squatter-rename   2024-09-19 13:44:23 1iejRst… <named list [39]>
#>  2 name-squatter-mv       2024-09-19 13:43:54 1lyFPxt… <named list [39]>
#>  3 chicken-mv-renamed.txt 2024-09-19 13:43:51 1y7LlJm… <named list [42]>
#>  4 name-squatter-upload   2024-09-18 13:45:31 1AT6KZa… <named list [39]>
#>  5 name-squatter-rename   2024-09-18 13:44:30 1DRTur5… <named list [39]>
#>  6 name-squatter-mv       2024-09-18 13:44:01 1kqe58I… <named list [39]>
#>  7 chicken-mv-renamed.txt 2024-09-18 13:43:56 1WMuZd4… <named list [42]>
#>  8 name-squatter-upload   2024-09-17 13:44:14 1LULuzk… <named list [39]>
#>  9 name-squatter-rename   2024-09-17 13:43:25 1om8oRG… <named list [39]>
#> 10 name-squatter-mv       2024-09-17 13:43:00 1ENeiEn… <named list [39]>
drive_reveal(files, "created_time")
#> # A dribble: 10 × 4
#>    name                   created_time        id       drive_resource   
#>    <chr>                  <dttm>              <drv_id> <list>           
#>  1 name-squatter-rename   2024-09-19 13:44:23 1iejRst… <named list [39]>
#>  2 name-squatter-mv       2024-09-19 13:43:54 1lyFPxt… <named list [39]>
#>  3 chicken-mv-renamed.txt 2024-09-19 13:43:50 1y7LlJm… <named list [42]>
#>  4 name-squatter-upload   2024-09-18 13:45:31 1AT6KZa… <named list [39]>
#>  5 name-squatter-rename   2024-09-18 13:44:30 1DRTur5… <named list [39]>
#>  6 name-squatter-mv       2024-09-18 13:44:01 1kqe58I… <named list [39]>
#>  7 chicken-mv-renamed.txt 2024-09-18 13:43:55 1WMuZd4… <named list [42]>
#>  8 name-squatter-upload   2024-09-17 13:44:14 1LULuzk… <named list [39]>
#>  9 name-squatter-rename   2024-09-17 13:43:25 1om8oRG… <named list [39]>
#> 10 name-squatter-mv       2024-09-17 13:43:00 1ENeiEn… <named list [39]>
drive_reveal(files, "owned_by_me")
#> # A dribble: 10 × 4
#>    name                   owned_by_me id       drive_resource   
#>    <chr>                  <lgl>       <drv_id> <list>           
#>  1 name-squatter-rename   TRUE        1iejRst… <named list [39]>
#>  2 name-squatter-mv       TRUE        1lyFPxt… <named list [39]>
#>  3 chicken-mv-renamed.txt TRUE        1y7LlJm… <named list [42]>
#>  4 name-squatter-upload   TRUE        1AT6KZa… <named list [39]>
#>  5 name-squatter-rename   TRUE        1DRTur5… <named list [39]>
#>  6 name-squatter-mv       TRUE        1kqe58I… <named list [39]>
#>  7 chicken-mv-renamed.txt TRUE        1WMuZd4… <named list [42]>
#>  8 name-squatter-upload   TRUE        1LULuzk… <named list [39]>
#>  9 name-squatter-rename   TRUE        1om8oRG… <named list [39]>
#> 10 name-squatter-mv       TRUE        1ENeiEn… <named list [39]>
drive_reveal(files, "size")
#> # A dribble: 10 × 4
#>    name                   size  id       drive_resource   
#>    <chr>                  <chr> <drv_id> <list>           
#>  1 name-squatter-rename   0     1iejRst… <named list [39]>
#>  2 name-squatter-mv       0     1lyFPxt… <named list [39]>
#>  3 chicken-mv-renamed.txt 165   1y7LlJm… <named list [42]>
#>  4 name-squatter-upload   0     1AT6KZa… <named list [39]>
#>  5 name-squatter-rename   0     1DRTur5… <named list [39]>
#>  6 name-squatter-mv       0     1kqe58I… <named list [39]>
#>  7 chicken-mv-renamed.txt 165   1WMuZd4… <named list [42]>
#>  8 name-squatter-upload   0     1LULuzk… <named list [39]>
#>  9 name-squatter-rename   0     1om8oRG… <named list [39]>
#> 10 name-squatter-mv       0     1ENeiEn… <named list [39]>
drive_reveal(files, "quota_bytes_used")
#> # A dribble: 10 × 4
#>    name                   quota_bytes_used id       drive_resource   
#>    <chr>                  <chr>            <drv_id> <list>           
#>  1 name-squatter-rename   0                1iejRst… <named list [39]>
#>  2 name-squatter-mv       0                1lyFPxt… <named list [39]>
#>  3 chicken-mv-renamed.txt 165              1y7LlJm… <named list [42]>
#>  4 name-squatter-upload   0                1AT6KZa… <named list [39]>
#>  5 name-squatter-rename   0                1DRTur5… <named list [39]>
#>  6 name-squatter-mv       0                1kqe58I… <named list [39]>
#>  7 chicken-mv-renamed.txt 165              1WMuZd4… <named list [42]>
#>  8 name-squatter-upload   0                1LULuzk… <named list [39]>
#>  9 name-squatter-rename   0                1om8oRG… <named list [39]>
#> 10 name-squatter-mv       0                1ENeiEn… <named list [39]>

# 'root' is a special file id that represents your My Drive root folder
drive_get(id = "root") %>%
  drive_reveal("path")
#> # A dribble: 1 × 4
#>   name     path  id                  drive_resource   
#>   <chr>    <chr> <drv_id>            <list>           
#> 1 My Drive ~/    0AO_RMaBzcP63Uk9PVA <named list [31]>