drive_rm {googledrive}	R Documentation

Delete files from Drive

Description

Caution: this will permanently delete your files! For a safer, reversible option, see drive_trash().

Usage

drive_rm(..., verbose = TRUE)

Arguments

...
One or more Drive files, specified in any valid way, i.e. as a dribble, by name or path, or by file id or URL marked with as_id(). Or any combination thereof. Elements are processed with as_dribble() and row-bound prior to deletion.

verbose
Logical, indicating whether to print informative messages (default TRUE).

Value

Logical vector, indicating whether the delete succeeded.

See Also

Wraps the files.delete endpoint:

https://developers.google.com/drive/v3/reference/files/delete

Examples

## Not run: 
## Create something to remove
drive_upload(drive_example("chicken.txt"), name = "chicken-rm.txt")

## Remove it by name
drive_rm("chicken-rm.txt")

## Create several things to remove
x1 <- drive_upload(drive_example("chicken.txt"), name = "chicken-abc.txt")
drive_upload(drive_example("chicken.txt"), name = "chicken-def.txt")
x2 <- drive_upload(drive_example("chicken.txt"), name = "chicken-ghi.txt")

## Remove them all at once, specified in different ways
drive_rm(x1, "chicken-def.txt", as_id(x2))

## End(Not run)

[Package googledrive version 1.0.0 Index]