Delete files from Drive

Caution: this will permanently delete your files! For a safer, reversible option, see drive_trash().

Usage

drive_rm(..., verbose = deprecated())

Arguments

...

One or more Drive files, specified in any valid way, i.e. as a dribble, by name or path, or by file id or URL marked with as_id(). Or any combination thereof. Elements are processed with as_dribble() and row-bound prior to deletion.

verbose

This logical argument to individual googledrive functions is deprecated. To globally suppress googledrive messaging, use options(googledrive_quiet = TRUE) (the default behaviour is to emit informational messages). To suppress messaging in a more limited way, use the helpers local_drive_quiet() or with_drive_quiet().

Value

Logical vector, indicating whether the delete succeeded.

See also

Examples

# Target one of the official example files to copy (then remove)
(src_file <- drive_example_remote("chicken.txt"))
#> # A dribble: 1 × 3
#>   name        id                                drive_resource   
#>   <chr>       <drv_id>                          <list>           
#> 1 chicken.txt 1wOLeWVRkTb6lDmLRiOhg9iKM7DlN762Y <named list [38]>

# Create a copy, then remove it by name
src_file %>%
  drive_cp(name = "chicken-rm.txt")
#> Original file:
#> • chicken.txt <id: 1wOLeWVRkTb6lDmLRiOhg9iKM7DlN762Y>
#> Copied to file:
#> • chicken-rm.txt <id: 136pP5qVTBopvVQGij0wDO7rNrrq-qyYv>
drive_rm("chicken-rm.txt")
#> File deleted:
#> • chicken-rm.txt <id: 136pP5qVTBopvVQGij0wDO7rNrrq-qyYv>

# Create several more copies
x1 <- src_file %>%
  drive_cp(name = "chicken-abc.txt")
#> Original file:
#> • chicken.txt <id: 1wOLeWVRkTb6lDmLRiOhg9iKM7DlN762Y>
#> Copied to file:
#> • chicken-abc.txt <id: 1pmOtpjRz0XL8PIekWQD6piZbi874EkaB>
drive_cp(src_file, name = "chicken-def.txt")
#> Original file:
#> • chicken.txt <id: 1wOLeWVRkTb6lDmLRiOhg9iKM7DlN762Y>
#> Copied to file:
#> • chicken-def.txt <id: 1KKl1ARIOSstIEZ2_6RS17SlhkhYvyT9m>
x2 <- src_file %>%
  drive_cp(name = "chicken-ghi.txt")
#> Original file:
#> • chicken.txt <id: 1wOLeWVRkTb6lDmLRiOhg9iKM7DlN762Y>
#> Copied to file:
#> • chicken-ghi.txt <id: 1MvThU5peaG9y0_VFnTA27-3B_nRUD0_3>

# Remove the copies all at once, specified in different ways
drive_rm(x1, "chicken-def.txt", as_id(x2))
#> Files deleted:
#> • chicken-abc.txt <id: 1pmOtpjRz0XL8PIekWQD6piZbi874EkaB>
#> • chicken-def.txt <id: 1KKl1ARIOSstIEZ2_6RS17SlhkhYvyT9m>
#> • chicken-ghi.txt <id: 1MvThU5peaG9y0_VFnTA27-3B_nRUD0_3>