drive_trash {googledrive}	R Documentation

Move Drive files to or from trash

Description

Move Drive files to or from trash

Usage

drive_trash(file, verbose = TRUE)

drive_untrash(file, verbose = TRUE)

Arguments

file
Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

verbose
Logical, indicating whether to print informative messages (default TRUE).

Value

An object of class dribble, a tibble with one row per item.

Examples

## Not run: 
## Create a file and put it in the trash.
file <- drive_upload(drive_example("chicken.txt"), "chicken-trash.txt")
drive_trash("chicken-trash.txt")

## Confirm it's in the trash
drive_find(trashed = TRUE)

## Remove it from the trash and confirm
drive_untrash("chicken-trash.txt")
drive_find(trashed = TRUE)

## Clean up
drive_rm("chicken-trash.txt")

## End(Not run)

[Package googledrive version 1.0.0 Index]