Move Drive files to or from trash

Move Drive files to or from trash

Usage

drive_trash(file, verbose = deprecated())

drive_untrash(file, verbose = deprecated())

Arguments

file

Something that identifies the file(s) of interest on your Google Drive. Can be a character vector of names/paths, a character vector of file ids or URLs marked with as_id(), or a dribble.

verbose

This logical argument to individual googledrive functions is deprecated. To globally suppress googledrive messaging, use options(googledrive_quiet = TRUE) (the default behaviour is to emit informational messages). To suppress messaging in a more limited way, use the helpers local_drive_quiet() or with_drive_quiet().

Value

An object of class dribble, a tibble with one row per file.

Examples

# Create a file and put it in the trash.
file <- drive_example_remote("chicken.txt") %>%
  drive_cp("chicken-trash.txt")
#> Original file:
#> • chicken.txt <id: 1wOLeWVRkTb6lDmLRiOhg9iKM7DlN762Y>
#> Copied to file:
#> • chicken-trash.txt <id: 1depIQFz2yL_Gji3lZOhsUADd_w_hytoK>
drive_trash("chicken-trash.txt")
#> File trashed:
#> • chicken-trash.txt <id: 1depIQFz2yL_Gji3lZOhsUADd_w_hytoK>

# Confirm it's in the trash
drive_find(trashed = TRUE)
#> # A dribble: 69 × 3
#>    name                 id                                drive_resource
#>    <chr>                <drv_id>                          <list>        
#>  1 chicken-trash.txt    1depIQFz2yL_Gji3lZOhsUADd_w_hytoK <named list>  
#>  2 name-squatter-rename 1om8oRGwjHq4rdvtRohLKsjX52X_rNHb0 <named list>  
#>  3 name-squatter-upload 1MYmq8Vvip-aZaOAnrkG6FNvO2depL216 <named list>  
#>  4 name-squatter-rename 1g1Ys_0pufYGiWkVypUMfmeyGNZQodidc <named list>  
#>  5 name-squatter-upload 1gBB0-FUsVRS5lO1wNr5NQzmrD76oG_jG <named list>  
#>  6 name-squatter-rename 1DgYczbHkrae7a7POdkgd3BSuPtMHpNO9 <named list>  
#>  7 name-squatter-upload 1EWVz4VVOvITwDai5uMVjRQtLFLZmd43e <named list>  
#>  8 name-squatter-rename 1z7Bb5xNpDsVLbQTocIgPqxOgoarm6oCE <named list>  
#>  9 name-squatter-upload 1w2A8R8I_bHJgjHrqgpzsDeQ183dKqM5X <named list>  
#> 10 name-squatter-rename 1V8KxrYN-SI62U6C-mS20u8KPumVoh13y <named list>  
#> # ℹ 59 more rows

# Remove it from the trash and confirm
drive_untrash("chicken-trash.txt")
#> File untrashed:
#> • chicken-trash.txt <id: 1depIQFz2yL_Gji3lZOhsUADd_w_hytoK>
drive_find(trashed = TRUE)
#> # A dribble: 68 × 3
#>    name                 id                                drive_resource
#>    <chr>                <drv_id>                          <list>        
#>  1 name-squatter-rename 1om8oRGwjHq4rdvtRohLKsjX52X_rNHb0 <named list>  
#>  2 name-squatter-upload 1MYmq8Vvip-aZaOAnrkG6FNvO2depL216 <named list>  
#>  3 name-squatter-rename 1g1Ys_0pufYGiWkVypUMfmeyGNZQodidc <named list>  
#>  4 name-squatter-upload 1gBB0-FUsVRS5lO1wNr5NQzmrD76oG_jG <named list>  
#>  5 name-squatter-rename 1DgYczbHkrae7a7POdkgd3BSuPtMHpNO9 <named list>  
#>  6 name-squatter-upload 1EWVz4VVOvITwDai5uMVjRQtLFLZmd43e <named list>  
#>  7 name-squatter-rename 1z7Bb5xNpDsVLbQTocIgPqxOgoarm6oCE <named list>  
#>  8 name-squatter-upload 1w2A8R8I_bHJgjHrqgpzsDeQ183dKqM5X <named list>  
#>  9 name-squatter-rename 1V8KxrYN-SI62U6C-mS20u8KPumVoh13y <named list>  
#> 10 name-squatter-upload 1Oi-APcgWSF8jDKVfNxX15uEeCesE94lS <named list>  
#> # ℹ 58 more rows

# Clean up
drive_rm("chicken-trash.txt")
#> File deleted:
#> • chicken-trash.txt <id: 1depIQFz2yL_Gji3lZOhsUADd_w_hytoK>