Update an existing Drive file

Update an existing Drive file id with new content ("media" in Drive API-speak), new metadata, or both.

drive_update(file, media = NULL, ..., verbose = TRUE)

Arguments

file
Something that identifies the file of interest on your Google Drive. Can be a name or path, a file id or URL marked with as_id(), or a dribble.

media
Character, path to the local file to upload.

...
Named parameters to pass along to the Drive API. You can affect the metadata of the target file by specifying properties of the Files resource via .... See the "Request body" section of the Drive API docs for the associated endpoint.

verbose
Logical, indicating whether to print informative messages (default TRUE).

Value

An object of class dribble, a tibble with one row per item.

See also

Wraps the files.update endpoint:

https://developers.google.com/drive/v3/reference/files/update

This function supports media upload:

https://developers.google.com/drive/v3/web/manage-uploads

Examples

# NOT RUN {
## Create a new file, so we can update it
x <- drive_upload(R.home("doc/NEWS"))

## Update the file with new media
x <- x %>%
  drive_update(R.home("doc/NEWS.1"))

## Update the file with new metadata.
## Notice here `name` is not an argument of `drive_update()`, we are passing
## this to the API via the `...``
x <- x %>%
  drive_update(name = "NEWS-1")

## We can add a parent folder by passing `addParents` via `...`.
folder <- drive_mkdir("second-parent-folder")
x <- x %>%
  drive_update(addParents = folder$id)
## Verify the file now has multiple parents
x$drive_resource[[1]]$parents

## Update the file with new media AND new metadata
x <- x %>%
  drive_update(R.home("doc/NEWS.2"), name = "NEWS-2")

## Clean up
drive_rm(x, folder)
# }