googledrive-deprecated {googledrive}	R Documentation

Deprecated googledrive functions

Description

Usage

drive_auth_config(active, app, path, api_key)

drive_oauth_app()

drive_example(path = NULL)

Arguments

app
Replaced by the client argument.

path
JSON downloaded from Google Cloud Console, containing a client id and secret, in one of the forms supported for the txt argument of jsonlite::fromJSON() (typically, a file path or JSON string).

api_key
API key.

drive_auth_config()

This function is defunct.

Use drive_auth_configure() to configure your own OAuth client or API key.

Use drive_deauth() to go into a de-authorized state.

Use drive_oauth_client() to retrieve a user-configured client, if it exists.

Use drive_api_key() to retrieve a user-configured API key, if it exists.

drive_oauth_app()

In light of the new gargle::gargle_oauth_client() constructor and class of the same name, drive_oauth_app() is being replaced by drive_oauth_client().

drive_example()

This function is defunct. Access example files with drive_examples_local(), drive_example_local(), drive_examples_remote(), and drive_example_remote().

[Package googledrive version 2.1.1 Index]