googledrive-deprecated {googledrive}	R Documentation

Deprecated googledrive functions

Description

Deprecated googledrive functions

Usage

drive_auth_config(active, app, path, api_key)

drive_example(path = NULL)

Arguments

app
OAuth app, in the sense of httr::oauth_app().

path
JSON downloaded from Google Cloud Platform Console, containing a client id (aka key) and secret, in one of the forms supported for the txt argument of jsonlite::fromJSON() (typically, a file path or JSON string).

api_key
API key.

[Package googledrive version 2.0.0 Index]