shared_drive_update {googledrive}	R Documentation

Update a shared drive

Description

Update the metadata of an existing shared drive, e.g. its background image or theme.

A shared drive supports files owned by an organization rather than an individual user. Shared drives follow different sharing and ownership models from a specific user's "My Drive". Shared drives are the successors to the earlier concept of Team Drives. Learn more about shared drives.

Usage

shared_drive_update(shared_drive, ...)

Arguments

shared_drive
Anything that identifies one specific shared drive: its name, its id or URL marked with as_id(), or a dribble. The value provided to shared_drive is pre-processed with as_shared_drive(). Read more about shared drives.

...
Properties to set in name = value form. See the "Request body" section of the Drive API docs for this endpoint.

Value

An object of class dribble, a tibble with one row per shared drive.

See Also

Wraps the drives.update endpoint:

https://developers.google.com/drive/api/v3/reference/drives/update

Examples

## Not run: 
# create a shared drive
sd <- shared_drive_create("I love themes!")

# see the themes available to you
themes <- drive_about()$driveThemes
purrr::map_chr(themes, "id")

# cycle through various themes for this shared drive
sd <- shared_drive_update(sd, themeId = "bok_choy")
sd <- shared_drive_update(sd, themeId = "cocktails")

# clean up
shared_drive_rm(sd)

## End(Not run)

[Package googledrive version 2.0.0 Index]