team_drive_find {googledrive}	R Documentation

Find Team Drives

Description

This is the closest googledrive function to what you get from visiting https://drive.google.com and clicking "Team Drives".

Note: Team Drives are only available to users of certain enhanced Google services, such as G Suite Enterprise, G Suite Business, or G Suite for Education.

Usage

team_drive_find(pattern = NULL, n_max = Inf, ..., verbose = TRUE)

Arguments

pattern
Character. If provided, only the items whose names match this regular expression are returned. This is implemented locally on the results returned by the API.

n_max
Integer. An upper bound on the number of items to return. This applies to the results requested from the API, which may be further filtered locally, via the pattern argument.

...
Other parameters to pass along in the request, such as pageSize.

verbose
Logical, indicating whether to print informative messages (default TRUE).

Value

An object of class dribble, a tibble with one row per item.

See Also

Wraps the teamdrives.list endpoint::

https://developers.google.com/drive/v3/reference/teamdrives/list

Examples

## Not run: 
team_drive_find()

## End(Not run)

[Package googledrive version 1.0.0 Index]