team_drive_update {googledrive}	R Documentation

Update an existing Team Drive

Description

Update the metadata of an existing Team Drive, e.g. its background image or theme.

Note: Team Drives are only available to users of certain enhanced Google services, such as G Suite Enterprise, G Suite Business, or G Suite for Education.

Usage

team_drive_update(team_drive, ..., verbose = TRUE)

Arguments

team_drive
Anything that identifies one specific Team Drive: its name, its id or URL marked with as_id(), or a dribble. Is pre-processed with as_team_drive(). Read more about Team Drives.

...
Named parameters to pass along to the Drive API. See the "Request body" section of the Drive API docs for the associated endpoint.

verbose
Logical, indicating whether to print informative messages (default TRUE).

Value

An object of class dribble, a tibble with one row per item.

See Also

Wraps the teamdrives.update endpoint:

https://developers.google.com/drive/v3/reference/teamdrives/update

Examples

## Not run: 
## create a Team Drive
td <- team_drive_create("I love themes!")

## see the themes available to you
themes <- drive_user(fields = "teamDriveThemes")$teamDriveThemes
purrr::map_chr(themes, "id")

## cycle through various themes for this Team Drive
td <- team_drive_update(td, themeId = "bok_choy")
td <- team_drive_update(td, themeId = "cocktails")

## clean up
team_drive_rm(td)

## End(Not run)

[Package googledrive version 1.0.0 Index]